General schemes

Arbitrary schemes over a base ring $\mathbb k$ which are given by means of their affine patches and glueings are instances of

Scheme — Type

Scheme{BaseRingType<:Ring, BaseRingElemType<:RingElement}

A scheme over a ring $𝕜$ of type BaseRingType with elements of type BaseRingElemType.

Affine schemes over noetherian base rings

Let $\mathbb k$ be a commutative noetherian base ring (in practice: an algebraic extension of $\mathbb Q$ or $\mathbb F_p$). An affine scheme $X$ over $\mathbb k$ is given as

\[ X = \mathrm{Spec} \left(\mathbb k[x_1,\dots,x_n]/I\right)\]

for some polynomial ring $R = \mathbb k[x_1,\dots,x_n]$ and an ideal $I \subset R$. More generally, we can form the Spec of any localized affine algebra $L$ as an instance of

Spec — Type

Spec{BRT, BRET, RT, RET, MST} <: Scheme{BRT, BRET}

An affine scheme $X = Spec ((R/I)[S⁻¹])$ with $R = k[x₁,…,xₙ]$ a free polynomial algebra of type RT over a base ring $k$ of type BRT, $I ⊂ R$ a finitely generated ideal with elements of type RET, and $S$ a multiplicative set in $R$ of type MST.

Internally, this merely stores an instance $L$ of MPolyQuoLocalizedRing. This ring can be obtained using

OO — Method

OO(X::Spec)

For $X = Spec ((𝕜[x₁,…,xₙ]/I)[S⁻¹])$ this returns $(𝕜[x₁,…,xₙ]/I)[S⁻¹]$.

One of the main reasons to allow such general schemes is that principal open subsets $U = \{ f \neq 0\}$ of affine schemes $X$, $f \in \mathcal O_X(X)$ are again affine; basically due to Rabinowitsch's trick:

\[ U = \mathrm{Spec} \left(\mathcal O_X(X)[f^{-1}]\right) = \mathrm{Spec} \left(\mathcal O_X(X)[t]/\langle 1 - t\cdot f \rangle\right).\]

This flexibility allows us to introduce and handle subschemes using the natural relations on their rings. Thus, we can, for instance, define subschemes via

subscheme — Method

subscheme(X::Spec{BRT, BRET, RT, RET, MST}, f::RET) where {BRT, BRET, RT, RET, MST}

For a scheme $X = Spec ((𝕜[x₁,…,xₙ]/I)[S⁻¹])$ and an element $f ∈ 𝕜[x₁,…,xₙ]$ this returns the closed subscheme defined by the ideal $I' = I + ⟨f⟩$.

hypersurface_complement — Method

hypersurface_complement(X::Spec{BRT, BRET, RT, RET, MST}, f::RET) where {BRT, BRET, RT, RET, MST<:MPolyPowersOfElement{BRT, BRET, RT, RET}}

For a scheme $X = Spec ((𝕜[x₁,…,xₙ]/I)[S⁻¹])$ and an element $f ∈ 𝕜[x₁,…,xₙ]$ this returns the open subscheme $U = X ∖ V(f)$ defined by the complement of the vanishing locus of $f$.

Containments can be checked using

issubset — Method

issubset(
  X::Spec{BRT, BRET, RT, RET, MST1}, 
  Y::Spec{BRT, BRET, RT, RET, MST2}
) where {BRT, BRET, RT, RET, MST1<:MPolyPowersOfElement{BRT, BRET, RT, RET}, MST2<:MPolyPowersOfElement{BRT, BRET, RT, RET}}

Checks whether $X$ is a subset of $Y$ based on the comparison of their coordinate rings.

is_closed_embedding — Method

is_closed_embedding(
  X::Spec{BRT, BRET, RT, RET, MST1}, 
  Y::Spec{BRT, BRET, RT, RET, MST2}
) where {BRT, BRET, RT, RET, MST1<:MPolyPowersOfElement{BRT, BRET, RT, RET}, MST2<:MPolyPowersOfElement{BRT, BRET, RT, RET}}

Checks whether $X$ is closed embedded in $Y$.

is_open_embedding — Method

is_open_embedding(
  X::Spec{BRT, BRET, RT, RET, MST1}, 
  Y::Spec{BRT, BRET, RT, RET, MST2}
) where {BRT, BRET, RT, RET, MST1<:MPolyPowersOfElement{BRT, BRET, RT, RET}, MST2<:MPolyPowersOfElement{BRT, BRET, RT, RET}}

Checks whether $X$ is openly embedded in $Y$.

The closure of a subscheme can be computed via

closure — Method

closure(
  X::Spec{BRT, BRET, RT, RET, MST1}, 
  Y::Spec{BRT, BRET, RT, RET, MST2}
) where {BRT, BRET, RT, RET, MST1<:MPolyPowersOfElement{BRT, BRET, RT, RET}, MST2<:MPolyPowersOfElement{BRT, BRET, RT, RET}}

Returns the closure of $X$ in $Y$.

Among the basic functionality for affine schemes we have

product — Method

product(X::Spec{BRT, BRET, RT, RET, MST}, Y::Spec{BRT, BRET, RT, RET, MST}) where {BRT, BRET, RT, RET, MST<:MPolyPowersOfElement}

Returns a triple $(X×Y, p₁, p₂)$ consisting of the product $X×Y$ and the two projections $p₁ : X×Y → X$ and $p₂ : X×Y → Y$.