Projective

AbsProjectiveRationalPoint — Type

AbsProjectiveRationalPoint

A rational point $P$ of a projective scheme $X$. We refer to $X$ as the parent of $P$.

Let $k$ be a field. A rational point is an element of $\mathbb{P}^n(k) = k^{n+1} \setminus \{0\} / k^*$ where two vectors $v,w$ in $k^{n+1} \setminus \{0\}$ are identified if $v = \alpha w$ for a non-zero scalar $\alpha \in k^*$.

Let $X \subseteq \mathbb{P}^n_k$ be an algebraic set or more generally a closed subscheme defined by the homogeneous ideal $I = (f_1, \dots f_r)$. Then a rational point of $X$ is $p \in \mathbb{P}^n(k)$ such that $f_1(p) = \dots = f_n(p) = 0$.

This type includes points in weighted projective space.

ProjectiveRationalPoint — Type

ProjectiveRationalPoint{CoeffType<:RingElem, ParentType<:AbsProjectiveScheme}

Type for rational points in projective varieties.

Examples

julia> P2 = projective_space(QQ, 2);

julia> P2([4, 0 , 2//3])
Projective rational point
  of Projective 2-space over QQ with coordinates [s0, s1, s2]
with coordinates (4 : 0 : 2//3)

coordinates — Method

coordinates(p::AbsProjectiveRationalPoint{S,T}) -> Vector{S}

Return the homogeneous coordinates of the rational point p.

ideal — Method

ideal(P::AbsProjectiveRationalPoint)

Return the homogeneous ideal associated to P in the homogeneous coordinate ring of its ambient space.

ideal(O::NfRelOrd, M::PMat; check::Bool = true, M_in_hnf::Bool = false) -> NfRelOrdIdl

Creates the ideal of $\mathcal O$ with basis pseudo-matrix $M$. If check is set, then it is checked whether $M$ defines an ideal. If M_in_hnf is set, then it is assumed that $M$ is already in lower left pseudo HNF.

ideal(O::NfRelOrd, M::Generic.Mat; check::Bool = true) -> NfRelOrdIdl

Creates the ideal of $\mathcal O$ with basis matrix $M$. If check is set, then it is checked whether $M$ defines an ideal.

ideal(O::NfRelOrd{T, S}, x::NfRelElem{T}, y::NfRelElem{T}, a::S, b::S; check::Bool = true) -> NfRelOrdIdl{T, S}

Creates the ideal $x\cdot a + y\cdot b$ of $\mathcal O$. If check is set, then it is checked whether these elements define an ideal.

ideal(O::NfRelOrd{T, S}, x::NfRelOrdElem{T}) -> NfRelOrdIdl{T, S}
*(O::NfRelOrd{T, S}, x::NfRelOrdElem{T}) -> NfRelOrdIdl{T, S}
*(x::NfRelOrdElem{T}, O::NfRelOrd{T, S}) -> NfRelOrdIdl{T, S}

Creates the ideal $x\cdot \mathcal O$ of $\mathcal O$.

ideal(O::NfRelOrd{T, S}, a::S; check::Bool = true) -> NfRelOrdIdl{T, S}

Creates the ideal $a \cdot \mathcal O$ of $\mathcal O$. If check is set, then it is checked whether $a$ defines an (integral) ideal.

ideal(O::AlgAssAbsOrd, x::AbsAlgAssElem) -> AlgAssAbsOrdIdl
ideal(O::AlgAssAbsOrd, x::AlgAssAbsOrdElem) -> AlgAssAbsOrdIdl

Returns the twosided principal ideal of $O$ generated by $x$.

ideal(O::AlgAssAbsOrd, x::AbsAlgAssElem, side::Symbol) -> AlgAssAbsOrdIdl
ideal(O::AlgAssAbsOrd, x::AlgAssAbsOrdElem, side::Symbol) -> AlgAssAbsOrdIdl

Returns the ideal $O \cdot x$ if side == :left, and $x \cdot O$ if side == :right.

ideal(A::AbsAlgAss, M::PMat; M_in_hnf::Bool = false) -> AlgAssRelOrdIdl

Returns the ideal in $A$ with basis pseudo-matrix $M$. If M_in_hnf == true, it is assumed that $M$ is already in lower left pseudo HNF.

ideal(A::AbsAlgAss, O::AlgAssRelOrd, M::PMat; side::Symbol = :nothing,
      M_in_hnf::Bool = false)
  -> AlgAssRelOrdIdl

Returns the ideal of $O$ in $A$ with basis pseudo-matrix $M$ (in the basis of $A$). If the ideal is known to be a right/left/twosided ideal of $O$, side may be set to :right/:left/:twosided respectively. If M_in_hnf == true, it is assumed that $M$ is already in lower left pseudo HNF.

ideal(O::AlgAssRelOrd, x::AbsAlgAssElem) -> AlgAssRelOrdIdl
ideal(O::AlgAssRelOrd, x::AlgAssRelOrdElem) -> AlgAssRelOrdIdl

Returns the twosided principal ideal of $O$ generated by $x$.

ideal(O::AlgAssRelOrd, x::AbsAlgAssElem, side::Symbol) -> AlgAssRelOrdIdl
ideal(O::AlgAssRelOrd, x::AlgAssRelOrdElem, side::Symbol) -> AlgAssRelOrdIdl

Returns the ideal $O \cdot x$ if side == :left, and $x \cdot O$ if side == :right.

ideal(O::AlgAssRelOrd, a::NfOrdFracIdl) -> AlgAssRelOrdIdl
ideal(O::AlgAssRelOrd, a::NfRelOrdFracIdl) -> AlgAssRelOrdIdl

Returns the ideal $a \cdot O$ where $a$ is a fractional ideal of base_ring(O).

scheme — Method

scheme(P::AbsProjectiveRationalPoint) -> AbsProjectiveScheme

Return the rational point $P$ viewed as a reduced, projective subscheme of its ambient projective space.

normalize! — Method

normalize!(a::AbsProjectiveRationalPoint{<:FieldElem})

Normalize a such that its first non-zero coordinate is one.

normalize! — Method

normalize!(a::AbsProjectiveRationalPoint{ZZRingElem})

Normalize a such that its first non-zero coordinate is positive.