Adjunction Process for Surfaces

A surface in this section is a smooth projective surface over $\mathbb C$.

Blowing up a surface in a point means to replace the point by an exceptional curve. Each such curve $E$ is a smooth, rational curve with self-intersection number $E^{2}=-1$. We speak of a $(-1)$-curve. A surface is minimal if it contains no $(-1)$-curves. That is, the surface cannot be obtained by blowing up a point on another surface. A surface $X_{\text{min}}$ is called a minimal model of a surface $X$ if $X_{\text{min}}$ is minimal and $X$ can be obtained from $X_{\text{min}}$ by repeatedly blowing up a point. Each surface $X$ has a minimal model which is unique if $X$ has non-negative Kodaira dimension. The Enriques-Kodaira classification classifies surfaces according to their minimal models. See [BHPV-D-V04] for more on this.

Given a surface, we may apply the adjunction process of Van de Ven and Sommese [SV-D-V87] to discover a minimal model. To describe this process, consider a surface $X \subset \mathbb P^{n}$ of codimension $c$. Let $S$ and $S_{X}$ denote the homogeneous coordinate rings of $\mathbb P^{n}$ and $X$, respectively. Consider $\omega_{X}=\text{Ext}^{c}_{S}(S_{X},S(-n-1)),$ the graded dualizing module of $S_{X}$. A basis of the graded piece $(\omega_{X})_{{1}}$ corresponds to the linear system $|K_X +H|$, where $K_X$ is a canonical divisor on $X$ and $H$ is the hyperplane class. Except for some exceptional cases, this linear system defines a birational morphism $\varphi_{|K_X+H|}\colon X \to X'$ onto another smooth projective surface $X'$ such that $\varphi_{|K_X+H|}$ blows down precisely all $(-1)$-lines on $X$. As shown by Van de Ven and Sommese, in the exceptional cases,

$X$ is a linearly or quadratically embedded $ \mathbb P^{2}$ or $X$ is ruled by lines, in which case $|K_X+H| = \emptyset$,

$X$ is an anti-canonically embedded del Pezzo surface, in which case $\varphi_{|K_X+H|}$ maps $X$ to a point,

$X$ is a conic bundle, in which case $\varphi_{|K_{X}+H|}\colon X \to B$ maps $X$ to a curve $B$ such that the fibers of $\varphi_{|K_{X}+H|}$ are the conics, or

$X$ is a surface in one of four explicit families identified by Sommese and Van de Ven, and $\varphi_{|K_X+H|}\colon X \to X'$ is not birational, but finite to one.

If we are not in one of these cases, a $(-1)$-conic $C$ in $X$ is mapped to a $(-1)$-line in $X'$ since $(K_X+H)\;\!. \;\! C=-1+2=1$. Thus, the adjunction process, which consists of applying the adjunction maps $\varphi_{|K_X+H|}$, $\varphi_{|K_{X'}+H'|}$ and so on, yields finitely many surfaces $X \rightarrow X^{\prime} \rightarrow X^{\prime\prime} \rightarrow \dots$ which are called the adjoint surfaces of $X$. The last adjoint surface is either minimal or belongs to one of the exceptional cases. In particular, if $X$ has non-negative Kodaira dimension, the adjunction process yields the uniquely determined minimal model of $X$.

Note

If $X$ is rational, the last adjoint surface is either $\mathbb P^{2}$, the Veronese surface, a Hirzebruch surface, a Del Pezzo surface, a conic bundle, or one of the four explicit families identified by Sommese and Van de Ven.

Note

In explicit computations, we consider surfaces which are defined by polynomial equations with coefficients in a subfield of $\mathbb C$ which can be handled by OSCAR.

Note

The surfaces in the examples below are taken from the OSCAR data base of nongeneral type surfaces in $\mathbb P^4$. To ease subsequent computations, the surfaces in the data base where constructed over finite fields. Note, however, that the recipes used in the constructions also work in characteristic zero. So all computations can be confirmed in characteristic zero, although this may be time consuming.

Adjunction Process

What we describe here goes back to joint work of Wolfram Decker and Frank-Olaf Schreyer. See [DES93], [DS00].

adjunction_process — Function

adjunction_process(X::AbsProjectiveVariety, steps::Int=0)

Given a smooth surface X and a non-negative integer steps, return data which describes the adjunction process for X: If steps == 0, carry out the complete process. Otherwise, carry out the indicated number of steps only.

More precisely, if $X^{(0)} = X \rightarrow X^{(1)}\rightarrow \dots \rightarrow X^{(r)}$ is the sequence of successive adjunction maps and adjoint surfaces in the completed adjunction process, return a quadruple L, say, where:

L[1] is a vector of tuples of numerical data: For each step $X^{(i)}\rightarrow X^{(i+1)}$, return the tuple $(n^{(i)}, d^{(i)}, \pi^{(i)}, s^{(i)}),$ where $n^{(i)}$ is the dimension of the ambient projective space of $X^{(i)}$, $d^{(i)}$ is the degree of $X^{(i)}$, $\pi^{(i)}$ is the sectional genus of $X^{(i)}$, and $s^{(i)}$ is the number of exceptional $(-1)$-lines on $X^{(i)}$ which are blown down to points in $ X^{(i+1)}$.

L[2] is a vector of adjoint matrices: For each step $X^{(i)}\rightarrow X^{(i+1)}$, return a presentation matrix of $S_X^{(i)}(1)$ considered as a module over $S_X^{(i+1)}$, where the $S_X^{(i)}$ are the homogeneous coordinate rings of the $X^{(i)}$. If X is rational, these matrices can be used to compute a rational parametrization of X.

L[3] is a vector of zero-dimensional projective algebraic sets: For each step $X^{(i)}\rightarrow X^{(i+1)}$, return the union of points in $ X^{(i+1)}$ which are obtained by blowing down the exceptional $(-1)$-lines on $X^{(i)}$.

L[4] is a projective variety: Return the last adjoint surface $X^{(r)}$.

Note

The function does not check whether X is smooth. If you are uncertain, enter is_smooth(X) first.

Warning

At current state, the adjunction process is only implemented for rational and Enriques surfaces which are linearly normal in the given embedding. The function does not check whether X is rational or an Enriques surface. In fact, at current state, OSCAR does not offer direct checks for this. Note, however, that the adjunction process will give an answer to this question a posteriori in cases where it terminates with a surface which is known to be rational or an Enriques surface.

Examples

julia> X = bordiga()
Projective variety
  in projective 4-space over GF(31991) with coordinates [x, y, z, u, v]
defined by ideal with 4 generators

julia> dim(X)
2

julia> codim(X)
2

julia> L = adjunction_process(X);

julia> L[1]
2-element Vector{NTuple{4, ZZRingElem}}:
 (4, 6, 3, 0)
 (2, 1, 0, 10)

julia> L[4]
Projective variety
  in projective 2-space over GF(31991) with coordinates [z[1], z[2], z[3]]
defined by ideal (0)

julia> L[3][1]
Projective algebraic set
  in projective 2-space over GF(31991) with coordinates [z[1], z[2], z[3]]
defined by ideal with 5 generators

julia> dim(L[3][1])
0

julia> degree(L[3][1])
10

julia> X = rational_d9_pi7();

julia> L = adjunction_process(X);

julia> L[1]
3-element Vector{NTuple{4, ZZRingElem}}:
 (4, 9, 7, 0)
 (6, 9, 4, 6)
 (3, 3, 1, 3)

Note

Inspecting the returned numerical data in the first example above, we see that the Bordiga surface is the blow-up of the projective plane in 10 points, embedded into projective 4-space by the linear system $H = 4L -\sum_{i=1}^{10} E_i$. Here, $L$ is the preimage of a line and the $E_i$ are the exceptional divisors. In the second example, we see from the output that the terminal object of the adjunction process is a Del Pezzo surface in projective 3-space, that is, the blow-up of the projective plane in 6 points. In sum, we see that X is the blow-up of the projective plane in 15 points, embedded into projective 4-space by the linear system $H = 9L - \sum_{i=1}^{6} 3E_i - \sum_{i=7}^{9} 2E_i - \sum_{i=10}^{15} E_i$.

Contact

Please direct questions about this part of OSCAR to the following people:

Wolfram Decker.

You can ask questions in the OSCAR Slack.

Alternatively, you can raise an issue on github.