Nongeneral Type Surfaces in $\mathbb P^4$

Every smooth, projective surface can be embedded in $\mathbb P^5$, but there are constraints on the numerical invariants of a smooth surface in $\mathbb P^4$: The invariants of each such surface $S$ satisfy the double point formula

\[d^2-5d-10(\pi-1)+2(\chi(\mathcal O_S)-K_S^2) = 0.\]

Here, $d$ is the degree of $S$, $\pi$ its sectional genus, $\chi(\mathcal O_S)$ its Euler-Poincare characteristic, and $K_S$ its canonical class. The double point formula is a key ingredient in the proof of a theorem of Ellingsrud and Peskine which states that there are only finitely many families of smooth surfaces in $\mathbb P^4$ which are not of general type. That is, the degree of such surfaces in bounded from above. The best bound known so far is $52$, while examples exist up to degree $15$.

For details, we refer to

[DES93]

[Pop93]

[DS00]

and the references given there.

Below, we present functions which return one hard coded example for each family presented in the first two papers above. Based on these papers, the existence of further families has been shown. Hard coded OSCAR examples for these surfaces are under construction.

Note

To ease subsequent computations, all hard coded examples are defined over a finite prime field.

Rational Surfaces

A Rational Surface with $d=3$, $\pi=0$

cubic_scroll — Method

cubic_scroll()

Return a smooth rational surface in $\mathbb P^4$ with degree 3 and sectional genus 0.

The returned surface is defined over a prime field of characteristic 31991.

A Rational Surface with $d=4$, $\pi=0$

veronese — Method

veronese()

Return a smooth rational surface in $\mathbb P^4$ with degree 4 and sectional genus 0.

The returned surface is defined over a prime field of characteristic 31991.

A Rational Surface with $d=5$, $\pi=2$

castelnuovo — Method

castelnuovo()

Return a smooth rational surface in $\mathbb P^4$ with degree 5 and sectional genus 2.

The returned surface is defined over a prime field of characteristic 31991.

A Rational Surface with $d=6$, $\pi=3$

bordiga — Method

bordiga()

Return a smooth rational surface in $\mathbb P^4$ with degree 6 and sectional genus 3.

The returned surface is defined over a prime field of characteristic 31991.

A Rational Surface with $d=7$, $\pi=4$

rational_d7_pi4 — Method

rational_d7_pi4()

Return a smooth rational surface in $\mathbb P^4$ with degree 7 and sectional genus 4.

The returned surface is defined over a prime field of characteristic 31991.

A Rational Surface with $d=8$, $\pi=5$

rational_d8_pi5 — Method

rational_d8_pi5()

Return a smooth rational surface in $\mathbb P^4$ with degree 8 and sectional genus 5.

The returned surface is defined over a prime field of characteristic 31991.

A Rational Surface with $d=8$, $\pi=6$

rational_d8_pi6 — Method

rational_d8_pi6()

Return a smooth rational surface in $\mathbb P^4$ with degree 8 and sectional genus 6.

The returned surface is defined over a prime field of characteristic 31991.

A Rational Surface with $d=9$, $\pi=6$

rational_d9_pi6 — Method

rational_d9_pi6()

Return a smooth rational surface in $\mathbb P^4$ with degree 9 and sectional genus 6.

The returned surface is defined over a prime field of characteristic 31991.

A Rational Surface with $d=9$, $\pi=7$

rational_d9_pi7 — Method

rational_d9_pi7()

Return a smooth rational surface in $\mathbb P^4$ with degree 9 and sectional genus 7.

The returned surface is defined over a prime field of characteristic 31991.

A Rational Surface with $d=10$, $\pi=8$

rational_d10_pi8 — Method

rational_d10_pi8()

Return a smooth rational surface in $\mathbb P^4$ with degree 10 and sectional genus 8.

The returned surface is defined over a prime field of characteristic 31991.

A Rational Surface with $d=10$, $\pi=9$ which is Contained in one Quartic

rational_d10_pi9_quart_1 — Method

rational_d10_pi9_quart_1()

Return a smooth rational surface in $\mathbb P^4$ with degree 10 and sectional genus 9 which is contained in precisely one quartic.

The returned surface is defined over a prime field of characteristic 31991.

A Rational Surface with $d=10$, $\pi=9$ which is Contained in a Pencil of Quartics

rational_d10_pi9_quart_2 — Method

rational_d10_pi9_quart_2()

Return a smooth rational surface in $\mathbb P^4$ with degree 10 and sectional genus 9 which is contained in a pencil of quartics.

The returned surface is defined over a prime field of characteristic 31991.

A Rational Surface with $d=11$, $\pi=11$, and no 6-Secant

rational_d11_pi11_ss_0 — Method

rational_d11_pi11_ss_0()

Return a smooth rational surface in $\mathbb P^4$ with degree 11, sectional genus 11, and no 6-secant.

The returned surface is defined over a prime field of characteristic 31991.

A Rational Surface with $d=11$, $\pi=11$, and one 6-Secant

rational_d11_pi11_ss_1 — Method

rational_d11_pi11_ss_1()

Return a smooth rational surface in $\mathbb P^4$ with degree 11, sectional genus 11, and one 6-secant.

The returned surface is defined over a prime field of characteristic 31991.

A Rational Surface with $d=11$, $\pi=11$, and Infinitely Many 6-Secants

rational_d11_pi11_ss_inf — Method

rational_d11_pi11_ss_inf()

Return a smooth rational surface in $\mathbb P^4$ with degree 11, sectional genus 11, and infinitely many 6-secants.

The returned surface is defined over a prime field of characteristic 31991.

Ruled Surfaces

A Ruled Surface with $d=5$, $\pi=1$

quintic_elliptic_scroll — Method

quintic_elliptic_scroll()

Return a smooth ruled surface in $\mathbb P^4$ with degree 5 and sectional genus 1.

The returned surface is defined over a prime field of characteristic 31991.

Enriques Surfaces

An Enriques Surface with $d=9$, $\pi=6$

enriques_d9_pi6 — Method

enriques_d9_pi6()

Return a smooth Enriques surface in $\mathbb P^4$ with degree 9 and sectional genus 6.

The returned surface is defined over a prime field of characteristic 31991.

An Enriques Surface with $d=10$, $\pi=8$

enriques_d10_pi8 — Method

enriques_d10_pi8()

Return a smooth Enriques surface in $\mathbb P^4$ with degree 10 and sectional genus 8.

The returned surface is defined over a prime field of characteristic 31991.

An Enriques Surface with $d=11$, $\pi=10$

enriques_d11_pi10 — Method

enriques_d11_pi10()

Return a smooth Enriques surface in $\mathbb P^4$ with degree 11 and sectional genus 10.

The returned surface is defined over a prime field of characteristic 43.

An Enriques Surface with $d=13$, $\pi=16$

enriques_d13_pi16 — Method

enriques_d13_pi16()

Return a smooth Enriques surface in $\mathbb P^4$ with degree 13 and sectional genus 16.

The returned surface is defined over a prime field of characteristic 31991.

An Enriques Surface with $d=13$, $\pi=16$

enriques_d13_pi16_two — Method

enriques_d13_pi16_two()

Return a smooth Enriques surface in $\mathbb P^4$ with degree 13 and sectional genus 16.

The returned surface is defined over a prime field of characteristic 31991.

K3 Surfaces

A K3 Surface with $d=7$, $\pi=5$

k3_d7_pi5 — Function

k3_d7_pi5()

Return a smooth K3 surface in $\mathbb P^4$ with degree 7 and sectional genus 5.

The returned surface is defined over a prime field of characteristic 31991.

A K3 Surface with $d=8$, $\pi=6$

k3_d8_pi6 — Function

k3_d8_pi6()

Return a smooth K3 surface in $\mathbb P^4$ with degree 8 and sectional genus 6.

The returned surface is defined over a prime field of characteristic 31991.

A K3 Surface with $d=9$, $\pi=8$

k3_d9_pi8 — Function

k3_d9_pi8()

Return a smooth K3 surface in $\mathbb P^4$ with degree 9 and sectional genus 8.

The returned surface is defined over a prime field of characteristic 31991.

A K3 Surface with $d=10$, $\pi=9$ which is Contained in one Quartic

k3_d10_pi9_quart_1 — Method

k3_d10_pi9_quart_1()

Return a smooth K3 surface in $\mathbb P^4$ with degree 10 and sectional genus 9 which is contained in precisely one quartic.

The returned surface is defined over a prime field of characteristic 31991.

A K3 Surface with $d=10$, $\pi=9$ which is Contained in a Pencil of Quartics

k3_d10_pi9_quart_2 — Method

k3_d10_pi9_quart_2()

Return a smooth K3 surface in $\mathbb P^4$ with degree 10 and sectional genus 9 which is contained in a pencil of quartics.

The returned surface is defined over a prime field of characteristic 31991.

A K3 Surface with $d=11$, $\pi=11$ and no 6-Secant

k3_d11_pi11_ss_0 — Method

k3_d11_pi11_ss_0()

Return a smooth K3 surface in $\mathbb P^4$ with degree 11, sectional genus 11, and no 6-secant.

The returned surface is defined over a prime field of characteristic 31991.

A K3 Surface with $d=11$, $\pi=11$ and one 6-Secant

k3_d11_pi11_ss_1 — Method

k3_d11_pi11_ss_1()

Return a smooth K3 surface in $\mathbb P^4$ with degree 11, sectional genus 11, and one 6-secant.

The returned surface is defined over a prime field of characteristic 31991.

A K3 Surface with $d=11$, $\pi=11$ and two 6-Secants

k3_d11_pi11_ss_2 — Method

k3_d11_pi11_ss_2()

Return a smooth K3 surface in $\mathbb P^4$ with degree 11, sectional genus 11, and two 6-secants.

The returned surface is defined over a prime field of characteristic 31991.

A K3 Surface with $d=11$, $\pi=11$ and three 6-Secants

k3_d11_pi11_ss_3 — Method

k3_d11_pi11_ss_3()

Return a smooth K3 surface in $\mathbb P^4$ with degree 11, sectional genus 11, and three 6-secants.

The returned surface is defined over a prime field of characteristic 31991.

A K3 Surface with $d=11$, $\pi=12$

k3_d11_pi12 — Method

k3_d11_pi12()

Return a smooth K3 surface in $\mathbb P^4$ with degree 11 and sectional genus 12.

The returned surface is defined over a prime field of characteristic 31991.

A K3 Surface with $d=12$, $\pi=14$

k3_d12_pi14 — Method

k3_d12_pi14()

Return a smooth K3 surface in $\mathbb P^4$ with degree 12 and sectional genus 14.

The returned surface is defined over a prime field of characteristic 31991.

A K3 Surface with $d=13$, $\pi=16$

k3_d13_pi16 — Method

k3_d13_pi16()

Return a smooth K3 surface in $\mathbb P^4$ with degree 13 and sectional genus 16.

The returned surface is defined over a prime field of characteristic 31991.

A K3 Surface with $d=14$, $\pi=19$

k3_d14_pi19 — Method

k3_d14_pi19()

Return a smooth K3 surface in $\mathbb P^4$ with degree 14 and sectional genus 19.

The returned surface is defined over a prime field of characteristic 31991.

Bielliptic Surfaces

A Bielliptic Surface with $d=10$, $\pi=6$

bielliptic_d10_pi6 — Method

bielliptic_d10_pi6()

Return a smooth bielliptic surface in $\mathbb P^4$ with degree 10 and sectional genus 6.

The returned surface is defined over a prime field of characteristic 911.

A Bielliptic Surface with $d=15$, $\pi=21$

bielliptic_d15_pi21 — Method

bielliptic_d15_pi21()

Return a smooth bielliptic surface in $\mathbb P^4$ with degree 15 and sectional genus 21.

The returned surface is defined over a prime field of characteristic 911.

Abelian Surfaces

An Abelian Surface with $d=10$, $\pi=6$

abelian_d10_pi6 — Method

abelian_d10_pi6()

Return a smooth abelian surface in $\mathbb P^4$ with degree 10 and sectional genus 6.

The returned surface is defined over a prime field of characteristic 31991.

An Abelian Surface with $d=15$, $\pi=21$ which is Contained in a Net of Quintics

abelian_d15_pi21_quintic_3 — Method

abelian_d15_pi21_quintic_3()

Return a smooth abelian surface in $\mathbb P^4$ with degree 15 and sectional genus 21 which is contained in a net of quintics.

The returned surface is defined over a prime field of characteristic 31991.

An Abelian Surface with $d=15$, $\pi=21$ which is Contained in one Quintic

abelian_d15_pi21_quintic_1 — Method

abelian_d15_pi21_quintic_1()

Return a smooth abelian surface in $\mathbb P^4$ with degree 15 and sectional genus 21 which is contained in precisely one quintic.

The returned surface is defined over a prime field of characteristic 31991.

Elliptic Surfaces

An Elliptic Surface with $d=7$, $\pi=6$

elliptic_d7_pi6 — Method

elliptic_d7_pi6()

Return a smooth elliptic surface in $\mathbb P^4$ with degree 7 and sectional genus 6.

The returned surface is defined over a prime field of characteristic 31991.

An Elliptic Surface with $d=8$, $\pi=7$

elliptic_d8_pi7 — Method

elliptic_d8_pi7()

Return a smooth elliptic surface in $\mathbb P^4$ with degree 8 and sectional genus 7.

The returned surface is defined over a prime field of characteristic 31991.

An Elliptic Surface with $d=9$, $\pi=7$

elliptic_d9_pi7 — Method

elliptic_d9_pi7()

Return a smooth elliptic surface in $\mathbb P^4$ with degree 9 and sectional genus 7.

The returned surface is defined over a prime field of characteristic 31991.

An Elliptic Surface with $d=10$, $\pi=9$

elliptic_d10_pi9 — Method

elliptic_d10_pi9()

Return a smooth elliptic surface in $\mathbb P^4$ with degree 10 and sectional genus 9.

The returned surface is defined over a prime field of characteristic 31991.

An Elliptic Surface with $d=10$, $\pi=10$

elliptic_d10_pi10 — Method

elliptic_d10_pi10()

Return a smooth elliptic surface in $\mathbb P^4$ with degree 10 and sectional genus 10.

The returned surface is defined over a prime field of characteristic 31991.

An Elliptic Surface with $d=11$, $\pi=12$

elliptic_d11_pi12 — Method

elliptic_d11_pi12()

Return a smooth elliptic surface in $\mathbb P^4$ with degree 11 and sectional genus 12.

The returned surface is defined over a prime field of characteristic 31991.

An Elliptic Surface with $d=12$, $\pi=13$

elliptic_d12_pi13 — Method

elliptic_d12_pi13()

Return a smooth elliptic surface in $\mathbb P^4$ with degree 12 and sectional genus 13.

The returned surface is defined over a prime field of characteristic 31991.

An Elliptic Surface with $d=12$, $\pi=14$ and no 6-Secant

elliptic_d12_pi14_ss_0 — Method

elliptic_d12_pi14_ss_0()

Return a smooth elliptic surface in $\mathbb P^4$ with degree 12, sectional genus 14, and no 6-secant.

The returned surface is defined over a prime field of characteristic 31991.

An Elliptic Surface with $d=12$, $\pi=14$, and Infinitely Many 6-Secants

elliptic_d12_pi14_ss_inf — Method

elliptic_d12_pi14_ss_inf()

Return a smooth elliptic surface in $\mathbb P^4$ with degree 12, sectional genus 14, and infinitely many 6-secants.

The returned surface is defined over a prime field of characteristic 31991.

Contact

Please direct questions about this part of OSCAR to the following people:

Wolfram Decker.

You can ask questions in the OSCAR Slack.

Alternatively, you can raise an issue on github.