Affine Algebras

Homomorphisms of Affine Algebras

Noether Normalization

Normalization of Rings

Integral Bases

Tests on Affine Algebras

Hilbert Series and Hilbert Polynomial

Affine Algebras

An affine algebra over a field $K$, or an affine $K$-algebra, or simply an affine algebra, is the quotient $A=R/I$ of a multivariate polynomial ring $R$ over $K$ modulo an ideal $I$ of $R$. The general constructor for quotients of multivariate polynomial rings modulo ideals allows one to create such algebras.

Homomorphisms of Affine Algebras

Constructors

AlgebraHomomorphism — Method

AlgebraHomomorphism(D::U, C::W, V::Vector{X}) where 
{T, S <: MPolyElem{T},
U <: Union{MPolyRing{T}, MPolyQuo{S}},
W <: Union{MPolyRing{T}, MPolyQuo{S}},
X <: Union{S, MPolyQuoElem{S}}}

Create the algebra homomorphism $D \rightarrow C$ defined by sending the $i$th generator of $D$ to the $i$th element of $V$. Allow types MPolyRing and MPolyQuo for $C$ and $D$ as well as entries of type MPolyElem and MPolyQuoElem for X. Alternatively, use hom(D::U, C::W, V::Vector{X}).

Data Associated to Homomorphisms of Affine Algebras

domain — Method

function domain(F::AlgHom)

Return the domain of F.

codomain — Method

function codomain(F::AlgHom)

Return the codomain of F.

preimage — Method

preimage(F::AlgHom, I::U) where U <: Union{MPolyIdeal, MPolyQuoIdeal}

Return the preimage of the ideal $I$ under the algebra homomorphism $F$.

kernel — Method

kernel(F::AlgHom)

Return the kernel of the algebra homomorphism $F$.

Examples

julia> D1, (w, x, y, z) = PolynomialRing(QQ, ["w", "x", "y", "z"])
(Multivariate Polynomial Ring in w, x, y, z over Rational Field, fmpq_mpoly[w, x, y, z])

julia> C1, (s,t) = PolynomialRing(QQ, ["s", "t"])
(Multivariate Polynomial Ring in s, t over Rational Field, fmpq_mpoly[s, t])

julia> V1 = [s^3, s^2*t, s*t^2, t^3]
4-element Array{fmpq_mpoly,1}:
 s^3
 s^2*t
 s*t^2
 t^3

julia> para = hom(D1, C1, V1)
Algebra homomorphism with

domain: Multivariate Polynomial Ring in w, x, y, z over Rational Field

codomain: Multivariate Polynomial Ring in s, t over Rational Field

defining images of generators: fmpq_mpoly[s^3, s^2*t, s*t^2, t^3]

julia> twistedCubic = kernel(para)
ideal generated by: -x*z + y^2, -w*z + x*y, -w*y + x^2

julia> C2, _ = quo(D1, twistedCubic)
(Quotient of Multivariate Polynomial Ring in w, x, y, z over Rational Field by ideal generated by: -x*z + y^2, -w*z + x*y, -w*y + x^2, Map from
Multivariate Polynomial Ring in w, x, y, z over Rational Field to Quotient of Multivariate Polynomial Ring in w, x, y, z over Rational Field by ideal generated by: -x*z + y^2, -w*z + x*y, -w*y + x^2 defined by a julia-function with inverse
)

julia> D2, (a, b, b) = PolynomialRing(QQ, ["a", "b", "c"])
(Multivariate Polynomial Ring in a, b, c over Rational Field, fmpq_mpoly[a, b, c])

julia> V2 = [w-y, x, z]
3-element Array{fmpq_mpoly,1}:
 w - y
 x
 z

julia> proj = hom(D2, C2, V2)
Algebra homomorphism with

domain: Multivariate Polynomial Ring in a, b, c over Rational Field

codomain: Quotient of Multivariate Polynomial Ring in w, x, y, z over Rational Field by ideal generated by: -x*z + y^2, -w*z + x*y, -w*y + x^2

defining images of generators: fmpq_mpoly[w - y, x, z]

julia> nodalCubic = kernel(proj)
ideal generated by: -a^2*c + b^3 - 2*b^2*c + b*c^2

julia> D3,y = PolynomialRing(QQ, :y => 1:3)
(Multivariate Polynomial Ring in y[1], y[2], y[3] over Rational Field, fmpq_mpoly[y[1], y[2], y[3]])

julia> C3, x = PolynomialRing(QQ, :x => 1:3)
(Multivariate Polynomial Ring in x[1], x[2], x[3] over Rational Field, fmpq_mpoly[x[1], x[2], x[3]])

julia> V3 = [x[1]*x[2], x[1]*x[3], x[2]*x[3]]
3-element Array{fmpq_mpoly,1}:
 x[1]*x[2]
 x[1]*x[3]
 x[2]*x[3]

julia> F3 = hom(D3, C3, V3)
Algebra homomorphism with

domain: Multivariate Polynomial Ring in y[1], y[2], y[3] over Rational Field

codomain: Multivariate Polynomial Ring in x[1], x[2], x[3] over Rational Field

defining images of generators: fmpq_mpoly[x[1]*x[2], x[1]*x[3], x[2]*x[3]]

julia> sphere = ideal(C3, [x[1]^3 + x[2]^3  + x[3]^3 - 1])
ideal generated by: x[1]^3 + x[2]^3 + x[3]^3 - 1

julia> steinerRomanSurface = preimage(F3, sphere)
ideal generated by: y[1]^6*y[2]^6 + 2*y[1]^6*y[2]^3*y[3]^3 + y[1]^6*y[3]^6 + 2*y[1]^3*y[2]^6*y[3]^3 + 2*y[1]^3*y[2]^3*y[3]^6 - y[1]^3*y[2]^3*y[3]^3 + y[2]^6*y[3]^6

Tests on Homomorphisms of Affine Algebras

isinjective — Method

isinjective(F::AlgHom)

Return true if F is injective, false otherwise.

issurjective — Method

issurjective(F::AlgHom)

Return true if F is surjective, false otherwise.

isbijective — Method

isbijective(F::AlgHom)

Return true if F is bijective, false otherwise.

isfinite — Method

isfinite(F::AlgHom)

Return true if F is finite, false otherwise.

Examples

julia> D, (x, y, z) = PolynomialRing(QQ, ["x", "y", "z"])
(Multivariate Polynomial Ring in x, y, z over Rational Field, fmpq_mpoly[x, y, z])

julia> S, (a, b, c) = PolynomialRing(QQ, ["a", "b", "c"])
(Multivariate Polynomial Ring in a, b, c over Rational Field, fmpq_mpoly[a, b, c])

julia> C, p = quo(S, ideal(S, [c-b^3]))
(Quotient of Multivariate Polynomial Ring in a, b, c over Rational Field by ideal generated by: -b^3 + c, Map from
Multivariate Polynomial Ring in a, b, c over Rational Field to Quotient of Multivariate Polynomial Ring in a, b, c over Rational Field by ideal generated by: -b^3 + c defined by a julia-function with inverse
)

julia> V = [p(2*a + b^6), p(7*b - a^2), p(c^2)]
3-element Array{MPolyQuoElem{fmpq_mpoly},1}:
 2*a + b^6
 -a^2 + 7*b
 c^2

julia> F = hom(D, C, V)
Algebra homomorphism with

domain: Multivariate Polynomial Ring in x, y, z over Rational Field

codomain: Quotient of Multivariate Polynomial Ring in a, b, c over Rational Field by ideal generated by: -b^3 + c

defining images of generators: MPolyQuoElem{fmpq_mpoly}[2*a + b^6, -a^2 + 7*b, c^2]

julia> issurjective(F)
true

julia> D1, _ = quo(D, kernel(F))
(Quotient of Multivariate Polynomial Ring in x, y, z over Rational Field by ideal generated by: x^12 - 12*x^11*z + 24*x^10*y + 66*x^10*z^2 - 240*x^9*y*z - 220*x^9*z^3 + 240*x^8*y^2 + 1080*x^8*y*z^2 + 495*x^8*z^4 - 1920*x^7*y^2*z - 2880*x^7*y*z^3 - 792*x^7*z^5 + 1280*x^6*y^3 + 6720*x^6*y^2*z^2 + 5040*x^6*y*z^4 + 924*x^6*z^6 - 7680*x^5*y^3*z - 13440*x^5*y^2*z^3 - 6048*x^5*y*z^5 - 792*x^5*z^7 + 3840*x^4*y^4 + 19200*x^4*y^3*z^2 + 16800*x^4*y^2*z^4 + 5040*x^4*y*z^6 + 495*x^4*z^8 - 15360*x^3*y^4*z - 25600*x^3*y^3*z^3 - 13440*x^3*y^2*z^5 - 2880*x^3*y*z^7 - 220*x^3*z^9 + 6144*x^2*y^5 + 23040*x^2*y^4*z^2 + 19200*x^2*y^3*z^4 + 6720*x^2*y^2*z^6 + 1080*x^2*y*z^8 + 66*x^2*z^10 - 12288*x*y^5*z - 15360*x*y^4*z^3 - 7680*x*y^3*z^5 - 1920*x*y^2*z^7 - 240*x*y*z^9 - 12*x*z^11 + 4096*y^6 + 6144*y^5*z^2 + 3840*y^4*z^4 + 1280*y^3*z^6 + 240*y^2*z^8 + 24*y*z^10 + z^12 - 481890304*z, Map from
Multivariate Polynomial Ring in x, y, z over Rational Field to Quotient of Multivariate Polynomial Ring in x, y, z over Rational Field by ideal generated by: x^12 - 12*x^11*z + 24*x^10*y + 66*x^10*z^2 - 240*x^9*y*z - 220*x^9*z^3 + 240*x^8*y^2 + 1080*x^8*y*z^2 + 495*x^8*z^4 - 1920*x^7*y^2*z - 2880*x^7*y*z^3 - 792*x^7*z^5 + 1280*x^6*y^3 + 6720*x^6*y^2*z^2 + 5040*x^6*y*z^4 + 924*x^6*z^6 - 7680*x^5*y^3*z - 13440*x^5*y^2*z^3 - 6048*x^5*y*z^5 - 792*x^5*z^7 + 3840*x^4*y^4 + 19200*x^4*y^3*z^2 + 16800*x^4*y^2*z^4 + 5040*x^4*y*z^6 + 495*x^4*z^8 - 15360*x^3*y^4*z - 25600*x^3*y^3*z^3 - 13440*x^3*y^2*z^5 - 2880*x^3*y*z^7 - 220*x^3*z^9 + 6144*x^2*y^5 + 23040*x^2*y^4*z^2 + 19200*x^2*y^3*z^4 + 6720*x^2*y^2*z^6 + 1080*x^2*y*z^8 + 66*x^2*z^10 - 12288*x*y^5*z - 15360*x*y^4*z^3 - 7680*x*y^3*z^5 - 1920*x*y^2*z^7 - 240*x*y*z^9 - 12*x*z^11 + 4096*y^6 + 6144*y^5*z^2 + 3840*y^4*z^4 + 1280*y^3*z^6 + 240*y^2*z^8 + 24*y*z^10 + z^12 - 481890304*z defined by a julia-function with inverse
)

julia> F1 = hom(D1, C, V)
Algebra homomorphism with

domain: Quotient of Multivariate Polynomial Ring in x, y, z over Rational Field by ideal generated by: x^12 - 12*x^11*z + 24*x^10*y + 66*x^10*z^2 - 240*x^9*y*z - 220*x^9*z^3 + 240*x^8*y^2 + 1080*x^8*y*z^2 + 495*x^8*z^4 - 1920*x^7*y^2*z - 2880*x^7*y*z^3 - 792*x^7*z^5 + 1280*x^6*y^3 + 6720*x^6*y^2*z^2 + 5040*x^6*y*z^4 + 924*x^6*z^6 - 7680*x^5*y^3*z - 13440*x^5*y^2*z^3 - 6048*x^5*y*z^5 - 792*x^5*z^7 + 3840*x^4*y^4 + 19200*x^4*y^3*z^2 + 16800*x^4*y^2*z^4 + 5040*x^4*y*z^6 + 495*x^4*z^8 - 15360*x^3*y^4*z - 25600*x^3*y^3*z^3 - 13440*x^3*y^2*z^5 - 2880*x^3*y*z^7 - 220*x^3*z^9 + 6144*x^2*y^5 + 23040*x^2*y^4*z^2 + 19200*x^2*y^3*z^4 + 6720*x^2*y^2*z^6 + 1080*x^2*y*z^8 + 66*x^2*z^10 - 12288*x*y^5*z - 15360*x*y^4*z^3 - 7680*x*y^3*z^5 - 1920*x*y^2*z^7 - 240*x*y*z^9 - 12*x*z^11 + 4096*y^6 + 6144*y^5*z^2 + 3840*y^4*z^4 + 1280*y^3*z^6 + 240*y^2*z^8 + 24*y*z^10 + z^12 - 481890304*z

codomain: Quotient of Multivariate Polynomial Ring in a, b, c over Rational Field by ideal generated by: -b^3 + c

defining images of generators: MPolyQuoElem{fmpq_mpoly}[2*a + b^6, -a^2 + 7*b, c^2]

julia> isbijective(F1)
true

julia> R, (x, y, z) = PolynomialRing(QQ, [ "x", "y", "z"])
(Multivariate Polynomial Ring in x, y, z over Rational Field, fmpq_mpoly[x, y, z])

julia> C, (s, t) = PolynomialRing(QQ, ["s", "t"])
(Multivariate Polynomial Ring in s, t over Rational Field, fmpq_mpoly[s, t])

julia> V = [s*t, t, s^2]
3-element Array{fmpq_mpoly,1}:
 s*t
 t
 s^2

julia> paraWhitneyUmbrella = hom(R, C, V)
Algebra homomorphism with

domain: Multivariate Polynomial Ring in x, y, z over Rational Field

codomain: Multivariate Polynomial Ring in s, t over Rational Field

defining images of generators: fmpq_mpoly[s*t, t, s^2]

julia> D, _ = quo(R, kernel(paraWhitneyUmbrella))
(Quotient of Multivariate Polynomial Ring in x, y, z over Rational Field by ideal generated by: -x^2 + y^2*z, Map from
Multivariate Polynomial Ring in x, y, z over Rational Field to Quotient of Multivariate Polynomial Ring in x, y, z over Rational Field by ideal generated by: -x^2 + y^2*z defined by a julia-function with inverse
)

julia> isfinite(hom(D, C, V))
true

Composition of Homomorphisms of Affine Algebras

compose — Method

compose(F::AlgHom{T}, G::AlgHom{T}) where T

Return the algebra homomorphism $H = G\circ F: domain(F) \rightarrow codomain(G)$.

Noether Normalization

noether_normalization — Method

noether_normalization(A::MPolyQuo)

Given an affine algebra $A=R/I$ over a field $K$, the return value is a triple $(L,f,g)$, such that: $L$ is an array of $d=\dim A$ elements of $A$, all represented by linear forms in $R$, and such that $K[L]\rightarrow A$ is a Noether normalization for $A$; $f: A \rightarrow A$ is an automorphism of $A$, induced by a linear change of coordinates of $R$, and mapping the $f_i$ to the last $d$ variables of $A$; and $g = f^{-1}$.

CAVEAT: The algorithm may not terminate over a small finite field. If it terminates, the result is correct.

Example

julia> R, (x, y, z) = PolynomialRing(QQ, ["x", "y", "z"])
(Multivariate Polynomial Ring in x, y, z over Rational Field, fmpq_mpoly[x, y, z])

julia> A, _ = quo(R, ideal(R, [x*y, x*z]))
(Quotient of Multivariate Polynomial Ring in x, y, z over Rational Field by ideal generated by: x*y, x*z, Map from
Multivariate Polynomial Ring in x, y, z over Rational Field to Quotient of Multivariate Polynomial Ring in x, y, z over Rational Field by ideal generated by: x*y, x*z defined by a julia-function with inverse
)

julia> L = noether_normalization(A);

julia> L[1]
2-element Array{MPolyQuoElem{fmpq_mpoly},1}:
 -3*x + y
 21*x - 8*y + z

julia> L[2]
Algebra homomorphism with

domain: Quotient of Multivariate Polynomial Ring in x, y, z over Rational Field by ideal generated by: x*y, x*z

codomain: Quotient of Multivariate Polynomial Ring in x, y, z over Rational Field by ideal generated by: x*y, x*z

defining images of generators: MPolyQuoElem{fmpq_mpoly}[x, 3*x + y, 3*x + 8*y + z]

julia> L[3]
Algebra homomorphism with

domain: Quotient of Multivariate Polynomial Ring in x, y, z over Rational Field by ideal generated by: x*y, x*z

codomain: Quotient of Multivariate Polynomial Ring in x, y, z over Rational Field by ideal generated by: x*y, x*z

defining images of generators: MPolyQuoElem{fmpq_mpoly}[x, -3*x + y, 21*x - 8*y + z]

Normalization of Rings

normalization — Method

normalization(A::MPolyQuo; alg=:equidimDec)

Find the normalization of a reduced affine algebra over a perfect field $K$. That is, given the quotient $A=R/I$ of a multivariate polynomial ring $R$ over $K$ modulo a radical ideal $I$, compute the integral closure $\overline{A}$ of $A$ in its total ring of fractions $Q(A)$, together with the embedding $f: A \rightarrow \overline{A}$.

Implemented Algorithms and how to Read the Output

The function relies on the algorithm of Greuel, Laplagne, and Seelisch which proceeds by finding a suitable decomposition $I=I_1\cap\dots\cap I_r$ into radical ideals $I_k$, together with the normalization maps $f_k: R/I_k \rightarrow A_k=\overline{R/I_k}$, such that

\[f=f_1\times \dots\times f_r: A \rightarrow A_1\times \dots\times A_r=\overline{A}\]

is the normalization map of $A$. For each $k$, the function specifies two representations of $A_k$: It returns an array of triples $(A_k, f_k, \mathfrak a_k)$, where $A_k$ is represented as an affine $K$-algebra, and $f_k$ as a map of affine $K$-algebras. The third entry $\mathfrak a_k$ is a tuple $(d_k, J_k)$, consisting of an element $d_k\in A$ and an ideal $J_k\subset A$, such that $\frac{1}{d_k}J_k = A_k$ as $A$-submodules of the total ring of fractions of $A$.

By default, as a first step on its way to find the decomposition $I=I_1\cap\dots\cap I_r$, the algorithm computes an equidimensional decomposition of the radical ideal $I$. Alternatively, if specified by alg=:primeDec, the algorithm computes $I=I_1\cap\dots\cap I_r$ as the prime decomposition of the radical ideal $I$.

See Gert-Martin Greuel , Santiago Laplagne , Frank Seelisch (2010).

CAVEAT: The function does not check whether $A$ is reduced. Use isreduced(A) in case you are unsure (this may take some time).

normalization_with_delta — Method

normalization_with_delta(A::MPolyQuo; alg=:equidimDec)

Compute the normalization

\[f=f_1\times \dots\times f_r: A \rightarrow A_1\times \dots\times A_r=\overline{A}\]

of $A$ as does normalize(A), but return additionally the delta invariant of $A$, that is, the dimension

\[\dim_K(\overline{A}/A)\]

.

How to Read the Output

The return value is a tuple whose first element is normalize(A), whose second element is an array containing the delta invariants of the $A_k$, and whose third element is the (total) delta invariant of $A$. The return value -1 in the third element indicates that the delta invariant is infinite.

Examples

julia> R, (x, y) = PolynomialRing(QQ, ["x", "y"])
(Multivariate Polynomial Ring in x, y over Rational Field, fmpq_mpoly[x, y])

julia> A, _ = quo(R, ideal(R, [(x^2-y^3)*(x^2+y^2)*x]))
(Quotient of Multivariate Polynomial Ring in x, y over Rational Field by ideal generated by: x^5 - x^3*y^3 + x^3*y^2 - x*y^5, Map from
Multivariate Polynomial Ring in x, y over Rational Field to Quotient of Multivariate Polynomial Ring in x, y over Rational Field by ideal generated by: x^5 - x^3*y^3 + x^3*y^2 - x*y^5 defined by a julia-function with inverse
)

julia> L = normalization(A)
2-element Array{Tuple{MPolyQuo{fmpq_mpoly},Oscar.AlgHom{fmpq},Tuple{MPolyQuoElem{fmpq_mpoly},MPolyQuoIdeal{fmpq_mpoly}}},1}:
 (Quotient of Multivariate Polynomial Ring in T(1), T(2), T(3), x, y over Rational Field by ideal generated by: 2*T(2)*y + T(3)*y + x, T(1)*x - T(2)*y - T(3)*y, -T(1)*y + T(2)*x - y, T(1)*y^2 - T(1)*y - T(3)*x - y, T(2)*y^3 + T(2)*y^2 - T(3)*x^2 + T(3)*y^3, T(1)^2 + T(1), T(1)*T(2), T(1) + T(2)^2 + 1, T(1)*T(3) + T(2) + T(3), -T(1) + T(2)*T(3) - 1, T(1)*y + T(1) + T(3)^2 + 1, -x^4 + x^2*y^3 - x^2*y^2 + y^5, Algebra homomorphism with

domain: Quotient of Multivariate Polynomial Ring in x, y over Rational Field by ideal generated by: x^5 - x^3*y^3 + x^3*y^2 - x*y^5

codomain: Quotient of Multivariate Polynomial Ring in T(1), T(2), T(3), x, y over Rational Field by ideal generated by: 2*T(2)*y + T(3)*y + x, T(1)*x - T(2)*y - T(3)*y, -T(1)*y + T(2)*x - y, T(1)*y^2 - T(1)*y - T(3)*x - y, T(2)*y^3 + T(2)*y^2 - T(3)*x^2 + T(3)*y^3, T(1)^2 + T(1), T(1)*T(2), T(1) + T(2)^2 + 1, T(1)*T(3) + T(2) + T(3), -T(1) + T(2)*T(3) - 1, T(1)*y + T(1) + T(3)^2 + 1, -x^4 + x^2*y^3 - x^2*y^2 + y^5

defining images of generators: MPolyQuoElem{fmpq_mpoly}[x, y]
, (-2*x^3 + x*y^3 - x*y^2, MPolyQuoIdeal(x^3 + x*y^2, -x^2*y + y^4, x^2*y^2 + x^2*y, -2*x^3 + x*y^3 - x*y^2)))
 (Quotient of Multivariate Polynomial Ring in x, y over Rational Field by ideal generated by: x, Algebra homomorphism with

domain: Quotient of Multivariate Polynomial Ring in x, y over Rational Field by ideal generated by: x^5 - x^3*y^3 + x^3*y^2 - x*y^5

codomain: Quotient of Multivariate Polynomial Ring in x, y over Rational Field by ideal generated by: x

defining images of generators: MPolyQuoElem{fmpq_mpoly}[x, y]
, (1, MPolyQuoIdeal(1)))

julia> R, (x, y, z) = PolynomialRing(QQ, ["x", "y", "z"])
(Multivariate Polynomial Ring in x, y, z over Rational Field, fmpq_mpoly[x, y, z])

julia> A, _ = quo(R, ideal(R, [z^3-x*y^4]))
(Quotient of Multivariate Polynomial Ring in x, y, z over Rational Field by ideal generated by: -x*y^4 + z^3, Map from
Multivariate Polynomial Ring in x, y, z over Rational Field to Quotient of Multivariate Polynomial Ring in x, y, z over Rational Field by ideal generated by: -x*y^4 + z^3 defined by a julia-function with inverse
)

julia> L = normalization(A)
1-element Array{Tuple{MPolyQuo{fmpq_mpoly},Oscar.AlgHom{fmpq},Tuple{MPolyQuoElem{fmpq_mpoly},MPolyQuoIdeal{fmpq_mpoly}}},1}:
 (Quotient of Multivariate Polynomial Ring in T(1), T(2), x, y, z over Rational Field by ideal generated by: T(1)*y - T(2)*z, T(2)*y - z, -T(1)*z + x*y^2, T(1)^2 - x*z, T(1)*T(2) - x*y, -T(1) + T(2)^2, x*y^4 - z^3, Algebra homomorphism with

domain: Quotient of Multivariate Polynomial Ring in x, y, z over Rational Field by ideal generated by: -x*y^4 + z^3

codomain: Quotient of Multivariate Polynomial Ring in T(1), T(2), x, y, z over Rational Field by ideal generated by: T(1)*y - T(2)*z, T(2)*y - z, -T(1)*z + x*y^2, T(1)^2 - x*z, T(1)*T(2) - x*y, -T(1) + T(2)^2, x*y^4 - z^3

defining images of generators: MPolyQuoElem{fmpq_mpoly}[x, y, z]
, (z^2, MPolyQuoIdeal(x*y^2*z, x*y^3, z^2)))

julia> R, (x, y) = PolynomialRing(QQ, ["x", "y"])
(Multivariate Polynomial Ring in x, y over Rational Field, fmpq_mpoly[x, y])

julia> A, _ = quo(R, ideal(R, [(x^2-y^3)*(x^2+y^2)*x]))
(Quotient of Multivariate Polynomial Ring in x, y over Rational Field by ideal generated by: x^5 - x^3*y^3 + x^3*y^2 - x*y^5, Map from
Multivariate Polynomial Ring in x, y over Rational Field to Quotient of Multivariate Polynomial Ring in x, y over Rational Field by ideal generated by: x^5 - x^3*y^3 + x^3*y^2 - x*y^5 defined by a julia-function with inverse
)

julia> L = normalization_with_delta(A)
(Tuple{MPolyQuo{fmpq_mpoly},Oscar.AlgHom{fmpq},Tuple{MPolyQuoElem{fmpq_mpoly},MPolyQuoIdeal{fmpq_mpoly}}}[(Quotient of Multivariate Polynomial Ring in T(1), x, y over Rational Field by ideal generated by: -T(1)*y + x, -T(1)*x + y^2, T(1)^2 - y, -x^2 + y^3, Algebra homomorphism with

domain: Quotient of Multivariate Polynomial Ring in x, y over Rational Field by ideal generated by: x^5 - x^3*y^3 + x^3*y^2 - x*y^5

codomain: Quotient of Multivariate Polynomial Ring in T(1), x, y over Rational Field by ideal generated by: -T(1)*y + x, -T(1)*x + y^2, T(1)^2 - y, -x^2 + y^3

defining images of generators: MPolyQuoElem{fmpq_mpoly}[x, y]
, (y, MPolyQuoIdeal(x, y))), (Quotient of Multivariate Polynomial Ring in T(1), x, y over Rational Field by ideal generated by: -T(1)*y + x, T(1)*x + y, T(1)^2 + 1, x^2 + y^2, Algebra homomorphism with

domain: Quotient of Multivariate Polynomial Ring in x, y over Rational Field by ideal generated by: x^5 - x^3*y^3 + x^3*y^2 - x*y^5

codomain: Quotient of Multivariate Polynomial Ring in T(1), x, y over Rational Field by ideal generated by: -T(1)*y + x, T(1)*x + y, T(1)^2 + 1, x^2 + y^2

defining images of generators: MPolyQuoElem{fmpq_mpoly}[x, y]
, (y, MPolyQuoIdeal(x, y))), (Quotient of Multivariate Polynomial Ring in x, y over Rational Field by ideal generated by: x, Algebra homomorphism with

domain: Quotient of Multivariate Polynomial Ring in x, y over Rational Field by ideal generated by: x^5 - x^3*y^3 + x^3*y^2 - x*y^5

codomain: Quotient of Multivariate Polynomial Ring in x, y over Rational Field by ideal generated by: x

defining images of generators: MPolyQuoElem{fmpq_mpoly}[x, y]
, (1, MPolyQuoIdeal(1)))], [1, 1, 0], 13)

Integral Bases

integral_basis — Method

integral_basis(f::MPolyElem, i::Int)

Given a polynomial $f$ in two variables with rational coefficients and an integer $i\in\{1,2\}$ specifying one of the variables, $f$ must be irreducible and monic in the specified variable: Say, $f\in\mathbb Q[x,y]$ is monic in $y$. Then the normalization of $A = Q[x,y]/\langle f \rangle$, that is, the integral closure $\overline{A}$ of $A$ in its quotient field, is a free module over $K[x]$ of finite rank, and any set of free generators for $\overline{A}$ over $K[x]$ is called an integral basis for $\overline{A}$ over $K[x]$. Relying on the algorithm by B"ohm, Decker, Laplagne, and Pfister, the function returns a pair $(d, V)$, where $d$ is an element of $A$, and $V$ is a vector of elements in $A$, such that the fractions $v/d, v\in V$, form an integral basis for $\overline{A}$ over $K[x]$.

NOTE: The conditions on $f$ are automatically checked.

Example

julia> R, (x, y) = PolynomialRing(QQ, ["x", "y"])
(Multivariate Polynomial Ring in x, y over Rational Field, fmpq_mpoly[x, y])

julia> f = (y^2-2)^2 + x^5
x^5 + y^4 - 4*y^2 + 4

julia> integral_basis(f, 2)
(x^2, fmpq_mpoly[x^2, x^2*y, y^2 - 2, y^3 - 2*y])

Tests on Affine Algebras

Subalgebra Membership

subalgebra_membership — Method

subalgebra_membership(f::S, v::Vector{S}) where S <: Union{MPolyElem{U}, MPolyQuoElem{U}} where U

If f is contained in the subalgebra of A generated by the elements contained in v, return (true, h), where h is giving the polynomial relation. Otherwise, return (false, 0).

julia> R, x = PolynomialRing(QQ, :x => 1:3)
(Multivariate Polynomial Ring in x[1], x[2], x[3] over Rational Field, fmpq_mpoly[x[1], x[2], x[3]])

julia> f = x[1]^6*x[2]^6-x[1]^6*x[3]^6;

julia> v = [x[1]^3*x[2]^3-x[1]^3*x[3]^3, x[1]^3*x[2]^3+x[1]^3*x[3]^3]
2-element Array{fmpq_mpoly,1}:
 x[1]^3*x[2]^3 - x[1]^3*x[3]^3
 x[1]^3*x[2]^3 + x[1]^3*x[3]^3

julia> subalgebra_membership(f,v)
(true, t_1*t_2)

Normality Test

isnormal — Method

isnormal(A::MPolyQuo)

Given an affine algebra A over a perfect field, return true if A is normal, false otherwise.

CAVEAT: The function computes the normalization of A. This may take some time.

Example

julia> R, (x, y, z) = PolynomialRing(QQ, ["x", "y", "z"])
(Multivariate Polynomial Ring in x, y, z over Rational Field, fmpq_mpoly[x, y, z])

julia> A, _ = quo(R, ideal(R, [z^2-x*y]))
(Quotient of Multivariate Polynomial Ring in x, y, z over Rational Field by ideal generated by: -x*y + z^2, Map from
Multivariate Polynomial Ring in x, y, z over Rational Field to Quotient of Multivariate Polynomial Ring in x, y, z over Rational Field by ideal generated by: -x*y + z^2 defined by a julia-function with inverse
)

julia> isnormal(A)
false

Cohen-Macaulayness Test

iscohenmacaulay(R)

Example

Hilbert Series and Hilbert Polynomial

Let $R=K[x_1, \dots x_n]$ be a polynomial ring in $n$ variables over a field $K$. Assign positive integer weights $w_i$ to the variables $x_i$, and grade $R=\bigoplus_{d\geq 0} R_d$ according to the corresponding weighted degree. Let $I$ be an ideal of $R$ which is homogeneous with respect to this grading. Then the affine $K$-algebra $A=R/I$ inherits the grading: $A = \bigoplus_{d\geq 0} A_d$, where each graded piece $A_d$ is a finite dimensional $K$-vector space. In this situation, the Hilbert function of $A$ is the function

\[H(A, \underline{\phantom{d}}): \N \rightarrow \N, d \mapsto \dim_K(d).\]

The Hilbert series of $A$ is the generating function

\[H_A(t)=\sum_{d\geq 0} H(A, d) t^d.\]

It can be written as a rational function in $t$, say, with denominator

\[(1-t^{w_1})\cdots (1-t^{w_n}).\]

Now suppose that the weights on the variables are all 1. Then we also have the Hilbert polynomial $P_A(t)\in\mathbb{Q}[t]$ which satisfies $H(M,d)=P_M(d)$ for all $d \gg 0$. Furthermore, the degree of $A$ is defined as the dimension of $A$ over $K$ if this dimension is finite, and as the integer $d$ such that the leading term of the Hilbert polynomial has the form $d t^e/e!$, otherwise.

CAVEAT: Currently only implemented in the case where the weights on the variables are all 1.

hilbert_series — Method

hilbert_series(A::Union{MPolyRing, MPolyQuo})

Given a graded affine algebra $A$ over a field $K$, return a pair $(p,q)$, say, of univariate polynomials in $t$ with integer coefficients such that: If $A = R/I$, where $R$ is a multivariate polynomial ring in $n$ variables over $K$, with positive integer weights $w_1, \dots, w_n$ assigned to the variables, and $I$ is a homogeneous ideal of $R$ when we grade $R$ according to the corresponding weighted degree, then $p/q$ represents the Hilbert series of $A$ as a rational function with denominator $q = (1-t^{w_1})\cdots (1-t^{w_n})$.

See also hilbert_series_reduced.

hilbert_series_reduced — Method

hilbert_series_reduced(A::Union{MPolyRing, MPolyQuo})

Given a graded affine algebra $A$ over a field $K$, return a pair $(p,q)$, say, of univariate polynomials in $t$ with integer coefficients such that: If $A = R/I$, where $R$ is a multivariate polynomial ring in $n$ variables over $K$, with positive integer weights $w_1, \dots, w_n$ assigned to the variables, and $I$ is a homogeneous ideal of $R$ when we grade $R$ according to the corresponding weighted degree, then $p/q$ represents the Hilbert series of $A$ as a rational function written in lowest terms.

See also hilbert_series.

hilbert_series_expanded — Method

hilbert_series_expanded(A::Union{MPolyRing, MPolyQuo}, d::Int)

Given a graded affine algebra $A = R/I$ over a field $K$ and an integer $d\geq 0$, return the Hilbert series of $A$ to precision $d$.

hilbert_function — Method

hilbert_function(A::Union{MPolyRing, MPolyQuo}, d::Int)

Given a graded affine algebra $A = R/I$ over a field $K$ and an integer $d\geq 0$, return the value $H(A, d)$, where $H(A, \underline{\phantom{d}}): \N \rightarrow \N, d \mapsto \dim_K A_d$ is the Hilbert function of $A$.

hilbert_polynomial — Method

 hilbert_polynomial(A::Union{MPolyRing, MPolyQuo})

Given a graded affine algebra $A = R/I$ over a field $K$ such that the weights on the variables are all 1, return the Hilbert polynomial of $A$.

degree — Method

degree(A::Union{MPolyRing, MPolyQuo})

Given a graded affine algebra $A = R/I$ over a field $K$ such that the weights on the variables are all 1, return the degree of $A$.

Examples

julia> R, (w, x, y, z) = GradedPolynomialRing(QQ, ["w", "x", "y", "z"])
(Multivariate Polynomial Ring in w, x, y, z over Rational Field graded by
  w -> [1]
  x -> [1]
  y -> [1]
  z -> [1], MPolyElem_dec{fmpq,fmpq_mpoly}[w, x, y, z])

julia> A, _ = quo(R, ideal(R, [w*y-x^2, w*z-x*y, x*z-y^2]));

julia> hilbert_series(A)
(2*t^3 - 3*t^2 + 1, t^4 - 4*t^3 + 6*t^2 - 4*t + 1)

julia> hilbert_series_reduced(A)
(2*t + 1, t^2 - 2*t + 1)

julia> hilbert_series_expanded(A, 7)
1 + 4*t + 7*t^2 + 10*t^3 + 13*t^4 + 16*t^5 + 19*t^6 + 14*t^7 + O(t^7)

julia> hilbert_function(A,7)
14

julia> hilbert_polynomial(A)
3*t + 1

julia> degree(A)
3