Binomial Primary Decomposition

Introduction

Basic Tests

Cellular Decomposition

Primary Decomposition of Cellular Ideals

Primary Decomposition of Binomial Ideals

Binomial Primary Decomposition

Introduction

A binomial is a polynomial consisting of at most two terms. A binomial ideal is an ideal which can be generated by binomials. A binomial primary decomposition is a primary decomposition of a binomial ideal into primary ideals which are binomial as well. In this section, we discuss functionality for computing such decompositions.

The crucial first step of the binomial primary decomposition algorithm is to decompose a given binomial ideal into cellular binomial ideals. Here, recall that a proper binomial ideal $I$ of a polynomial ring $R = K[x_1, \dots, x_n]$ is called cellular if each variable $x_i$ is either a nonzerodivisor or nilpotent modulo $I$.

Note

Current status: characteristic zero, pure binomial ideals

Papers offering details on theory and algorithms as well as examples include:

David Eisenbud , Bernd Sturmfels (1996)

Ignacio Ojeda {MartÍnez Castilla} , Ramón Peidra Sánchez (2000)

Thomas Kahle (2010)

Zekiye Sahin Eser , Laura Felicia Matusevich (2016)

Zekiye Sahin Eser , Laura Felicia Matusevich (2019)

Basic Tests

Binomiality Test

isbinomial — Method

isbinomial(f::MPolyElem)

Return true if f consists of at most 2 terms, false otherwise.

isbinomial — Method

isbinomial(I::MPolyIdeal)

Return true I can be generated by polynomials consisting of at most 2 terms, false otherwise.

julia> R, (x, y, z) = PolynomialRing(QQ, ["x", "y", "z"])
(Multivariate Polynomial Ring in x, y, z over Rational Field, fmpq_mpoly[x, y, z])

julia> f = 2*x+y
2*x + y

julia> isbinomial(f)
true

julia> J = ideal(R, [x^2-y^3, z^2])
ideal generated by: x^2 - y^3, z^2

julia> isbinomial(J)
true

Cellularity Test

iscellular — Method

iscellular(I::MPolyIdeal)

Given a binomial ideal I, return true together with the indices of the cellular variables if I is cellular. Return false together with the index of a variable which is a zerodivisor but not nilpotent modulo I, otherwise.

Examples

julia> R, x = PolynomialRing(QQ, "x" => 1:4)
(Multivariate Polynomial Ring in x[1], x[2], x[3], x[4] over Rational Field, fmpq_mpoly[x[1], x[2], x[3], x[4]])

julia> I = ideal(R, [x[1]^3*x[3]-x[1]^3, x[1]^4, x[1]^2*x[2]*x[4]-x[1]^2*x[2], x[2]^2, x[4]^3-1])
ideal generated by: x[1]^3*x[3] - x[1]^3, x[1]^4, x[1]^2*x[2]*x[4] - x[1]^2*x[2], x[2]^2, x[4]^3 - 1

julia> iscellular(I)
(true, [3, 4])

julia> R, x = PolynomialRing(QQ, "x" => 1:5)
(Multivariate Polynomial Ring in x[1], x[2], x[3], x[4], x[5] over Rational Field, fmpq_mpoly[x[1], x[2], x[3], x[4], x[5]])

julia> I = ideal(R, [x[1]*x[4]^2-x[2]*x[5]^2, x[1]^3*x[3]^3-x[2]^4*x[4]^2, x[2]*x[4]^8-x[3]^3*x[5]^6])
ideal generated by: x[1]*x[4]^2 - x[2]*x[5]^2, x[1]^3*x[3]^3 - x[2]^4*x[4]^2, x[2]*x[4]^8 - x[3]^3*x[5]^6

julia> iscellular(I)
(false, [1])

Unitality Test

isunital — Method

isunital(I::MPolyIdeal)

Given a binomial ideal I, return true if I can be generated by differences of monic monomials and monomials.

Examples

julia> R, (x, y, z) = PolynomialRing(QQ, ["x", "y", "z"])
(Multivariate Polynomial Ring in x, y, z over Rational Field, fmpq_mpoly[x, y, z])

julia> I = ideal(R, [x+y])
ideal generated by: x + y

julia> isunital(I)
false

julia> J = ideal(R, [x^2-y^3, z^2])
ideal generated by: x^2 - y^3, z^2

julia> isunital(J)
true

Cellular Decomposition

cellular_decomposition — Method

cellular_decomposition(I::MPolyIdeal)

Given a binomial ideal I, return a cellular decomposition of I.

julia> R, x = PolynomialRing(QQ, "x"=>1:5)
(Multivariate Polynomial Ring in x[1], x[2], x[3], x[4], x[5] over Rational Field, fmpq_mpoly[x[1], x[2], x[3], x[4], x[5]])

julia> I = ideal(R, [x[1]*x[4]^2-x[2]*x[5]^2, x[1]^3*x[3]^3-x[2]^4*x[4]^2, x[2]*x[4]^8-x[3]^3*x[5]^6])
ideal generated by: x[1]*x[4]^2 - x[2]*x[5]^2, x[1]^3*x[3]^3 - x[2]^4*x[4]^2, x[2]*x[4]^8 - x[3]^3*x[5]^6

julia> cellular_decomposition(I)
2-element Array{MPolyIdeal{fmpq_mpoly},1}:
 ideal generated by: x[1]*x[4]^2 - x[2]*x[5]^2, x[1]^3*x[3]^3 - x[2]^4*x[4]^2, -x[1]^2*x[3]^3*x[5]^2 + x[2]^3*x[4]^4, -x[1]*x[3]^3*x[5]^4 + x[2]^2*x[4]^6, x[2]*x[4]^8 - x[3]^3*x[5]^6
 ideal generated by: x[1]^2, x[1]*x[4]^2 - x[2]*x[5]^2, x[1]*x[2]*x[5]^2, x[2]^5, x[5]^6, x[2]^2*x[5]^4, x[2]^4*x[4]^2, x[4]^8, x[2]*x[4]^6*x[5]^2

Primary Decomposition of Cellular Ideals

cellular_hull — Method

cellular_hull(I::MPolyIdeal)

Given a cellular binomial ideal I, return the intersection of the minimal primary components of I.

cellular_associated_primes — Function

cellular_associated_primes(I::MPolyIdeal)

Given a cellular binomial ideal I, return the associated primes of I.

cellular_minimal_associated_primes — Function

cellular_minimal_associated_primes(I::MPolyIdeal)

Given a cellular binomial ideal I, return the minimal associated primes of I.

cellular_primary_decomposition — Method

cellular_primary_decomposition(I::MPolyIdeal)

Given a cellular binomial ideal I, return a binomial primary decomposition of I.

Primary Decomposition of Binomial Ideals

binomial_primary_decomposition — Method

binomial_primary_decomposition(I::MPolyIdeal)

Given a binomial ideal I, return a binomial primary decomposition of I.