Quotient Rings of Polynomial Rings

Constructor

Data Associated to Quotient Rings of Polynomial Rings

Tests on Quotient Rings of Polynomial Rings

Elements of Quotient Rings

Ideals in Quotient Rings of Polynomial Rings

Quotient Rings of Polynomial Rings

We discuss functionality for working with quotient rings of multivariate polynomial rings which may or may not have an assigned grading.

Constructor

quo — Method

quo(R::MPolyRing, I::MPolyIdeal) -> MPolyQuoRing, Map

Create the quotient ring $R/I$ and return the new ring as well as the projection map $R\rightarrow R/I$.

quo(R::MPolyRing, V::Vector{MPolyElem}) -> MPolyQuoRing, Map

As above, where $I\subset R$ is the ideal generated by the polynomials in $V$.

Examples

julia> R, (x, y) = PolynomialRing(QQ, ["x", "y"])
(Multivariate Polynomial Ring in x, y over Rational Field, fmpq_mpoly[x, y])

julia> Q, _ = quo(R, ideal(R, [x^2-y^3, x-y]))
(Quotient of Multivariate Polynomial Ring in x, y over Rational Field by ideal(x^2 - y^3, x - y), Map from
Multivariate Polynomial Ring in x, y over Rational Field to Quotient of Multivariate Polynomial Ring in x, y over Rational Field by ideal(x^2 - y^3, x - y) defined by a julia-function with inverse
)

julia> typeof(x)
fmpq_mpoly

julia> Q, q = quo(R, ideal(R, [x^2-y^3, x-y]))
(Quotient of Multivariate Polynomial Ring in x, y over Rational Field by ideal(x^2 - y^3, x - y), Map from
Multivariate Polynomial Ring in x, y over Rational Field to Quotient of Multivariate Polynomial Ring in x, y over Rational Field by ideal(x^2 - y^3, x - y) defined by a julia-function with inverse
)

julia> q
Map from
Multivariate Polynomial Ring in x, y over Rational Field to Quotient of Multivariate Polynomial Ring in x, y over Rational Field by ideal(x^2 - y^3, x - y) defined by a julia-function with inverse

julia> q(x)
x

julia> typeof(q(x))
MPolyQuoElem{fmpq_mpoly}

julia> S, (x, y, z) = GradedPolynomialRing(QQ, ["x", "y", "z"]);

julia> P, p = quo(S, ideal(S, [x^2*z-y^3, x-y]))
(Quotient of Multivariate Polynomial Ring in x, y, z over Rational Field graded by
  x -> [1]
  y -> [1]
  z -> [1] by ideal(x^2*z - y^3, x - y), Map from
Multivariate Polynomial Ring in x, y, z over Rational Field graded by
  x -> [1]
  y -> [1]
  z -> [1] to Quotient of Multivariate Polynomial Ring in x, y, z over Rational Field graded by
  x -> [1]
  y -> [1]
  z -> [1] by ideal(x^2*z - y^3, x - y) defined by a julia-function with inverse
)

Note

With or without an assigned grading, the return types of quo are subtypes of MPolyQuo.

Note

In Oscar, elements of quotient rings are not necessarily reduced with regard to the modulus of the quotient ring. Operations involving Gröbner basis computations may lead to partial reductions. Full reductions, depending on the choice of a monomial ordering, are achieved by explicitly computing normal forms. The functions simplify and simplify! discussed in the sections below implements this.

Data Associated to Quotient Rings of Polynomial Rings

Basic Data

If Q=R/I is the quotient of a multivariate polynomial ring R modulo an ideal I of R, then

base_ring(Q) refers to R,

modulus(Q) to I,

gens(Q) to the generators of Q, and

ngens(Q) to the number of these generators.

Examples

julia> R, (x, y, z) = PolynomialRing(QQ, ["x", "y", "z"])(Multivariate Polynomial Ring in x, y, z over Rational Field, fmpq_mpoly[x, y, z])
julia> Q, _ = quo(R, ideal(R, [y-x^2, x-z^3]))(Quotient of Multivariate Polynomial Ring in x, y, z over Rational Field by ideal(-x^2 + y, x - z^3), Map from
Multivariate Polynomial Ring in x, y, z over Rational Field to Quotient of Multivariate Polynomial Ring in x, y, z over Rational Field by ideal(-x^2 + y, x - z^3) defined by a julia-function with inverse
)
julia> base_ring(Q)Multivariate Polynomial Ring in x, y, z over Rational Field
julia> modulus(Q)ideal(-x^2 + y, x - z^3)
julia> gens(Q)3-element Vector{MPolyQuoElem{fmpq_mpoly}}:
 x
 y
 z
julia> ngens(Q)3

Dimension

dim — Method

dim(Q::MPolyQuo)

Return the Krull dimension of the quotient ring Q.

Examples

julia> R, (x, y, z) = PolynomialRing(QQ, ["x", "y", "z"])
(Multivariate Polynomial Ring in x, y, z over Rational Field, fmpq_mpoly[x, y, z])

julia> Q, _ = quo(R, ideal(R, [y-x^2, x-z^3]))
(Quotient of Multivariate Polynomial Ring in x, y, z over Rational Field by ideal(-x^2 + y, x - z^3), Map from
Multivariate Polynomial Ring in x, y, z over Rational Field to Quotient of Multivariate Polynomial Ring in x, y, z over Rational Field by ideal(-x^2 + y, x - z^3) defined by a julia-function with inverse
)

julia> dim(Q)
1

Tests on Quotient Rings of Polynomial Rings

Reducedness Test

isreduced — Method

isreduced(Q::MPolyQuo)

Given a quotient ring Q = R/I, return true if Q is reduced, false otherwise.

CAVEAT: The function computes the radical of I. This may take some time.

Examples

julia> R, (x,) = PolynomialRing(QQ, ["x"])
(Multivariate Polynomial Ring in x over Rational Field, fmpq_mpoly[x])

julia> Q, _ = quo(R, ideal(R, [x^4]))
(Quotient of Multivariate Polynomial Ring in x over Rational Field by ideal(x^4), Map from
Multivariate Polynomial Ring in x over Rational Field to Quotient of Multivariate Polynomial Ring in x over Rational Field by ideal(x^4) defined by a julia-function with inverse
)

julia> isreduced(Q)
false

Elements of Quotient Rings

Reducing Elements of Quotient Rings

simplify — Method

simplify(f::MPolyQuoElem)

Reduce f with regard to the modulus of the quotient ring.

simplify!(f::MPolyQuoElem)

Reduce f with regard to the modulus of the quotient ring, and replace f by the reduction.

Examples

julia> R, (x,) = PolynomialRing(QQ, ["x"]);

julia> Q, q = quo(R, ideal(R, [x^4]));

julia> f = q(x-x^6)
-x^6 + x

julia> simplify(f)
x

julia> f
-x^6 + x

julia> simplify!(f)
x

julia> f
x

Tests on Elements of Quotient Rings

== — Method

==(f::MPolyQuoElem, g::MPolyQuoElem)

Return true if f is equal to g, false otherwise.

Examples

julia> R, (x,) = PolynomialRing(QQ, ["x"]);

julia> Q, q = quo(R, ideal(R, [x^4]));

julia> f = q(x-x^6)
-x^6 + x

julia> g = q(x)
x

julia> f == g
true

Ideals in Quotient Rings of Polynomial Rings

Constructors

ideal — Method

ideal(Q::MPolyQuo{T}, V::Vector{T}) where T <: MPolyElem

Given a (graded) quotient ring Q=R/I and a vector V of (homogeneous) polynomials in R, create the ideal of Q which is generated by the images of the entries of V.

ideal(Q::MPolyQuo{T}, V::Vector{MPolyQuoElem{T}}) where T <: MPolyElem

Given a (graded) quotient ring Q and a vector V of (homogeneous) elements of Q, create the ideal of Q which is generated by the entries of V.

Reducing Ideals in Quotient Rings

simplify — Method

simplify(a::MPolyQuoIdeal)

Reduce a with regard to the modulus of the quotient ring.

simplify!(a::MPolyQuoIdeal)

Reduce a with regard to the modulus of the quotient ring, and replace a by the reduction.

CAVEAT: The implementation proceeds by computing a Groebner basis of the quotient ideal first. This may take some time.

Examples

julia> R, (x, y) = PolynomialRing(QQ, ["x", "y"], ordering=:degrevlex);

julia> Q, q = quo(R, ideal(R, [x^3*y^2-y^3*x^2, x*y^4-x*y^2]));

julia> a = ideal(Q,[x^3*y^4-x+y, x*y+y^2*x])
ideal(x^3*y^4 - x + y, x*y^2 + x*y)

julia> simplify(a);

julia> a
ideal(x^3*y^4 - x + y, x*y^2 + x*y)

julia> simplify!(a);

julia> a
ideal(x^2*y^3 - x + y, x*y^2 + x*y)

Data Associated to Ideals in Quotient Rings

Basic Data

If a is an ideal of the quotient ring Q, then

base_ring(a) refers to Q,

gens(a) to the generators of a, and

ngens(a) to the number of these generators.

Dimension of Ideals in Quotient Rings

dim — Method

dim(a::MPolyQuoIdeal)

Return the Krull dimension of a.

Examples

julia> R, (x, y, z) = PolynomialRing(QQ, ["x", "y", "z"], ordering=:degrevlex)
(Multivariate Polynomial Ring in x, y, z over Rational Field, fmpq_mpoly[x, y, z])

julia> Q, q = quo(R, ideal(R, [x^3*y^2-y^3*x^2, x*y^4-x*y^2]))
(Quotient of Multivariate Polynomial Ring in x, y, z over Rational Field by ideal(x^3*y^2 - x^2*y^3, x*y^4 - x*y^2), Map from
Multivariate Polynomial Ring in x, y, z over Rational Field to Quotient of Multivariate Polynomial Ring in x, y, z over Rational Field by ideal(x^3*y^2 - x^2*y^3, x*y^4 - x*y^2) defined by a julia-function with inverse
)

julia> a = ideal(Q,[x^3*y^4-x+y, x*y+y^2*x])
ideal(x^3*y^4 - x + y, x*y^2 + x*y)

julia> dim(a)
1

Operations on Ideals in Quotient Rings

Simple Ideal Operations in Quotient Rings

Powers of Ideal

^ — Method

:^(a::MPolyQuoIdeal, m::Int)

Return the m-th power of a.

Sum of Ideals

+ — Method

:+(a::MPolyQuoIdeal, b::MPolyQuoIdeal)

Return the sum of a and b.

Product of Ideals

* — Method

:*(a::MPolyQuoIdeal, b::MPolyQuoIdeal)

Return the product of a and b.

Intersection of Ideals

intersect — Method

intersect(a::MPolyQuoIdeal, bs::MPolyQuoIdeal...)

Return the intersection of two or more ideals.

Examples

julia> R, (x, y) = PolynomialRing(QQ, ["x", "y"])
(Multivariate Polynomial Ring in x, y over Rational Field, fmpq_mpoly[x, y])

julia> A, _ = quo(R, ideal(R, [x^2-y^3, x-y]))
(Quotient of Multivariate Polynomial Ring in x, y over Rational Field by ideal(x^2 - y^3, x - y), Map from
Multivariate Polynomial Ring in x, y over Rational Field to Quotient of Multivariate Polynomial Ring in x, y over Rational Field by ideal(x^2 - y^3, x - y) defined by a julia-function with inverse
)

julia> a = ideal(A, [y^2])
ideal(y^2)

julia> b = ideal(A, [x])
ideal(x)

julia> intersect(a,b)
ideal(x*y)

Ideal Quotients

quotient — Method

quotient(a::MPolyQuoIdeal, b::MPolyQuoIdeal)

Return the ideal quotient of a by b. Alternatively, use a:b.

Examples

julia> R, (x, y) = PolynomialRing(QQ, ["x", "y"]);

julia> Q, _ = quo(R, ideal(R, [x^2-y^3, x-y]))
(Quotient of Multivariate Polynomial Ring in x, y over Rational Field by ideal(x^2 - y^3, x - y), Map from
Multivariate Polynomial Ring in x, y over Rational Field to Quotient of Multivariate Polynomial Ring in x, y over Rational Field by ideal(x^2 - y^3, x - y) defined by a julia-function with inverse
)

julia> a = ideal(Q, [y^2])
ideal(y^2)

julia> b = ideal(Q, [x])
ideal(x)

julia> a:b
ideal(y)

Tests on Ideals in Quotient Rings

Basic Tests

iszero — Method

iszero(a::MPolyQuoIdeal)

Return true if a is the zero ideal, false otherwise.

Equality of Ideals in Quotient Rings

== — Method

:(==)(a::MPolyQuoIdeal, b::MPolyQuoIdeal)

Return true if a=b, false otherwise.

Examples

julia> R, (x, y) = PolynomialRing(QQ, ["x", "y"], ordering=:degrevlex)
(Multivariate Polynomial Ring in x, y over Rational Field, fmpq_mpoly[x, y])

julia> Q, q = quo(R, ideal(R, [x^3*y^2-y^3*x^2, x*y^4-x*y^2]))
(Quotient of Multivariate Polynomial Ring in x, y over Rational Field by ideal(x^3*y^2 - x^2*y^3, x*y^4 - x*y^2), Map from
Multivariate Polynomial Ring in x, y over Rational Field to Quotient of Multivariate Polynomial Ring in x, y over Rational Field by ideal(x^3*y^2 - x^2*y^3, x*y^4 - x*y^2) defined by a julia-function with inverse
)

julia> a = ideal(Q,[x^3*y^4-x+y, x*y+y^2*x])
ideal(x^3*y^4 - x + y, x*y^2 + x*y)

julia> b = ideal(Q,[x^3*y^3-x+y, x^2*y+y^2*x])
ideal(x^3*y^3 - x + y, x^2*y + x*y^2)

julia> a == b
false

Containment of Ideals in Quotient Rings

issubset — Method

issubset(a::MPolyQuoIdeal, b::MPolyQuoIdeal)

Return true if a is contained in b, false otherwise.

Examples

julia> R, (x, y) = PolynomialRing(QQ, ["x", "y"], ordering=:degrevlex)
(Multivariate Polynomial Ring in x, y over Rational Field, fmpq_mpoly[x, y])

julia> Q, q = quo(R, ideal(R, [x^3*y^2-y^3*x^2, x*y^4-x*y^2]))
(Quotient of Multivariate Polynomial Ring in x, y over Rational Field by ideal(x^3*y^2 - x^2*y^3, x*y^4 - x*y^2), Map from
Multivariate Polynomial Ring in x, y over Rational Field to Quotient of Multivariate Polynomial Ring in x, y over Rational Field by ideal(x^3*y^2 - x^2*y^3, x*y^4 - x*y^2) defined by a julia-function with inverse
)

julia> a = ideal(Q,[x^3*y^4-x+y, x*y+y^2*x])
ideal(x^3*y^4 - x + y, x*y^2 + x*y)

julia> b = ideal(Q,[x^3*y^3-x+y, x^2*y+y^2*x])
ideal(x^3*y^3 - x + y, x^2*y + x*y^2)

julia> issubset(a,b)
false

julia> issubset(b,a)
true