Polynomial Rings

Coefficients

Constructions

Gradings

Polynomial Rings

Referring to the sections on rings and fields for more details, we summarize how to construct multivariate polynomial rings. Also, we show how to decorate such rings with gradings.

Coefficients

Multivariate polynomial rings over coefficient rings from the list below allow for Gröbner basis computations :

The field of rational numbers $\mathbb{Q}$

julia> QQ
Rational Field

Finite fields $\mathbb{F_p}$, $p$ a prime

julia> GF(3)
Galois field with characteristic 3

julia> GF(ZZ(2)^127 - 1)
Galois field with characteristic 170141183460469231731687303715884105727

Finite fields $\mathbb{F}_{p^n}$ with $p^n$ elements, $p$ a prime

julia> FiniteField(2, 70, "a")
(Finite field of degree 70 over F_2, a)

Simple algebraic extensions of $\mathbb{Q}$ or $\mathbb{F}_p$

julia> Qx, x = QQ["x"]
(Univariate Polynomial Ring in x over Rational Field, x)

julia> K, a = NumberField(x^2 + 1, "a")
(Number field over Rational Field with defining polynomial x^2 + 1, a)

julia> F = GF(3)
Galois field with characteristic 3

julia> Fx, x = F["x"]
(Univariate Polynomial Ring in x over Galois field with characteristic 3, x)

julia> K, a = FiniteField(x^2 + 1, "a")
(Finite field of degree 2 over F_3, a)

Purely transcendental extensions of $\mathbb{Q}$ or $\mathbb{F}_p$

julia> T, t = PolynomialRing(QQ, :t=>1:2);

julia> QT = FractionField(T)
Fraction field of Multivariate Polynomial Ring in t[1], t[2] over Rational Field

julia> parent(t[1])
Multivariate Polynomial Ring in t[1], t[2] over Rational Field

julia> parent(1//t[1])
Fraction field of Multivariate Polynomial Ring in t[1], t[2] over Rational Field

julia> U, u = PolynomialRing(GF(3), :u=>1:2);

julia> QU = FractionField(U)
Fraction field of Multivariate Polynomial Ring in u[1], u[2] over Galois field with characteristic 3

The ring of integers $\mathbb{ZZ}$

julia> ZZ
Integer Ring

Constructions

The basic constructor for multivariate polynomial rings reads as follows:

PolynomialRing(C::Ring, v::Vector{String}; ordering=:lex)

Its use and that of additional constructors which allow for indexed variables is indicated below.

Examples

julia> R, (x, y, z) = PolynomialRing(QQ, ["x", "y", "z"])
(Multivariate Polynomial Ring in x, y, z over Rational Field, fmpq_mpoly[x, y, z])

julia> base_ring(R)
Rational Field

julia> gens(R)
3-element Array{fmpq_mpoly,1}:
 x
 y
 z

julia> gen(R, 2)
y

julia> R[3]
z

julia> ngens(R)
3

julia> f = x^2+y*z
x^2 + y*z

julia> parent(f)
Multivariate Polynomial Ring in x, y, z over Rational Field

julia> typeof(f)
fmpq_mpoly

julia> S, (a,b) = PolynomialRing(ZZ, ["a", "b"])
(Multivariate Polynomial Ring in a, b over Integer Ring, fmpz_mpoly[a, b])

julia> g = a*b
a*b

julia> typeof(g)
fmpz_mpoly

julia> v = ["x[1]", "x[2]"]
2-element Array{String,1}:
 "x[1]"
 "x[2]"

julia> T, x = PolynomialRing(GF(3), v)
(Multivariate Polynomial Ring in x[1], x[2] over Galois field with characteristic 3, gfp_mpoly[x[1], x[2]])

julia> x
2-element Array{gfp_mpoly,1}:
 x[1]
 x[2]

julia> R, x = PolynomialRing(QQ, 2, "x");

julia> x
2-element Array{fmpq_mpoly,1}:
 x1
 x2

julia> x1
ERROR: UndefVarError: x1 not defined

julia> x[1]^2-x[2]
x1^2 - x2

julia> S, a, b, c= PolynomialRing(QQ, "a"=>1:3, "b"=>1:3, "c"=>1:5:10)
(Multivariate Polynomial Ring in 8 variables a[1], a[2], a[3], b[1], ..., c[6] over Rational Field, fmpq_mpoly[a[1], a[2], a[3]], fmpq_mpoly[b[1], b[2], b[3]], fmpq_mpoly[c[1], c[6]])

julia> a, b, c
(fmpq_mpoly[a[1], a[2], a[3]], fmpq_mpoly[b[1], b[2], b[3]], fmpq_mpoly[c[1], c[6]])

julia> R, x = @PolynomialRing(QQ, x[1:3]);

julia> x
3-element Array{fmpq_mpoly,1}:
 x[1]
 x[2]
 x[3]

julia> x[1]^2-x[2]
x[1]^2 - x[2]

julia> R, x = @PolynomialRing(QQ, x[1:3, 1:3]);

julia> x
3×3 Array{fmpq_mpoly,2}:
 x[1,1]  x[1,2]  x[1,3]
 x[2,1]  x[2,2]  x[2,3]
 x[3,1]  x[3,2]  x[3,3]

julia> R, x, y = @PolynomialRing(QQ, x[1:3], y[1:2]);

julia> x, y
(fmpq_mpoly[x[1], x[2], x[3]], fmpq_mpoly[y[1], y[2]])

julia> I = [1, 3, 5];

julia> R, x = @PolynomialRing(QQ, x[I]);

julia> x
3-element Array{fmpq_mpoly,1}:
 x[1]
 x[3]
 x[5]

Gradings

grade — Method

grade(R::MPolyRing, v::Array{Int, 1})

Grade R by assigning weights to the variables according to the entries of v.

grade(R::MPolyRing)

Grade R by assigning weight 1 to each variable.

Examples

julia> R, (x, y, z) = PolynomialRing(QQ, ["x", "y", "z"])
(Multivariate Polynomial Ring in x, y, z over Rational Field, fmpq_mpoly[x, y, z])

julia> v = [1, 2, 3]
3-element Array{Int64,1}:
 1
 2
 3

julia> S, (x, y, z) = grade(R, v)
(Multivariate Polynomial Ring in x, y, z over Rational Field graded by 
  x -> [1]
  y -> [2]
  z -> [3], Oscar.MPolyElem_dec{fmpq,fmpq_mpoly}[x, y, z])

julia> T, (x, y, z) = grade(R)
(Multivariate Polynomial Ring in x, y, z over Rational Field graded by 
  x -> [1]
  y -> [1]
  z -> [1], Oscar.MPolyElem_dec{fmpq,fmpq_mpoly}[x, y, z])