Gröbner Bases

Computing Gröbner Bases

Leading Ideals

Normal Forms

Syzygies

Gröbner Bases

Computing Gröbner Bases

groebner_basis — Method

groebner_basis(I::MPolyIdeal;
  ordering::MonomialOrdering = default_ordering(base_ring(I)),
  complete_reduction::Bool = false)

Given an ideal I and optional parameters monomial ordering ordering and complete_reduction, compute a Groebner basis (if complete_reduction = true the reduced Groebner basis) of I w.r.t. the given monomial ordering ordering (as default degree reverse lexicographical).

Examples

julia> R, (x, y) = PolynomialRing(QQ, ["x", "y"])
(Multivariate Polynomial Ring in x, y over Rational Field, fmpq_mpoly[x, y])

julia> I = ideal([x*y-3*x,y^3-2*x^2*y])
ideal(x*y - 3*x, -2*x^2*y + y^3)

julia> H = groebner_basis(I, ordering=lex(R))
Gröbner basis with elements
1 -> y^4 - 3*y^3
2 -> x*y - 3*x
3 -> 6*x^2 - y^3
with respect to the ordering
lex([x, y])

standard_basis — Method

standard_basis(I::MPolyIdeal;
  ordering::MonomialOrdering = default_ordering(base_ring(I)),
  complete_reduction::Bool = false)

Compute a standard basis of I for the monomial ordering ordering.

Note: Since there is no general notion of complete reduction for non-global orderings, this option is not available for this command; instead, use groebner_basis directly.

Examples

julia> R,(x,y) = PolynomialRing(QQ, ["x","y"])
(Multivariate Polynomial Ring in x, y over Rational Field, fmpq_mpoly[x, y])

julia> I = ideal([x*(x+1), x^2-y^2+(x-2)*y])
ideal(x^2 + x, x^2 + x*y - y^2 - 2*y)

julia> standard_basis(I, ordering=negdegrevlex(R))
Standard basis with elements
1 -> x
2 -> y
with respect to the ordering
negdegrevlex([x, y])

See e.g. Gert-Martin Greuel, Gerhard Pfister (2008) for the theoretical background on Gröbner- and standard bases.

Gröbner Bases with transformation matrix

groebner_basis_with_transformation_matrix — Method

groebner_basis_with_transformation_matrix(I::MPolyIdeal;
  ordering::MonomialOrdering = default_ordering(base_ring(I)),
  complete_reduction::Bool=false)

Return a pair G, m where G is a Groebner basis of the ideal I with respect to the monomial ordering ordering (as default degree reverse lexicographical), and m is a transformation matrix from gens(I)toG. If complete_reductionis set to true (as default 'false') then G will be the reduced Groebner basis.

Examples

julia> R,(x,y) = PolynomialRing(QQ,["x","y"])
(Multivariate Polynomial Ring in x, y over Rational Field, fmpq_mpoly[x, y])

julia> I = ideal([x*y^2-1,x^3+y^2+x*y])
ideal(x*y^2 - 1, x^3 + x*y + y^2)

julia> G,m = groebner_basis_with_transformation_matrix(I)
(fmpq_mpoly[x*y^2 - 1, x^3 + x*y + y^2, x^2 + y^4 + y], [1 0 -x^2-y; 0 1 y^2])

julia> gens(I)*m == G
true

fglm

Gröbner walks

Hilbert-driven

Expert functions for Gröbner bases

The following functions are low-level implementations of various Gröbner basis algorithms with many adjustable arguments. Only use these functions directly if you know what you are doing.

f4 — Method

f4(I::MPolyIdeal, <keyword arguments>)

Compute a Groebner basis of the given ideal I w.r.t. to the degree reverse lexicographical monomial ordering using Faugère's F4 algorithm. See Jean-Charles Faugère (1999) for more information.

Note: At the moment only ground fields of characteristic p, p prime, p < 2^{31} are supported.

Arguments

Ì::MPolyIdeal: input ideal.

initial_hts::Int=17: initial hash table size log_2.

nr_thrds::Int=1: number of threads for parallel linear algebra.

max_nr_pairs::Int=0: maximal number of pairs per matrix, only bounded by minimal degree if 0.

la_option::Int=2: linear algebra option: exact sparse-dense (1), exact sparse (2, default), probabilistic sparse-dense (42), probabilistic sparse(44).

eliminate::Int=0: size of first block of variables to be eliminated.

complete_reduction::Bool=true: compute a reduced Gröbner basis for I

info_level::Int=0: info level printout: off (0, default), summary (1), detailed (2).

Examples

julia> R,(x,y,z) = PolynomialRing(GF(101), ["x","y","z"], ordering=:degrevlex)
(Multivariate Polynomial Ring in x, y, z over Galois field with characteristic 101, gfp_mpoly[x, y, z])

julia> I = ideal(R, [x+2*y+2*z-1, x^2+2*y^2+2*z^2-x, 2*x*y+2*y*z-y])
ideal(x + 2*y + 2*z + 100, x^2 + 2*y^2 + 2*z^2 + 100*x, 2*x*y + 2*y*z + 100*y)

 
julia> f4(I)
Gröbner basis with elements
1 -> x + 2*y + 2*z + 100
2 -> y*z + 82*z^2 + 10*y + 40*z
3 -> y^2 + 60*z^2 + 20*y + 81*z
4 -> z^3 + 28*z^2 + 64*y + 13*z
with respect to the ordering
degrevlex([x, y, z])

julia> isdefined(I, :gb)
true

Gröbner Bases over the integers

Over the integers the coefficients of the polynomials are not invertible, thus their handling when computing Gröbner bases and normal forms plays an important role. This is done when computing strong Gröbner bases which ensure the following property: For any element of an ideal its leading term is divisible by a leading term of an element of a corresponding strong Gröbner basis.

The textbook William W. Adams, Philippe Loustaunau (1994) provides details on theory and algorithms as well as references.

julia> R, (x,y) = PolynomialRing(ZZ, ["x","y"])(Multivariate Polynomial Ring in x, y over Integer Ring, fmpz_mpoly[x, y])
julia> I = ideal(R, [2x,3x,4y])ideal(2*x, 3*x, 4*y)
julia> H = groebner_basis(I)Gröbner basis with elements
1 -> 4*y
2 -> x
with respect to the ordering
degrevlex([x, y])

Leading Ideals

leading_ideal — Method

leading_ideal(g::Vector{T}; ordering::MonomialOrdering) where { T <: MPolyElem }

Return the ideal generated by the leading monomials of the given polynomials w.r.t. the given monomial ordering.

Examples

julia> R, (x, y) = PolynomialRing(QQ, ["x", "y"])
(Multivariate Polynomial Ring in x, y over Rational Field, fmpq_mpoly[x, y])

julia> L = leading_ideal([x*y^2-3*x, x^3-14*y^5], ordering=degrevlex(R))
ideal(x*y^2, y^5)

julia> L = leading_ideal([x*y^2-3*x, x^3-14*y^5], ordering=lex(R))
ideal(x*y^2, x^3)

leading_ideal — Method

leading_ideal(I::MPolyIdeal; ordering::MonomialOrdering)

Given a multivariate polynomial ideal Ì this function returns the leading ideal for I. This is done w.r.t. the given monomial ordering.

Examples

julia> R, (x, y) = PolynomialRing(QQ, ["x", "y"])
(Multivariate Polynomial Ring in x, y over Rational Field, fmpq_mpoly[x, y])

julia> I = ideal(R,[x*y^2-3*x, x^3-14*y^5])
ideal(x*y^2 - 3*x, x^3 - 14*y^5)

julia> L = leading_ideal(I, ordering=degrevlex(R))
ideal(x*y^2, x^4, y^5)

julia> L = leading_ideal(I, ordering=lex(R))
ideal(y^7, x*y^2, x^3)

Normal Forms

normal_form — Method

normal_form(f::T, J::MPolyIdeal) where { T <: MPolyElem }

Compute the normal form of the polynomial f w.r.t. a Groebner basis of J.

CAVEAT: This computation needs a Groebner basis of J. If this Groebner basis is not available, one is computed automatically. This may take some time.

Examples

julia> R,(a,b,c) = PolynomialRing(QQ,["a","b","c"])
(Multivariate Polynomial Ring in a, b, c over Rational Field, fmpq_mpoly[a, b, c])

julia> J = ideal(R,[-1+c+b,-1+b+c*a+2*a*b])
ideal(b + c - 1, 2*a*b + a*c + b - 1)

julia> gens(groebner_basis(J))
2-element Vector{fmpq_mpoly}:
 b + c - 1
 a*c - 2*a + c

julia> normal_form(-1+c+b+a^3, J)
a^3

normal_form — Method

normal_form(A::Vector{T}, J::MPolyIdeal) where { T <: MPolyElem }

Compute the normal form of the elements of the array A w.r.t. a Groebner basis of J.

CAVEAT: This computation needs a Groebner basis of J. If this Groebner basis is not available, one is computed automatically. This may take some time.

Examples

julia> R,(a,b,c) = PolynomialRing(QQ,["a","b","c"])
(Multivariate Polynomial Ring in a, b, c over Rational Field, fmpq_mpoly[a, b, c])

julia> A = [-1+c+b+a^3,-1+b+c*a+2*a^3,5+c*b+c^2*a]
3-element Vector{fmpq_mpoly}:
 a^3 + b + c - 1
 2*a^3 + a*c + b - 1
 a*c^2 + b*c + 5

julia> J = ideal(R,[-1+c+b,-1+b+c*a+2*a*b])
ideal(b + c - 1, 2*a*b + a*c + b - 1)

julia> gens(groebner_basis(J))
2-element Vector{fmpq_mpoly}:
 b + c - 1
 a*c - 2*a + c

julia> normal_form(A, J)
3-element Vector{fmpq_mpoly}:
 a^3
 2*a^3 + 2*a - 2*c
 4*a - 2*c^2 - c + 5

Syzygies

syzygy_generators — Method

syzygy_generators(a::Vector{<:MPolyElem})

Return generators for the syzygies on the given polynomials.

Examples

julia> R, (x, y) = PolynomialRing(QQ, ["x", "y"])
(Multivariate Polynomial Ring in x, y over Rational Field, fmpq_mpoly[x, y])

julia> S = syzygy_generators([x^3+y+2,x*y^2-13*x^2,y-14])
3-element Vector{FreeModElem{fmpq_mpoly}}:
 (-y + 14)*e[2] + (-13*x^2 + x*y^2)*e[3]
 (-169*y + 2366)*e[1] + (-13*x*y + 182*x - 196*y + 2744)*e[2] + (13*x^2*y^2 - 2548*x^2 + 196*x*y^2 + 169*y + 338)*e[3]
 (-13*x^2 + 196*x)*e[1] + (-x^3 - 16)*e[2] + (x^4*y + 14*x^4 + 13*x^2 + 16*x*y + 28*x)*e[3]