Localizations of commutative rings

Suppose $R$ is a commutative ring with unit and $S \subset R$ is a multiplicatively closed set containing $1 \in R$. Then we can form the localization of $R$ at $S$

\[ R[S^{-1}] = \left\{ \frac{p}{q} : p,q \in R, \, q \in S \right\},\]

with its standard arithmetic for fractions. See, for instance, [Eis95] for an account on localizations.

Oscar provides a general framework for such localizations, originally intended to be used with multivariate polynomial rings $R$ over some base field $\mathbb k$, but also applicable to more general commutative rings.

In the case of polynomials, the localization framework provides the structure for certain algorithms using standard bases. Note that, in general, localizations of polynomial algebras are not finitely generated as algebras over $\mathbb k$; for instance when localizing at some maximal ideal $\mathfrak m \subset R$. However, many ideal- and module-theoretic questions in the localization $R[S^{-1}]$, such as e.g. the ideal membership, can be transformed to questions on ideals and modules over the base ring $R$ and then solved using Groebner- or standard-basis techniques. This makes it important to regard localizations $R[S^{-1}]$ as rings with a history of creation from the original pair $S \subset R$.

The localization interface

Localized rings

The interface that needs to be implemented for any concrete instance of localized rings is the following. Multiplicatively closed sets are derived from the abstract type

AbsMultSet — Type

AbsMultSet{RingType, RingElemType}

Abstract type for a multiplicatively closed set in a commutative (Noetherian) ring R of type RingType with elements of type RingElemType.

The basic functionality that has to be implemented for any concrete type derived from this is to be able to check containment of elements via

in — Method

in(f::RingElemType, S::AbsMultSet{RingType, RingElemType}) where {RingType, RingElemType}

Return true if f belongs to S; false otherwise.

Note: If this routine is not implemented, the function call will default to the execution of an error message.

This is supposed to be an extension of the methods of the function Base.in.

A localized ring should then be derived from

AbsLocalizedRing — Type

AbsLocalizedRing{RingType, RingElemType, MultSetType}

The localization R[S⁻¹] of a ring R of type RingType with elements of type RingElemType at a multiplicatively closed set S of type MultSetType.

In general, the arithmetic of such a localized ring R[S⁻¹] should be implemented using fractions of elements in the original ring R. The methods provided for the multiplicatively closed set S can be used to check whether a given denominator is admissible for the specific localization.

Depending on the actual type of R and S, further functionality can then be provided using different Groebner-basis driven backends.

The basic way to construct localized rings is to first specify a multiplicative set S and then call

Localization — Method

Localization(S::AbsMultSet)

Return the localization of the ambient_ring of S at S itself.

This method must be implemented with a dispatch depending on the concrete type of S.

For any concrete instance of type AbsLocalizedRing the following methods must be implemented:

base_ring — Method

base_ring(W::AbsLocalizedRing)

Return the base ring R for a localized ring of the form W = R[S⁻¹].

inverted_set — Method

inverted_set(W::AbsLocalizedRing)

Return the set S of at which has been localized for a localized ring W = R[S⁻¹].

Also, conversion of fractions to elements of localized rings must be implemented in the form (W::AbsLocalizedRing{RingType, RingElemType, MultSetType})(a::RingElemType) where {RingType, RingElemType, MultSetType}, taking $a$ to the element $\frac{a}{1}$. For more general fractions one needs (W::AbsLocalizedRing{RingType, RingElemType, MultSetType})(a::RingElemType, b::RingElemType) where {RingType, RingElemType, MultSetType}, mapping a pair $(a, b)$ to the fraction $\frac{a}{b}$.

The elements of localized rings must be derived from

AbsLocalizedRingElem — Type

AbsLocalizedRingElem{RingType, RingElemType, MultSetType}

The abstract type of an element of the localization R[S⁻¹] of a commutative ring R of type RingType with elements of type RingElemType at a multiplicatively closed set S of type MultSetType.

For any concrete instance F of AbsLocalizedRingElem there must be the following methods:

numerator — Method

numerator(f::AbsLocalizedRingElem)

Return the numerator of f.

denominator — Method

denominator(f::AbsLocalizedRingElem)

Return the denominator of f.

parent — Method

parent(f::AbsLocalizedRingElem)

Return the parent ring R[S⁻¹] of f.

A default version of the arithmetic is implemented on the generic level using the above functionality and the arithmetic for the original ring. Depending on the actual concrete instance, one might wish to provide more fine-tuned methods, starting e.g. by implementing

reduce_fraction — Method

reduce_fraction(a::AbsLocalizedRingElem)

Reduce the fraction a = p/q. Warning: The catchall-implementation does nothing!

Note that this is called after every arithmetic operation (addition, multiplication,...), so the computations carried out here should be computationally cheap.

Two toy examples for implementations of this interface for localizations of the integers $\mathbb Z$ and rings of the form $\mathbb Z/n\mathbb Z$ can be found in the test files test/Rings/integer-localizations.jl and test/Rings/nmod-localizations.jl.

Note: Any concrete type for localized rings is also required to implement the general Ring Interface of Oscar! This has not been done to a full extent for the previous two examples, but for MPolyLocalizedRing; see below.

Homomorphisms for localized rings

Homomorphisms from localized rings to arbitrary algebras are of type

AbsLocalizedRingHom — Type

AbsLocalizedRingHom{
    DomainType<:AbsLocalizedRing,
    CodomainType<:Ring,
    RestrictedMapType
  } <: Map{
    DomainType,
    CodomainType,
    SetMap,
    AbsLocalizedRingHom
  }

Homomorphism $ϕ : R[U⁻¹] → S$ from the localization $R[U⁻¹]$ of type $DomainType$ to an arbitrary ring S of type CodomainType. Such a homomorphism is completely determined by its 'restriction' $ϕ' : R → R[U⁻¹] → S$ to the base_ring $R$ before localization and the type parameter RestrictedMapType is reserved for that map.

Note that, in order to be well-defined, we must have that $\phi'(u) \in S$ must be a unit for every element $u \in U$.

The getters associated to this type which need to be implemented are

domain — Method

domain(f::AbsLocalizedRingHom)

Return the domain of definition of f.

codomain — Method

codomain(f::AbsLocalizedRingHom)

Return the codomain of f.

restricted_map — Method

restricted_map(f::AbsLocalizedRingHom)

For a ring homomorphism $ϕ : R[U⁻¹] → S$ return the underlying restriction $ϕ' : R → S$.

Any concrete instance f of AbsLocalizedRingHom can then be applied to elements a of domain(f) by calling f(a).

Ideals in localized rings

One of the main reasons to implement localizations in the first place is that this process preserves the property of a ring to be Noetherian; which is crucial for computer algebra. In this regard, we have

AbsLocalizedIdeal — Type

AbsLocalizedIdeal{RingType, RingElemType, MultSetType}

Abstract type for finitely generated ideals $I ⊂ R[S⁻¹]$ in localized rings.

The required getter methods are

gens — Method

gens(M::SubQuo{T}) where T

Return the generators of M.

gens(I::sideal)

Return the generators in the internal representation of the ideal $I$ as an array.

gens(A::AbsAlgAss, return_full_basis::Typel{Val{T}} = Val{false};
     thorough_search::Bool = false) where T
  -> Vector{AbsAlgAssElem}

Returns a subset of basis(A), which generates $A$ as an algebra over base_ring(A). If return_full_basis is set to Val{true}, the function also returns a Vector{AbsAlgAssElem} containing a full basis consisting of monomials in the generators and a Vector{Vector{Tuple{Int, Int}}} containing the information on how these monomials are built. E. g.: If the function returns g, full_basis and v, then we have full_basis[i] = prod( g[j]^k for (j, k) in v[i] ). If thorough_search is true, the number of returned generators is possibly smaller. This will in general increase the runtime. It is not guaranteed that the number of generators is minimal in any case.

gens(A::AlgGrp, return_full_basis::Type{Val{T}} = Val{false})
  -> Vector{AlgGrpElem}

Returns a subset of basis(A), which generates $A$ as an algebra over base_ring(A). If return_full_basis is set to Val{true}, the function also returns a Vector{AbsAlgAssElem} containing a full basis consisting of monomials in the generators and a Vector{Vector{Tuple{Int, Int}}} containing the information on how these monomials are built. E. g.: If the function returns g, full_basis and v, then we have full_basis[i] = prod( g[j]^k for (j, k) in v[i] ).

base_ring — Method

base_ring(I::MPolyIdeal)

Return the ambient ring of I.

Examples

julia> R, (x, y) = PolynomialRing(QQ, ["x", "y"])
(Multivariate Polynomial Ring in x, y over Rational Field, fmpq_mpoly[x, y])

julia> I = ideal(R, [x, y])^2
ideal(x^2, x*y, y^2)

julia> base_ring(I)
Multivariate Polynomial Ring in x, y over Rational Field

base_ring(M::PMat)

The PMat $M$ defines an $R$-module for some maximal order $R$. This function returns the $R$ that was used to defined $M$.

The constructors to be implemented are

   ideal(W::AbsLocalizedRing{RingType, RingElemType, MultSetType}, f::AbsLocalizedRingElem{RingType, RingElemType, MultSetType}) where {RingType, RingElemType, MultSetType}
   ideal(W::AbsLocalizedRing{RingType, RingElemType, MultSetType}, f::RingElemType) where {RingType, RingElemType, MultSetType}
   ideal(W::AbsLocalizedRing{RingType, RingElemType, MultSetType}, v::Vector{AbsLocalizedRingElem{RingType, RingElemType, MultSetType}}) where {RingType, RingElemType, MultSetType}
   ideal(W::AbsLocalizedRing{RingType, RingElemType, MultSetType}, v::Vector{RingElemType}) where {RingType, RingElemType, MultSetType}

for a single and a list of generators from both the base_ring of W and from W itself.

The minimal functionality which should be implemented for ideals is the test for ideal membership

Base.in(
    f::AbsLocalizedRingElem{RingType, RingElemType, MultSetType}, 
    I::AbsLocalizedIdeal{RingType, RingElemType, MultSetType}
  ) where {RingType, RingElemType, MultSetType}

and again the same for elements f of type RingElemType.

Basic operations on ideals which are already implemented on the generic level are

Base.:*(I::T, J::T) where {T<:AbsLocalizedIdeal}
Base.:+(I::T, J::T) where {T<:AbsLocalizedIdeal}

Everything else, such as e.g. intersections of ideals, has to be implemented for the specific types by the user.

Localizations of multivariate polynomial rings

Various primitive types of multiplicative sets are available, such as

MPolyComplementOfPrimeIdeal — Type

MPolyComplementOfPrimeIdeal{
    BaseRingType, 
    BaseRingElemType,
    RingType,
    RingElemType
  } <: AbsMPolyMultSet{
    BaseRingType, 
    BaseRingElemType,
    RingType,
    RingElemType
  }

The complement of a prime ideal $P ⊂ 𝕜[x₁,…,xₙ]$ in a multivariate polynomial ring with elements of type RingElemType over a base ring $𝕜$ of type BaseRingType.

MPolyComplementOfKPointIdeal — Type

MPolyComplementOfKPointIdeal{
    BaseRingType,
    BaseRingElemType, 
    RingType,
    RingElemType
  } <: AbsMPolyMultSet{
    BaseRingType,
    BaseRingElemType, 
    RingType, 
    RingElemType
  }

Complement of a maximal ideal $𝔪 = ⟨x₁-a₁,…,xₙ-aₙ⟩⊂ 𝕜[x₁,…xₙ]$ with $aᵢ∈ 𝕜$.

MPolyPowersOfElement — Type

MPolyPowersOfElement{
    BaseRingType,
    BaseRingElemType, 
    RingType,
    RingElemType
  } <: AbsMPolyMultSet{
    BaseRingType,
    BaseRingElemType, 
    RingType, 
    RingElemType
  }

The set S = { aᵏ : k ∈ ℕ₀ } for some $a ∈ R$ with $R$ of type BaseRingType.

Moreover, such types can be combined to products:

Definition (Products of multiplicative sets): Let $T$ and $U$ be multiplicative sets in a commutative ring $R$. The product of $T$ and $U$ is defined as

\[ T\cdot U = \left\{ f\cdot g : f \in T \textnormal{ and }g \in U \right\}.\]

A product of multiplicative sets $U = U_1 \cdot \dots \cdot U_r$ is called interreduced if neither one of the factors $U_i$ is contained in one of the others $U_j, j \neq i$.

Note that any product of multiplicative sets may be replaced by an interreduced one. However, such an interreduced multiplicative set is not unique as the following example shows:

Example (interreduction of products of multiplicative sets): An interreduced factorization of a product of multiplicative sets may not be unique: Consider the ring $\mathbb Z[x]$ and the multiplicative sets

\[ T = \left\{(5x)^k : k \in \mathbb N_0\right\}, \quad T' = \left\{ x^k : k \in \mathbb N_0\right\},\quad S = \left\{ c_0 \cdot x^0 : c_0 \notin 7 \mathbb Z\right\}.\]

Then $T\cdot S = \left\{ a⋅x^k : a \notin 7\mathbb Z, k \in \mathbb N_0 \right\} = T'\cdot S$.

Upshot: Whenever a product is taken, some interreduced form of the entire product is returned. Besides the obvious simplification in case all factors are contained in a single one, it is difficult to determine which interreduction is the best one. Localizations of multivariate polynomial rings are of type

The type for storing general products is

MPolyProductOfMultSets — Type

MPolyProductOfMultSets{
    BaseRingType,
    BaseRingElemType, 
    RingType,
    RingElemType
  } <: AbsMPolyMultSet{
    BaseRingType,
    BaseRingElemType, 
    RingType, 
    RingElemType
  }

A finite product T⋅U = { a⋅b : a ∈ T, b∈ U} of arbitrary other multiplicative sets in a multivariate polynomial ring.

and products can be taken using the usual arithmetic

product — Method

product(T::AbsMPolyMultSet, U::AbsMPolyMultSet)

Returns the product of the multiplicative sets T and U.

Note: The methods of this function naturally attempt to return a primitive type of multiplicative sets whenever possible. Hence, they are not type-stable.

Localizations of polynomial rings are of type

MPolyLocalizedRing — Type

MPolyLocalizedRing{
    BaseRingType,
    BaseRingElemType,
    RingType,
    RingElemType,
    MultSetType
  } <: AbsLocalizedRing{
    RingType,
    RingType,
    MultSetType
  }

The localization of a multivariate polynomial ring $R = 𝕜[x₁,…,xₙ]$ over a base field $𝕜$ of type BaseRingType and with elements of type RingElemType at a multiplicative set $S ⊂ R$ of type MultSetType.

with elements of type

MPolyLocalizedRingElem — Type

MPolyLocalizedRingElem{
    BaseRingType, 
    BaseRingElemType,
    RingType,
    RingElemType, 
    MultSetType
  } <: AbsLocalizedRingElem{
    RingType,
    RingElemType, 
    MultSetType
  }

Elements of localizations of polynomial rings.

Ideals in localized polynomial rings are of type

MPolyLocalizedIdeal — Type

MPolyLocalizedIdeal{BRT, BRET, RT, RET, MST} <: AbsLocalizedIdeal{RT, RET, MST}

Ideals in localizations of polynomial rings.

Recall (see e.g. [Eis95]) that if $\mathbb k$ is a Noetherian ring, any localization $W = R[U^{-1}]$ of a multivariate polynomial ring $R = \mathbb k[x_1,\dots,x_n]$ is again Noetherian and any ideal $I \subset W$ is of the form $I = I'\cdot W$ for some ideal $I' \subset R$. This correspondence is not 1:1 but for any ideal $I \subset W$ we always have that

\[ J = \left\{ x\in R : \exists u \in U : u\cdot x \in I \right\}\]

is the unique element which is maximal among all ideals $I'$ in $R$ for which $I = I'\cdot W$. We call this the saturated ideal of the localization and it can be obtained using

saturated_ideal — Method

saturated_ideal(I::MPolyLocalizedIdeal)

For an ideal $I ⊂ R[S⁻¹]$ in a localized polynomial ring this returns the unique ideal $J ⊂ R$ which is maximal among all those ideals $I' ⊂ R$ for which $I' ⋅ S⁻¹ = I$.

Groebner bases for the saturated ideal can be used to bring the numerators of any fraction $\frac{a}{b} \in R[S^{-1}]$ into normal form and check for ideal membership and/or equality of elements modulo ideals in $R[S^{-1}]$. But for some cases, e.g. when using local orderings for localizations at $\mathbb k$-points, it is desirable, to have the groebner- and standard basis functionality available directly in the localized ring. To this end we have

LocalizedBiPolyArray — Type

LocalizedBiPolyArray{BRT, BRET, RT, RET, MST}

Main workhorse for binding of ideals in localizations $R[S⁻¹]$ of multivariate polynomial rings $R = 𝕜[x₁,…,xₙ]$ to Singular. To a set of elements $f₁/g₁,…, fᵣ/gᵣ ∈ R[S⁻¹]$ this associates the numerators $ϕ(f₁),…,ϕ(fᵣ)$ as polynomials in Singular, possibly after applying a shift $ϕ : xᵢ ↦ xᵢ- aᵢ$ with constants $aᵢ∈ 𝕜$ depending on the type MST of the multiplicative set.

Note: The optional coordinate shift is to make local orderings available for localizations at arbitrary $𝕜$-points.

which has a monomial ordering and a Singular ring associated to it.

Note: Transfering an element $\frac{a}{b} \in R[S^{-1}]$ of a localized ring to the Singular-side drops all denominators and only the numerators appear as polynomials in Singular! Hence, a LocalizedBiPolyArray is not really a 1:1-correspondence of elements and in particular, the Oscar fractions can not be recovered from the Singular side.

This is also the type returned by any Groebner- or standard basis computation. We make the following convention:

Definition: Let $\mathbb k[x_1,\dots,x_n][S^{-1}]$ be a localized polynomial ring with $R = \mathbb k[x_1,\dots,x_n]$. A monomial ordering $\geq$ is compatible with the localization, if every unit in the localization $R_{\geq}$ is also a unit in $R[S^{-1}]$.

For an ideal $I \subset R[S^{-1}]$ and a compatible monomial ordering $\geq$ we say that a set of elements $\frac{g_1}{1},\dots,\frac{g_r}{1} \in R[S^{-1}]$ is a groebner/standard basis for $I$ if the elements $g_1,\dots,g_r$ are a standard basis for the saturated ideal $J$ of $I$ in $R$.

Note: When localizing at $\mathbb k$-points $a = (a_1,\dots,a_n) \in \mathbb k^n$ outside the origin, the transfer of polynomials from the Oscar to the Singular side in LocalizedBiPolyArray shifts the coordinates such that $a$ becomes zero. A monomial ordering is always considered after application of such shifts.

Groebner and standard bases of ideals can be computed for explicit orderings using

    groebner_basis(I::MPolyLocalizedIdeal, ord::Symbol)

Note that depending on the type parameters of I, this method is dispatched differently, which will also lead to different interpretations of the ordering. For instance, for multiplicative sets of type MPolyComplementOfKPointIdeal, a shift of variables taking the geometric point to the origin is applied to all polynomials when passing to the singular side.

If the second argument is omitted, a default ordering will be chosen, depending on the type of the multiplicative set.

Remark: Why bother introducing Groebner and standard basis for localized ideals in the first place and not only work with the saturated ideal? The main reason is that for localizations at $\mathbb k$-points, the computation of the saturated ideal is quite expensive: It involves a primary decomposition using a global ordering and discarding components outside the point at which has been localized. Using local orderings, on the other hand, we can decide ideal membership or equality of elements without computing the saturated ideal explicitly.

The following method might be of practical interest:

as_affine_algebra — Method

as_affine_algebra(
  L::MPolyLocalizedRing{BRT, BRET, RT, RET, 
  MPolyPowersOfElement{BRT, BRET, RT, RET}}; 
  inverse_name::String="θ"
) where {BRT, BRET, RT, RET}

For a localized polynomial ring $L = 𝕜[x₁,…,xₙ][f⁻¹]$ this returns a quintuple $(A, I, d, ϕ, θ)$ consisting of

an AffineAlgebra $A = 𝕜[x₁,…,xₙ,θ]/⟨1 - θ⋅d⟩$ isomorphic to $L$

the ideal $⟨1 - θ⋅d⟩$

an element $d ∈ 𝕜[x₁,…,xₙ]$ at which has been localized

the natural inclusion $ϕ : 𝕜[x₁,…,xₙ] ↪ A$

the localization variable $θ$ corresponding to $d⁻¹$.

Localizations of affine algebras

Let $R = 𝕜[x₁,…,xₘ]$ be a polynomial ring, $I ⊂ R$ some ideal and $P = R/I$ its quotient. Then $P$ is naturally an $R$-module and localization of $P$ as a ring coincides with localization as an $R$-module in the sense that for every multiplicative set $T ⊂ R$ there is a commutative diagram

\[\begin{matrix} R & → & P = R/I\\ ↓ & & ↓ \\ W = R[T⁻¹] & → & P[T⁻¹]. \end{matrix}\]

Observe that, moreover, for every multiplicative set $T' ⊂ P$ the preimage $T$ of $T'$ in $R$ is also a multiplicative set.

We may therefore treat localizations of polynomial algebras as localizations of modules over free polynomial rings: and apply the following

Convention: For localizations of affine algebras $L = (𝕜[x₁,…,xₙ]/I)[S⁻¹]$

ideals in $L$ are given by ideals in $W = 𝕜[x₁,…,xₙ][S⁻¹]$ containing $I\cdot S^{-1}$.

the available multiplicative sets for $L$ are exclusively those for $𝕜[x₁,…,xₙ]$.

Note that this leads to the following differences compared to the standard usage of the localization interface:

The base_ring returns neither $P$, nor $W$, but $R$.

The BaseRingType is the type of $R$ and similar for the other ring-based type parameters.

This is to make the data structure most accessible for the computational backends.

The type returned by numerator and denominator on an element of type MPolyQuoLocalizedRingElem is not RingElemType, but the type of $P$.

This is to comply with the purely mathematical viewpoint where elements of localized rings are fractions of residue classes rather than residue classes of fractions.

Localizations of affine algebras are realized by

MPolyQuoLocalizedRing — Type

MPolyQuoLocalizedRing{
    BaseRingType,
    BaseRingElemType,
    RingType,
    RingElemType,
    MultSetType <: AbsMultSet{RingType, RingElemType}
  } <: AbsLocalizedRing{
    RingType,
    RingElemType,
    MultSetType
  }

Localization $L = (𝕜[x₁,…,xₙ]/I)[S⁻¹]$ of a quotient $𝕜[x₁,…,xₙ]/I$ of a polynomial ring $P = 𝕜[x₁,…,xₙ]$ of type RingType over a base ring $𝕜$ of type BaseRingType at a multiplicative set $S ⊂ P$ of type MultSetType.

which have the additional methods

quotient_ring — Method

quotient_ring(L::MPolyQuoLocalizedRing)

For $L = (𝕜[x₁,…,xₙ]/I)[S⁻¹]$ this returns $𝕜[x₁,…,xₙ]/I$.

localized_ring — Method

localized_ring(L::MPolyQuoLocalizedRing)

For $L = (𝕜[x₁,…,xₙ]/I)[S⁻¹]$ this returns $𝕜[x₁,…,xₙ][S⁻¹]$.

returning the other two computationally important rings in its construction.

Elements of such rings are of the form

MPolyQuoLocalizedRingElem — Type

MPolyQuoLocalizedRingElem{
  BaseRingType, 
  BaseRingElemType,
  RingType,
  RingElemType, 
  MultSetType
} <: AbsLocalizedRingElem{
  RingType,
  RingElemType, 
  MultSetType
}

Elements $a//b$ of localizations $L = (𝕜[x₁,…,xₙ]/I)[S⁻¹]$ of type MPolyQuoLocalizedRing{BaseRingType, BaseRingElemType, RingType, RingElemType, MultSetType}.

In contrast to ordinary elements of a localized ring, they have the additional methods

lifted_numerator — Method

lifted_numerator(a::MPolyQuoLocalizedRingElem)

For $A//B ∈ (𝕜[x₁,…,xₙ]/I)[S⁻¹]$ this returns a representative $a ∈ 𝕜[x₁,…,xₙ]$ of the numerator.

lifted_denominator — Method

lifted_denominator(a::MPolyQuoLocalizedRingElem)

For $A//B ∈ (𝕜[x₁,…,xₙ]/I)[S⁻¹]$ this returns a representative $b ∈ 𝕜[x₁,…,xₙ]$ of the denominator.

fraction — Method

fraction(a::MPolyQuoLocalizedRingElem)

For $A//B ∈ (𝕜[x₁,…,xₙ]/I)[S⁻¹]$ this returns a representative $a//b ∈ Quot(𝕜[x₁,…,xₙ])$ of the fraction.