$G_4$-Fluxes

In F-theory compactifications on Calabi–Yau 4-folds, $G_4$-fluxes are essential ingredients for defining consistent and physically meaningful vacua. They affect various physical aspects, such as the chiral spectrum in the 4D effective theory, tadpole cancellation, and moduli stabilization. This page provides an accessible introduction to working with $G_4$-fluxes in FTheoryTools, including background, construction methods, and available functionality.

Background: What Is a $G_4$-Flux?

Let $\widehat{Y}_4$ be a smooth Calabi–Yau 4-fold obtained via crepant resolution of a singular elliptically fibered 4-fold $Y_4 \to B_3$. A $G_4$-flux is a cohomology class:

\[G_4 \in H^{2, 2}(\widehat{Y}_4, \mathbb{R}) \equiv H^{2, 2}(\widehat{Y}_4, \mathbb{C}) \cap H^4(\widehat{Y}_4, \mathbb{R})\]

subject to quantization, transversality, and several additional consistency conditions:

Quantization (cf. Witten 1997): $G_4 + \frac{1}{2} c_2(\widehat{Y}_4) \in H^4(\widehat{Y}_4, \mathbb{Z})$

Transversality: The flux must be orthogonal to both the base and the fiber in a precise cohomological sense.

Additional constraints, such as D3-tadpole cancellation, self-duality, and primitivity, are also typically imposed. One often requires that the $G_4$-flux preserves the non-abelian gauge symmetry encoded in the fibration.

These conditions ensure compatibility with fundamental physical principles like anomaly cancellation and D3-brane charge quantization.

For further theoretical background, see the exposition in Weigand 2018.

Modeling $G_4$-Fluxes

In practice, FTheoryTools focuses on candidate $G_4$-fluxes: elements of the rational cohomology group $H^{2,2}(\widehat{Y}_4, \mathbb{Q})$, i.e.

\[G_4 \in H^{2, 2}(\widehat{Y}_4, \mathbb{Q}) \equiv H^{2, 2}(\widehat{Y}_4, \mathbb{C}) \cap H^4(\widehat{Y}_4, \mathbb{Q})\,.\]

We work with rational coefficients due to both the quantization condition–-which implies that valid fluxes must have rational coefficients–-and the practical advantages this offers in computation. Currently, FTheoryTools cannot fully verify the quantization condition, so we adopt the term candidate flux. We elaborate on our approach to the quantization condition below.

The space $H^{2,2}(\widehat{Y}_4, \mathbb{C})$ admits an orthogonal decomposition:

\[H^{2, 2}(\widehat{Y}_4, \mathbb{C}) = H^{2, 2}_\text{hor}(\widehat{Y}_4, \mathbb{C}) \oplus H^{2, 2}_\text{vert}(\widehat{Y}_4, \mathbb{C}) \oplus H^{2, 2}_\text{rem}(\widehat{Y}_4, \mathbb{C})\,.\]

Horizontal fluxes arise from variations of the complex structure,

Vertical fluxes are built from wedge products of (1,1)-forms,

Remainder fluxes account for the rest.

FTheoryTools focuses on the vertical part.

Standing Assumption

FTheoryTools assumes that $\widehat{Y}_4$ is defined as a hypersurface in a complete, smooth toric variety $X_\Sigma$. (In principle, it would suffice for the hypersurface to be smooth and the ambient space to be simplicial. However, we compute characteristic classes of the ambient space and restrict them to the hypersurface to obtain $c_2(\widehat{Y}_4)$, and this requires the ambient space to be smooth.)

Workflow

Under this assumption, we can access a subspace of the vertical cohomology:

\[\left. H^{2,2}(X_\Sigma, \mathbb{Q}) \right|_{\widehat{Y}_4} \subseteq H^{2,2}_{\text{vert}}(\widehat{Y}_4, \mathbb{Q}) \,.\]

We refer to elements in this subspace as ambient vertical fluxes.

Theorem 9.3.2 of Cox, Little, Schenck 2011 states that for complete, simplicial toric varieties, $H^{p, q}(X_\Sigma, \mathbb{Q}) = 0$ whenever $p \ne q$. Consequently, $H^{2, 2}(X_\Sigma, \mathbb{Q}) = H^4(X_\Sigma, \mathbb{Q})$. Furthermore, Theorem 12.4.1 of Cox, Little, Schenck 2011 establishes that $H^4(X_\Sigma, \mathbb{Q})$ is isomorphic to $R_{\mathbb{Q}}(\Sigma)_2$, the space of degree-2 monomials (under the standard grading) in the cohomology ring $R_{\mathbb{Q}}(\Sigma)$ of $X_\Sigma$. This cohomology ring is the quotient of $\mathbb{Q}[x_1, \dots, x_r]$, where $r$ is the number of rays in the polyhedral fan $\Sigma$ of the toric ambient space $X_\Sigma$, by the sum of the Stanley–Reisner ideal and the ideal of linear relations.

To construct a generating set of ambient candidate $G_4$-fluxes, we proceed as follows:

Construct the cohomology ring $R_{\mathbb{Q}}(\Sigma)$.

Identify a basis of degree-2 monomials in $R_{\mathbb{Q}}(\Sigma)$.

Restrict these monomials to $\widehat{Y}_4$.

While performing the restriction to $\widehat{Y}_4$, dependencies may arise:

Some monomials vanish identically,

Others may become linearly dependent.

FTheoryTools currently detects only obviously trivial restrictions (e.g., cycles whose intersection with $\widehat{Y}_4$ is empty). It does not eliminate all possible dependencies–-hence the result is a generating set, not a minimal basis. This pragmatic choice enables efficient and scalable construction of vertical flux candidates across a wide range of F-theory geometries.

FTheoryTools includes specialized algorithms to perform steps 1 through 3. These algorithms–-described in detail in the appendix of BMT25–-are especially effective for large geometries such as those studied in Taylor, Wang 2015, where traditional Gröbner basis techniques are computationally impractical. Even for more manageable QSM geometries (cf. Cvetič, Halverson, Ling, Liu, Tian 2019), we observe a significant performance boost.

Necessary Conditions for Flux Quantization

Verifying the flux quantization condition (cf. Witten 1997) is generally difficult. In practice, FTheoryTools performs necessary consistency checks to gather evidence for (or against) proper quantization.

Let $H$ denote the hypersurface divisor class defining $\widehat{Y}_4$ inside the ambient toric variety $X_\Sigma$. Let $\{D_i\}$ be a basis for the toric divisor classes of $X_\Sigma$, and let $\hat{c}_2 \in H^{2,2}(X_\Sigma, \mathbb{Q})$ be a class whose restriction gives the second Chern class of the Calabi–Yau hypersurface, i.e., $\left. \hat{c}_2 \right|_{\widehat{Y}_4} = c_2(\widehat{Y}_4)$.

As elaborated above, in FTheoryTools, a $G_4$-flux candidate is modeled by a class $g \in H^{2,2}(X_\Sigma, \mathbb{Q})$, and its restriction to $\widehat{Y}_4$ defines the actual flux candidate. A necessary condition for quantization of this flux $G_4$ –- modelled by $g \in H^{2,2}(X_\Sigma, \mathbb{Q})$ –- is that, for all pairs of indices $i$, $j$, the following integral evaluates to an integer:

\[\int_{X_\Sigma}{\left(g + \frac{1}{2} \hat{c}_2 \right) \wedge [H] \wedge [D_i] \wedge [D_j]}\,,\]

where $[D]$ indicates the Poincaré dual cohomology class to the divisor $D$. It is these conditions that we check in FTheoryTools.

Even these elementary checks can be computationally expensive on large geometries. Therefore, FTheoryTools allows users to skip quantization checks during flux construction using the keyword argument check = false. This can be useful in early-stage explorations or for performance optimization.

Working with $G_4$-Fluxes

Constructing $G_4$-Fluxes

In FTheoryTools, $G_4$-fluxes are created from cohomology classes associated with a given resolved F-theory model.

g4_flux — Method

g4_flux(model::AbstractFTheoryModel, class::CohomologyClass; check::Bool = true)

Construct a candidate $G_4$-flux for a resolved F-theory model from a given cohomology class on the toric ambient space.

Requirements

The model must be defined via a Weierstrass, global Tate, or hypersurface construction.

The ambient space must a smooth and complete toric variety.

Parameters

check (default: true): Perform basic consistency and necessary quantization checks.

Examples

julia> qsm_model = literature_model(arxiv_id = "1903.00009", model_parameters = Dict("k" => 4))
Hypersurface model over a concrete base

julia> cohomology_ring(ambient_space(qsm_model), check = false);

julia> g4_class = cohomology_class(anticanonical_divisor_class(ambient_space(qsm_model)), quick = true)^2;

julia> g4f = g4_flux(qsm_model, g4_class, check = false)
G4-flux candidate
  - Elementary quantization checks: not executed
  - Transversality checks: not executed
  - Non-abelian gauge group: breaking pattern not analyzed
  - Tadpole cancellation check: not computed

Experimental

This function is part of the experimental code in Oscar. Please read here for more details.

For the special class of quadrillion Standard models (QSMs) (cf. Cvetič, Halverson, Ling, Liu, Tian 2019), the chosen flux can be obtained conveniently.

qsm_flux — Method

qsm_flux(qsm_model::AbstractFTheoryModel)

Return the $G_4$-flux associated with one of the Quadrillion F-theory Standard models, as described in CHLLT19.

This flux has been pre-validated to pass essential consistency checks.

Examples

julia> qsm_model = literature_model(arxiv_id = "1903.00009", model_parameters = Dict("k" => 4))
Hypersurface model over a concrete base

julia> qsm_flux(qsm_model)
G4-flux candidate
  - Elementary quantization checks: satisfied
  - Transversality checks: satisfied
  - Non-abelian gauge group: unbroken
  - Tadpole cancellation check: not computed

Experimental

This function is part of the experimental code in Oscar. Please read here for more details.

Attributes of $G_4$-Fluxes

You can inspect a $G_4$-flux using the following attributes:

model — Method

model(gf::G4Flux)

Return the F-theory model used to construct the G_4-flux candidate.

julia> qsm_model = literature_model(arxiv_id = "1903.00009", model_parameters = Dict("k" => 4))
Hypersurface model over a concrete base

julia> g4_candidate = qsm_flux(qsm_model)
G4-flux candidate
  - Elementary quantization checks: satisfied
  - Transversality checks: satisfied
  - Non-abelian gauge group: unbroken
  - Tadpole cancellation check: not computed

julia> model(g4_candidate)
Hypersurface model over a concrete base

julia> model(g4_candidate) == qsm_model
true

Experimental

This function is part of the experimental code in Oscar. Please read here for more details.

cohomology_class — Method

cohomology_class(gf::G4Flux)

Return the ambient space cohomology class which defines the $G_4$-flux candidate.

julia> qsm_model = literature_model(arxiv_id = "1903.00009", model_parameters = Dict("k" => 4))
Hypersurface model over a concrete base

julia> cohomology_ring(ambient_space(qsm_model), check = false);

julia> g4_class = cohomology_class(anticanonical_divisor_class(ambient_space(qsm_model)), quick = true)^2;

julia> g4f = g4_flux(qsm_model, g4_class, check = false)
G4-flux candidate
  - Elementary quantization checks: not executed
  - Transversality checks: not executed
  - Non-abelian gauge group: breaking pattern not analyzed
  - Tadpole cancellation check: not computed

julia> cohomology_class(g4f);

Experimental

This function is part of the experimental code in Oscar. Please read here for more details.

d3_tadpole_constraint — Method

d3_tadpole_constraint(gf::G4Flux; check::Bool = true)

Return the d3-tapdole of a G4-flux, that is compute and return the quantity $- \frac{1}{2} \cdot \int_{\widehat{Y_4}}{G_4 \wedge G_4} + \frac{1}{24} \cdot \chi(\widehat{Y}_4)$.

julia> qsm_model = literature_model(arxiv_id = "1903.00009", model_parameters = Dict("k" => 4))
Hypersurface model over a concrete base

julia> g4 = qsm_flux(qsm_model)
G4-flux candidate
  - Elementary quantization checks: satisfied
  - Transversality checks: satisfied
  - Non-abelian gauge group: unbroken
  - Tadpole cancellation check: not computed

julia> d3_tadpole_constraint(g4, check = false)
12

julia> qsm_model = literature_model(arxiv_id = "1903.00009", model_parameters = Dict("k" => 2021))
Hypersurface model over a concrete base

julia> gfs = special_flux_family(qsm_model, check = false)
Family of G4 fluxes:
  - Elementary quantization checks: satisfied
  - Transversality checks: satisfied
  - Non-abelian gauge group: breaking pattern not analyzed

julia> g4_2 = random_flux_instance(gfs, check = false)
G4-flux candidate
  - Elementary quantization checks: satisfied
  - Transversality checks: satisfied
  - Non-abelian gauge group: breaking pattern not analyzed
  - Tadpole cancellation check: not computed

julia> dv2 = d3_tadpole_constraint(g4_2, check = false);

julia> int_comb = integral_coefficients(g4_2);

julia> rat_comb = rational_coefficients(g4_2);

julia> g4_3 = flux_instance(gfs, int_comb, rat_comb, check = false)
G4-flux candidate
  - Elementary quantization checks: satisfied
  - Transversality checks: satisfied
  - Non-abelian gauge group: breaking pattern not analyzed
  - Tadpole cancellation check: not computed

julia> d3_tadpole_constraint(gfs, check = false);

julia> dv3 = d3_tadpole_constraint(g4_3, check = false);

julia> dv2 == dv3
true

Experimental

This function is part of the experimental code in Oscar. Please read here for more details.

Properties of $G_4$-Fluxes

You can check whether a $G_4$-flux satisfies common physical constraints using:

is_well_quantized — Method

is_well_quantized(gf::G4Flux)

G4-fluxes are subject to the quantization condition [Wit97] $G_4 + \frac{1}{2} c_2(Y) \in H^{(2,2)}(Y, \mathbb{Z})$. It is hard to verify that this condition is met. However, we can execute a number of simple consistency checks, by verifying that $\int_{Y}{G_4 \wedge [D_1] \wedge [D_2]} \in \mathbb{Z}$ for any two toric divisors $D_1$, $D_2$. If all of these simple consistency checks are met, this method will return true and otherwise false.

It is worth mentioning that currently (August 2024), we only support this check for $G_4$-fluxes defined on Weierstrass, global Tate and hypersurface models. If this condition is not met, this method will return an error.

julia> qsm_model = literature_model(arxiv_id = "1903.00009", model_parameters = Dict("k" => 4))
Hypersurface model over a concrete base

julia> cohomology_ring(ambient_space(qsm_model), check = false);

julia> g4_class = cohomology_class(anticanonical_divisor_class(ambient_space(qsm_model)))^2;

julia> g4 = g4_flux(qsm_model, g4_class, check = false)
G4-flux candidate
  - Elementary quantization checks: not executed
  - Transversality checks: not executed
  - Non-abelian gauge group: breaking pattern not analyzed
  - Tadpole cancellation check: not computed

julia> is_well_quantized(g4)
true

julia> g4
G4-flux candidate
  - Elementary quantization checks: satisfied
  - Transversality checks: not executed
  - Non-abelian gauge group: breaking pattern not analyzed
  - Tadpole cancellation check: not computed

Experimental

This function is part of the experimental code in Oscar. Please read here for more details.

passes_transversality_checks — Method

passes_transversality_checks(gf::G4Flux)

G4-fluxes are subject to transversality conditions (cf. [Wei18]). If these conditions are met, this method will return true and otherwise false

julia> qsm_model = literature_model(arxiv_id = "1903.00009", model_parameters = Dict("k" => 4))
Hypersurface model over a concrete base

julia> divs = torusinvariant_prime_divisors(ambient_space(qsm_model));

julia> e1 = cohomology_class(divs[35]);e2 = cohomology_class(divs[32]);e4 = cohomology_class(divs[34]);

julia> u = cohomology_class(divs[33]);v = cohomology_class(divs[30]);pb_Kbar = cohomology_class(sum([divs[k] for k in 1:29]));

julia> g4_class = (-3) // kbar3(qsm_model) * (5 * e1 * e4 + pb_Kbar * (-3 * e1 - 2 * e2 - 6 * e4 + pb_Kbar - 4 * u + v));

julia> g4 = g4_flux(qsm_model, g4_class, check = false)
G4-flux candidate
  - Elementary quantization checks: not executed
  - Transversality checks: not executed
  - Non-abelian gauge group: breaking pattern not analyzed
  - Tadpole cancellation check: not computed

julia> passes_transversality_checks(g4)
true

julia> g4
G4-flux candidate
  - Elementary quantization checks: not executed
  - Transversality checks: satisfied
  - Non-abelian gauge group: breaking pattern not analyzed
  - Tadpole cancellation check: not computed

Experimental

This function is part of the experimental code in Oscar. Please read here for more details.

passes_tadpole_cancellation_check — Method

passes_tadpole_cancellation_check(gf::G4Flux)

G4-fluxes are subject to the D3-tadpole cancellation condition described in [Wei18]. This check verifies that $euler_characteristic(Y)/24 - 1/2 * \int_{Y}{G_4 \wedge G_4}$ is a non-negative integer.

julia> qsm_model = literature_model(arxiv_id = "1903.00009", model_parameters = Dict("k" => 4))
Hypersurface model over a concrete base

julia> divs = torusinvariant_prime_divisors(ambient_space(qsm_model));

julia> e1 = cohomology_class(divs[35]);e2 = cohomology_class(divs[32]);e4 = cohomology_class(divs[34]);

julia> u = cohomology_class(divs[33]);v = cohomology_class(divs[30]);pb_Kbar = cohomology_class(sum([divs[k] for k in 1:29]));

julia> g4_class = (-3) // kbar3(qsm_model) * (5 * e1 * e4 + pb_Kbar * (-3 * e1 - 2 * e2 - 6 * e4 + pb_Kbar - 4 * u + v));

julia> g4 = g4_flux(qsm_model, g4_class, check = false)
G4-flux candidate
  - Elementary quantization checks: not executed
  - Transversality checks: not executed
  - Non-abelian gauge group: breaking pattern not analyzed
  - Tadpole cancellation check: not computed

julia> passes_tadpole_cancellation_check(g4)
true

julia> g4
G4-flux candidate
  - Elementary quantization checks: not executed
  - Transversality checks: not executed
  - Non-abelian gauge group: breaking pattern not analyzed
  - Tadpole cancellation check: satisfied

Experimental

This function is part of the experimental code in Oscar. Please read here for more details.

breaks_non_abelian_gauge_group — Method

breaks_non_abelian_gauge_group(gf::G4Flux)

G4-fluxes may break the non-abelian gauge group (cf. [Wei18]). This function verifies if this is the case for the given G4-flux. If it does not break any non-abelian gauge factor, we return true and otherwise false

julia> qsm_model = literature_model(arxiv_id = "1903.00009", model_parameters = Dict("k" => 4))
Hypersurface model over a concrete base

julia> divs = torusinvariant_prime_divisors(ambient_space(qsm_model));

julia> e1 = cohomology_class(divs[35]);e2 = cohomology_class(divs[32]);e4 = cohomology_class(divs[34]);

julia> u = cohomology_class(divs[33]);v = cohomology_class(divs[30]);pb_Kbar = cohomology_class(sum([divs[k] for k in 1:29]));

julia> g4_class = (-3) // kbar3(qsm_model) * (5 * e1 * e4 + pb_Kbar * (-3 * e1 - 2 * e2 - 6 * e4 + pb_Kbar - 4 * u + v));

julia> g4 = g4_flux(qsm_model, g4_class, check = false)
G4-flux candidate
  - Elementary quantization checks: not executed
  - Transversality checks: not executed
  - Non-abelian gauge group: breaking pattern not analyzed
  - Tadpole cancellation check: not computed

julia> breaks_non_abelian_gauge_group(g4)
false

julia> g4
G4-flux candidate
  - Elementary quantization checks: not executed
  - Transversality checks: not executed
  - Non-abelian gauge group: unbroken
  - Tadpole cancellation check: not computed

Experimental

This function is part of the experimental code in Oscar. Please read here for more details.

Working with Families of G_4-Fluxes

In many F-theory applications, it is useful to study entire families of fluxes rather than individual ones. FTheoryTools supports both the construction and analysis of such flux families.

Constructing a Flux Family from Parametrization and Offset

You can construct a flux family explicitly by specifying its parametrization in terms of the ambient G_4-flux generators:

chosen_g4_flux_gens — Method

chosen_g4_flux_gens(m::AbstractFTheoryModel; check::Bool = true)::Vector{CohomologyClass}

Given an F-theory model m defined as a hypersurface in a simplicial and complete toric base, this method computes a basis of $H^{2,2}(X, \mathbb{Q})$ (using the method basis_of_h22) and then filters out certain basis elements whose restriction to the hypersurface in question is trivial. The criteria for "certain" elements are explained in the documentation above this method.

Note: Checking whether a toric variety $X$ is complete and simplicial can be computationally expensive. The optional argument check can be set to false to skip these tests.

Examples

julia> B3 = projective_space(NormalToricVariety, 3)
Normal toric variety

julia> Kbar = anticanonical_divisor_class(B3)
Divisor class on a normal toric variety

julia> t = literature_model(arxiv_id = "1109.3454", equation = "3.1", base_space = B3, defining_classes = Dict("w"=>Kbar))
Construction over concrete base may lead to singularity enhancement. Consider computing singular_loci. However, this may take time!

Global Tate model over a concrete base -- SU(5)xU(1) restricted Tate model based on arXiv paper 1109.3454 Eq. (3.1)

julia> g4_basis = chosen_g4_flux_gens(t);

julia> length(g4_basis)
2

julia> g4_basis[1]
G4-flux candidate
  - Elementary quantization checks: not executed
  - Transversality checks: not executed
  - Non-abelian gauge group: breaking pattern not analyzed
  - Tadpole cancellation check: not computed

julia> cohomology_class(g4_basis[1])
Cohomology class on a normal toric variety given by y^2

julia> cohomology_class(g4_basis[2])
Cohomology class on a normal toric variety given by z^2

julia> qsm_model = literature_model(arxiv_id = "1903.00009", model_parameters = Dict("k" => 8))
Hypersurface model over a concrete base

julia> g4_basis = chosen_g4_flux_gens(qsm_model, check = false);

julia> cohomology_class(g4_basis[1])
Cohomology class on a normal toric variety given by x15*e2

julia> length(g4_basis) == 172
true

Experimental

This function is part of the experimental code in Oscar. Please read here for more details.

To define a flux family in this way, you must provide:

Integral coefficients: Rational combinations of generators whose \mathbb{Z}-span lies within the family,

Rational coefficients: Rational combinations of generators forming the \mathbb{Q}-span of the family,

Offset vector: Typically representing the shift by \frac{1}{2} c_2(\widehat{Y}_4) from the quantization condition.

The constructor is:

family_of_g4_fluxes — Method

family_of_g4_fluxes(m::AbstractFTheoryModel, mat_int::QQMatrix, mat_rat::QQMatrix; check::Bool = true)

Given an F-theory model with a toric ambient space, we can identify ambient space candidates of G4-fluxes. In terms of these candidates, we can define a family of G4-fluxes as:

$\mathbb{Z}$-linear combinations, provided by a matrix $mat_int$,

$\mathbb{Q}$-linear combinations, provided by a matrix $mat_rat$,

a shift (invoked for instance by the appearance of $\frac{1}{2} \cdot c_2$ in the quantization condition), provided by a vector $offset$.

For convenience we also allow to only provide $mat_int$, $mat_rat$. In this case, the shift is taken to be zero.

An example is in order.

Examples

julia> qsm_model = literature_model(arxiv_id = "1903.00009", model_parameters = Dict("k" => 2021))
Hypersurface model over a concrete base

julia> mat_int = zero_matrix(QQ, 37, 1);

julia> mat_int[1,1] = 1;

julia> mat_rat = zero_matrix(QQ, 37, 1);

julia> mat_rat[2,1] = 1;

julia> shift = [zero(QQ) for k in 1:37];

julia> family_of_g4_fluxes(qsm_model, mat_int, mat_rat, shift, check = false)
Family of G4 fluxes:
  - Elementary quantization checks: not executed
  - Transversality checks: not executed
  - Non-abelian gauge group: breaking pattern not analyzed

Experimental

This function is part of the experimental code in Oscar. Please read here for more details.

Once a flux family is constructed, you can extract its defining data:

matrix_integral — Method

matrix_integral(gf::FamilyOfG4Fluxes)

Return the matrix whose columns specify those combinations of ambient space G4-flux candidates, of which integral linear combinations are contained in this family of G4-fluxes.

julia> qsm_model = literature_model(arxiv_id = "1903.00009", model_parameters = Dict("k" => 2021))
Hypersurface model over a concrete base

julia> mat_int = zero_matrix(QQ, 37, 1);

julia> mat_int[1,1] = 1;

julia> mat_rat = zero_matrix(QQ, 37, 1);

julia> mat_rat[2,1] = 1;

julia> shift = [zero(QQ) for k in 1:37];

julia> f_gs = family_of_g4_fluxes(qsm_model, mat_int, mat_rat, shift, check = false)
Family of G4 fluxes:
  - Elementary quantization checks: not executed
  - Transversality checks: not executed
  - Non-abelian gauge group: breaking pattern not analyzed

julia> matrix_integral(f_gs) == mat_int
true

Experimental

This function is part of the experimental code in Oscar. Please read here for more details.

matrix_rational — Method

matrix_rational(gf::FamilyOfG4Fluxes)

Return the matrix whose columns specify those combinations of ambient space G4-flux candidates, of which rational linear combinations are contained in this family of G4-fluxes.

julia> qsm_model = literature_model(arxiv_id = "1903.00009", model_parameters = Dict("k" => 2021))
Hypersurface model over a concrete base

julia> mat_int = zero_matrix(QQ, 37, 1);

julia> mat_int[1,1] = 1;

julia> mat_rat = zero_matrix(QQ, 37, 1);

julia> mat_rat[2,1] = 1;

julia> shift = [zero(QQ) for k in 1:37];

julia> f_gs = family_of_g4_fluxes(qsm_model, mat_int, mat_rat, shift, check = false)
Family of G4 fluxes:
  - Elementary quantization checks: not executed
  - Transversality checks: not executed
  - Non-abelian gauge group: breaking pattern not analyzed

julia> matrix_rational(f_gs) == mat_rat
true

Experimental

This function is part of the experimental code in Oscar. Please read here for more details.

offset — Method

offset(gf::FamilyOfG4Fluxes)

Return the vector whose entries specify the offset by which fluxes in this family of G4-fluxes are shifted.

julia> qsm_model = literature_model(arxiv_id = "1903.00009", model_parameters = Dict("k" => 2021))
Hypersurface model over a concrete base

julia> mat_int = zero_matrix(QQ, 37, 1);

julia> mat_int[1,1] = 1;

julia> mat_rat = zero_matrix(QQ, 37, 1);

julia> mat_rat[2,1] = 1;

julia> shift = [zero(QQ) for k in 1:37];

julia> f_gs = family_of_g4_fluxes(qsm_model, mat_int, mat_rat, shift, check = false)
Family of G4 fluxes:
  - Elementary quantization checks: not executed
  - Transversality checks: not executed
  - Non-abelian gauge group: breaking pattern not analyzed

julia> offset(f_gs) == shift
true

Experimental

This function is part of the experimental code in Oscar. Please read here for more details.

Constructing a Flux Family Based on Physical Conditions

A common use case is to construct the family of all G_4-fluxes satisfying key physical constraints:

Passing the necessary quantization checks (as discussed earlier),

Satisfying all transversality conditions,

Optionally preserving the non-abelian gauge symmetry.

This method is less flexible than direct parametrization, but very convenient for typical model-building workflows.

!!! warning This operation may be computationally expensive for large or complex geometries.

special_flux_family — Method

special_flux_family(m::AbstractFTheoryModel; not_breaking::Bool = false, check::Bool = true, algorithm::String = "default")

Compute a family of G4-fluxes with specified properties for a given F-theory model m, defined as a hypersurface in a simplicial and complete toric ambient space.

Description

This method models the G4-flux family using the restriction of cohomology classes on the toric ambient space to the hypersurface. In the toric ambient space, these classes are vertical, meaning they are of the form $a \wedge b$, where $a, b \in H^(1,1)(X_\Sigma)$, with $X_\Sigma$ denoting the toric ambient space.

The resulting family is subjected to consistency conditions to ensure it satisfies elementary quantization requirements and transversality conditions. By default, the method returns ambient space G4-flux candidates that meet these conditions.

Optional Arguments

not_breaking = true: Ensures the flux family preserves the non-abelian gauge group.

check = false: Skips computational checks for whether the toric ambient space $X_\Sigma$ is complete and simplicial, which can be resource-intensive.

Notes

Warning

This method assumes that $c_2( \widehat{Y}_4)$ is even. No checks or errors are implemented for this condition, so use cautiously.

This assumption relates to the quantization condition, which requires verifying if the twist of a given $G_4$-flux by $1/2 \cdot c_2( \widehat{Y}_4)$ is even. For many F-theory models, such as compactifications on smooth Calabi-Yau 4-folds with globally defined Weierstrass models ([CS12]), $c_2( \widehat{Y}_4)$ is known to be even. This also applies to all F-theory QSMs ([CHLLT19]).

Computational Details

The method internally identifies two matrices related to the family of fluxes:

matrix_integral: Specifies rational combinations of ambient space G4-flux candidates that can only form $\mathbb{Z}$-linear combinations without violating elementary flux quantization conditions.

matrix_rational: Specifies rational combinations of flux candidates for which any rational linear combination satisfies the elementary flux quantization conditions.

These matrices are accessible for further analysis.

Examples

julia> qsm_model = literature_model(arxiv_id = "1903.00009", model_parameters = Dict("k" => 2021))
Hypersurface model over a concrete base

julia> fg = special_flux_family(qsm_model, check = false)
Family of G4 fluxes:
  - Elementary quantization checks: satisfied
  - Transversality checks: satisfied
  - Non-abelian gauge group: breaking pattern not analyzed

julia> g4_tester = random_flux_instance(fg, check = false)
G4-flux candidate
  - Elementary quantization checks: satisfied
  - Transversality checks: satisfied
  - Non-abelian gauge group: breaking pattern not analyzed
  - Tadpole cancellation check: not computed

julia> g4_tester_double = g4_flux(qsm_model, cohomology_class(g4_tester), check = false);

julia> is_well_quantized(g4_tester_double)
true

julia> passes_transversality_checks(g4_tester_double)
true

julia> c = [60, 51, 90, 0, 24, 51, -24, 45, 30, 0, -48, 90, -57, 60, 30, 15, 120, 0, -60, 0, -720, -420, -270, -60, -2190];

julia> qsm_g4_flux = flux_instance(fg, transpose(matrix(ZZ, [c])), zero_matrix(QQ, 0, 1), check = false)
G4-flux candidate
  - Elementary quantization checks: satisfied
  - Transversality checks: satisfied
  - Non-abelian gauge group: breaking pattern not analyzed
  - Tadpole cancellation check: not computed

julia> passes_transversality_checks(qsm_g4_flux)
true

julia> is_well_quantized(qsm_g4_flux)
true

Finally, we demonstrate the computation of the well-quantized $G_4$-fluxes which pass the transversality checks, and which in addition do not break the non-abelian gauge group.

julia> qsm_model = literature_model(arxiv_id = "1903.00009", model_parameters = Dict("k" => 2021))
Hypersurface model over a concrete base

julia> fg = special_flux_family(qsm_model, not_breaking = true, check = false)
Family of G4 fluxes:
  - Elementary quantization checks: satisfied
  - Transversality checks: satisfied
  - Non-abelian gauge group: unbroken

julia> g4_tester = random_flux_instance(fg, check = false)
G4-flux candidate
  - Elementary quantization checks: satisfied
  - Transversality checks: satisfied
  - Non-abelian gauge group: unbroken
  - Tadpole cancellation check: not computed

julia> g4_tester_double = g4_flux(qsm_model, cohomology_class(g4_tester), check = false);

julia> is_well_quantized(g4_tester_double)
true

julia> passes_transversality_checks(g4_tester_double)
true

julia> breaks_non_abelian_gauge_group(g4_tester_double)
false

julia> c = [3];

julia> qsm_g4_flux = flux_instance(fg, matrix(ZZ, [[3]]), zero_matrix(QQ, 0, 1), check = false)
G4-flux candidate
  - Elementary quantization checks: satisfied
  - Transversality checks: satisfied
  - Non-abelian gauge group: unbroken
  - Tadpole cancellation check: not computed

julia> is_well_quantized(qsm_g4_flux)
true

julia> passes_transversality_checks(qsm_g4_flux)
true

julia> breaks_non_abelian_gauge_group(qsm_g4_flux)
false

julia> qsm_g4_flux == qsm_flux(qsm_model)
true

Experimental

This function is part of the experimental code in Oscar. Please read here for more details.

Attributes of a Flux Family

You can access key data about a flux family using the following attributes:

model — Method

model(gf::FamilyOfG4Fluxes)

Return the F-theory model for which this family of $G_4$-flux candidates is defined.

julia> qsm_model = literature_model(arxiv_id = "1903.00009", model_parameters = Dict("k" => 2021))
Hypersurface model over a concrete base

julia> mat_int = zero_matrix(QQ, 37, 1);

julia> mat_int[1,1] = 1;

julia> mat_rat = zero_matrix(QQ, 37, 1);

julia> mat_rat[2,1] = 1;

julia> shift = [zero(QQ) for k in 1:37];

julia> f_gs = family_of_g4_fluxes(qsm_model, mat_int, mat_rat, shift, check = false)
Family of G4 fluxes:
  - Elementary quantization checks: not executed
  - Transversality checks: not executed
  - Non-abelian gauge group: breaking pattern not analyzed

julia> model(f_gs) == qsm_model
true

Experimental

This function is part of the experimental code in Oscar. Please read here for more details.

d3_tadpole_constraint — Method

d3_tadpole_constraint(fgs::FamilyOfG4Fluxes; check::Bool = true)

Return the d3-tapdole constraint of Family of G4-fluxes. Recall that for a given $G_4$-flux, this constraint is $- \frac{1}{2} \cdot G_4^2 + \frac{1}{24} \cdot \chi(\widehat{Y}_4) \stackrel{!}{\geq} 0$.

Note that the family of fluxes is specified by linear combination of cohomology classes, some with rational and some with integral coefficients. In terms of these coefficients the d3-tadpole constraint is the demand that a quadratic polynomial in the coefficients evaluates to a non-negative number. This method returns said polynomial in the cofficients. In order to evaluate the D3-tadpole for a particular $G_4$-flux, one has to evaluate this polynomial for the numeric coefficient values that correspond to the given $G_4$-flux.

We use symbols $a_i$ to indicate integral coefficients and $r_i$ to indicate rational coefficients.

julia> qsm_model = literature_model(arxiv_id = "1903.00009", model_parameters = Dict("k" => 2021))
Hypersurface model over a concrete base

julia> fgs = special_flux_family(qsm_model, check = false)
Family of G4 fluxes:
  - Elementary quantization checks: satisfied
  - Transversality checks: satisfied
  - Non-abelian gauge group: breaking pattern not analyzed

julia> d3_tadpole_constraint(fgs);

Experimental

This function is part of the experimental code in Oscar. Please read here for more details.

Properties of a Flux Family

To test whether a family satisfies standard physical consistency conditions, use:

is_well_quantized — Method

is_well_quantized(fgs::FamilyOfG4Fluxes; check::Bool = true)

Executes elementary tests (they are necessary but not sufficient) to tell if Family of $G_4$-fluxes is well-quantized. In case any of these tests fails, we know that this family of $G_4$-fluxes is definitely not well-quantized. This method then returns false.

In the opposite case that all elementary tests pass, this method returns true. Note however that this does not imply that the family is well-quantized. At least, according to the current theoretical understanding it does not. Rather, this means that to the best of our current means, this family of fluxes appears to be well-quantized.

julia> qsm_model = literature_model(arxiv_id = "1903.00009", model_parameters = Dict("k" => 2021))
Hypersurface model over a concrete base

julia> gf = special_flux_family(qsm_model, check = false)
Family of G4 fluxes:
  - Elementary quantization checks: satisfied
  - Transversality checks: satisfied
  - Non-abelian gauge group: breaking pattern not analyzed

julia> is_well_quantized(gf)
true

julia> m1 = matrix_integral(gf);

julia> m2 = matrix_rational(gf);

julia> gf2 = family_of_g4_fluxes(qsm_model, m1, m2, check = false)
Family of G4 fluxes:
  - Elementary quantization checks: not executed
  - Transversality checks: not executed
  - Non-abelian gauge group: breaking pattern not analyzed

julia> is_well_quantized(gf2, check = false)
true

Experimental

This function is part of the experimental code in Oscar. Please read here for more details.

passes_transversality_checks — Method

passes_transversality_checks(fgs::FamilyOfG4Fluxes; check::Bool = true)

Check if the given family of $G_4$-fluxes passes the transversality checks. If so, this method returns true and otherwise false.

julia> qsm_model = literature_model(arxiv_id = "1903.00009", model_parameters = Dict("k" => 2021))
Hypersurface model over a concrete base

julia> gf = special_flux_family(qsm_model, check = false)
Family of G4 fluxes:
  - Elementary quantization checks: satisfied
  - Transversality checks: satisfied
  - Non-abelian gauge group: breaking pattern not analyzed

julia> passes_transversality_checks(gf, check = false)
true

julia> m1 = matrix_integral(gf);

julia> m2 = matrix_rational(gf);

julia> gf3 = family_of_g4_fluxes(qsm_model, m1, m2, check = false)
Family of G4 fluxes:
  - Elementary quantization checks: not executed
  - Transversality checks: not executed
  - Non-abelian gauge group: breaking pattern not analyzed

julia> passes_transversality_checks(gf3)
true

Experimental

This function is part of the experimental code in Oscar. Please read here for more details.

breaks_non_abelian_gauge_group — Method

breaks_non_abelian_gauge_group(fgs::FamilyOfG4Fluxes; check::Bool = true)

Check if Family of G4-fluxes breaks the non-abelian gauge group. If so, this method returns true and otherwise false.

julia> qsm_model = literature_model(arxiv_id = "1903.00009", model_parameters = Dict("k" => 2021))
Hypersurface model over a concrete base

julia> gf = special_flux_family(qsm_model, check = false)
Family of G4 fluxes:
  - Elementary quantization checks: satisfied
  - Transversality checks: satisfied
  - Non-abelian gauge group: breaking pattern not analyzed

julia> breaks_non_abelian_gauge_group(gf)
true

julia> gf3 = special_flux_family(qsm_model, not_breaking = true, check = false)
Family of G4 fluxes:
  - Elementary quantization checks: satisfied
  - Transversality checks: satisfied
  - Non-abelian gauge group: unbroken

julia> breaks_non_abelian_gauge_group(gf3)
false

julia> m1 = matrix_integral(gf3);

julia> m2 = matrix_rational(gf3);

julia> gf4 = family_of_g4_fluxes(qsm_model, m1, m2, check = false)
Family of G4 fluxes:
  - Elementary quantization checks: not executed
  - Transversality checks: not executed
  - Non-abelian gauge group: breaking pattern not analyzed

julia> breaks_non_abelian_gauge_group(gf4, check = false)
false

Experimental

This function is part of the experimental code in Oscar. Please read here for more details.

Constructing Flux Instances from a Flux Family

You can generate specific fluxes within a family using:

flux_instance — Method

flux_instance(fgs::FamilyOfG4Fluxes, int_combination::ZZMatrix, rat_combination::QQMatrix; check::Bool = true)

Create an element of a family of G4-fluxes.

Examples

julia> qsm_model = literature_model(arxiv_id = "1903.00009", model_parameters = Dict("k" => 2021))
Hypersurface model over a concrete base

julia> mat_int = zero_matrix(QQ, 37, 1);

julia> mat_int[1,1] = 1;

julia> mat_rat = zero_matrix(QQ, 37, 1);

julia> mat_rat[2,1] = 1;

julia> shift = [zero(QQ) for k in 1:37];

julia> fgs = family_of_g4_fluxes(qsm_model, mat_int, mat_rat, shift, check = false)
Family of G4 fluxes:
  - Elementary quantization checks: not executed
  - Transversality checks: not executed
  - Non-abelian gauge group: breaking pattern not analyzed

julia> int_combination = matrix(ZZ, [[3]])
[3]

julia> rat_combination = matrix(QQ, [[5//2]])
[5//2]

julia> flux_instance(fgs, int_combination, rat_combination, check = false)
G4-flux candidate
  - Elementary quantization checks: not executed
  - Transversality checks: not executed
  - Non-abelian gauge group: breaking pattern not analyzed
  - Tadpole cancellation check: not computed

julia> flux_instance(fgs, Int[], [], check = false)
G4-flux candidate
  - Elementary quantization checks: not executed
  - Transversality checks: not executed
  - Non-abelian gauge group: breaking pattern not analyzed
  - Tadpole cancellation check: not computed

julia> flux_instance(fgs, [3], [], check = false)
G4-flux candidate
  - Elementary quantization checks: not executed
  - Transversality checks: not executed
  - Non-abelian gauge group: breaking pattern not analyzed
  - Tadpole cancellation check: not computed

julia> flux_instance(fgs, [], [5//2], check = false)
G4-flux candidate
  - Elementary quantization checks: not executed
  - Transversality checks: not executed
  - Non-abelian gauge group: breaking pattern not analyzed
  - Tadpole cancellation check: not computed

julia> flux_instance(fgs, [3], [5//2], check = false)
G4-flux candidate
  - Elementary quantization checks: not executed
  - Transversality checks: not executed
  - Non-abelian gauge group: breaking pattern not analyzed
  - Tadpole cancellation check: not computed

Experimental

This function is part of the experimental code in Oscar. Please read here for more details.

random_flux_instance — Method

random_flux_instance(fgs::FamilyOfG4Fluxes; check::Bool = true)

Create a random element of a family of G4-fluxes.

Examples

julia> qsm_model = literature_model(arxiv_id = "1903.00009", model_parameters = Dict("k" => 2021))
Hypersurface model over a concrete base

julia> mat_int = zero_matrix(QQ, 37, 1);

julia> mat_int[1,1] = 1;

julia> mat_rat = zero_matrix(QQ, 37, 1);

julia> mat_rat[2,1] = 1;

julia> shift = [zero(QQ) for k in 1:37];

julia> fgs = family_of_g4_fluxes(qsm_model, mat_int, mat_rat, shift, check = false)
Family of G4 fluxes:
  - Elementary quantization checks: not executed
  - Transversality checks: not executed
  - Non-abelian gauge group: breaking pattern not analyzed

julia> random_flux_instance(fgs, check = false)
G4-flux candidate
  - Elementary quantization checks: not executed
  - Transversality checks: not executed
  - Non-abelian gauge group: breaking pattern not analyzed
  - Tadpole cancellation check: not computed

Experimental

This function is part of the experimental code in Oscar. Please read here for more details.

For convenience, a random physically consistent flux can also be generated directly from a model:

random_flux — Method

random_flux(m::AbstractFTheoryModel; not_breaking::Bool = false, check::Bool = true)

Create a random $G_4$-flux on a given F-theory model.

Examples

julia> qsm_model = literature_model(arxiv_id = "1903.00009", model_parameters = Dict("k" => 2021))
Hypersurface model over a concrete base

julia> rf = random_flux(qsm_model, check = false)
G4-flux candidate
  - Elementary quantization checks: satisfied
  - Transversality checks: satisfied
  - Non-abelian gauge group: breaking pattern not analyzed
  - Tadpole cancellation check: not computed

Experimental

This function is part of the experimental code in Oscar. Please read here for more details.

Recovering a Flux Family from a Flux Instance

Starting from a single G_4-flux, you can find the full family of fluxes that share its physical properties:

g4_flux_family — Method

g4_flux_family(gf::G4Flux; check::Bool = true)

Return the family of $G_4$-fluxes that possesses, such that all fluxes in this family share the following properties with the given $G_4$-flux: Transversality and breaking of the non-Abelian gauge group.

julia> qsm_model = literature_model(arxiv_id = "1903.00009", model_parameters = Dict("k" => 2021))
Hypersurface model over a concrete base

julia> g4 = qsm_flux(qsm_model)
G4-flux candidate
  - Elementary quantization checks: satisfied
  - Transversality checks: satisfied
  - Non-abelian gauge group: unbroken
  - Tadpole cancellation check: not computed

julia> g4_flux_family(g4, check = false)
Family of G4 fluxes:
  - Elementary quantization checks: satisfied
  - Transversality checks: satisfied
  - Non-abelian gauge group: unbroken

Experimental

This function is part of the experimental code in Oscar. Please read here for more details.

To determine the location of the original flux within that family, retrieve its defining coefficients and offset:

integral_coefficients — Method

integral_coefficients(gf::G4Flux)

Return the integral coefficients of a given $G_4$-flux. If these coefficients are not known, an error is raised.

julia> qsm_model = literature_model(arxiv_id = "1903.00009", model_parameters = Dict("k" => 4))
Hypersurface model over a concrete base

julia> gfs = special_flux_family(qsm_model, check = false, algorithm = "special")
Family of G4 fluxes:
  - Elementary quantization checks: satisfied
  - Transversality checks: satisfied
  - Non-abelian gauge group: breaking pattern not analyzed

julia> g4 = random_flux_instance(gfs, check = false)
G4-flux candidate
  - Elementary quantization checks: satisfied
  - Transversality checks: satisfied
  - Non-abelian gauge group: breaking pattern not analyzed
  - Tadpole cancellation check: not computed

julia> integral_coefficients(g4);

Experimental

This function is part of the experimental code in Oscar. Please read here for more details.

rational_coefficients — Method

rational_coefficients(gf::G4Flux)

Return the integral coefficients of a given $G_4$-flux. If these coefficients are not known, an error is raised.

julia> qsm_model = literature_model(arxiv_id = "1903.00009", model_parameters = Dict("k" => 4))
Hypersurface model over a concrete base

julia> gfs = special_flux_family(qsm_model, check = false, algorithm = "special")
Family of G4 fluxes:
  - Elementary quantization checks: satisfied
  - Transversality checks: satisfied
  - Non-abelian gauge group: breaking pattern not analyzed

julia> g4 = random_flux_instance(gfs, check = false)
G4-flux candidate
  - Elementary quantization checks: satisfied
  - Transversality checks: satisfied
  - Non-abelian gauge group: breaking pattern not analyzed
  - Tadpole cancellation check: not computed

julia> rational_coefficients(g4);

Experimental

This function is part of the experimental code in Oscar. Please read here for more details.

offset — Method

offset(gf::G4Flux)

Return the offset of a given $G_4$-flux. If this shift is not known, an error is raised.

julia> qsm_model = literature_model(arxiv_id = "1903.00009", model_parameters = Dict("k" => 2021))
Hypersurface model over a concrete base

julia> gfs = special_flux_family(qsm_model, check = false)
Family of G4 fluxes:
  - Elementary quantization checks: satisfied
  - Transversality checks: satisfied
  - Non-abelian gauge group: breaking pattern not analyzed

julia> g4 = random_flux_instance(gfs, check = false)
G4-flux candidate
  - Elementary quantization checks: satisfied
  - Transversality checks: satisfied
  - Non-abelian gauge group: breaking pattern not analyzed
  - Tadpole cancellation check: not computed

julia> offset(g4);

Experimental

This function is part of the experimental code in Oscar. Please read here for more details.

Advanced Methods (typically run in the background)

These methods support internal computations, such as the construction of cohomology bases and dictionary converters. Most users will not need to call these directly.

basis_of_h22_ambient — Method

basis_of_h22_ambient(m::AbstractFTheoryModel; check::Bool = true)

By virtue of Theorem 12.4.1 in [CLS11], one can compute a monomial basis of $H^4(X, \mathbb{Q})$ for a simplicial, complete toric variety $X$ by truncating its cohomology ring to degree $2$. Inspired by this, this method identifies a basis of $H^{(2,2)}(X, \mathbb{Q})$ by multiplying pairs of cohomology classes associated with toric coordinates.

By definition, $H^{(2,2)}(X, \mathbb{Q})$ is a subset of $H^{4}(X, \mathbb{Q})$. However, by Theorem 9.3.2 in [CLS11], for complete and simplicial toric varieties and $p \neq q$ it holds $H^{(p,q)}(X, \mathbb{Q}) = 0$. It follows that for such varieties $H^{(2,2)}(X, \mathbb{Q}) = H^4(X, \mathbb{Q})$ and the vector space dimension of those spaces agrees with the Betti number $b_4(X)$.

This method computes a basis of $H^{(2,2)}(X, \mathbb{Q})$ as vector of cohomology classes.

Note that it can be computationally very demanding to check if a toric variety $X$ is complete (and simplicial). The optional argument check can be set to false to skip these tests.

Examples

julia> qsm_model = literature_model(arxiv_id = "1903.00009", model_parameters = Dict("k" => 283))
Hypersurface model over a concrete base

julia> length(basis_of_h22_ambient(qsm_model, check = false)) == betti_number(ambient_space(qsm_model), 4)
true

Experimental

This function is part of the experimental code in Oscar. Please read here for more details.

gens_of_h22_hypersurface — Method

gens_of_h22_hypersurface(m::AbstractFTheoryModel; check::Bool = true)

By virtue of Theorem 12.4.1 in [CLS11], one can compute a monomial basis of $H^4(X, \mathbb{Q})$ for a simplicial, complete toric variety $X$ by truncating its cohomology ring to degree $2$. Inspired by this, this method identifies a basis of $H^{(2,2)}(X, \mathbb{Q})$ by multiplying pairs of cohomology classes associated with toric coordinates.

By definition, $H^{(2,2)}(X, \mathbb{Q})$ is a subset of $H^{4}(X, \mathbb{Q})$. However, by Theorem 9.3.2 in [CLS11], for complete and simplicial toric varieties and $p \neq q$ it holds $H^{(p,q)}(X, \mathbb{Q}) = 0$. It follows that for such varieties $H^{(2,2)}(X, \mathbb{Q}) = H^4(X, \mathbb{Q})$ and the vector space dimension of those spaces agrees with the Betti number $b_4(X)$.

Once restricted to the hypersurface of the model, these elements are generators of a subset $S \subseteq H^(2,2)(Y, \mathbb{Q})$, where $Y$ is the hypersurface of interest in $X$. This method computes this generating set in terms of ambient space cohomology classes.

Note that it can be computationally very demanding to check if a toric variety $X$ is complete (and simplicial). The optional argument check can be set to false to skip these tests.

Examples

julia> qsm_model = literature_model(arxiv_id = "1903.00009", model_parameters = Dict("k" => 283))
Hypersurface model over a concrete base

julia> length(gens_of_h22_hypersurface(qsm_model, check = false))
25

Experimental

This function is part of the experimental code in Oscar. Please read here for more details.

basis_of_h22_ambient_indices — Method

basis_of_h22_ambient_indices(m::AbstractFTheoryModel; check::Bool = true)

By virtue of Theorem 12.4.1 in [CLS11], one can compute a monomial basis of $H^4(X, \mathbb{Q})$ for a simplicial, complete toric variety $X$ by truncating its cohomology ring to degree $2$. Inspired by this, this method identifies a basis of $H^{(2,2)}(X, \mathbb{Q})$ by multiplying pairs of cohomology classes associated with toric coordinates.

By definition, $H^{(2,2)}(X, \mathbb{Q})$ is a subset of $H^{4}(X, \mathbb{Q})$. However, by Theorem 9.3.2 in [CLS11], for complete and simplicial toric varieties and $p \neq q$ it holds $H^{(p,q)}(X, \mathbb{Q}) = 0$. It follows that for such varieties $H^{(2,2)}(X, \mathbb{Q}) = H^4(X, \mathbb{Q})$ and the vector space dimension of those spaces agrees with the Betti number $b_4(X)$.

This method computes a vector of tuples of two integers. The tuple $(a,b)$ states that the product of the a-th and b-th variables in the Cox ring is an element of the basis of $H^(2,2)(X, \mathbb{Q})$, that we have chosen.

Note that it can be computationally very demanding to check if a toric variety $X$ is complete (and simplicial). The optional argument check can be set to false to skip these tests.

Examples

julia> qsm_model = literature_model(arxiv_id = "1903.00009", model_parameters = Dict("k" => 283))
Hypersurface model over a concrete base

julia> bhi = basis_of_h22_ambient_indices(qsm_model, check = false);

julia> length(bhi) == betti_number(ambient_space(qsm_model), 4)
true

Experimental

This function is part of the experimental code in Oscar. Please read here for more details.

gens_of_h22_hypersurface_indices — Method

gens_of_h22_hypersurface_indices(m::AbstractFTheoryModel; check::Bool = true)

By virtue of Theorem 12.4.1 in [CLS11], one can compute a monomial basis of $H^4(X, \mathbb{Q})$ for a simplicial, complete toric variety $X$ by truncating its cohomology ring to degree $2$. Inspired by this, this method identifies a basis of $H^{(2,2)}(X, \mathbb{Q})$ by multiplying pairs of cohomology classes associated with toric coordinates.

By definition, $H^{(2,2)}(X, \mathbb{Q})$ is a subset of $H^{4}(X, \mathbb{Q})$. However, by Theorem 9.3.2 in [CLS11], for complete and simplicial toric varieties and $p \neq q$ it holds $H^{(p,q)}(X, \mathbb{Q}) = 0$. It follows that for such varieties $H^{(2,2)}(X, \mathbb{Q}) = H^4(X, \mathbb{Q})$ and the vector space dimension of those spaces agrees with the Betti number $b_4(X)$.

Once restricted to the hypersurface of the model, these elements are generators of a subset $S \subseteq H^(2,2)(Y, \mathbb{Q})$, where $Y$ is the hypersurface of interest in $X$. This method computes a vector of tuples, each tuple encodes on generator of $S \subseteq H^(2,2)(Y, \mathbb{Q})$. Specifically, the tuple $(a,b)$ indicates that the product of the $a$-th and the $b$-th variable in the Cox ring encodes an element of the cohomology ring of $X$, whose restriction to $Y$ is one of our generators.

Note that it can be computationally very demanding to check if a toric variety $X$ is complete (and simplicial). The optional argument check can be set to false to skip these tests.

Examples

julia> qsm_model = literature_model(arxiv_id = "1903.00009", model_parameters = Dict("k" => 283))
Hypersurface model over a concrete base

julia> length(gens_of_h22_hypersurface_indices(qsm_model, check = false))
25

Experimental

This function is part of the experimental code in Oscar. Please read here for more details.

converter_dict_h22_ambient — Method

converter_dict_h22_ambient(m::AbstractFTheoryModel; check::Bool = true)

By virtue of Theorem 12.4.1 in [CLS11], one can compute a monomial basis of $H^4(X, \mathbb{Q})$ for a simplicial, complete toric variety $X$ by truncating its cohomology ring to degree $2$. Inspired by this, this method identifies a basis of $H^{(2,2)}(X, \mathbb{Q})$ by multiplying pairs of cohomology classes associated with toric coordinates.

By definition, $H^{(2,2)}(X, \mathbb{Q})$ is a subset of $H^{4}(X, \mathbb{Q})$. However, by Theorem 9.3.2 in [CLS11], for complete and simplicial toric varieties and $p \neq q$ it holds $H^{(p,q)}(X, \mathbb{Q}) = 0$. It follows that for such varieties $H^{(2,2)}(X, \mathbb{Q}) = H^4(X, \mathbb{Q})$ and the vector space dimension of those spaces agrees with the Betti number $b_4(X)$.

This method computes a dictionary which allows to map any element of $H^(2,2)(X, \mathbb{Q})$ into the basis found/chosen for $H^(2,2)(X, \mathbb{Q})$.

Note that it can be computationally very demanding to check if a toric variety $X$ is complete (and simplicial). The optional argument check can be set to false to skip these tests.

Examples

julia> qsm_model = literature_model(arxiv_id = "1903.00009", model_parameters = Dict("k" => 283))
Hypersurface model over a concrete base

julia> cdh22 = converter_dict_h22_ambient(qsm_model, check = false);

julia> length(collect(keys(cdh22)))
81

Experimental

This function is part of the experimental code in Oscar. Please read here for more details.

converter_dict_h22_hypersurface — Method

converter_dict_h22_hypersurface(m::AbstractFTheoryModel; check::Bool = true)

By virtue of Theorem 12.4.1 in [CLS11], one can compute a monomial basis of $H^4(X, \mathbb{Q})$ for a simplicial, complete toric variety $X$ by truncating its cohomology ring to degree $2$. Inspired by this, this method identifies a basis of $H^{(2,2)}(X, \mathbb{Q})$ by multiplying pairs of cohomology classes associated with toric coordinates.

By definition, $H^{(2,2)}(X, \mathbb{Q})$ is a subset of $H^{4}(X, \mathbb{Q})$. However, by Theorem 9.3.2 in [CLS11], for complete and simplicial toric varieties and $p \neq q$ it holds $H^{(p,q)}(X, \mathbb{Q}) = 0$. It follows that for such varieties $H^{(2,2)}(X, \mathbb{Q}) = H^4(X, \mathbb{Q})$ and the vector space dimension of those spaces agrees with the Betti number $b_4(X)$. Once restricted to the hypersurface $Y$ in question, we obtain a generating set of a subset $S \subseteq H^{(2,2)}(Y, \mathbb{Q})$.

This method computes a dictionary which allows to map any element of $H^(2,2)(X, \mathbb{Q})$ into our chosen generating set of $S \subseteq H^{(2,2)}(Y, \mathbb{Q})$.

Note that it can be computationally very demanding to check if a toric variety $X$ is complete (and simplicial). The optional argument check can be set to false to skip these tests.

Examples

julia> qsm_model = literature_model(arxiv_id = "1903.00009", model_parameters = Dict("k" => 283))
Hypersurface model over a concrete base

julia> cdh22 = converter_dict_h22_hypersurface(qsm_model, check = false);

julia> length(collect(keys(cdh22)))
81

Experimental

This function is part of the experimental code in Oscar. Please read here for more details.