Miscellaneous functions

Utilities

is_orthogonally_stable — Function

is_orthogonally_stable(chi::GAPGroupClassFunction; check::Bool = true)

Return nothing if the indicator of some irreducible constituent of chi is not known; this can happen only if chi has characteristic 2.

Otherwise return true if chi is orthogonally stable, and false otherwise.

A character is called orthogonally stable if

chi is orthogonal, that is, chi is real, and all its absolutely irreducible constituents of indicator - have even multiplicity and

all its absolutely irreducible constituents of indicator + have even degree.

If we know that chi is orthogonal then we can set check to false; in this case, some nothing results can be avoided.

Examples

julia> t = character_table("A6");

julia> println(map(is_orthogonally_stable, t))
Bool[0, 0, 0, 1, 1, 0, 1]

julia> println(map(is_orthogonally_stable, mod(t, 3)))
Bool[0, 0, 0, 1, 0]

show_with_ODs — Function

show_with_ODs(tbl::Oscar.GAPGroupCharacterTable)

Show tbl with 2nd indicators, known ODs, and degrees of character fields. (See Base.show(io::IO, ::MIME"text/plain", tbl::Oscar.GAPGroupCharacterTable) for ways to modify what is shown.)

Examples

julia> t = character_table("A5");

julia> Oscar.OrthogonalDiscriminants.show_with_ODs(t)
A5

          2  2  2  .  .  .
          3  1  .  1  .  .
          5  1  .  .  1  1
                          
            1a 2a 3a 5a 5b
         2P 1a 1a 3a 5b 5a
         3P 1a 2a 1a 5b 5a
         5P 1a 2a 3a 1a 1a
    d OD  2               
X_1 1     +  1  1  1  1  1
X_2 2     +  3 -1  .  A A*
X_3 2     +  3 -1  . A*  A
X_4 1  5  +  4  .  1 -1 -1
X_5 1     +  5  1 -1  .  .

A = z_5^3 + z_5^2 + 1
A* = -z_5^3 - z_5^2

show_OD_info — Function

show_OD_info(tbl::Oscar.GAPGroupCharacterTable)
show_OD_info(name::String)

Show an overview of known information about the ordinary and modular orthogonal discriminants for tbl or for the character table with identifier name.

Examples

julia> show_OD_info("A5")
A5:  2^2*3*5
------------

i|chi|K|disc| 2| 3|       5
-+---+-+----+--+--+--------
4| 4a|Q|   5|4a|4a|(def. 1)
 |   | |    |O-|O-|        

Functions related to Specht modules

dimension_specht_module — Function

dimension_specht_module(mu::Partition{T}) where T <: IntegerUnion -> ZZRingElem

Return the dimension of the Specht module for mu.

Examples

julia> print([dimension_specht_module(p) for p in partitions(4)])
ZZRingElem[1, 3, 2, 3, 1]

gram_determinant_specht_module — Function

gram_determinant_specht_module(mu::Partition{T}) where T <: IntegerUnion

Return the determinant of the Gram matrix for the Specht module for mu, in factorized collected form.

Examples

julia> print(gram_determinant_specht_module(partition([4, 3, 2, 1])))
Vector{ZZRingElem}[[3, 1152], [5, 768], [7, 384]]