Basics

Elements of groups

Operations

Properties of groups

Attributes of groups

Basics

Elements of groups

Given a group G, it is always possible to have access to some particular elements.

GAPGroup — Type

GAPGroup <: AbstractAlgebra.Group

The idea is that each object of the abstract type GAPGroup stores a group object from the GAP system, and thus can delegate questions about this group to GAP.

Concrete subtypes of GAPGroup are PermGroup, FPGroup, PcGroup, and MatrixGroup.

BasicGAPGroupElem — Type

BasicGAPGroupElem{T<:GAPGroup}

The type BasicGAPGroupElem gathers all types of group elements described only by an underlying GAP object.

If $x$ is an element of the group G of type T, then the type of $x$ is BasicGAPGroupElem{T}.

elem_type — Method

elem_type(::Type{T}) where T <: GAPGroup
elem_type(::T) where T <: GAPGroup

elem_type maps (the type of) a group to the type of its elements. For now, a group of type T has elements of type BasicGAPGroupElem{T}. So we provide it mostly for consistency with other parts of OSCAR. In the future, a more elaborate setup for group element types might also be needed.

one — Method

one(G::GAPGroup) -> elem_type(G)

Return the identity of the group G.

one — Method

one(x::GAPGroupElem{T}) -> GAPGroupElem{T}

Return the identity of the parent group of x.

isfinite_order — Method

isfinite_order(g::GAPGroupElem) -> Bool

Return true if g has finite order, and false otherwise.

Examples

julia> isfinite_order(gen(symmetric_group(5), 1))
true

julia> isfinite_order(gen(free_group(2), 1))
false

gens — Method

gens(G::Group)

Return a vector of generators of the group G. To get the i-th generator, use G[i] or gen(G,i) (see gen) instead of gens(G)[i], as that is more efficient.

Examples

julia> g = symmetric_group(5);  gens(g)
2-element Vector{PermGroupElem}:
 (1,2,3,4,5)
 (1,2)

julia> g[2]
(1,2)

Note

The output of gens(G) is not, in general, the minimal list of generators for G.

ngens — Method

ngens(G::GAPGroup) -> Int

Return the length of the vector gens(G).

WARNING:

this is NOT, in general, the minimum number of generators for G.

gen — Method

gen(G::GAPGroup, i::Integer)

Return the i-th element of the vector gens(G). This is equivalent to G[i], and returns gens(G)[i] but may be more efficient than the latter.

An exception is thrown if i is larger than the length of gens(G).

rand — Method

rand(rng::Random.AbstractRNG = Random.GLOBAL_RNG, G::Group)

Return a random element of G, using the random number generator rng.

rand_pseudo — Method

rand_pseudo(G::Group)

Return a pseudo random element of G. This works faster than rand, but the returned elements are not necessarily uniformly distributed.

It is also possible to obtain the generators of G by typing

f1,f2,f3 = gens(G)

This is equivalent to

f1=G[1]; f2=G[2]; f3=G[3];

For a group G that has been created as a subgroup of another group, generated by a list L of elements, gens(G) is equal to L.

Operations

Oscar supports the following operations and functions on group elements.

*, multiplication between two elements in a group.

inv(x) and x^-1, the inverse of x.

x/y, the element x y^-1.

x^n, the n-th power of x; if n == 0, the identity of the group is returned; if n < 0, the -n-th power of the inverse of x is returned.

isone(x) returns whether x is the identity of the group.

conj(x,y) and x^y, the conjugate of x by y, i. e., the element y^-1 x y.

comm(x,y), the commutator of x and y, i. e., the element x^-1 y^-1 x y.

Note

In Oscar, the expression x^y^z is equivalent to x^(y^z). In other words, conjugations are evaluated from the right to the left.

comm — Method

comm(x::GAPGroupElem, y::GAPGroupElem)

Return the commutator of x and y, which is defined as x^-1*y^-1*x*y, and usually denoted as [x,y] in the literature.

nilpotency_class — Method

nilpotency_class(G::GAPGroup) -> Int

Return the nilpotency class of G, i.e., the smallest integer $n$ such that G has a central series of length $n$.

An exception is thrown if G is not nilpotent.

Properties of groups

isfinite — Method

isfinite(G::GAPGroup) -> Bool

Return true if G is finite, and false otherwise.

Examples

julia> isfinite(symmetric_group(5))
true

julia> isfinite(free_group(2))
false

isabelian — Method

isabelian(G::Group)

Return true if G is abelian (commutative), that is, $x*y = y*x$ holds for all elements $x, y$ in G.

ispgroup — Function

ispgroup(G)

Return (true, nothing) if G is the trivial group, (true, p) if the order of every element in G is a power of a prime p, and (false, nothing) otherwise.

For finite groups G, the first return value is true if and only if the order of G is a prime power.

isnilpotent — Function

isnilpotent(G::GAPGroup)

Return whether G is nilpotent, i. e., whether the lower central series of G reaches the trivial subgroup in a finite number of steps.

issupersolvable — Function

issupersolvable(G::GAPGroup)

Return whether G is supersolvable, i. e., G is finite and has a normal series with cyclic factors.

issolvable — Function

issolvable(G::GAPGroup)

Return whether G is solvable, i. e., whether derived_series(G) reaches the trivial subgroup in a finite number of steps.

isperfect — Function

isperfect(G)

Return whether G is a perfect group, i.e., equal to its derived subgroup.

issimple — Function

issimple(G)

Return whether G is a simple group, i.e., G is not trivial and has no non-trivial normal subgroups.

isalmostsimple — Function

isalmostsimple(G)

Return whether G is an almost simple group, i. e., G is isomorphic to a group $H$ with the property $S \leq H \leq Aut(S)$, for some non-abelian simple group $S$.

Attributes of groups

order — Method

order(::Type{T} = fmpz, x::Union{GAPGroupElem, GAPGroup}) where T <: Union{Integer, fmpz}

Return the order of x, as an instance of T.

For a group element x in the group G, the order of x is the smallest positive integer n such that x^n is the identity of G. For a group x, the order of x is the number of elements in x.

An exception is thrown if the order of x is infinite, use isfinite in order to check for finiteness.

exponent — Method

exponent(::Type{T} = fmpz, G::GAPGroup) where T <: Union{Integer, fmpz}

Return the exponent of G, as an instance of T, i. e., the smallest positive integer $e$ such that $g^e$ is the identity of G for every $g$ in G.