Groups

Basics

Subgroups

Quotients

Permutation groups

Polycyclic groups

Homomorphisms

Automorphisms groups

Groups

Oscar supports the following types of groups:

PermGroup = groups of permutations

MatrixGroup = groups of matrices

FPGroup = finitely presented groups

PcGroup = polycyclic groups

DirectProductOfGroups = direct product of two groups

AutomorphismGroup = group of automorphisms over a group

If x is an element of the group G of type T, then the type of x is GAPGroupElement{T}.

Basics

Elements of groups

Given a group G, it is always possible to have access to some particular elements.

Base.one — Method

one(G::Group) -> x::GAPGroupElem{typeof(G)}

Return the identity of the group G.

Base.rand — Method

rand(G::Group)

Return a random element of the group G.

Oscar.rand_pseudo — Method

rand_pseudo(G::Group)

Return a random element of the group G. It works faster than rand, but the elements are not necessarily equally distributed.

AbstractAlgebra.gens — Method

gens(G::Group)

Return an array of generators of the group G. To access the array, it can be used the shorter notation G[i] instead of gens(G)[i].

AbstractAlgebra.gens — Method

gens(G::Group, i::Integer)

Return the i-th element of the array gens(G). It is equivalent to G[i] and gens(G)[i]. If i is greater than the length of gens(G), an ERROR is returned.

It is also possible to obtain the generators of G by typing

f1,f2,f3 = gens(G)

This is equivalent to

f1=G[1]; f2=G[2]; f3=G[3];

If the group G has been created as the subgroup of another group generated by a list of elements L, then the generating set returned by the function gens corresponds to L.

Note

The output of gens(G) is not, in general, the minimal list of generators for G.

Operations

Oscar supports the following operations and functions on group elements.

* = multiplication between two elements in a group.

inv(x) = $x^{-1}$ the inverse of x.

x/y = the element $xy^{-1}$.

x^n = the $n$-th power of x. If $n = 0$, the identity of the group is returned; if $n < 0$, the $-n$-th power of the inverse of x is returned.

isone(x): returns whether x is the identity of the group.

conj(x,y) = x^y = the conjugate of x by y, i.e. the element $y^{-1}xy$.

comm(x,y) = the commutator of x and y, i.e. the element $x^{-1}y^{-1}xy$.

Note

In Oscar, the expression x^y^z is equivalent to x^(y^z). In other words, conjugations are evaluated from the right to the left.

Subgroups

The subgroup of a group G generated by the elements x,y,... is defined by the following instruction:

sub(G, [x,y,...]) ;

sub(x,y,...).

This function returns two objects: a group H, that is the subgroup of G generated by the elements x,y,..., and the embedding homomorphism of H into G. The object H has the same type of G, and it has no memory of the "parent" group G: it is an independent group.

Examples:

julia> G = symmetric_group(4); H = sub(G,[cperm([1,2,3]),cperm([2,3,4])]);
julia> H[1] == alternating_group(4)
true

The following functions are available in Oscar for subgroup properties:

Hecke.issubgroup — Function

issubgroup(G::T, H::T) where T <: GAPGroup

Return (true,f) if H is a subgroup of G, where f is the embedding homomorphism of H into G, otherwise return (false,Nothing).

Oscar.embedding — Method

embedding(G::T, H::T) where T <: GAPGroup

Return the embedding morphism of H into G. It throws ERROR if H is not a subgroup of G.

Hecke.index — Method

index(G::T, H::T) where T <: GAPGroup

Return the index of H in G.

Hecke.isnormal — Method

isnormal(G::T, H::T) where T <: GAPGroup

Return whether the subgroup H is normal in G.

Hecke.ischaracteristic — Method

ischaracteristic(G::T, H::T) where T <: GAPGroup

Return whether the subgroup H is characteristic in G.

Standard subgroups

The following functions are available in Oscar to obtain standard subgroups of a group G. Every such function returns a tuple (H,f), where H is a group of the same type of G and f is the embedding homomorphism of H into G.

Oscar.trivial_subgroup — Function

trivial_subgroup(G::GAPGroup)

Return the trivial subgroup of G.

Oscar.centre — Function

centre(G::Group)

Return the centre of G, i.e. the subgroup of all x in G such that xy=yx for every y in G, together with its embedding morphism into G.

Oscar.sylow_subgroup — Method

sylow_subgroup(G::Group, p::Int64)

Return a Sylow p-subgroup of G.

Oscar.hall_subgroup — Method

hall_subgroup(G::Group, P::Array{Int64})

Return a Hall P-subgroup of G. It works only if G is solvable.

Oscar.derived_subgroup — Function

derived_subgroup(G::GAPGroup)

Return the derived subgroup of G, i.e. the subgroup generated by all commutators of G.

Oscar.fitting_subgroup — Method

fitting_subgroup(G::GAPGroup)

Return the Fitting subgroup of G, the largest nilpotent normal subgroup of G.

Oscar.frattini_subgroup — Method

frattini_subgroup(G::GAPGroup)

Return the Frattini subgroup of G, the intersection of all maximal subgroups of G.

Oscar.radical_subgroup — Method

radical_subgroup(G::GAPGroup)

Return the radical subgroup of G, the largest solvable normal subgroup of G.

Oscar.socle — Method

socle(G::GAPGroup)

Return the socle of G, the subgroup generated by all minimal normal subgroups of G.

The following functions return a list of subgroups.

Hecke.subgroups — Method

subgroups(G::Group)

Return the list of subgroups of G.

Oscar.normal_subgroups — Function

normal_subgroups(G::Group)

Return the list of normal subgroups of G.

Oscar.maximal_subgroups — Function

maximal_subgroups(G::Group)

Return the list of maximal subgroups of G.

Oscar.maximal_normal_subgroups — Function

maximal_normal_subgroups(G::Group)

Return the list of maximal normal subgroups of G.

Oscar.minimal_normal_subgroups — Function

minimal_normal_subgroups(G::Group)

Return the list of minimal normal subgroups of G.

Oscar.characteristic_subgroups — Function

characteristic_subgroups(G::Group)

Return the list of characteristic subgroups of G, i.e. the subgroups that are invariant under all automorphisms of G.

Hecke.derived_series — Function

derived_series(G::GAPGroup)

Return the list [G_1, G_2, G_3, ... ], where G_1=G and G_{i+1} = derived_subgroup(G_i).

Oscar.sylow_system — Function

sylow_system(G::Group)

Return an array of Sylow $p$-subgroups of G, where $p$ runs over the prime factors of |G|, such that every two such subgroups commute each other (as subgroups). It works only if G is solvable.

Oscar.hall_system — Function

hall_system(G::Group)

Return an array of $P$-Hall subgroups of G, where $P$ runs over the subsets of prime factors of |G|. It works only if G is solvable.

Note

When a function returns a list of subgroups, the output consists in the subgroups only; the embeddings are not returned as well. To get the embedding homomorphism of the subgroup H in G, one has to type embedding(G,H)

Quotients

Quotient groups in Oscar can be defined using the instruction quo in two ways.

Quotients by normal subgroups.

AbstractAlgebra.quo — Method

quo(G::T, H::T)

Return the quotient group G/H of type PcGroup (if the quotient group is solvable) or PermGroup (otherwise), together with the projection G -> G/H.

Quotients by elements.

AbstractAlgebra.quo — Method

quo(G::T, elements::Vector{S})

Return the quotient group G/H of type FPGroup (if T=FPGroup), PcGroup (if the quotient group is solvable) or PermGroup (otherwise), where H is the normal closure of elements in G.

This is the typical way to build finitely presented groups.

Example:

julia> F=free_group(2);
julia> (f1,f2)=gens(F);
julia> G=quo(F,[f1^2,f2^3,(f1*f2)^2])[1];
julia> isisomorphic(G,symmetric_group(3))[1]
true

Similarly to the subgroups, the output consists of a pair (Q,p), where Q is the quotient group and p is the projection homomorphism of G into Q.

Permutation groups

Permutation groups can be defined as symmetric groups, alternating groups or their subgroups.

Oscar.symmetric_group — Function

symmetric_group(n::Int64)
symmetric_group(::Type{T}, n::Int)

Return the full symmetric group over a set of n elements. The group is returned of type T for T in {PermGroup, PcGroup}. If T is not specified, then T is set as PermGroup.

Oscar.alternating_group — Function

alternating_group(n::Int64)
alternating_group(::Type{T}, n::Int)

Return the full alternating group over a set of n elements. The group is returned of type T for T in {PermGroup, PcGroup}. If T is not specified, then T is set as PermGroup.

In Oscar, every permutation group has a degree n, that corresponds to the size of the set on which G acts. This can be displayed by typing

degree(G)

Note

The degree of a group of permutations is not necessarily equal to the largest moved point of the group G. For example, the trivial subgroup of symmetric_group(n) has degree n even though it fixes n.

Permutations

Permutations in Oscar are displayed as products of disjoint cycles, as in GAP. An explicit permutation can be built using the functions perm, gap_perm and cperm.

AbstractAlgebra.Generic.perm — Function

perm(G::PermGroup, L::AbstractVector{<:Integer})
(G::PermGroup)(L::AbstractVector{<:Integer})

Return the permutation x which maps every i from 1 to length(L) to L[i]. L must contain every integer from 1 to length(L) exactly, otherwise an exception is thrown. The parent of x is G. If x is not contained in G, an ERROR is returned. For gap_perm, the parent group of x is set as Sym(n), where n is the largest moved point of x. Example:

julia> perm(symmetric_group(6),[2,4,6,1,3,5])
(1,2,4)(3,6,5)

Oscar.gap_perm — Function

gap_perm(L::AbstractVector{<:Integer})

Return the permutation x which maps every i from 1 to length(L) to L[i]. L must contain every integer from 1 to length(L) exactly, otherwise an exception is thrown. The parent of x is set as Sym(n).

julia> gap_perm([2,4,6,1,3,5])
(1,2,4)(3,6,5)

Oscar.cperm — Function

cperm(L::AbstractVector{<:Integer}...)
cperm(G::PermGroup, L::AbstractVector{<:Integer}...)

For given lists of positive integers [a_1, a_2, ..., a_n],[b_1, b_2, ... , b_m], ... return the permutation x = (a_1,a_2,...,a_n)(b_1,b_2,...,b_m).... The array [n,n+1,...,n+k] can be replaced by n:n+k.

If a list is empty or contains duplicates, it fails. The parent of x is G. If x is not contained in G, an ERROR is returned. If G is not specified, then the parent of x is set as Sym(n), where n is the largest moved point of x. Example:

julia> cperm([1,2,3],4:7)
(1,2,3)(4,5,6,7)

julia> cperm([1,2],[2,3])
(1,3,2)

Example:

julia> perm(symmetric_group(6),[2,4,6,1,3,5])
(1,2,4)(3,6,5)

julia> cperm([1,2,3],4:7)
(1,2,3)(4,5,6,7)

julia> cperm([1,2],[2,3])
(1,3,2)

At the moment, the input vectors of the function cperm need not to be disjoint.

Warning

If the function perm is evaluated in a vector of integers without specifying the group G, then the returned value is an element of the AbstractAlgebra.jl type Perm{Int}. For this reason, if one wants a permutation of type GAPGroupElem{PermGroup} without specifying a parent, one has to use the function gap_perm.

Every permutation has always a permutation group as a parent. Two permutations coincide if, and only if, they move the same points and their parent groups have the same degree.

julia> G=symmetric_group(5);
julia> A=alternating_group(5);
julia> x=cperm(G,[1,2]);
julia> y=cperm(A,[1,2]);
julia> z=cperm([1,2]);
julia> x==y
true

julia> x==z
false

In the example above, x and y are equal because both act on a set of cardinality 5, while x and z are different because x belongs to Sym(5) and z belongs to Sym(2).

If G is a group and x is a permutation, it is possible to set G as parent of x simply typing G(x). This returns the permutation x as element of G (or ERROR if x does not embed into G).

julia> G=symmetric_group(5);

julia> x=cperm([1,2,3]);

julia> y=G(x);

julia> parent(x)
Sym( [ 1 .. 3 ] )

julia> parent(y)
Sym( [ 1 .. 5 ] )

The function listperm works in the opposite way with respect to perm:

Oscar.listperm — Function

listperm(x::PermGroupElem)

Return the list L defined by L = [ x(i) for i in 1:n ], where n is the degree of parent(x).

Permutations as functions

A permutation can be viewed as a function on the set {1,...,n}, hence it can be evaluated on integers.

Note

The multiplication between permutations works from the left to the right. So, if x and y are permutations and n is an integer, then (x*y)(n) = (y(x(n)), NOT x(y(n)).

julia> x = cperm([1,2,3,4,5]);
julia> x(2)
3

This works also if the argument is not in the range 1:n; in such a case, the output coincides with the input.

Polycyclic groups

Julia has the following functions that allow to generate polycyclic groups:

Hecke.abelian_group — Method

abelian_group(::Type{T}, v::Vector{Int}) where T <: Group -> PcGroup

Return the direct product of cyclic groups of order v[1] x v[2] x ... x v[n], as group of type T. Here, T must be of type PermGroup, FPGroup or PcGroup.

Oscar.cyclic_group — Function

cyclic_group(::Type{T}, n::Int)

Return the cyclic group of order n and type T. If the type is not specified, the group is returned of type PcGroup.

Oscar.dihedral_group — Function

dihedral_group(n::Int)
dihedral_group(::Type{T}, n::Int)

Return the dihedral group of order n of type T, where T is in {PcGroup,PermGroup,FPGroup}. In the first case, the type is set as PcGroup.

Oscar.quaternion_group — Function

quaternion_group(n::Int)
quaternion_group(::Type{T}, n::Int)

Return the quaternion group of order n of type T, where T is in {PcGroup,PermGroup,FPGroup,MatrixGroup}. In the first case, the type is set as PcGroup.

Warning

The type need to be specified in the input of the function abelian_group, otherwise a group of type GrpAbFinGen is returned, which is not a GAP group type. In future versions of Oscar, this may change.

The generators of a polycyclic group are displayed as f1, f2, f3, etc., and every element of a polycyclic group is displayed as product of such generators.

Example:

julia> G=abelian_group(PcGroup, [2,4]);
julia> G[1], G[2]
(f1, f2)
julia> G[2]*G[1]
f1*f2

Note that this does not define Julia variables named f1, f2, etc.! To get the generators of the group G, use gens(G); for convenience they can also be accessed as G[1], G[2], as shown in Section Elements of groups.

Homomorphisms

In Oscar, a group homomorphism from G to H is an object of parametric type GAPGroupHomomorphism{S,T}, where S and T are the types of G and H respectively.

A homomorphism from G to H can be defined in two ways.

Writing explicitly the images of the generators of G:

f = hom(G,H,[x1,x2,...],[y1,y2,...])

Here, [x1,x2,...] must be a generating set for G (not necessarily minimal) and [y1,y2,...] is a vector of elements of H of the same length of [x1,x2,...]. This assigns to f the value of the group homomorphism sending x_i into y_i. If such a homomorphism does not exist, an error is returned.

Taking an existing function g satisfying the group homomorphism properties:

f = hom(G,H,g)

If the function g does not satisfy the group homomorphism properties, an error is returned.

Example: The following procedures define the same homomorphism (conjugation by x) in the two ways explained above.

julia> S=symmetric_group(4);
julia> x=S[1];
julia> f=hom(S,S,gens(S),[S[1]^x,S[2]^x]);
julia> g=hom(S,S,y->y^x);
julia> f==g
true

Oscar has also the following standard homomorphism.

Hecke.id_hom — Function

id_hom(G::GAPGroup)

Return the identity homomorphism on the group G.

Oscar.trivial_morphism — Function

trivial_morphism(G::GAPGroup, H::GAPGroup)

Return the homomorphism from G to H sending every element of G into the identity of H. If H is not specified, it is taken equal to G.

To evaluate the homomorphism f in the element x of G, it is possible to use the instruction

image(f,x)

or the more compact notations f(x) and x^f.

Example:

julia> S=symmetric_group(4);
julia> f=hom(S,S,x->x^S[1]);
julia> x=cperm(S,[1,2]);
julia> image(f,x)
(2,3)
julia> f(x)
(2,3)
julia> x^f
(2,3)

A sort of "inverse" of the evaluation is the following

Hecke.haspreimage — Method

haspreimage(f::GAPGroupHomomorphism, x::GAPGroupElem)

Return (true,y) if there exists y such that f(y) = x; otherwise, return (false,1).

Example:

julia> S=symmetric_group(4);
julia> f=hom(S,S,x->x^S[1]);
julia> x=cperm(S,[1,2]);
julia> haspreimage(f,x)
(true,(1,4))

Warning

Do not confuse haspreimage with the function has_preimage, which works on variable of type GrpGenToGrpGenMor.

A further function that produces isomorphisms is isisomorphic:

AbstractAlgebra.Generic.isisomorphic — Method

isisomorphic(G::Group, H::Group)

Return (true,f) if G and H are isomorphic groups, where f is a group isomorphism. Otherwise, return (false,f), where f is the trivial homomorphism.

Operations on homomorphisms

Oscar supports the following operations on homomorphisms.

inv(f) = the inverse of f. An error is returned if f is not bijective.

f^n = the homomorphism f composed n times with itself. An error is returned if the domain and the codomain of f do not coincide (unless n=1). If n is negative, the result is the inverse of f composed n times with itself.

compose(g,f) = composition of g and f. This works only if the codomain of g coincide with the domain of f. Shorter equivalent expressions are g*f and f(g).

Example:

julia> S=symmetric_group(4);
julia> f=hom(S,S,x->x^S[1]);
julia> g=hom(S,S,x->x^S[2]);
julia> f*g==hom(S,S,x->x^(S[1]*S[2]))
true
julia> f==f^-3
true

Note

The composition operation * has to be read from the right to the left. So, (f*g)(x) is equivalent to g(f(x)).

Properties of homomorphisms

Oscar implements the following attributes of homomorphisms.

AbstractAlgebra.Generic.domain — Function

domain(f::GAPGroupHomomorphism)

Return the domain of f.

AbstractAlgebra.Generic.codomain — Function

codomain(f::GAPGroupHomomorphism)

Return the codomain of f.

Hecke.isinjective — Method

isinjective(f::GAPGroupHomomorphism)

Return whether f is injective.

Hecke.issurjective — Method

issurjective(f::GAPGroupHomomorphism)

Return whether f is surjective.

Hecke.isbijective — Method

isbijective(f::GAPGroupHomomorphism)

Return whether f is bijective.

Hecke.isinvertible — Method

isinvertible(f::GAPGroupHomomorphism)

Return whether f is invertible.

Oscar.isinvariant — Method

isinvariant(f::GAPGroupHomomorphism, H::Group)
isinvariant(f::GAPGroupElem{AutomorphismGroup{T}}, H::T)

Return whether f(H) == H.

Subgroups described by homomorphisms

The following functions compute subgroups or quotients of either the domain or the codomain. Analogously to the functions described in Sections Subgroups and Quotients, the output consists of a pair (H, g), where H is a subgroup (resp. quotient) and g is its embedding (resp. projection) homomorphism.

AbstractAlgebra.Generic.kernel — Method

kernel(f::GAPGroupHomomorphism)

Return the kernel of f, together with its embedding into domain(f).

AbstractAlgebra.Generic.image — Method

image(f::GAPGroupHomomorphism)

Return the image of f as subgroup of codomain(f), together with the embedding homomorphism.

AbstractAlgebra.Generic.image — Method

image(f::GAPGroupHomomorphism{S, T}, H::S) where S <: GAPGroup where T <: GAPGroup
(f::GAPGroupHomomorphism{S, T})(H::S)

Return f(H), together with the embedding homomorphism into codomain(f).

Hecke.cokernel — Method

cokernel(f::GAPGroupHomomorphism)

Return the cokernel of f, that is, the quotient of the codomain of f by the normal closure of the image.

AbstractAlgebra.Generic.preimage — Method

preimage(f::GAPGroupHomomorphism{S, T}, H::T) where S <: GAPGroup where T <: GAPGroup

If H is a subgroup of the codomain of f, return the subgroup f^-1(H), together with its embedding homomorphism into the domain of f.

Automorphisms groups

Groups of automorphisms over a group G have parametric type AutomorphismGroup{T}, where T is the type of G. The group of automorphisms over a group G is defined by the following instruction:

automorphism_group(G)

The evaluation of the automorphism f in the element x is analogous to the homomorphism evaluation: it can be obtained by typing either f(x) or x^f.

It is possible to turn an automorphism f into a homomorphism by typing hom(f). The viceversa is also possible: if g is a bijective homomorphism from the group G to itself and A is the automorphism group of G, then the instruction A(g) returns g as automorphism of G. This is the standard way to build explicitly an automorphism (another way, available for inner automorphisms, is shown in Section Inner_automorphisms).

Example:

julia> S=symmetric_group(4);
julia> A=automorphism_group(S);
julia> g=hom(S,S,x->x^S[1]);
julia> g in A
false
julia> au=A(g);
julia> au in A
true
julia> g=hom(au)
true
julia> x=cperm(S,[1,2,3]);
julia> au(x)
(2,3,4)
julia> g(x)==au(x)
true

In Oscar it is possible to multiply homomorphisms and automorphisms (whenever it makes sense); in such cases, the output is always a variable of type GAPGroupHomomorphism{S,T}.

julia> S=symmetric_group(4);

julia> A=automorphism_group(S);

julia> g=hom(S,S,x->x^S[1]);

julia> f=A(g);

julia> typeof(g*f)
Oscar.GAPGroupHomomorphism{PermGroup,PermGroup}

The following functions are available for automorphisms, some of them similar to the corresponding functions for homomorphisms of groups.

Oscar.isinvariant — Method

isinvariant(f::GAPGroupElem{AutomorphismGroup{T}}, H::T)

Return whether f(H) == H.

Oscar.restrict_automorphism — Method

restrict_automorphism(f::GAPGroupElem{AutomorphismGroup{T}}, H::T)

If H is invariant under f, returns the restriction of f to H as automorphism of H; otherwise it returns ERROR.

Oscar.induced_automorphism — Method

induced_automorphism(f::GAPGroupHomomorphism, g::GAPGroupHomomorphism)
induced_automorphism(f::GAPGroupHomomorphism, g::GAPGroupElem{AutomorphismGroup{T}})

Return the automorphism h of the image of f such that h(f) == f(g), where g is an automorphism of a group G and f is a group homomorphism defined over G such that the kernel of f is invariant under g

Inner automorphisms

Oscar disposes of the following functions to handle inner automorphisms of a group.

Oscar.inner_automorphism — Method

inner_automorphism(g::GAPGroupElem)

Return the inner automorphism in automorphism_group(parent(g)) defined by x -> x^g.

Oscar.isinner_automorphism — Method

isinner_automorphism(f::GAPGroupHomomorphism)
isinner_automorphism(f::GAPGroupElem{AutomorphismGroup{T}})

Return whether f is an inner automorphism.

Oscar.inner_automorphisms_group — Method

inner_automorphisms_group(A::AutomorphismGroup{T})

Return the subgroup of A of the inner automorphisms.