Permutation groups

Permutations

Permutations as functions

Permutation groups

Permutation groups can be defined as symmetric groups, alternating groups or their subgroups.

PermGroup — Type

PermGroup

Groups of permutations. Every group of this type is the subgroup of Sym(n) for some n.

Examples

symmetric_group(n::Int): the symmetric group Sym(n)

alternating_group(n::Int): the alternating group Alt(n)

subgroups of Sym(n)

dihedral_group(PermGroup, n::Int): the dihedral group of order n as a group of permutations. Same holds replacing dihedral_group by quaternion_group

PermGroupElem — Type

PermGroupElem

Element of a group of permutations. It is displayed as product of disjoint cycles.

Assumptions:

for x,y in Sym(n), the product xy is read from left to right;

for x in Sym(n) and i in {1,...,n}, i^x and x(i) return the image of i under the action of x.

symmetric_group — Function

symmetric_group(::Type{T} = PermGroup, n::Int)

Return the full symmetric group on the set {1, 2, ..., n}, as an instance of T, where T is in {PermGroup, PcGroup}.

isnatural_symmetric_group — Method

isnatural_symmetric_group(G::GAPGroup)

Return true if G is a permutation group acting as the symmetric group on its moved points, and false otherwise.

isisomorphic_with_symmetric_group — Method

isisomorphic_with_symmetric_group(G::GAPGroup)

Return true if G is isomorphic with a symmetric group, and false otherwise.

alternating_group — Function

alternating_group(::Type{T} = PermGroup, n::Int)

Return the full alternating group on the set {1, 2, ..., n}, as an instance of T, where T is in {PermGroup, PcGroup}.

isnatural_alternating_group — Method

isnatural_alternating_group(G::GAPGroup)

Return true if G is a permutation group acting as the alternating group on its moved points, and false otherwise.

isisomorphic_with_alternating_group — Method

isisomorphic_with_alternating_group(G::GAPGroup)

Return true if G is isomorphic with an alternating group, and false otherwise.

In Oscar, every permutation group has a degree n, that corresponds to the size of the set on which G acts.

degree — Method

degree(G::PermGroup) -> Int

Return the degree of G as a permutation group, that is, an integer n that is stored in G, with the following meaning.

G embeds into symmetric_group(n).

Two permutation groups of different degrees are regarded as not equal, even if they contain the same permutations.

Subgroups constructed with derived_subgroup, sylow_subgroup, etc., get the same degree as the given group.

The range 1:degree(G) is used as the default set of points on which G and its element acts.

Note

The degree of a group of permutations is not necessarily equal to the largest moved point of the group G. For example, the trivial subgroup of symmetric_group(n) has degree n even though it fixes n.

Examples

julia> degree(symmetric_group(4))
4

julia> t4 = trivial_subgroup(symmetric_group(4))[1];

julia> degree(t4)
4

julia> t4 == trivial_subgroup(symmetric_group(5))[1]
false

julia> show(Vector(gen(symmetric_group(4), 2)))
[2, 1, 3, 4]
julia> show(Vector(gen(symmetric_group(5), 2)))
[2, 1, 3, 4, 5]

Permutations

Permutations in Oscar are displayed as products of disjoint cycles, as in GAP. An explicit permutation can be built using the functions perm, gap_perm and cperm.

perm — Function

perm(G::PermGroup, L::AbstractVector{<:Integer})
(G::PermGroup)(L::AbstractVector{<:Integer})

Return the permutation $x$ which maps every i from 1 to $n$= length(L) to L$[i]$. The parent of $x$ is G. An exception is thrown if $x$ is not contained in G or L does not contain every integer from 1 to $n$ exactly once.

For gap_perm, the parent group of $x$ is set to symmetric_group$(n)$.

Examples

julia> perm(symmetric_group(6),[2,4,6,1,3,5])
(1,2,4)(3,6,5)

gap_perm — Function

gap_perm(L::AbstractVector{<:Base.Integer})

Return the permutation $x$ which maps every $i$ from 1 to $n$= length(L) to L$[i]$. The parent of $x$ is set to symmetric_group$(n)$. An exception is thrown if L does not contain every integer from 1 to $n$ exactly once.

Examples

julia> gap_perm([2,4,6,1,3,5])
(1,2,4)(3,6,5)

cperm — Function

cperm(L::AbstractVector{<:T}...) where T <: Union{Integer, fmpz}
cperm(G::PermGroup, L::AbstractVector{<:T}...)

For given lists $[a_1, a_2, \ldots, a_n], [b_1, b_2, \ldots , b_m], \ldots$ of positive integers, return the permutation $x = (a_1, a_2, \ldots, a_n) * (b_1, b_2, \ldots, b_m) * \ldots$. Arrays of the form [n, n+1, ..., n+k] can be replaced by n:n+k.

The parent of $x$ is G. If G is not specified then the parent of $x$ is set to symmetric_group$(n)$, where $n$ is the largest integer that occurs in an entry of L.

An exception is thrown if $x$ is not contained in G or one of the given vectors is empty or contains duplicates.

Examples

julia> cperm([1,2,3],4:7)
(1,2,3)(4,5,6,7)

julia> cperm([1,2],[2,3])
(1,3,2)

julia> p = cperm([1,2,3],[7])
(1,2,3)

julia> degree(parent(p))
7

At the moment, the input vectors of the function cperm need not be disjoint.

Warning

If the function perm is evaluated in a vector of integers without specifying the group G, then the returned value is an element of the AbstractAlgebra.jl type Perm{Int}. For this reason, if one wants a permutation of type GAPGroupElem{PermGroup} without specifying a parent, one has to use the function gap_perm.

Every permutation has always a permutation group as a parent. Two permutations coincide if, and only if, they move the same points and their parent groups have the same degree.

julia> G=symmetric_group(5);

julia> A=alternating_group(5);

julia> x=cperm(G,[1,2,3]);

julia> y=cperm(A,[1,2,3]);

julia> z=cperm([1,2,3]); parent(z)
Sym( [ 1 .. 3 ] )

julia> x==y
true

julia> x==z
false

In the example above, x and y are equal because both act on a set of cardinality 5, while x and z are different because x belongs to Sym(5) and z belongs to Sym(3).

If G is a group and x is a permutation, G(x) returns a permutation x with parent G; an exception is thrown if x does not embed into G.

julia> G=symmetric_group(5);
julia> x=cperm([1,2,3]);
julia> y=G(x);
julia> parent(x)Sym( [ 1 .. 3 ] )
julia> parent(y)Sym( [ 1 .. 5 ] )

The function Vector{T} works in the opposite way with respect to perm:

Vector — Type

Vector{T}(x::PermGroupElem, n::Int = x.parent.deg) where T <: Union{Integer, fmpz}
Vector(x::PermGroupElem, n::Int = x.parent.deg)

Return the list of length n that contains x(i) at position i. If not specified, T is set as Int.

Examples

julia> pi = cperm(1:3)
(1,2,3)
julia> Vector(pi)
3-element Vector{Int64}:
 2
 3
 1
julia> Vector(pi, 2)
2-element Vector{Int64}:
 2
 3
julia> Vector(pi, 4)
4-element Vector{Int64}:
 2
 3
 1
 4
julia> Vector{fmpz}(pi, 2)
2-element Vector{fmpz}:
 2
 3

Permutations as functions

A permutation can be viewed as a function on the set {1,...,n}, hence it can be evaluated on integers.

Note

The multiplication between permutations works from the left to the right. So, if x and y are permutations and n is an integer, then (x*y)(n) = (y(x(n)), NOT x(y(n)). This works also if the argument is not in the range 1:n; in such a case, the output coincides with the input.

julia> x = cperm([1,2,3,4,5]);

julia> x(2)
3

julia> x(6)
6