Quotients

Quotients

Quotient groups in Oscar can be defined using the instruction quo in two ways.

Quotients by normal subgroups.

quo — Method

quo(G::T, H::T)

Return the quotient group G/H of type PcGroup (if the quotient group is solvable) or PermGroup (otherwise), together with the projection G -> G/H.

Quotients by elements.

quo — Method

quo(G::T, elements::Vector{S})

Return the quotient group G/H of type FPGroup (if T=FPGroup), PcGroup (if the quotient group is solvable) or PermGroup (otherwise), where H is the normal closure of elements in G.

This is the typical way to build finitely presented groups.

Example:

julia> F=free_group(2);

julia> (f1,f2)=gens(F);

julia> G,_=quo(F,[f1^2,f2^3,(f1*f2)^2]);

julia> isfinite(G)
true

julia> isisomorphic(G,symmetric_group(3))[1]
true

Similarly to the subgroups, the output consists of a pair (Q,p), where Q is the quotient group and p is the projection homomorphism of G into Q.