Quadratic and hermitian lattices

Creation of lattices

quadratic_lattice — Method

quadratic_lattice(K::NumField, B::MatElem; gram_ambient_space = F) -> QuadLat

Given a number field $K$ and a matrix $B$, returns the quadratic lattice spanned by the rows of $B$ inside the quadratic space over $K$ with gram matrix $F$.

If $F$ is not supplied, the gram matrix of the ambient space will be the identity matrix.

quadratic_lattice — Method

quadratic_lattice(K::NumField, B::PMat; gram_ambient_space = F) -> QuadLat

Given a number field $K$ and a pseudo-matrix $B$, returns the quadratic lattice spanned by the pseudo-matrix $B$ inside the quadratic space over $K$ with gram matrix $F$.

If $F$ is not supplied, the gram matrix of the ambient space will be the identity matrix.

hermitian_lattice — Method

hermitian_lattice(K::NumField, B::MatElem; gram_ambient_space = F) -> HermLat

Given a number field $K$ and a matrix $B$, returns the hermitian lattice spanned by the rows of $B$ inside the hermitian space over $K$ with gram matrix $F$.

If $F$ is not supplied, the gram matrix of the ambient space will be the identity matrix.

hermitian_lattice — Method

hermitian_lattice(K::NumField, B::PMat; gram_ambient_space = F) -> HermLat

Given a number field $K$ and a pseudo-matrix $B$, returns the hermitian lattice spanned by the pseudo-matrix $B$ inside the hermitian space over $K$ with gram matrix $F$.

If $F$ is not supplied, the gram matrix of the ambient space will be the identity matrix.

Basic invariants

ambient_space — Method

ambient_space(L::AbsLat) -> AbsSpace

Returns the ambient space of $L$. If the ambient space is not known, an error is raised.

rational_span — Method

rational_span(L::AbsLat) -> AbsSpace

Returns the rational span of $L$.

diagonal — Method

diagonal(A::Mat{T}) -> Vector{T}

Returns the diagonal of A as an array.

fixed_field — Method

fixed_field(L::AbsLat) -> NumField

Returns the fixed field of the involution of $L$.

involution — Method

involution(L::AbsLat) -> Map

Returns the involution of the rational span of $L$.

rank — Method

rank(L::AbsLat) -> Int

Returns the rank of $L$, that is the dimension of the rational span of $L$.

degree — Method

degree(L::AbsLat) -> Int

Returns the dimension of the ambient space of $L$.

generators — Method

generators(L::AbsLat; minimal = false) -> Vector{Vector}

Returns a set of generators of $L$ over the base ring of $L$.

If minimal == true, the number of generators is minimal. Note that computing minimal generators is expensive.

discriminant — Method

discriminant(L::AbsLat) -> NfOrdFracIdl

Returns the discriminant of $L$, that is, the generalized index ideal $[L^\# : L]$.

pseudo_matrix — Method

pseudo_matrix(L::AbsLat) -> PMat

Returns the basis pseudo-matrix of $L$.

coefficient_ideals — Method

coefficient_ideals(L::AbsLat) -> Vector{NfOrdIdl}

Returns the coefficient ideals of the pseudo-basis of $L$.

absolute_basis — Method

absolute_basis(L::AbsLat) -> Vector

Returns a $\mathbf{Z}$-basis of $L$.

absolute_basis_matrix — Method

absolute_basis_matrix(L::AbsLat) -> MatElem

Returns a $\mathbf{Z}$-basis matrix of $L$.

Rational invariants

hasse_invariant — Method

hasse_invariant(L::AbsLat, p::Union{InfPlc, NfOrdIdl}) -> Int

Returns the Hasse invariant of the rational span of $L$ at $p$. The lattice must be quadratic.

witt_invariant — Method

witt_invariant(L::AbsLat, p::Union{InfPlc, NfOrdIdl}) -> Int

Returns the Witt invariant of the rational span of $L$ at $p$. The lattice must be quadratic.

isrationally_equivalent — Method

isrationally_equivalent(L::AbsLat, M::AbsLat, p::Union{InfPlc, NfOrdIdl})
                                                                     -> Bool

Returns whether the rational spans of $L$ and $M$ are equivalent over the completion at $\mathfrak p$.

isrationally_equivalent — Method

isrationally_equivalent(L::AbsLat, M::AbsLat)
                                        -> Bool

Returns whether the rational spans of $L$ and $M$ are equivalent.

Definiteness

ispositive_definite — Method

ispositive_definite(L::AbsLat) -> Bool

Returns whether the rational span of $L$ is positive definite.

isnegative_definite — Method

isnegative_definite(L::AbsLat) -> Bool

Returns whether the rational span of $L$ is negative definite.

isdefinite — Method

isdefinite(L::AbsLat) -> Bool

Returns whether the rational span of $L$ is definite.

can_scale_totally_positive — Method

can_scale_totally_positive(L::AbsLat) -> Bool, NumFieldElem

Returns whether there is a totally positive rescaled lattice of $L$. If so, the second return value is an element $a$ such that $L^a$ is totally positive.

Module operations

* — Method

*(a::NumFieldElem, M::AbsLat) -> AbsLat

Returns the lattice $aM$ inside the ambient space of $M$.

* — Method

*(a::NfOrdIdl, M::AbsLat) -> AbsLat

Returns the lattice $aM$ inside the ambient space of $M$.

* — Method

*(a::NfOrdFracIdl, M::AbsLat) -> AbsLat

Returns the lattice $aM$ inside the ambient space of $M$.

rescale — Method

rescale(L::AbsLat, a::NumFieldElem) -> AbsLat

Returns the rescaled lattice $L^a$. Note that this has a different ambient space than $L$.

dual — Method

dual(L::AbsLat) -> AbsLat

Returns the dual lattice of $L$.

Invariants

norm — Method

norm(L::AbsLat) -> NfOrdFracIdl

Returns the norm of $L$. This is a fractional ideal of the fixed field of $L$.

scale — Method

scale(L::AbsLat) -> NfOrdFracIdl

Returns the scale of $L$.

isintegral — Method

isintegral(L::AbsLat) -> Bool

Returns whether the lattice $L$ is integral.

volume — Method

volume(L::AbsLat) -> NfOrdFracIdl

Returns the volume of $L$.

ismodular — Method

ismodular(L::AbsLat) -> Bool, NfOrdFracIdl

Returns whether $L$ is modular. In this case, the second return value is a fractional ideal $\mathfrak a$ such that $\mathfrak a L^\# = L$, where $L^\#$ is the dual of $L$.

Local properties

local_basis_matrix — Method

local_basis_matrix(L::AbsLat, p::NfOrdIdl; type = :any) -> MatElem

Given a prime ideal $\mathfrak p$ and a lattice $L$, this function returns a basis matrix of lattice $M$ such that $M_{\mathfrak{p}} = L_{\mathfrak{p}}$.

If type == :submodule, the lattice $L$ will be a sublattice of $M$.

If type == :supermodule, the lattice $L$ will be a superlattice of $M$.

If type == :any, there may not be any containment relation between $M$ and $L$.

jordan_decomposition — Method

jordan_decomposition(L::AbsLat, p::NfOrdIdl)
                            -> Vector{MatElem}, Vector{MatElem}, Vector{Int}

Returns a Jordan decomposition of the completion of $L$ at $\mathfrak p$.

The return value consists of three lists $(M_i)_i$, $(G_i)_i$ and $(s_i)_i$ of the same length $r$. The completions of the row spans of the matrices $M_i$ yield a Jordan decomposition of $L_{\mathfrak{p}}$ into modular sublattices $L_i$ with gram matrices $G_i$ and scale of $\mathfrak{p}$-adic valuation $s_i$.

islocally_isometric — Method

islocally_isometric(L::AbsLat, M::AbsLat, p::NfOrdIdl) -> Bool

Returns whether the completions of $L$ and $M$ at the prime ideal $\mathfrak{p}$ are locally isometric.

Genera

Creation of genera from lattices

genus — Method

genus(L::HermLat, p::NfOrdIdl) -> LocalGenusHerm

Returns the genus of $L$ at the prime ideal $\mathfrak p$.

See [Kir16, Definition 8.3.1].

genus — Method

genus(L::HermLat) -> GenusHerm

Given a Hermitian lattice, return the genus it belongs to.

Properties of genera

rank — Method

rank(G::LocalGenusHerm) -> Int

Given a genus symbol $G$, return the rank of a lattice in $G$.

rank — Method

rank(G::LocalGenusHerm, i::Int)

Given a genus symbol for Hermitian lattices over $E/K$, return the rank of the $i$th Jordan block of $G$.

ranks — Method

ranks(G::LocalGenusHerm)

Given a genus symbol for Hermitian lattices over $E/K$, return the ranks of the Jordan blocks of $G$.

det — Method

det(G::LocalGenusHerm) -> Int

Given a genus symbol $G$, return the determinant of a lattice in $G$. This will be 1 or -1 depending on whether the determinant is a local norm or not.

det_representative — Method

det_representative(G::LocalGenusHerm) -> NumFieldElem

Return a representative for the norm class of the determinant of $G$.

gram_matrix — Method

gram_matrix(G::LocalGenusHerm)

Return a matrix $M$, such that a lattice with Gram matrix $M$ is an element of the given genus.

primes — Method

primes(G::GenusHerm) -> Vector{NfOrdIdl}

Return the primes of a global genus symbol.

Check if lattice is contained in genus

in — Method

in(L::HermLat, G::LocalGenusHerm) -> Bool

Test if the lattice $L$ is contained in the local genus $G$.

in — Method

in(L::HermLat, G::GenusHerm) -> Bool

Test if the lattice $L$ is contained in the genus $G$.

Creating representatives

representative — Method

representative(G::LocalGenusHerm) -> HermLat

Given a local genus, return a Hermitian lattice contained in this genus.