Sparse linear algebra

Introduction

This chapter deals with sparse linear algebra over commutative rings and fields.

Sparse linear algebra, that is, linear algebra with sparse matrices, plays an important role in various algorithms in algebraic number theory. For example, it is one of the key ingredients in the computation of class groups and discrete logarithms using index calculus methods.

Sparse rows

Building blocks for sparse matrices are sparse rows, which are modelled by objects of type \texttt{SRow}. More precisely, the type is of parametrized form objects of type SRow. More precisely, the type is of parametrized form SRow{T}, where T is the element type of the base ring $R$. For example, SRow{fmpz} is the type for sparse rows over the integers.

It is important to note that sparse rows do not have a fixed number of columns, that is, they represent elements of $\{ (x_i)_i \in R^{\mathbb{N}} \mid x_i = 0 \text{ for almost all }i\}$. In particular any two sparse rows over the same base ring can be added.

Creation

sparse_row — Method

sparse_row(R::Ring, J::Vector{Tuple{Int, T}}) -> SRow{T}

Constructs the sparse row $(a_i)_i$ with $a_{i_j} = x_j$, where $J = (i_j, x_j)_j$. The elements $x_i$ must belong to the ring $R$.

sparse_row — Method

sparse_row(R::Ring, J::Vector{Tuple{Int, T}}) -> SRow{T}

Constructs the sparse row $(a_i)_i$ with $a_{i_j} = x_j$, where $J = (i_j, x_j)_j$. The elements $x_i$ must belong to the ring $R$.

sparse_row(R::Ring, J::Vector{Tuple{Int, Int}}) -> SRow

Constructs the sparse row $(a_i)_i$ over $R$ with $a_{i_j} = x_j$, where $J = (i_j, x_j)_j$.

sparse_row — Method

sparse_row(R::Ring, J::Vector{Int}, V::Vector{T}) -> SRow{T}

Constructs the sparse row $(a_i)_i$ over $R$ with $a_{i_j} = x_j$, where $J = (i_j)_j$ and $V = (x_j)_j$.

Basic operations

Rows support the usual operations:

+, -, ==

multiplication by scalars

div, divexact

getindex — Method

getindex(A::SRow, j::Int) -> RingElem

Given a sparse row $(a_i)_{i}$ and an index $j$ return $a_j$.

add_scaled_row — Method

add_scaled_row(A::SRow{T}, B::SRow{T}, c::T) -> SRow{T}

Returns the row $c A + B$.

add_scaled_row — Method

add_scaled_row(A::SRow{T}, B::SRow{T}, c::T) -> SRow{T}

Returns the row $c A + B$.

transform_row — Method

transform_row(A::SRow{T}, B::SRow{T}, i::Int, j::Int, a::T, b::T, c::T, d::T)

Returns the tuple $(aA + bB, cA + dB)$.

length — Method

length(A::SRow)

Returns the number of nonzero entries of $A$.

Change of base ring

change_base_ring — Method

change_base_ring(R::Ring, A::SRow) -> SRow

Create a new sparse row by coercing all elements into the ring $R$.

Maximum, minimum and 2-norm

maximum — Method

maximum(A::SRow{T}) -> T

Returns the largest entry of $A$.

maximum — Method

maximum(A::SRow{T}) -> T

Returns the largest entry of $A$.

minimum — Method

  minimum(A::NfRelOrdIdl) -> NfOrdIdl
  minimum(A::NfRelOrdIdl) -> NfRelOrdIdl

Returns the ideal $A \cap O$ where $O$ is the maximal order of the coefficient ideals of $A$.

minimum(A::SRow{T}) -> T

Returns the smallest entry of $A$.

minimum — Method

minimum(A::SRow{T}) -> T

Returns the smallest entry of $A$.

norm2 — Method

norm2(A::SRow{T} -> T

Returns $A \cdot A^t$.

Functionality for integral sparse rows

lift — Method

lift(A::SRow{nmod}) -> SRow{fmpz}

Return the sparse row obtained by lifting all entries in $A$.

mod! — Method

mod!(A::SRow{fmpz}, n::fmpz) -> SRow{fmpz}

Inplace reduction of all entries of $A$ modulo $n$ to the positive residue system.

mod_sym! — Method

mod_sym!(A::SRow{fmpz}, n::fmpz) -> SRow{fmpz}

Inplace reduction of all entries of $A$ modulo $n$ to the symmetric residue system.

mod_sym! — Method

mod_sym!(A::SRow{fmpz}, n::Integer) -> SRow{fmpz}

Inplace reduction of all entries of $A$ modulo $n$ to the symmetric residue system.

maximum — Method

maximum(abs, A::SRow{fmpz}) -> fmpz

Returns the largest, in absolute value, entry of $A$.

Sparse matrices

Let $R$ be a commutative ring. Sparse matrices with base ring $R$ are modelled by objects of type SMat. More precisely, the type is of parametrized form SRow{T}, where T is the element type of the base ring. For example, SMat{fmpz} is the type for sparse matrices over the integers.

In constrast to sparse rows, sparse matrices have a fixed number of rows and columns, that is, they represent elements of the matrices space $\mathrm{Mat}_{n\times m}(R)$. Internally, sparse matrices are implemented as an array of sparse rows. As a consequence, unlike their dense counterparts, sparse matrices have a mutable number of rows and it is very performant to add additional rows.

Construction

sparse_matrix — Method

sparse_matrix(R::Ring) -> SMat

Return an empty sparse matrix with base ring $R$.

Sparse matrices can also be created from dense matrices as well as from julia arrays:

sparse_matrix — Method

sparse_matrix(A::MatElem; keepzrows::Bool = true)

Constructs the sparse matrix corresponding to the dense matrix $A$. If keepzrows is false, then the constructor will drop any zero row of $A$.

sparse_matrix — Method

sparse_matrix(R::Ring, A::Matrix{T}) -> SMat

Constructs the sparse matrix over $R$ corresponding to $A$.

sparse_matrix — Method

sparse_matrix(R::Ring, A::Matrix{T}) -> SMat

Constructs the sparse matrix over $R$ corresponding to $A$.

The normal way however, is to add rows:

push! — Method

push!(A::SMat{T}, B::SRow{T}) where T

Appends the sparse row B to A.

Sparse matrices can also be concatenated to form larger ones:

vcat! — Method

vcat!(A::SMat, B::SMat) -> SMat

Vertically joins $A$ and $B$ inplace, that is, the rows of $B$ are appended to $A$.

vcat — Method

vcat(A::SMat, B::SMat) -> SMat

Vertically joins $A$ and $B$.

hcat! — Method

hcat!(A::SMat, B::SMat) -> SMat

Horizontally concatenates $A$ and $B$, inplace, changing $A$.

hcat — Method

hcat(A::SMat, B::SMat) -> SMat

Horizontally concatenates $A$ and $B$.

(Normal julia $cat$ is also supported)

There are special constructors:

identity_matrix — Method

identity_matrix(::Type{SMat}, R::Ring, n::Int)

Return a sparse $n$ times $n$ identity matrix over $R$.

zero_matrix — Method

zero_matrix(::Type{SMat}, R::Ring, n::Int)

Return a sparse $n$ times $n$ zero matrix over $R$.

zero_matrix — Method

zero_matrix(::Type{SMat}, R::Ring, n::Int, m::Int)

Return a sparse $n$ times $m$ zero matrix over $R$.

Slices:

sub — Method

sub(A::SMat, r::UnitRange, c::UnitRange) -> SMat

Return the submatrix of $A$, where the rows correspond to $r$ and the columns correspond to $c$.

Transpose:

transpose — Method

transpose(A::SMat) -> SMat

Returns the transpose of $A$.

Elementary Properties

sparsity — Method

sparsity(A::SMat) -> Float64

Return the sparsity of A, that is, the number of zero-valued elements divided by the number of all elements.

density — Method

density(A::SMat) -> Float64

Return the density of A, that is, the number of nonzero-valued elements divided by the number of all elements.

nnz — Method

nnz(A::SMat) -> Int

Return the number of non-zero entries of $A$.

nrows — Method

nrows(A::SMat) -> Int

Return the number of rows of $A$.

ncols — Method

ncols(A::SMat) -> Int

Return the number of columns of $A$.

isone — Method

isone(A::SMat)

Tests if $A$ is an identity matrix.

iszero — Method

iszero(A::SMat)

Tests if $A$ is a zero matrix.

isupper_triangular — Method

isupper_triangular(A::SMat)

Returns true if and only if $A$ is upper (right) triangular.

maximum — Method

maximum(A::SMat{T}) -> T

Finds the largest entry of $A$.

minimum — Method

minimum(A::SMat{T}) -> T

Finds the smallest entry of $A$.

maximum — Method

maximum(abs, A::SMat{fmpz}) -> fmpz

Finds the largest, in absolute value, entry of $A$.

elementary_divisors — Method

elementary_divisors(A::SMat{fmpz}) -> Vector{fmpz}

The elementary divisors of $A$, i.e. the diagonal elements of the Smith normal form of $A$.

solve_dixon_sf — Method

solve_dixon_sf(A::SMat{fmpz}, b::SRow{fmpz}, is_int::Bool = false) -> SRow{fmpz}, fmpz
solve_dixon_sf(A::SMat{fmpz}, B::SMat{fmpz}, is_int::Bool = false) -> SMat{fmpz}, fmpz

For a sparse square matrix $A$ of full rank and a sparse matrix (row), find a sparse matrix (row) $x$ and an integer $d$ s.th. $x A = bd$ holds. The algorithm is a Dixon-based linear p-adic lifting method. If \code{is_int} is given, then $d$ is assumed to be $1$. In this case rational reconstruction is avoided.

hadamard_bound2 — Method

hadamard_bound2(A::SMat{T}) -> T

The square of the product of the norms of the rows of $A$.

echelon_with_transform — Method

echelon_with_transform(A::SMat{nmod}) -> SMat, SMat

Find a unimodular matrix $T$ and an upper-triangular $E$ s.th. $TA = E$ holds.

reduce_full — Method

reduce_full(A::SMat{fmpz}, g::SRow{fmpz},
                      trafo = Val{false}) -> SRow{fmpz}, Vector{Int}

Reduces $g$ modulo $A$ and assumes that $A$ is upper triangular.

The second return value is the array of pivot elements of $A$ that changed.

If trafo is set to Val{true}, then additionally an array of transformations is returned.

hnf! — Method

hnf!(A::SMat{fmpz})

Inplace transform of $A$ into upper right Hermite normal form.

hnf — Method

hnf(A::SMat{fmpz}) -> SMat{fmpz}

Return the upper right Hermite normal form of $A$.

snf — Method

snf(A::SMat{fmpz})

The Smith normal form (snf) of $A$.

hnf_extend! — Method

hnf_extend!(A::SMat{fmpz}, b::SMat{fmpz}, offset::Int = 0) -> SMat{fmpz}

Given a matrix $A$ in HNF, extend this to get the HNF of the concatenation with $b$.

is_diagonal — Method

is_diagonal(A::SMat) -> Bool

True iff only the i-th entry in the i-th row is non-zero.

det — Method

det(A::SMat{fmpz})

The determinant of $A$ using a modular algorithm. Uses the dense (nmod_mat) determinant on $A$ for various primes $p$.

det_mc — Method

det_mc(A::SMat{fmpz})

Computes the determinant of $A$ using a LasVegas style algorithm, i.e. the result is not proven to be correct. Uses the dense (nmod_mat) determinant on $A$ for various primes $p$.

valence_mc — Method

valence_mc{T}(A::SMat{T}; extra_prime = 2, trans = Vector{SMatSLP_add_row{T}}()) -> T

Uses a Monte-Carlo algorithm to compute the valence of $A$. The valence is the valence of the minimal polynomial $f$ of $transpose(A)*A$, thus the last non-zero coefficient, typically $f(0)$.

The valence is computed modulo various primes until the computation stabilises for extra_prime many.

trans, if given, is a SLP (straight-line-program) in GL(n, Z). Then the valence of trans * $A$ is computed instead.

saturate — Method

saturate(A::SMat{fmpz}) -> SMat{fmpz}

Computes the saturation of $A$, that is, a basis for $\mathbf{Q}\otimes M \meet \mathbf{Z}^n$, where $M$ is the row span of $A$ and $n$ the number of rows of $A$.

Equivalently, return $TA$ for an invertible rational matrix $T$, such that $TA$ is integral and the elementary divisors of $TA$ are all trivial.

hnf_kannan_bachem — Method

hnf_kannan_bachem(A::SMat{fmpz}) -> SMat{fmpz}

Compute the Hermite normal form of $A$ using the Kannan-Bachem algorithm.

diagonal_form — Method

diagonal_form(A::SMat{fmpz}) -> SMat{fmpz}

A matrix $D$ that is diagonal and obtained via unimodular row and column operations. Like a snf without the divisibility condition.

Manipulation/ Access

getindex — Method

getindex(A::SMat, i::Int, j::Int)

Given a sparse matrix $A = (a_{ij})_{i, j}$, return the entry $a_{ij}$.

getindex — Method

getindex(A::SMat, i::Int) -> SRow

Given a sparse matrix $A$ and an index $i$, return the $i$-th row of $A$.

setindex! — Method

setindex!(A::SMat, b::SRow, i::Int)

Given a sparse matrix $A$, a sparse row $b$ and an index $i$, set the $i$-th row of $A$ equal to $b$.

swap_rows! — Method

swap_rows!(A::SMat{T}, i::Int, j::Int)

Swap the $i$-th and $j$-th row of $A$ inplace.

swap_cols! — Method

swap_cols!(A::SMat, i::Int, j::Int)

Swap the $i$-th and $j$-th column of $A$ inplace.

scale_row! — Method

scale_row!(A::SMat{T}, i::Int, c::T)

Multiply the $i$-th row of $A$ by $c$ inplace.

add_scaled_col! — Method

add_scaled_col!(A::SMat{T}, i::Int, j::Int, c::T)

Add $c$ times the $i$-th column to the $j$-th column of $A$ inplace, that is, $A_j \rightarrow A_j + c \cdot A_i$, where $(A_i)_i$ denote the columns of $A$.

add_scaled_row! — Method

add_scaled_row!(A::SMat{T}, i::Int, j::Int, c::T)

Add $c$ times the $i$-th row to the $j$-th row of $A$ inplace, that is, $A_j \rightarrow A_j + c \cdot A_i$, where $(A_i)_i$ denote the rows of $A$.

transform_row! — Method

transform_row!(A::SMat{T}, i::Int, j::Int, a::T, b::T, c::T, d::T)

Applies the transformation $(A_i, A_j) \rightarrow (aA_i + bA_j, cA_i + dA_j)$ to $A$, where $(A_i)_i$ are the rows of $A$.

diagonal — Method

diagonal(A::SMat) -> fmpz[]

The diagonal elements of $A$ in an array.

reverse_rows! — Method

reverse_rows!(A::SMat)

Inplace inversion of the rows of $A$.

mod_sym! — Method

mod_sym!(A::SMat{fmpz}, n::fmpz)

Inplace reduction of all entries of $A$ modulo $n$ to the symmetric residue system.

find_row_starting_with — Method

find_row_starting_with(A::SMat, p::Int) -> Int

Tries to find the index $i$ such that $A_{i,p} \neq 0$ and $A_{i, p-j} = 0$ for all $j > 1$. It is assumed that $A$ is upper triangular. If such an index does not exist, find the smallest index larger.

reduce — Method

reduce(A::SMat{fmpz}, g::SRow{fmpz}, m::fmpz) -> SRow{fmpz}

Given an upper triangular matrix $A$ over the integers, a sparse row $g$ and an integer $m$, this function reduces $g$ modulo $A$ and returns $g$ modulo $m$ with respect to the symmetric residue system.

reduce — Method

reduce(A::SMat{fmpz}, g::SRow{fmpz}) -> SRow{fmpz}

Given an upper triangular matrix $A$ over a field and a sparse row $g$, this function reduces $g$ modulo $A$.

reduce — Method

reduce(A::SMat{T}, g::SRow{T}) -> SRow{T}

Given an upper triangular matrix $A$ over a field and a sparse row $g$, this function reduces $g$ modulo $A$.

rand_row — Method

rand_row(A::SMat) -> SRow

Return a random row of the sparse matrix $A$.

Changing of the ring:

map_entries — Method

map_entries(f, A::SMat) -> SMat

Given a sparse matrix $A$ and a callable object $f$, this function will construct a new sparse matrix by applying $f$ to all elements of $A$.

change_base_ring — Method

change_base_ring(R::Ring, A::SMat)

Create a new sparse matrix by coercing all elements into the ring $R$.

Arithmetic

Matrices support the usual operations as well

+, -, ==

div, divexact by scalars

multiplication by scalars

Various products:

mul — Method

mul(A::SMat{T}, b::AbstractVector{T}) -> Vector{T}

Return the product $A \cdot b$ as a dense vector.

mul — Method

mul(A::SMat{T}, b::AbstractMatrix{T}) -> Matrix{T}

Return the product $A \cdot b$ as a dense array.

mul — Method

mul(A::SMat{T}, b::MatElem{T}) -> MatElem

Return the product $A \cdot b$ as a dense matrix.

mul — Method

mul(A::SRow, B::SMat) -> SRow

Return the product $A\cdot B$ as a sparse row.

Other:

sparse — Method

sparse(A::SMat) -> SparseMatrixCSC

The same matrix, but as a sparse matrix of julia type SparseMatrixCSC.

fmpz_mat — Method

fmpz_mat(A::SMat{fmpz})

The same matrix $A$, but as an fmpz_mat.

fmpz_mat — Method

fmpz_mat(A::SMat{T}) where {T <: Integer}

The same matrix $A$, but as an fmpz_mat. Requires a conversion from the base ring of $A$ to $\mathbb ZZ$.

Array — Method

Array(A::SMat{T}) -> Matrix{T}

The same matrix, but as a two-dimensional julia array.