Finite fields

Finite fields are provided in Nemo by Flint. This allows construction of finite fields of any characteristic and degree for which there are Conway polynomials. It is also possible for the user to specify their own irreducible polynomial generating a finite field.

Finite fields are constructed using the FlintFiniteField function. However, for convenience we define

FiniteField = FlintFiniteField

so that finite fields can be constructed using FiniteField rather than FlintFiniteField. Note that this is the name of the constructor, but not of finite field type.

The types of finite field elements in Nemo are given in the following table, along with the libraries that provide them and the associated types of the parent objects.

Library	Field	Element type	Parent type

Flint	$\mathbb{F}_{p^n}$ (small $p$)	fq_nmod	FqNmodFiniteField

Flint	$\mathbb{F}_{p^n}$ (large $p$)	fq	FqFiniteField

The only difference between the fq and fq_nmod types is the representation. The former is for finite fields with multiprecision characteristic and the latter is for characteristics that fit into a single unsigned machine word. The FlintFiniteField constructor automatically picks the correct representation for the user, and so the average user doesn't need to know about the actual types.

All the finite field types belong to the FinField abstract type and the finite field element types belong to the FinFieldElem abstract type.

Since all the functionality for the fq finite field type is identical to that provided for the fq_nmod finite field type, we simply document the former.

Finite field functionality

Finite fields in Nemo provide all the field functionality described in AbstractAlgebra:

https://nemocas.github.io/AbstractAlgebra.jl/latest/field

Below we describe the functionality that is provided in addition to this.

Constructors

In order to construct finite field elements in Nemo, one must first construct the finite field itself. This is accomplished with one of the following constructors.

FlintFiniteField — Function

FlintFiniteField(char::IntegerUnion, deg::Int, s::AbstractString; cached = true)

Returns a tuple $S, x$ consisting of a finite field parent object $S$ and generator $x$ for the finite field of the given characteristic and degree. The string $s$ is used to designate how the finite field generator will be printed. The characteristic must be prime. When a Conway polynomial is known, the field is generated using the Conway polynomial. Otherwise a random sparse, irreducible polynomial is used. The generator of the field is guaranteed to be a multiplicative generator only if the field is generated by a Conway polynomial. We require the degree to be positive.

FlintFiniteField(pol::Union{fmpz_mod_poly, gfp_fmpz_poly}, s::AbstractString; cached = true, check = true)

Returns a tuple $S, x$ consisting of a finite field parent object $S$ and generator $x$ for the finite field over $F_p$ defined by the given polynomial, i.e. $\mathbb{F}_p[t]/(pol)$. The characteristic is specified by the modulus of pol. The polynomial is required to be irreducible, but this is not checked. The base ring of the polynomial is required to be a field, which is checked by default. Use check = false to disable the check. The string $s$ is used to designate how the finite field generator will be printed. The generator will not be multiplicative in general.

Here are some examples of creating finite fields and making use of the resulting parent objects to coerce various elements into those fields.

Examples

R, x = FiniteField(7, 3, "x")
S, y = FiniteField(ZZ(12431351431561), 2, "y")
T, t = PolynomialRing(ResidueRing(ZZ, 12431351431561), "t")
U, z = FiniteField(t^2 + 7, "z")

a = R(5)
b = R(x)
c = S(ZZ(11))
d = U(7)

Basic manipulation

gen — Method

gen(a::FqFiniteField)

Return the generator of the finite field. Note that this is only guaranteed to be a multiplicative generator if the finite field is generated by a Conway polynomial automatically.

isgen — Method

isgen(a::fq)

Return true if the given finite field element is the generator of the finite field, otherwise return false.

coeff — Method

coeff(x::fq, n::Int)

Return the degree $n$ coefficient of the polynomial representing the given finite field element.

degree — Method

degree(a::FqFiniteField)

Return the degree of the given finite field.

modulus — Method

modulus(k::FqFiniteField, var::String="T")

Return the modulus defining the finite field $k$.

Examples

R, x = FiniteField(ZZ(7), 5, "x")

c = gen(R)
d = characteristic(R)
f = order(R)
g = degree(R)
n = isgen(x)

Special functions

Various special functions with finite field specific behaviour are defined.

tr — Method

tr(x::fq)

Return the trace of $x$. This is an element of $\mathbb{F}_p$, but the value returned is this value embedded in the original finite field.

norm — Method

norm(x::fq)

Return the norm of $x$. This is an element of $\mathbb{F}_p$, but the value returned is this value embedded in the original finite field.

frobenius — Method

frobenius(x::fq, n = 1)

Return the iterated Frobenius $\sigma_p^n(x)$ where $\sigma_p$ is the Frobenius map sending the element $a$ to $a^p$ in the finite field of characteristic $p$. By default the Frobenius map is applied $n = 1$ times if $n$ is not specified.

pth_root — Method

pth_root(x::fq)

Return the $p$-th root of $x$ in the finite field of characteristic $p$. This is the inverse operation to the Frobenius map $\sigma_p$.

Examples

R, x = FiniteField(ZZ(7), 5, "x")

a = x^4 + 3x^2 + 6x + 1

b = tr(a)
c = norm(a)
d = frobenius(a)
f = frobenius(a, 3)
g = pth_root(a)

Lift

lift — Method

lift(R::GFPFmpzPolyRing, x::fq)

Lift the finite field element x to a polynomial over the prime field.

Examples

R, x = FiniteField(23, 2, "x")
S, y = PolynomialRing(GF(23), "y")

f = 8x + 9

lift(S, f)

```