Free Associative Algebras

Two-sided ideals

Free Associative Algebras

Two-sided ideals

Types

The OSCAR type for two-sided ideals in a free associative algebra is FreeAssAlgIdeal{T}, where T is the element type of the algebra.

Constructors

ideal(R::FreeAssAlgebra, g::Vector{T}) where T <: FreeAssAlgElem
ideal(g::Vector{T}) where T <: FreeAssAlgElem

Ideal Membership

ideal_membership — Method

ideal_membership(a::FreeAssAlgElem, I::FreeAssAlgIdeal, deg_bound::Int)

Return true if calucations with intermediate degrees bounded by deg_bound prove that $a$ is in $I$. Otherwise, a return of false indicates an inconclusive answer, but larger deg_bounds give more confidence in a negative answer.