Visualization

Visualization

visualize — Method

visualize(P::Polyhedron)

Visualize a polyhedron of dimension at most four (in 3-space). In dimensions up to 3 a usual embedding is shown. Four-dimensional polytopes are visualized as a Schlegel diagram, which is a projection onto one of the facets; e.g., see Chapter 5 of Günter M. Ziegler (1995).

In higher dimensions there is no standard method; use projections to lower dimensions or try ideas from Ewgenij Gawrilow, Michael Joswig, Thilo Rörig, Nikolaus Witte (2010).