Introduction

Introduction

Halfspaces — Type

Halfspaces

Dummy type used for specifying the desired output format. Each halfspace H(a,b) is given by a vector a and a value b such that $H(a,b) = \{ x\ |\ ax ≤ b \}.$

faces — Function

faces(as::Type{T} = Polyhedron, P::Polyhedron, face_dim::Int)

Return the faces of P of dimension face_dim as an iterator over the type of object given by as.

Optional arguments for as include

Polyhedron/Polyhedra.

Examples

An Array containing the six sides of the 3-dimensional cube can be obtained via the following input:

julia> F = faces(Polyhedron, cube(3), 2)
Oscar.PolyhedronFacePolyhedronIterator(A polyhedron in ambient dimension 3, 2)

julia> collect(F)
6-element Array{Any,1}:
 A polyhedron in ambient dimension 3
 A polyhedron in ambient dimension 3
 A polyhedron in ambient dimension 3
 A polyhedron in ambient dimension 3
 A polyhedron in ambient dimension 3
 A polyhedron in ambient dimension 3

vertices — Function

vertices(as::Type{T} = Points, P::Polyhedron)

Return an iterator over the vertices of P in the format defined by as.

Optional arguments for as include

Points.

See also: vertices_as_point_matrix.

Examples

The following code computes the vertices of the Minkowski sum of a triangle and a square:

julia> P = simplex(2) + cube(2);

julia> collect(vertices(Points, P))
5-element Array{Polymake.Vector{Polymake.Rational},1}:
pm::Vector<pm::Rational>
-1 -1
pm::Vector<pm::Rational>
2 -1
pm::Vector<pm::Rational>
2 1
pm::Vector<pm::Rational>
-1 2
pm::Vector<pm::Rational>
1 2

vertices_as_point_matrix — Function

vertices_as_point_matrix(P::Polyhedron)

Return a matrix whose rows are the vertices of P.

Examples

The following code computes the vertices of the Minkowski sum of a triangle and a square:

julia> P = simplex(2) + cube(2);

julia> vertices_as_point_matrix(P)
pm::Matrix<pm::Rational>
-1 -1
2 -1
2 1
-1 2
1 2

nrays — Function

nrays(C)

Return the number of rays of the cone C.

Arguments

C::Cone: A cone.

Examples

Here a cone is constructed from three rays. Calling nrays reveals that one of these was redundant:

julia> R = [1 0; 0 1; 0 2];

julia> PO = positive_hull(R);

julia> nrays(PO)
2

nrays(P::Polyhedron)

Return the number of rays of P.

Examples

Reflecting the input, the upper half-plane indeed has one ray.

julia> UH = convex_hull([0 0],[0 1],[1 0]);

julia> nrays(UH)
1

nrays(PF::PolyhedralFan)

Returns the number of rays of a polyhedral fan.

nvertices — Function

nvertices(P::Polyhedron)

Return the number of vertices of P.

Examples

The 3-cube's number of vertices can be obtained with this input:

julia> C = cube(3);

julia> nvertices(C)
8

rays — Function

rays(C)

Return the rays of a cone.

Arguments

C::Cone: A cone.

Examples

Here a cone is constructed from three rays. Calling rays reveals that one of these was redundant:

julia> R = [1 0; 0 1; 0 2];

julia> PO = positive_hull(R);

julia> collect(rays(PO))
2-element Vector{Polymake.Vector{Polymake.Rational}}:
 pm::Vector<pm::Rational>
1 0
 pm::Vector<pm::Rational>
0 1

rays(as::Type{T} = Points, P::Polyhedron)

Return minimal set of generators of the cone of unbounded directions of P (i.e. its rays) in the format defined by as.

Optional arguments for as include

Points.

See also rays_as_point_matrix.

Examples

We can verify that the positive orthant of the plane is generated by the two rays in positive unit direction:

julia> PO = convex_hull([0 0], [1 0; 0 1]);
julia> collect(rays(Points, PO))
2-element Array{Polymake.Vector{Polymake.Rational},1}:
 pm::Vector<pm::Rational>
1 0
 pm::Vector<pm::Rational>
0 1

rays(P::Polyhedron)

Return minimal set of generators of the cone of unbounded directions of P (i.e. its rays) as points.

See also rays_as_point_matrix.

Examples

We can verify that the positive orthant of the plane is generated by the two rays in positive unit direction:

julia> PO = convex_hull([0 0], [1 0; 0 1]);
julia> collect(rays(PO))
2-element Array{Polymake.Vector{Polymake.Rational},1}:
 pm::Vector<pm::Rational>
1 0
 pm::Vector<pm::Rational>
0 1

rays(PF::PolyhedralFan)

Returns the rays of a polyhedral fan.

nfacets — Function

nfacets(P::Polyhedron)

Return the number of facets of P.

Examples

The number of facets of the 5-dimensional cross polytope can be retrieved via the following line:

julia> nfacets(cross(5))
32

facets — Function

facets(as::Type{T} = Halfspaces, P::Polyhedron)

Return the facets of P in the format defined by as.

The allowed values for as are

Halfspaces,

Polyhedron/Polyhedra.

See also facets_as_halfspace_matrix_pair.

Examples

We can retrieve the six facets of the 3-dimensional cube this way:

julia> C = cube(3);

julia> collect(facets(Polyhedron, C))
6-element Array{Any,1}:
A polyhedron in ambient dimension 3
A polyhedron in ambient dimension 3
A polyhedron in ambient dimension 3
A polyhedron in ambient dimension 3
A polyhedron in ambient dimension 3
A polyhedron in ambient dimension 3

julia> collect(facets(Halfspaces, C))
6-element Array{Tuple{Polymake.Vector{Polymake.Rational},Polymake.Rational},1}:
(pm::Vector<pm::Rational>
-1 0 0, 1)
(pm::Vector<pm::Rational>
1 0 0, 1)
(pm::Vector<pm::Rational>
0 -1 0, 1)
(pm::Vector<pm::Rational>
0 1 0, 1)
(pm::Vector<pm::Rational>
0 0 -1, 1)
(pm::Vector<pm::Rational>
0 0 1, 1)

facets(P::Polyhedron)

Return the facets of P as halfspaces.

See also facets_as_halfspace_matrix_pair.

Examples

We can retrieve the six facets of the 3-dimensional cube this way:

julia> C = cube(3);

julia> collect(facets(C))
6-element Array{Tuple{Polymake.Vector{Polymake.Rational},Polymake.Rational},1}:
(pm::Vector<pm::Rational>
-1 0 0, 1)
(pm::Vector<pm::Rational>
1 0 0, 1)
(pm::Vector<pm::Rational>
0 -1 0, 1)
(pm::Vector<pm::Rational>
0 1 0, 1)
(pm::Vector<pm::Rational>
0 0 -1, 1)
(pm::Vector<pm::Rational>
0 0 1, 1)

facets_as_halfspace_matrix_pair — Function

facets_as_halfspace_matrix_pair(P::Polyhedron)

Return (A,b) such that $P=P(A,b)$ where $P(A,b) = \{ x | Ax ≤ b \}.$

Examples

julia> C = cube(3);

julia> facets_as_halfspace_matrix_pair(C)
(A = pm::SparseMatrix<pm::Rational, pm::NonSymmetric>
(3) (0 -1)
(3) (0 1)
(3) (1 -1)
(3) (1 1)
(3) (2 -1)
(3) (2 1)
, b = pm::SparseVector<pm::Rational>
1 1 1 1 1 1)

volume — Function

volume(P::Polyhedron)

Return the (Euclidean) volume of P.

Examples

julia> C = cube(2);

julia> volume(C)
4

volume(L::AbsLat) -> NfOrdFracIdl

Returns the volume of $L$.

normalized_volume — Function

normalized_volume(P::Polyhedron)

Return the (normalized) volume of P`.

Examples

julia> C = cube(2);

julia> normalized_volume(C)
8

dim — Function

dim(I::MPolyIdeal)

Return the Krull dimension of I.

Examples

julia> R, (x, y, z) = PolynomialRing(QQ, ["x", "y", "z"])
(Multivariate Polynomial Ring in x, y, z over Rational Field, fmpq_mpoly[x, y, z])

julia> I = ideal(R, [y-x^2, x-z^3])
ideal generated by: -x^2 + y, x - z^3

julia> dim(I)
1

dim(a::MPolyQuoIdeal)

Return the Krull dimension of a.

Examples

julia> R, (x, y, z) = PolynomialRing(QQ, ["x", "y", "z"], ordering=:degrevlex)
(Multivariate Polynomial Ring in x, y, z over Rational Field, fmpq_mpoly[x, y, z])

julia> Q, q = quo(R, ideal(R, [x^3*y^2-y^3*x^2, x*y^4-x*y^2]))
(Quotient of Multivariate Polynomial Ring in x, y, z over Rational Field by ideal generated by: x^3*y^2 - x^2*y^3, x*y^4 - x*y^2, Map from
Multivariate Polynomial Ring in x, y, z over Rational Field to Quotient of Multivariate Polynomial Ring in x, y, z over Rational Field by ideal generated by: x^3*y^2 - x^2*y^3, x*y^4 - x*y^2 defined by a julia-function with inverse
)

julia> I = ideal(Q,[x^3*y^4-x+y, x*y+y^2*x])
MPolyQuoIdeal(x^3*y^4 - x + y, x*y^2 + x*y)

julia> dim(I)
1

dim(Q::MPolyQuo)

Return the Krull dimension of the quotient ring Q.

Examples

julia> R, (x, y, z) = PolynomialRing(QQ, ["x", "y", "z"])
(Multivariate Polynomial Ring in x, y, z over Rational Field, fmpq_mpoly[x, y, z])

julia> Q, _ = quo(R, ideal(R, [y-x^2, x-z^3]))
(Quotient of Multivariate Polynomial Ring in x, y, z over Rational Field by ideal generated by: -x^2 + y, x - z^3, Map from
Multivariate Polynomial Ring in x, y, z over Rational Field to Quotient of Multivariate Polynomial Ring in x, y, z over Rational Field by ideal generated by: -x^2 + y, x - z^3 defined by a julia-function with inverse
)

julia> dim(Q)
1

dim(C)

Return the dimension of a cone.

Arguments

C::Cone: A cone.

Examples

The cone C in this example is 2-dimensional within a 3-dimensional ambient space.

julia> C = Cone([1 0 0; 1 1 0; 0 1 0]);

julia> dim(C)
2

dim(P::Polyhedron)

Return the dimension of P.

Examples

julia> V = [1 2 3; 1 3 2; 2 1 3; 2 3 1; 3 1 2; 3 2 1];

julia> P = convex_hull(V);

julia> dim(P)
2

dim(PF::PolyhedralFan)

Returns the dimension of a polyhedral fan.

dim(Y::YoungTableau) -> BigInt

Return the dimension (using hook-length formula) of the irreducible representation of permutation group $S_n$ associated the partition Y.part.

Since the computation overflows easily BigInt is returned. You may perform the computation of the dimension in different type by calling dim(Int, Y).

Examples

julia> dim(YoungTableau([4,3,1]))
70

julia> dim(YoungTableau([3,1])) # the regular representation of S_4
3

dim(M::FreeModule{T}) where T <: FieldElement

Return the dimension of the given vector space.

dim(N::Submodule{T}) where T <: FieldElement

Return the dimension of the given vector subspace.

dim(N::QuotientModule{T}) where T <: FieldElement

Return the dimension of the given vector quotient space.

dim(V::AbsSpace) -> Int

Return the dimension of the space V.

dim(S::ZpGenus) -> fmpz

Return the dimension of this genus.

dim(G::ZGenus) -> Int

Return the dimension of this genus.

lattice_points — Function

lattice_points(P::Polyhedron)

Return the integer points contained in the bounded polyhedron P.

Examples

julia> S = 2 * simplex(2);

julia> lattice_points(S)
6-element Array{Polymake.VectorAllocated{Polymake.Integer},1}:
 pm::Vector<pm::Integer>
0 0
 pm::Vector<pm::Integer>
0 1
 pm::Vector<pm::Integer>
0 2
 pm::Vector<pm::Integer>
1 0
 pm::Vector<pm::Integer>
1 1
 pm::Vector<pm::Integer>
2 0

ambient_dim — Function

ambient_dim(C)

Return the ambient dimension of a cone.

Arguments

C::Cone: A cone.

Examples

The cone C in this example is 2-dimensional within a 3-dimensional ambient space.

julia> C = Cone([1 0 0; 1 1 0; 0 1 0]);

julia> ambient_dim(C)
3

ambient_dim(P::Polyhedron)

Return the ambient dimension of P.

Examples

julia> V = [1 2 3; 1 3 2; 2 1 3; 2 3 1; 3 1 2; 3 2 1];

julia> P = convex_hull(V);

julia> ambient_dim(P)
3

ambient_dim(PF::PolyhedralFan)

Returns the ambient dimension of a polyhedral fan, which is the dimension of the embedding space. This is equal to the dimension of the fan if the fan is full-dimensional.

codim — Function

codim(I::MPolyIdeal)

Return the codimension of I.

Examples

julia> R, (x, y, z) = PolynomialRing(QQ, ["x", "y", "z"])
(Multivariate Polynomial Ring in x, y, z over Rational Field, fmpq_mpoly[x, y, z])

julia> I = ideal(R, [y-x^2, x-z^3])
ideal generated by: -x^2 + y, x - z^3

julia> codim(I)
2

codim(C)

Returns the codimension of a cone.

Arguments

C::Cone: A cone.

Examples

The cone C in this example is 2-dimensional within a 3-dimensional ambient space.

julia> C = Cone([1 0 0; 1 1 0; 0 1 0]);

julia> codim(C)
1

codim(P::Polyhedron)

Return the codimension of P.

Examples

julia> V = [1 2 3; 1 3 2; 2 1 3; 2 3 1; 3 1 2; 3 2 1];

julia> P = convex_hull(V);

julia> codim(P)
1

lineality_space — Function

lineality_space(C)

Return a basis of the lineality space of a cone.

Arguments

C::Cone: A cone.

Examples

Three rays are used here to construct the upper half-plane. Actually, two of these rays point in opposite directions. This gives us a 1-dimensional lineality.

julia> UH = Cone([1 0; 0 1; -1 0]);

julia> lineality_space(UH)
pm::Matrix<pm::Rational>
1 0

lineality_space(P::Polyhedron)

Return a matrix whose row span is the lineality space of P.

Examples

Despite not being reflected in this construction of the upper half-plane, its lineality in $x$-direction is recognized:

julia> UH = convex_hull([0 0],[0 1; 1 0; -1 0]);

julia> lineality_space(UH)
pm::Matrix<pm::Rational>
1 0

lineality_space(PF::PolyhedralFan)

Returns the lineality_space of a polyhedral fan.

recession_cone — Function

recession_cone(P::Polyhedron)

Return the recession cone of P.

Examples

julia> P = Polyhedron([1 -2; -1 1; -1 0; 0 -1],[2,1,1,1]);

julia> collect(vertices(P))
3-element Array{Polymake.Vector{Polymake.Rational},1}:
 pm::Vector<pm::Rational>
0 -1
 pm::Vector<pm::Rational>
-1 0
 pm::Vector<pm::Rational>
-1 -1

julia> recession_cone(P)
A polyhedral cone in ambient dimension 2

julia> collect(rays(recession_cone(P)))
2-element Array{Polymake.Vector{Polymake.Rational},1}:
 pm::Vector<pm::Rational>
1 1/2
 pm::Vector<pm::Rational>
1 1

isfeasible — Function

isfeasible(P::Polyhedron)

Check whether P is feasible, i.e. non-empty.

Examples

julia> P = Polyhedron([1 -1; -1 1; -1 0; 0 -1],[-1,-1,1,1]);

julia> isfeasible(P)
false

issmooth — Function

issmooth(P::Polyhedron)

Check whether P is smooth.

Examples

julia> C = cube(8);

julia> issmooth(C)
true

issmooth(PF::PolyhedralFan)

Determine whether the fan is smooth.

issmooth(C::AffinePlaneCurve{S}, P::Point{S}) where S <: FieldElem

Throw an error if P is not a point of C, return false if P is a singular point of C, and true if P is a smooth point of C.

issmooth(C::ProjectivePlaneCurve{S}, P::Oscar.Geometry.ProjSpcElem{S}) where S <: FieldElem

Throw an error if P is not a point of C, return false if P is a singular point of C, and true if P is a smooth point of C.

isnormal — Function

isnormal(G::T, H::T) where T <: GAPGroup

Return whether the subgroup H is normal in G.

isnormal(A::MPolyQuo)

Given an affine algebra A over a perfect field, return true if A is normal, false otherwise.

CAVEAT: The function computes the normalization of A. This may take some time.

isnormal(P::Polyhedron)

Check whether P is normal.

Examples

The 3-cube is normal.

julia> C = cube(3)
A polyhedron in ambient dimension 3

julia> isnormal(C)
true

But this pyramid is not:

julia> P = convex_hull([0 0 0; 0 1 1; 1 1 0; 1 0 1]);

julia> isnormal(P)
false

 isnormal(K::AnticNumberField) -> Bool

Returns true if $K$ is a normal extension of $\mathbb Q$, false otherwise.

isnormal(C::ClassField) -> Bool

For a class field $C$ defined over a normal base field $k$, decide if $C$ is normal over $Q$.

isnormal(a::AlgAssAbsOrdIdl) -> Bool

Returns true if $a$ is a normal ideal and false otherwise.

isbounded — Function

isbounded(P::Polyhedron)

Check whether P is bounded.

Examples

julia> P = Polyhedron([1 -3; -1 1; -1 0; 0 -1],[1,1,1,1]);

julia> isbounded(P)
false

isfulldimensional — Function

isfulldimensional(C)

Determine whether the cone is full dimensional

Arguments

C::Cone: A cone.

Examples

The cone C in this example is 2-dimensional within a 3-dimensional ambient space.

julia> C = Cone([1 0 0; 1 1 0; 0 1 0]);

julia> isfulldimensional(C)
false

isfulldimensional(P::Polyhedron)

Check whether P is full dimensional.

Examples

julia> V = [1 2 3; 1 3 2; 2 1 3; 2 3 1; 3 1 2; 3 2 1];

julia> isfulldimensional(convex_hull(V))
false

f_vector — Function

f_vector(P::Polyhedron)

Compute the vector (f_1,f_2,...,f_(dim(P)-1)) where f_i is the number of faces of P of dimension i.

Examples

Here we compute the f-vector of the 5-cube:

julia> f_vector(cube(5))
5-element Array{Int64,1}:
 32
 80
 80
 40
 10

support_function — Function

support_function(P::Polyhedron; convention::Symbol = :max)

Produce a function $h(ω) = max\{dot(x,ω)\ |\ x \in P\}$. $max$ may be changed to $min$ by setting convention = :min.

Examples

julia> P = cube(3) + simplex(3);

julia> φ = support_function(P);

julia> φ([1,2,3])
9

julia> ψ = support_function(P, convention = :min);

julia> ψ([1,2,3])
-6

orbit_polytope — Function

orbit_polytope(V, G)

Construct the convex hull of the orbit of one or several points under the action of a permutation group.

Arguments

V::AbstractVecOrMat: Initial point(s).

P::PermGroup: A permutation group.

Examples

This will construct the $3$-dimensional permutahedron:

julia> V = [1 2 3];

julia> G = symmetric_group(3);

julia> P = orbit_polytope(V, G)
A polyhedron in ambient dimension 3

julia> collect(vertices(P))
6-element Vector{Polymake.Vector{Polymake.Rational}}:
 pm::Vector<pm::Rational>
1 2 3
 pm::Vector<pm::Rational>
1 3 2
 pm::Vector<pm::Rational>
2 1 3
 pm::Vector<pm::Rational>
2 3 1
 pm::Vector<pm::Rational>
3 1 2
 pm::Vector<pm::Rational>
3 2 1

cube — Function

cube(d , l = -1, u = 1)

Construct the $[l,u]$-cube in dimension $d$.

Arguments

d::Int: Dimension of the cube.

l::Rational: Lower bound for each coordinate.

u::Rational: Upper bound for each coordinate.

Examples

In this example the 5-dimensional unit cube is constructed to ask for one of its properties:

julia> C = cube(5,0,1);

julia> normalized_volume(C)
120

newton_polytope — Function

newton_polytope(poly)

Compute the Newton polytope of the given polynomial poly.

Arguments

poly::Polynomial: A multivariate polynomial.

Examples

julia> S, (x, y) = PolynomialRing(ZZ, ["x", "y"])
(Multivariate Polynomial Ring in x, y over Integer Ring, fmpz_mpoly[x, y])

julia> f = x^3*y + 3x*y^2 + 1
x^3*y + 3*x*y^2 + 1

julia> NP = newton_polytope(f)
A polyhedron in ambient dimension 2

julia> collect(vertices(NP))
3-element Array{Polymake.Vector{Polymake.Rational},1}:
 pm::Vector<pm::Rational>
3 1
 pm::Vector<pm::Rational>
1 2
 pm::Vector<pm::Rational>
0 0

intersect — Function

intersect(V::T...) where T <: Group
intersect(V::AbstractVector{T}) where T <: Group

If V = [G_1, ... , G_n], return the group intersection K of the groups G_1, ..., G_n, together with the embeddings K -> G_i.

intersect(V::AbstractVector{Union{T, GroupCoset, GroupDoubleCoset}}) where T <: GAPGroup

Return an array containing all elements belonging to all groups and cosets in the vector V.

intersect(I::MPolyIdeal, Js::MPolyIdeal...)

Return the intersection of two or more ideals.

Examples

julia> R, (x, y) = PolynomialRing(QQ, ["x", "y"])
(Multivariate Polynomial Ring in x, y over Rational Field, fmpq_mpoly[x, y])

julia> I = intersect(ideal(R, [x, y])^2, ideal(R, [y^2-x^3+x]))
ideal generated by: x^3*y - x*y - y^3, x^4 - x^2 - x*y^2

intersect(a::MPolyQuoIdeal, bs::MPolyQuoIdeal...)

Return the intersection of two or more ideals.

Examples

julia> R, (x, y) = PolynomialRing(QQ, ["x", "y"])
(Multivariate Polynomial Ring in x, y over Rational Field, fmpq_mpoly[x, y])

julia> A, _ = quo(R, ideal(R, [x^2-y^3, x-y]))
(Quotient of Multivariate Polynomial Ring in x, y over Rational Field by ideal generated by: x^2 - y^3, x - y, Map from
Multivariate Polynomial Ring in x, y over Rational Field to Quotient of Multivariate Polynomial Ring in x, y over Rational Field by ideal generated by: x^2 - y^3, x - y defined by a julia-function with inverse
)

julia> I = ideal(A, [y^2])
MPolyQuoIdeal(y^2)

julia> J = ideal(A, [x])
MPolyQuoIdeal(x)

julia> intersect(I,J)
MPolyQuoIdeal(x*y)

intersect(P, Q)

Intersect two polyhedra.

Arguments

P::Polyhedron: First polyhedron.

Q::Polyhedron: Second polyhedron.

Examples

The positive orthant of the plane is the intersection of the two halfspaces with $x>0$ and $y>0$ respectively.

julia> UH1 = convex_hull([0 0],[1 0],[0 1]);

julia> UH2 = convex_hull([0 0],[0 1],[1 0]);

julia> PO = intersect(UH1, UH2)
A polyhedron in ambient dimension 2

julia> collect(rays(PO))
2-element Vector{Polymake.Vector{Polymake.Rational}}:
 pm::Vector<pm::Rational>
1 0
 pm::Vector<pm::Rational>
0 1

Base.intersect(M::FPModule{T}, N::FPModule{T}) where T <: RingElement

Return the intersection of the modules $M$ as a submodule of $M$. Note that $M$ and $N$ must be (constructed as) submodules (transitively) of some common module $P$.

intersect( C::AffinePlaneCurve{S}, D::AffinePlaneCurve{S}) where S <: FieldElem

Return the variety defined by the intersection of C and D.

intersect(mG::GrpAbFinGenMap, mH::GrpAbFinGenMap) -> GrpAbFinGen, Map

Given two injective maps of abelian groups with the same codomain $G$, return the intersection of the images as a subgroup of $G$.

intersect(a::NfRelOrdIdl, b::NfRelOrdIdl) -> NfRelOrdIdl

Returns $a \cap b$.

intersect(x::NfOrdIdl, y::NfOrdIdl) -> NfOrdIdl
lcm(x::NfOrdIdl, y::NfOrdIdl) -> NfOrdIdl

Returns $x \cap y$.

intersect(a::ClassField, b::ClassField) -> ClassField

The intersection of $a$ and $b$ as a class field.

intersect(a::AlgAssAbsOrdIdl, b::AlgAssAbsOrdIdl) -> AlgAssAbsOrdIdl

Returns $a \cap b$.

intersect(a::AlgAssRelOrdIdl, b::AlgAssRelOrdIdl) -> AlgAssRelOrdIdl

Returns $a \cap b$.

intersect(L::AbsLat, M::AbsLat) -> AbsLat

Returns the intersection of $L$ and $M$.

The lattices $L$ and $M$ must have the same ambient space.

minkowski_sum — Function

minkowski_sum(P::Polyhedron, Q::Polyhedron)

Minkowski sum of two polyhedra.

Arguments

P::Polyhedron: First polyhedron.

Q::Polyhedron: Second polyhedron.

Examples

The Minkowski sum of a square and the 2-dimensional cross-polytope is an octagon:

julia> P = cube(2);

julia> Q = cross(2);

julia> M = minkowski_sum(P, Q)
A polyhedron in ambient dimension 2

julia> nvertices(M)
8

+ — Method

+(P::Polyhedron, Q::Polyhedron)

Minkowski sum of two polyhedra.

Arguments

P::Polyhedron: First polyhedron.

Q::Polyhedron: Second polyhedron.

Examples

The Minkowski sum of a square and the 2-dimensional cross-polytope is an octagon:

julia> P = cube(2);

julia> Q = cross(2);

julia> M = minkowski_sum(P, Q)
A polyhedron in ambient dimension 2

julia> nvertices(M)
8

* — Method

*(P::Polyhedron, k::Int)

Return the scaled polyhedron kP. Note that k*P = P*k.

Arguments

k::Int: Scaling factor.

Q::Polyhedron: A polyhedron.

Examples

Scaling an $n$-dimensional bounded polyhedron by the factor $k$ results in the volume being scaled by $k^n$. This example confirms the statement for the 6-dimensional cube and $k = 2$.

julia> C = cube(6);

julia> SC = C*2
A polyhedron in ambient dimension 6

julia> volume(SC)//volume(C)
64

* — Method

*(k::Int, Q::Polyhedron)

Return the scaled polyhedron kQ. Note that k*Q = Q*k.

Arguments

k::Int: Scaling factor.

Q::Polyhedron: A polyhedron.

Examples

Scaling an $n$-dimensional bounded polyhedron by the factor $k$ results in the volume being scaled by $k^n$. This example confirms the statement for the 6-dimensional cube and $k = 2$.

julia> C = cube(6);

julia> SC = 2*C
A polyhedron in ambient dimension 6

julia> volume(SC)//volume(C)
64

+ — Method

+(P::Polyhedron, v::AbstractVector)

Return the translation P+v of P by the vector v.

Arguments

P::Polyhedron: A polyhedron.

v::AbstractVector: A vector of the same dimension as the ambient space of P.

Examples

We construct a polyhedron from its $V$-description. Shifting it by the right vector reveals that its inner geometry corresponds to that of the 3-simplex.

julia> P = convex_hull([100 200 300; 101 200 300; 100 201 300; 100 200 301]);

julia> v = [-100, -200, -300];

julia> S = P + v
A polyhedron in ambient dimension 3

julia> collect(vertices(S))
4-element Vector{Polymake.Vector{Polymake.Rational}}:
 pm::Vector<pm::Rational>
0 0 0
 pm::Vector<pm::Rational>
1 0 0
 pm::Vector<pm::Rational>
0 1 0
 pm::Vector<pm::Rational>
0 0 1

+ — Method

+(v::AbstractVector,P::Polyhedron)

Return the translation v+P of P by the vector v.

Arguments

P::Polyhedron: A polyhedron.

v::AbstractVector: A vector of the same dimension as the ambient space of P.

Examples

We construct a polyhedron from its $V$-description. Shifting it by the right vector reveals that its inner geometry corresponds to that of the 3-simplex.

julia> P = convex_hull([100 200 300; 101 200 300; 100 201 300; 100 200 301]);

julia> v = [-100, -200, -300];

julia> S = v + P
A polyhedron in ambient dimension 3

julia> collect(vertices(S))
4-element Vector{Polymake.Vector{Polymake.Rational}}:
 pm::Vector<pm::Rational>
0 0 0
 pm::Vector<pm::Rational>
1 0 0
 pm::Vector<pm::Rational>
0 1 0
 pm::Vector<pm::Rational>
0 0 1

cross — Function

cross(d[,n])

Construct a $d$-dimensional cross polytope around origin with vertices located at $\pm e_i$ for each unit vector $e_i$ of $R^d$, scaled by $n$.

Arguments

d::Int: Dimension of the cross polytope (and its ambient space).

n::Scalar: Scaling factor.

Examples

Here we print the facets of a non-scaled and a scaled 3-dimensional cross polytope:

julia> C = cross(3)
A polyhedron in ambient dimension 3

julia> collect(facets(C))
8-element Vector{Tuple{Polymake.Vector{Polymake.Rational}, Polymake.Rational}}:
 (pm::Vector<pm::Rational>
1 1 1, 1)
 (pm::Vector<pm::Rational>
-1 1 1, 1)
 (pm::Vector<pm::Rational>
1 -1 1, 1)
 (pm::Vector<pm::Rational>
-1 -1 1, 1)
 (pm::Vector<pm::Rational>
1 1 -1, 1)
 (pm::Vector<pm::Rational>
-1 1 -1, 1)
 (pm::Vector<pm::Rational>
1 -1 -1, 1)
 (pm::Vector<pm::Rational>
-1 -1 -1, 1)

julia> D = cross(3, 2)
A polyhedron in ambient dimension 3

julia> collect(facets(D))
8-element Vector{Tuple{Polymake.Vector{Polymake.Rational}, Polymake.Rational}}:
 (pm::Vector<pm::Rational>
1 1 1, 2)
 (pm::Vector<pm::Rational>
-1 1 1, 2)
 (pm::Vector<pm::Rational>
1 -1 1, 2)
 (pm::Vector<pm::Rational>
-1 -1 1, 2)
 (pm::Vector<pm::Rational>
1 1 -1, 2)
 (pm::Vector<pm::Rational>
-1 1 -1, 2)
 (pm::Vector<pm::Rational>
1 -1 -1, 2)
 (pm::Vector<pm::Rational>
-1 -1 -1, 2)

archimedean_solid — Function

archimedean_solid(s)

Construct an Archimedean solid with the name given by String s from the list below. The polytopes are realized with floating point numbers and thus not exact; Vertex-facet-incidences are correct in all cases.

Arguments

s::String: The name of the desired Archimedean solid.

Possible values:

"truncated_tetrahedron" : Truncated tetrahedron.
    Regular polytope with four triangular and four hexagonal facets.
"cuboctahedron" : Cuboctahedron.
    Regular polytope with eight triangular and six square facets.
"truncated_cube" : Truncated cube.
    Regular polytope with eight triangular and six octagonal facets.
"truncated_octahedron" : Truncated Octahedron.
    Regular polytope with six square and eight hexagonal facets.
"rhombicuboctahedron" : Rhombicuboctahedron.
    Regular polytope with eight triangular and 18 square facets.
"truncated_cuboctahedron" : Truncated Cuboctahedron.
    Regular polytope with 12 square, eight hexagonal and six octagonal
    facets.
"snub_cube" : Snub Cube.
    Regular polytope with 32 triangular and six square facets.
    The vertices are realized as floating point numbers.
    This is a chiral polytope.
"icosidodecahedron" : Icosidodecahedon.
    Regular polytope with 20 triangular and 12 pentagonal facets.
"truncated_dodecahedron" : Truncated Dodecahedron.
    Regular polytope with 20 triangular and 12 decagonal facets.
"truncated_icosahedron" : Truncated Icosahedron.
    Regular polytope with 12 pentagonal and 20 hexagonal facets.
"rhombicosidodecahedron" : Rhombicosidodecahedron.
    Regular polytope with 20 triangular, 30 square and 12 pentagonal
    facets.
"truncated_icosidodecahedron" : Truncated Icosidodecahedron.
    Regular polytope with 30 square, 20 hexagonal and 12 decagonal
    facets.
"snub_dodecahedron" : Snub Dodecahedron.
    Regular polytope with 80 triangular and 12 pentagonal facets.
    The vertices are realized as floating point numbers.
    This is a chiral polytope.

Examples

julia> T = archimedean_solid("cuboctahedron")
A polyhedron in ambient dimension 3

julia> sum([nvertices(F) for F in faces(T, 2)] .== 3)
8

julia> sum([nvertices(F) for F in faces(T, 2)] .== 4)
6

julia> nfacets(T)
14

The polyhedral geometry part of OSCAR provides functionality for handling

convex polytopes, unbounded polyhedra and cones

linear programs

General textbooks offering details on theory and algorithms include:

Michael Joswig, Thorsten Theobald (2013)

Günter M. Ziegler (1995)