Polyhedral Fans

Introduction

Construction

Saving and loading

Polyhedral Fans

Introduction

Let $\mathbb{F}$ be an ordered field; the default is that $\mathbb{F}=\mathbb{Q}$ is the field of rational numbers and other fields are not yet supported everywhere in the implementation.

A nonempty finite collection $\mathcal{F}$ of (polyhedral) cones in $\mathbb{F}^n$, for $n$ fixed, is a (polyhedral) fan if

the set $\mathcal{F}$ is closed with respect to taking faces and

if $C,D\in\mathcal{F}$ then $C\cap D$ is a face of both, $C$ and $D$.

Construction

To construct a polyhedral fan, you must pass the rays of each cone in the fan, along with an IncidenceMatrix encoding which rays generate which cones.

PolyhedralFan — Method

PolyhedralFan(Rays, Cones)

Arguments

R::Matrix: Rays generating the cones of the fan; encoded row-wise as representative vectors.

Cones::IncidenceMatrix: An incidence matrix; there is a 1 at position (i,j) if cone i has ray j as extremal ray, and 0 otherwise.

A polyhedral fan formed from rays and cones made of these rays. The cones are given as an IncidenceMatrix, where the columns represent the rays and the rows represent the cones.

Examples

To obtain the upper half-space of the plane:

julia> R = [1 0; 1 1; 0 1; -1 0; 0 -1];

julia> IM=IncidenceMatrix([[1,2],[2,3],[3,4],[4,5],[1,5]]);

julia> PF=PolyhedralFan(R,IM)
A polyhedral fan in ambient dimension 2

normal_fan — Method

normal_fan(P::Polyhedron)

Return the normal fan of P. The maximal cones of the normal fan of P are dual to the edge cones at the vertices of P.

Examples

The rays of a normal fan of a cube point in every positive and negative unit direction.

julia> C = cube(3);

julia> NF = normal_fan(C)
A polyhedral fan in ambient dimension 3

julia> collect(rays(NF))
6-element Vector{Polymake.Vector{Polymake.Rational}}:
 pm::Vector<pm::Rational>
1 0 0
 pm::Vector<pm::Rational>
-1 0 0
 pm::Vector<pm::Rational>
0 1 0
 pm::Vector<pm::Rational>
0 -1 0
 pm::Vector<pm::Rational>
0 0 1
 pm::Vector<pm::Rational>
0 0 -1

face_fan — Method

face_fan(P::Polyhedron)

Return the face fan of P. The polytope P has to contain the origin, then the maximal cones of the face fan of P are the cones over the facets of P.

Examples

By definition, this bounded polyhedron's number of facets equals the amount of maximal cones of its face fan.

julia> C = cross(3);

julia> FF = face_fan(C)
A polyhedral fan in ambient dimension 3

julia> nmaximal_cones(FF) == nfacets(C)
true

Saving and loading

Objects of type PolyhedralFan can be saved to a file and loaded from a file in the following way:

julia> square = cube(2)A polyhedron in ambient dimension 2
julia> fan = normal_fan(square)A polyhedral fan in ambient dimension 2
julia> save_polyhedralfan(fan, "F.fan")
julia> f = load_polyhedralfan("F.fan")A polyhedral fan in ambient dimension 2
julia> collect(rays(f))4-element Vector{Polymake.Vector{Polymake.Rational}}:
 pm::Vector<pm::Rational>
1 0
 pm::Vector<pm::Rational>
-1 0
 pm::Vector<pm::Rational>
0 1
 pm::Vector<pm::Rational>
0 -1

The file is in JSON format and contains all previously gathered data belonging to the underlying polymake object. In particular, this file can now be read by both polymake and Oscar.

save_polyhedralfan — Method

save_polyhedralfan(PolyhedralFan, String)

Save a polyhedral fan to a file in JSON format. The first argument is the polyhedral fan, the second argument is the filename.

load_polyhedralfan — Method

load_polyhedralfan(String)

Load a polyhedral fan stored in JSON format, given the filename as input.