Polyhedra

Introduction

Construction

Polyhedron and polymake's Polytope

Polyhedra

Introduction

A subset $P \subseteq \mathbb{R}^n$ is called a (metric) polyhedron if it can be written as the intersection of finitely many closed affine halfspaces in $\mathbb{R}^n$. That is, there exists a matrix $A$ and a vector $b$ such that $P = P(A,b) = \{ x \in \mathbb{R}^n \mid Ax \leq b\}.$ Writing $P$ as above is called the $H$-representation of $P$.

Polyhedra are convex bodies. When a polyhedron $P \subset \mathbb{R}^n$ is bounded, it is called a polytope and the fundamental theorem of polytopes states that it may be written as the convex hull of finitely many points in $\mathbb{R}^n$. That is $P = \textrm{conv}(p_1,\ldots,p_N), p_i \in \mathbb{R}^n.$ Writing $P$ in this way is called its $V$-representation.

Since many polyhedral computations are done through polymake, the structure Polyhedron contains a pointer pm_polytope to the corresponding polymake object. Thus, all functionality of Polymake.jl is accessible by calling a Polymake.jl function on this pointer.

Warning: Polyhedra in polymake and Polymake.jl use homogeneous coordinates. The polyhedra in Oscar use affine coordinates.

Construction

Based on the definition of the $H$-representation, the constructor of Polyhedron can be used:

Polyhedron — Method

Polyhedron(A::Union{Oscar.MatElem,AbstractMatrix}, b)

The (metric) polyhedron defined by

\[P(A,b) = \{ x | Ax ≤ b \}.\]

see Def. 3.35 and Section 4.1. of Joswig, M. and Theobald, T. "Polyhedral and Algebraic Methods in Computational Geometry", Springer 2013.

Arguments

A::Matrix: Matrix corresponding to the linear coefficients of the inequalilites that describe P.

b::Vector: Vector corresponding to the constant term of the inequalilites that describe P.

Examples

The following lines define the square $[0,1]^2 \subset \mathbb{R}^2$:

A=[1 0; 0 1; -1 0 ; 0 -1];
b=[1,1,0,0];
Polyhedron(A,b)
A polyhedron in ambient dimension 2

A Polyhedron given by its $V$-representation can also be created by calling the following function:

convex_hull — Method

convex_hull(V [, R [, L]])

The polytope given as the convex hull of the rows of a set of points.

see Def. 2.11 and Def. 3.1 of Joswig, M. and Theobald, T. "Polyhedral and Algebraic Methods in Computational Geometry", Springer 2013.

Arguments

V::Matrix: Points whose convex hull is to be computed; encoded as row vectors.

R::Matrix: Rays completing the set of points; encoded row-wise as representative vectors.

L::Matrix: Generators of the Lineality space; encoded as row vectors.

Examples

The following lines define the square $[0,1]^2 \subset \mathbb{R}^2$:

Square = convex_hull([0 0; 0 1; 1 0; 1 1])
A polyhedron in ambient dimension 2

To construct the positive orthant, rays have to be passed:

V = [0 0];
R = [1 0; 0 1];
PO = convex_hull(V, R)
A polyhedron in ambient dimension 2

The closed-upper half plane can be constructed by passing rays and a lineality space:

V = [0 0];
R = [0 1];
L = [1 0];
UH = convex_hull(V, R, L)
A polyhedron in ambient dimension 2

To obtain the x-axis in $\mathbb{R}^2$:

V = [0 0];
R = [];
L = [1 0];
XA = convex_hull(V, R, L)
A polyhedron in ambient dimension 2

Note that both examples provided result in equal polyhedra.

Computing convex hulls

Polyhedron and polymake's Polytope

To allow both Oscar's and polymake's functionality to be applicable to a polyhedron object, it can be converted back and forth:

Polyhedron — Method

Polyhedron(P::Polymake.BigObject)

Construct a Polyhedron corresponding to a polymake Polytope.

pm_polytope — Method

pm_polytope(P::Polyhedron)

Get the underlying polymake Polytope.