Affine Rings

Quotient Rings

Predicates

Properties

Basic arithmetic

Affine Rings

general bla explaining what this is...

Quotient Rings

julia> R, x = PolynomialRing(QQ, :x=>1:3)
(Multivariate Polynomial Ring in x1, x2, x3 over Rational Field, fmpq_mpoly[x1, x2, x3])

julia> Q, mQ = quo(R, ideal([x[1]^2-2, x[2]^2-3, x[3]^2-5]))
(Quotient of Multivariate Polynomial Ring in x1, x2, x3 over Rational Field by ideal generated by: x1^2 - 2, x2^2 - 3, x3^2 - 5, Map from
Multivariate Polynomial Ring in x1, x2, x3 over Rational Field to Quotient of Multivariate Polynomial Ring in x1, x2, x3 over Rational Field by ideal generated by: x1^2 - 2, x2^2 - 3, x3^2 - 5 defined by a julia-function with inverse
)

AbstractAlgebra.quo — Method

quo(R::MPolyRing, I::MPolyIdeal) -> MPolyQuoRing, Map

Creates the affine ring $R$ modulo $I$ and return the new ring as well as the inclusion map from $R$

Elements

Elements are created by moving polynomials into the ring:

julia> R, x = PolynomialRing(QQ, :x=>1:3)
(Multivariate Polynomial Ring in x1, x2, x3 over Rational Field, fmpq_mpoly[x1, x2, x3])

julia> Q, mQ = quo(R, ideal([x[1]^2-2, x[2]^2-3, x[3]^2-5]))
(Quotient of Multivariate Polynomial Ring in x1, x2, x3 over Rational Field by ideal generated by: x1^2 - 2, x2^2 - 3, x3^2 - 5, Map from
Multivariate Polynomial Ring in x1, x2, x3 over Rational Field to Quotient of Multivariate Polynomial Ring in x1, x2, x3 over Rational Field by ideal generated by: x1^2 - 2, x2^2 - 3, x3^2 - 5 defined by a julia-function with inverse
)

julia> Q(x[2]+x[3])
x2 + x3

julia> Q(x[2])
x2

julia> ans^2
ERROR: MethodError: no method matching ^(::Oscar.MPolyQuoElem{fmpq_mpoly}, ::Int64)
Closest candidates are:
  ^(!Matched::Singular.n_Q, ::Int64) at /home/fieker/.julia/packages/Singular/KXbMt/src/number/n_Q.jl:224
  ^(!Matched::Nemo.nmod, ::Int64) at /home/fieker/.julia/packages/Nemo/mmz6V/src/flint/nmod.jl:219
  ^(!Matched::Oscar.StraightLinePrograms.Lazy, ::Integer) at /home/fieker/actions-runner/_work/Oscar.jl/Oscar.jl/StraightLinePrograms/src/lazy.jl:72
  ...

julia> ans == 3
false

As shown: elements are not automatically simplified.

Hecke.simplify — Method

simplify(a::MPolyQuoElem)

Use the relations of the parent ring to obtain a unique simplified representation.

Oscar.simplify! — Method

simplify!(a::MPolyQuoElem)

Use the relations of the parent ring to obtain a unique simplified representation inplace, ie. modify the representatio of $a$.

One can also construct the integer $0$ with the empty constructor:

julia> ZZ()
0

The following special constructors are also provided:

zero(ZZ)

one(ZZ)

julia> zero(ZZ)
0

julia> one(ZZ)
1

Predicates

iszero(n::Oscar.Integer) -> Bool

isone(n::Oscar.Integer) -> Bool

isprobable_prime(n::Oscar.Integer) -> Bool

Properties

sign(n::Oscar.Integer) -> Oscar.Integer

Returns the sign of n, i.e. $n/|n|$ if $n \neq 0$, or $0$ otherwise.

julia> sign(ZZ(23))
1

julia> sign(ZZ(0))
0

julia> sign(ZZ(-1))
-1

Basic arithmetic

Oscar provides the basic arithmetic operations +, - and * and comparison operators ==, !=, including mixed operations between Julia and Oscar integers. It also provides division and powering as described below.