Normal Toric Varities

Introduction

We introduce two main types of normal toric varieties, distinguishing between the affine and non-affine case:

AffineNormalToricVariety is the toric variety associated to a cone $\sigma$, denoted by $U_{\sigma}$ in David A. Cox, John B. Little, Henry K. Schenck (2011)

NormalToricVariety is the toric variety associated to a polyhedral fan $\Sigma$, denoted by $X_{\Sigma}$ in David A. Cox, John B. Little, Henry K. Schenck (2011)

Warning

The lattice is always assumed to be the standard lattice $\mathbb{Z}^n$. Transformations for non-standard lattices will have to be done by the user.

Construction

AffineNormalToricVariety — Method

AffineNormalToricVariety(C::Cone)

Construct the affine normal toric variety corresponding to a polyhedral cone C.

Example

Set C to be the positive orthant in two dimensions.

julia> C = positive_hull([1 0; 0 1])
A polyhedral cone in ambient dimension 2

julia> antv = AffineNormalToricVariety(C)
A normal toric variety corresponding to a polyhedral fan in ambient dimension 2

NormalToricVariety — Method

NormalToricVariety(C::Cone)

Construct the (affine) normal toric variety corresponding to a polyhedral cone C.

Example

Set C to be the positive orthant in two dimensions.

julia> C = positive_hull([1 0; 0 1])
A polyhedral cone in ambient dimension 2

julia> ntv = NormalToricVariety(C)
A normal toric variety corresponding to a polyhedral fan in ambient dimension 2

NormalToricVariety — Method

NormalToricVariety(PF::PolyhedralFan)

Construct the normal toric variety corresponding to a polyhedral fan PF.

Example

Take PF to be the normal fan of the square.

julia> square = cube(2)
A polyhedron in ambient dimension 2

julia> nf = normal_fan(square)
A polyhedral fan in ambient dimension 2

julia> ntv = NormalToricVariety(nf)
A normal toric variety corresponding to a polyhedral fan in ambient dimension 2

NormalToricVariety — Method

NormalToricVariety(P::Polyhedron)

Construct the normal toric variety corresponding to the normal fan of the given polyhedron P.

Note that this only coincides with the projective variety associated to P, if P is normal.

Example

Set P to be a square.

julia> square = cube(2)
A polyhedron in ambient dimension 2

julia> ntv = NormalToricVariety(square)
A normal toric variety corresponding to a polyhedral fan in ambient dimension 2

Auxiliary functions

isaffine — Method

isaffine(ntv::NormalToricVarieties)

Check whether a normal toric variety is affine, i.e. the associated polyhedral fan has exactly one maximal cone.

Examples

julia> C = positive_hull([-2 5; 1 0])
A polyhedral cone in ambient dimension 2

julia> ntv = NormalToricVariety(C)
A normal toric variety corresponding to a polyhedral fan in ambient dimension 2

julia> isaffine(ntv)
true

julia> square = cube(2)
A polyhedron in ambient dimension 2

julia> ntv = NormalToricVariety(square)
A normal toric variety corresponding to a polyhedral fan in ambient dimension 2

julia> isaffine(ntv)
false

iscomplete — Method

iscomplete(ntv::NormalToricVarieties)

Check whether a normal toric variety is complete, i.e. the associated polyhedral fan covers the whole space.

Examples

julia> C = positive_hull([-2 5; 1 0])
A polyhedral cone in ambient dimension 2

julia> ntv = NormalToricVariety(C)
A normal toric variety corresponding to a polyhedral fan in ambient dimension 2

julia> iscomplete(ntv)
false

julia> square = cube(2)
A polyhedron in ambient dimension 2

julia> ntv = NormalToricVariety(square)
A normal toric variety corresponding to a polyhedral fan in ambient dimension 2

julia> iscomplete(ntv)
true

isnormal_variety — Method

isnormal_variety(ntv::NormalToricVarieties)

Check whether a normal toric variety is normal, i.e. the semigroups generated by the Hilbert basis of the dual cones to the maximal cones of the associated polyhedral fan are all saturated.

This is trivially true for toric varieties coming from fans or cones.

isprojective — Method

isprojective(ntv::NormalToricVarieties)

Check whether a normal toric variety is projective.

issimplicial — Method

issimplicial(ntv::NormalToricVarieties)

Check whether a normal toric variety is simplicial.

This is trivially true for all two-dimensional normal toric varieties.

Examples

Take the cone over a square at height one.

julia> C = positive_hull([1 0 0; 1 1 0; 1 1 1; 1 0 1])
A polyhedral cone in ambient dimension 3

julia> antv = AffineNormalToricVariety(C)
A normal toric variety corresponding to a polyhedral fan in ambient dimension 3

julia> issimplicial(antv)
false

issmooth — Method

issmooth(ntv::NormalToricVarieties)

Check whether a normal toric variety is smooth, i.e. whether all the maximal cones of the associated polyhedral fan are simplicial and generated by a lattice basis.

Examples

julia> C = positive_hull([-2 5; 1 0])
A polyhedral cone in ambient dimension 2

julia> ntv = NormalToricVariety(C)
A normal toric variety corresponding to a polyhedral fan in ambient dimension 2

julia> issmooth(ntv)
false

julia> square = cube(2)
A polyhedron in ambient dimension 2

julia> ntv = NormalToricVariety(square)
A normal toric variety corresponding to a polyhedral fan in ambient dimension 2

julia> issmooth(ntv)
true

toric_ideal_binomial_generators — Method

toric_ideal_binomial_generators(antv::AffineNormalToricVariety)

Get the exponent vectors corresponding to the generators of the toric ideal associated to the affine normal toric variety antv.

Example

Take the cyclic quotient singularity corresponding to the pair of integers (2,5).

julia> C = positive_hull([-2 5; 1 0])
A polyhedral cone in ambient dimension 2

julia> antv = AffineNormalToricVariety(C)
A normal toric variety corresponding to a polyhedral fan in ambient dimension 2

julia> toric_ideal_binomial_generators(antv)
pm::Matrix<long>
-1 -1 2 1
-1 0 3 -1
0 -1 -1 2