Toric Line Bundles

Constructors

Properties

Attributes

Method

Toric Line Bundles

Constructors

Generic constructors

ToricLineBundle — Method

ToricLineBundle(v::AbstractNormalToricVariety, c::Vector{fmpz})

Construct the line bundle on the abstract normal toric variety v with class c.

ToricLineBundle — Method

ToricLineBundle(v::AbstractNormalToricVariety, c::Vector{Int})

Convenience method for ToricLineBundle(v::AbstractNormalToricVariety, c::Vector{fmpz}).

Examples

julia> v = projective_space(NormalToricVariety, 2)
A normal, non-affine, smooth, projective, gorenstein, fano, 2-dimensional toric variety without torusfactor

julia> l = ToricLineBundle( v, [ 2 ] )
A toric line bundle on a normal toric variety

ToricLineBundle — Method

ToricLineBundle(v::AbstractNormalToricVariety, d::ToricDivisor)

Construct the toric variety associated to a (Cartier) torus-invariant divisor d on the normal toric variety v.

Examples

julia> v = projective_space(NormalToricVariety, 2)
A normal, non-affine, smooth, projective, gorenstein, fano, 2-dimensional toric variety without torusfactor

julia> l = ToricLineBundle(v, ToricDivisor(v,[1,2,3]))
A toric line bundle on a normal toric variety

Tensor products

* — Method

Base.:*(l1::ToricLineBundle, l2::ToricLineBundle)

Return the tensor product of two line bundles l1 and l2.

Examples

julia> P2 = projective_space(NormalToricVariety, 2)
A normal, non-affine, smooth, projective, gorenstein, fano, 2-dimensional toric variety without torusfactor

julia> l1 = ToricLineBundle(P2, [1])
A toric line bundle on a normal toric variety

julia> l2 = ToricLineBundle(P2, [2])
A toric line bundle on a normal toric variety

julia> l1*l2
A toric line bundle on a normal toric variety

^ — Method

Base.:^(l::ToricLineBundle, p::fmpz)

Return the p-th tensor power of the toric line bundle l.

Examples

julia> P2 = projective_space(NormalToricVariety, 2)
A normal, non-affine, smooth, projective, gorenstein, fano, 2-dimensional toric variety without torusfactor

julia> l = ToricLineBundle(P2, [1])
A toric line bundle on a normal toric variety

julia> l^(-1)
A toric line bundle on a normal toric variety

inv — Method

Base.:inv(l::ToricLineBundle)

Returns the inverse of the toric line bundle l.

Examples

julia> P2 = projective_space(NormalToricVariety, 2)
A normal, non-affine, smooth, projective, gorenstein, fano, 2-dimensional toric variety without torusfactor

julia> l = ToricLineBundle(P2, [1])
A toric line bundle on a normal toric variety

julia> inv(l)
A toric line bundle on a normal toric variety

Equality

== — Method

Base.:(==)(l1::ToricLineBundle, l2::ToricLineBundle)

Returns true if the toric line bundles l1 and l2 are isomorphic and false otherwise.

Examples

julia> P2 = projective_space(NormalToricVariety, 2)
A normal, non-affine, smooth, projective, gorenstein, fano, 2-dimensional toric variety without torusfactor

julia> l1 = ToricLineBundle(P2, [0])
A toric line bundle on a normal toric variety

julia> l2 = l1^(-1)
A toric line bundle on a normal toric variety

julia> l1 == l2
true

Properties

istrivial — Method

istrivial(l::ToricLineBundle)

Returns "true" if the toric line bundle l is trivial and "false" otherwise.

Examples

julia> H = hirzebruch_surface(4)
A normal, non-affine, smooth, projective, gorenstein, non-fano, 2-dimensional toric variety without torusfactor

julia> istrivial(ToricLineBundle(H, [1,0]))
false

julia> istrivial(ToricLineBundle(H, [0,0]))
true

is_basepoint_free — Method

is_basepoint_free(l::ToricLineBundle)

Returns "true" if the toric line bundle l is basepoint free and "false" otherwise.

Examples

julia> H = hirzebruch_surface(4)
A normal, non-affine, smooth, projective, gorenstein, non-fano, 2-dimensional toric variety without torusfactor

julia> is_basepoint_free(ToricLineBundle(H, [1,0]))
true

isample — Method

isample(l::ToricLineBundle)

Returns "true" if the toric line bundle l is ample and "false" otherwise.

Examples

julia> H = hirzebruch_surface(4)
A normal, non-affine, smooth, projective, gorenstein, non-fano, 2-dimensional toric variety without torusfactor

julia> isample(ToricLineBundle(H, [1,0]))
false

is_very_ample — Method

is_very_ample(l::ToricLineBundle)

Returns "true" if the toric line bundle l is very ample and "false" otherwise.

Examples

julia> H = hirzebruch_surface(4)
A normal, non-affine, smooth, projective, gorenstein, non-fano, 2-dimensional toric variety without torusfactor

julia> is_very_ample(ToricLineBundle(H, [1,0]))
false

Attributes

degree — Method

degree(l::ToricLineBundle)

Returns the degree of the toric line bundle l.

divisor_class — Method

divisor_class(l::ToricLineBundle)

Return the divisor class which defines the toric line bundle l.

toric_divisor — Method

toric_divisor(l::ToricLineBundle)

Returns a divisor corresponding to the toric line bundle l.

toric_variety — Method

toric_variety(l::ToricLineBundle)

Return the toric variety over which the toric line bundle l is defined.

Method

We use cohomCalg to compute the dimension of line bundle cohomologies. This is achieved with the following methods.

all_cohomologies — Method

all_cohomologies(l::ToricLineBundle)

Computes the dimension of all sheaf cohomologies of the toric line bundle l by use of the cohomCalg algorithm Ralph Blumenhagen, Benjamin Jurke, Thorsten Rahn, Helmut Roschy (2010), [cohomCalg:Implementation(@cite), Helmut Roschy, Thorsten Rahn (2010), Shin-Yao Jow (2011), Ralph Blumenhagen, Benjamin Jurke, Thorsten Rahn, Helmut Roschy (2012).

Examples

julia> dP3 = del_pezzo(3)
A normal, non-affine, smooth, projective, gorenstein, fano, 2-dimensional toric variety without torusfactor

julia> all_cohomologies(ToricLineBundle(dP3, [1,2,3,4]))
3-element Vector{fmpz}:
 11
 0
 0

cohomology — Method

cohomology(l::ToricLineBundle, i::Int)

Computes the dimension of the i-th sheaf cohomology of the toric line bundle l by use of the cohomCalg algorithm Ralph Blumenhagen, Benjamin Jurke, Thorsten Rahn, Helmut Roschy (2010), [cohomCalg:Implementation(@cite), Helmut Roschy, Thorsten Rahn (2010), Shin-Yao Jow (2011), Ralph Blumenhagen, Benjamin Jurke, Thorsten Rahn, Helmut Roschy (2012).

Examples

julia> dP3 = del_pezzo(3)
A normal, non-affine, smooth, projective, gorenstein, fano, 2-dimensional toric variety without torusfactor

julia> cohomology(ToricLineBundle(dP3, [1,2,3,4]), 0)
11

We can also compute a basis of the global sections.

basis_of_global_sections_via_rational_functions — Method

basis_of_global_sections_via_rational_functions(l::ToricLineBundle)

Returns a basis of the global sections of the toric line bundle l.

basis_of_global_sections_via_homogeneous_component — Method

basis_of_global_sections_via_homogeneous_component(l::ToricLineBundle)

Returns a basis of the global sections of the toric line bundle l, which is computed from a homogeneous component of the cox ring.