Commutative Algebra

Introduction

This part of OSCAR is under development with regard to providing both the functionality and the documentation. At current state, the documentation serves as a guide on what to do and will be constantly updated.

General textbooks offering details on theory and algorithms include:

G. Greuel , G. Pfister (2007)

W. Decker , C. Lossen (2006)

W. Decker , G. Pfister (2013)

Ideal Constructors for Multivariate Polynomial Rings

Hecke.ideal — Method

ideal(g::Array{T, 1}) where {T <: MPolyElem}

Creates the ideal generated by the polynomials in g.

Example

julia> R, (x, y) = PolynomialRing(QQ, ["x", "y"])
(Multivariate Polynomial Ring in x, y over Rational Field, fmpq_mpoly[x, y])

julia> I = ideal([x*y-3*x,y^3-2*x^2*y])
ideal generated by: x*y - 3*x, -2*x^2*y + y^3

Hecke.ideal — Method

ideal(Rx::MPolyRing, g::Array{<:Any, 1})

Creates the ideal generated by the polynomials in g, specifying the ambient ring of the ideal.

Example

julia> R, (x, y) = PolynomialRing(QQ, ["x", "y"])
(Multivariate Polynomial Ring in x, y over Rational Field, fmpq_mpoly[x, y])

julia> J = ideal(R, [(x+y)^2])
ideal generated by: x^2 + 2*x*y + y^2

Gröbner Bases

Monomial Orders

Normal Forms

Computing Gröbner Bases

Oscar.groebner_basis — Method

groebner_basis(I::MPolyIdeal)

Given an ideal I this function computes a Groebner basis w.r.t. the given monomial ordering of the polynomial ring. The Groebner basis is then returned as an array of multivariate polynomials.

Example

julia> R, (x, y) = PolynomialRing(QQ, ["x", "y"], ordering=:degrevlex)
(Multivariate Polynomial Ring in x, y over Rational Field, fmpq_mpoly[x, y])

julia> I = ideal([x*y-3*x,y^3-2*x^2*y])
ideal generated by: x*y - 3*x, -2*x^2*y + y^3

julia> G = groebner_basis(I)
3-element Array{fmpq_mpoly,1}:
 x*y - 3*x
 y^3 - 6*x^2
 x^3 - 9//2*x

Oscar.groebner_basis — Method

groebner_basis(I::MPolyIdeal, ord::Symbol = :degrevlex; complete_reduction::Bool=false)

Given an ideal I, a monomial ordering ord and an optional parameter complete_reduction this function computes a Groebner basis (if complete_reduction = true the reduced Groebner basis) of I w.r.t. the given monomial ordering ord. The Groebner basis is then returned as an array of multivariate polynomials.

Example

julia> R, (x, y) = PolynomialRing(QQ, ["x", "y"], ordering=:degrevlex)
(Multivariate Polynomial Ring in x, y over Rational Field, fmpq_mpoly[x, y])

julia> I = ideal([x*y-3*x,y^3-2*x^2*y])
ideal generated by: x*y - 3*x, -2*x^2*y + y^3

julia> H = groebner_basis(I, ord=:lex)
3-element Array{fmpq_mpoly,1}:
 y^4 - 3*y^3
 x*y - 3*x
 -1//6*y^3 + x^2

Gröbner Bases with transformation matrix

Oscar.groebner_basis_with_transformation_matrix — Method

groebner_basis_with_transformation_matrix(I::MPolyIdeal; ord::Symbol = :degrevlex, complete_reduction::Bool=false)

Returns a pair G, m where G is a Groebner basis of the ideal I with respect to the monomial ordering ord, and m is a transformation matrix from gens(I) to G. If complete_reduction is set to true then G will be the reduced Groebner basis.

Example

julia> R,(x,y) = PolynomialRing(QQ,["x","y"])
(Multivariate Polynomial Ring in x, y over Rational Field, fmpq_mpoly[x, y])

julia> I = ideal([x*y^2-1,x^3+y^2+x*y])
ideal generated by: x*y^2 - 1, x^3 + x*y + y^2

julia> G,m = groebner_basis_with_transformation_matrix(I)
(fmpq_mpoly[x*y^2 - 1, x^3 + x*y + y^2, x^2 + y^4 + y], fmpq_mpoly[1 0; 0 1; -x^2 - y y^2])

julia> m * gens(I) == G
true

fglm

Gröbner walks

Hilbert-driven

Leading Ideals

Oscar.leading_ideal — Method

leading_ideal(g::Array{T, 1}, args...) where { T <: MPolyElem }

Given an array of multivariate polynomials this function returns the ideal generated by the leading monomials of the given array elements. If not otherwise given as a further argument this is done w.r.t. the degree reverse lexicographical monomial ordering.

Example

julia> R, (x, y) = PolynomialRing(QQ, ["x", "y"])
(Multivariate Polynomial Ring in x, y over Rational Field, fmpq_mpoly[x, y])

julia> L = leading_ideal([x*y^2-3*x, x^3-14*y^5])
ideal generated by: x*y^2, x^3

Oscar.leading_ideal — Method

leading_ideal(I::MPolyIdeal)

Given a multivariate polynomial ideal Ì this function returns the leading ideal for I. This is done w.r.t. the given monomial ordering in the polynomial ring of I.

Example

julia> R, (x, y) = PolynomialRing(QQ, ["x", "y"])
(Multivariate Polynomial Ring in x, y over Rational Field, fmpq_mpoly[x, y])

julia> I = ideal(R,[x*y^2-3*x, x^3-14*y^5])
ideal generated by: x*y^2 - 3*x, x^3 - 14*y^5

julia> L = leading_ideal(I)
ideal generated by: x*y^2, x^4, x^3

Gröbner Bases over the integers

....

Syzygies

Generators of syzygies

Oscar.syzygy_generators — Method

syzygy_generators(a::Array{<:MPolyElem, 1})

Given an array of multivariate polynomials this function returns the generators of the syzygy module of the ideal generated by the given array elements.

Example

julia> R, (x, y) = PolynomialRing(QQ, ["x", "y"])
(Multivariate Polynomial Ring in x, y over Rational Field, fmpq_mpoly[x, y])

julia> S = syzygy_generators([x^3+y+2,x*y^2-13*x^2,y-14])
3-element Array{AbstractAlgebra.Generic.FreeModuleElem{fmpq_mpoly},1}:
 (0, -y + 14, -13*x^2 + x*y^2)
 (-169*y + 2366, -13*x*y + 182*x - 196*y + 2744, 13*x^2*y^2 - 2548*x^2 + 196*x*y^2 + 169*y + 338)
 (-13*x^2 + 196*x, -x^3 - 16, x^4*y + 14*x^4 + 13*x^2 + 16*x*y + 28*x)

Ideals in Multivariate Polynomial Rings

Introduction

Distinction: ideals <–> homogeneous ideals. To be discussed.

Data Associated to Ideals

AbstractAlgebra.base_ring — Method

base_ring(I::MPolyIdeal{S}) where {S}

Returns the ambient ring of I.

Example

julia> R, (x, y) = PolynomialRing(QQ, ["x", "y"])
(Multivariate Polynomial Ring in x, y over Rational Field, fmpq_mpoly[x, y])

julia> I = ideal(R, [x, y])^2
ideal generated by: x^2, x*y, y^2

julia> base_ring(I)
Multivariate Polynomial Ring in x, y over Rational Field

Number of Generators

AbstractAlgebra.Generic.ngens — Method

ngens(I::MPolyIdeal)

Returns the number of generators of I.

Example

julia> R, (x, y) = PolynomialRing(QQ, ["x", "y"])
(Multivariate Polynomial Ring in x, y over Rational Field, fmpq_mpoly[x, y])

julia> I = ideal(R, [x, y])^2
ideal generated by: x^2, x*y, y^2

julia> ngens(I)
3

Generators

AbstractAlgebra.gens — Method

gens(I::MPolyIdeal)

Returns the generators of I as an array of multivariate polynomials.

Example

julia> R, (x, y) = PolynomialRing(QQ, ["x", "y"])
(Multivariate Polynomial Ring in x, y over Rational Field, fmpq_mpoly[x, y])

julia> I = ideal(R, [x, y])^2
ideal generated by: x^2, x*y, y^2

julia> gens(I)
3-element Array{fmpq_mpoly,1}:
 x^2
 x*y
 y^2

Dimension

AbstractAlgebra.Generic.dim — Method

dim(I::MPolyIdeal)

Returns the Krull dimension of I.

Example

julia> R, (x, y, z) = PolynomialRing(QQ, ["x", "y", "z"])
(Multivariate Polynomial Ring in x, y, z over Rational Field, fmpq_mpoly[x, y, z])

julia> I = ideal(R, [y-x^2, x-z^3])
ideal generated by: -x^2 + y, x - z^3

julia> dim(I)
1

Codimension

Oscar.codim — Method

codim(I::MPolyIdeal)

Returns the codimension of I.

Example

julia> R, (x, y, z) = PolynomialRing(QQ, ["x", "y", "z"])
(Multivariate Polynomial Ring in x, y, z over Rational Field, fmpq_mpoly[x, y, z])

julia> I = ideal(R, [y-x^2, x-z^3])
ideal generated by: -x^2 + y, x - z^3

julia> codim(I)
2

Data Associated to Homogeneous Ideals

                         min_base(I)   (or so)
 
degree:                  degree(I)                             (type integer)
Hilbert function:        hilbert_function(d,I)                 (type integer)
Hilbert series:          hilbert_series(I)                     (type univariate rational function)  numerator, denominator
reduced Hilbert series:  reduced_Hilbert_series(I)             (type univariate rational function)
Hilbert polynomial:      hilbert_polynomial(I)                 (type univariate polynomial, direkt in Julia von hilbert_series(I))

Ideal Operations for Multivariate Polynomial Rings

Simple ideal Operations

Powers of Ideal

Base.:^ — Method

:^(I::MPolyIdeal, m::Int)

Returns the m-th power of I.

Example

julia> R, (x, y, z) = PolynomialRing(QQ, ["x", "y", "z"])
(Multivariate Polynomial Ring in x, y, z over Rational Field, fmpq_mpoly[x, y, z])

julia> I = ideal(R, [x, y])
ideal generated by: x, y

julia> I^3
ideal generated by: x^3, x^2*y, x*y^2, y^3

Sum of Ideals

Base.:+ — Method

:+(I::MPolyIdeal, J::MPolyIdeal)

Returns the sum of I and J.

Example

julia> R, (x, y, z) = PolynomialRing(QQ, ["x", "y", "z"])
(Multivariate Polynomial Ring in x, y, z over Rational Field, fmpq_mpoly[x, y, z])

julia> I = ideal(R, [x, y])
ideal generated by: x, y

julia> J = ideal(R, [z^2])
ideal generated by: z^2

julia> I+J
ideal generated by: x, y, z^2

Product of Ideals

Base.:* — Method

:*(I::MPolyIdeal, J::MPolyIdeal)

Returns the product of I and J.

Example

julia> R, (x, y, z) = PolynomialRing(QQ, ["x", "y", "z"])
(Multivariate Polynomial Ring in x, y, z over Rational Field, fmpq_mpoly[x, y, z])

julia> I = ideal(R, [x, y])
ideal generated by: x, y

julia> J = ideal(R, [z^2])
ideal generated by: z^2

julia> I*J
ideal generated by: x*z^2, y*z^2

Intersection of Ideals

Base.intersect — Method

intersect(I::MPolyIdeal, Js::MPolyIdeal...)

Returns the intersection of two or more ideals.

Example

julia> R, (x, y) = PolynomialRing(QQ, ["x", "y"])
(Multivariate Polynomial Ring in x, y over Rational Field, fmpq_mpoly[x, y])

julia> I = intersect(ideal(R, [x, y])^2, ideal(R, [y^2-x^3+x]))
ideal generated by: x^3*y - x*y - y^3, x^4 - x^2 - x*y^2

Ideal Quotients

Given two ideals $I, J$ of a ring $R$, the ideal quotient of $I$ by $J$ is defined to be the ideal

\[I:J= \bigl\{f \in R\:\big|\: f J \subset I\bigr\}\subset R.\]

Hecke.quotient — Method

quotient(I::MPolyIdeal, J::MPolyIdeal)

Returns the ideal quotient of I by J. Alternatively, use I:J.

Example

julia> R, (x, y, z) = PolynomialRing(QQ, ["x", "y", "z"])
(Multivariate Polynomial Ring in x, y, z over Rational Field, fmpq_mpoly[x, y, z])

julia> I = ideal(R, [x^4+x^2*y*z+y^3*z, y^4+x^3*z+x*y^2*z, x^3*y+x*y^3])
ideal generated by: x^4 + x^2*y*z + y^3*z, x^3*z + x*y^2*z + y^4, x^3*y + x*y^3

julia> J = ideal(R,[x,y,z])^2
ideal generated by: x^2, x*y, x*z, y^2, y*z, z^2

julia> L = quotient(I,J)
ideal generated by: x^3*z + x*y^2*z + y^4, x^3*y + x*y^3, x^4 + x^2*y*z + y^3*z, x^3*z^2 - x^2*y*z^2 + x*y^2*z^2 - y^3*z^2, x^2*y^2*z - x^2*y*z^2 - y^3*z^2, x^3*z^2 + x^2*y^3 - x^2*y^2*z + x*y^2*z^2

julia> I:J
ideal generated by: x^3*z + x*y^2*z + y^4, x^3*y + x*y^3, x^4 + x^2*y*z + y^3*z, x^3*z^2 - x^2*y*z^2 + x*y^2*z^2 - y^3*z^2, x^2*y^2*z - x^2*y*z^2 - y^3*z^2, x^3*z^2 + x^2*y^3 - x^2*y^2*z + x*y^2*z^2

Saturation

Given two ideals $I, J$ of a ring $R$, the saturation of $I$ with respect to $J$ is defined to be the ideal

\[I:J^{\infty} = \bigl\{ f \in R \:\big|\: f J^m \!\subset I {\text{ for some }}m\geq 1 \bigr\} = \textstyle{\bigcup\limits_{m=1}^{\infty} (I:J^m)}.\]

Oscar.saturation — Method

saturation(I::MPolyIdeal, J::MPolyIdeal)

Returns the saturation of I with respect to J.

Example

julia> R, (x, y, z) = PolynomialRing(QQ, ["x", "y", "z"])
(Multivariate Polynomial Ring in x, y, z over Rational Field, fmpq_mpoly[x, y, z])

julia> I = ideal(R, [x^4+x^2*y*z+y^3*z, y^4+x^3*z+x*y^2*z, x^3*y+x*y^3])
ideal generated by: x^4 + x^2*y*z + y^3*z, x^3*z + x*y^2*z + y^4, x^3*y + x*y^3

julia> J = ideal(R, [x,y,z])^2
ideal generated by: x^2, x*y, x*z, y^2, y*z, z^2

julia> L = saturation(I,J)
ideal generated by: x^3 - x^2*y + x*y^2 - y^3, x^2*y*z + y^4 + y^3*z, x^2*y^2 + y^4

Elimination

Oscar.eliminate — Method

eliminate(I::MPolyIdeal, polys::Array{MPolyElem, 1})

Given a list of polynomials which are variables, these variables are eliminated from I. That is, the function returns the ideal of all polynomials in I which only depend on the remaining variables.

Example

julia> R, (t, x, y, z) = PolynomialRing(QQ, ["t", "x", "y", "z"])
(Multivariate Polynomial Ring in t, x, y, z over Rational Field, fmpq_mpoly[t, x, y, z])

julia> I = ideal(R, [t-x, t^2-y, t-z^3])
ideal generated by: t - x, t^2 - y, t - z^3

julia> A = [t]
1-element Array{fmpq_mpoly,1}:
 t

julia> TC = eliminate(I,A)
ideal generated by: x^2 - y, -x + z^3

Oscar.eliminate — Method

eliminate(I::MPolyIdeal, polys::AbstractArray{Int, 1})

Given a list of indices which specify variables, these variables are eliminated from I. That is, the function returns the ideal of all polynomials in I which only depend on the remaining variables.

Example

julia> R, (t, x, y, z) = PolynomialRing(QQ, ["t", "x", "y", "z"])
(Multivariate Polynomial Ring in t, x, y, z over Rational Field, fmpq_mpoly[t, x, y, z])

julia> I = ideal(R, [t-x, t^2-y, t-z^3])
ideal generated by: t - x, t^2 - y, t - z^3

julia> A = [1]
1-element Array{Int64,1}:
 1

julia> TC = eliminate(I,A)
ideal generated by: x^2 - y, -x + z^3

Homogenization and Dehomogenization

homogenize(I,t)   CAVEAT: Als Ideal! Auch poly, vector, etc. Siehe M2.

dehomogenize(I,t)

Tests on Ideals

Equality of Ideals

Base.:== — Method

:(==)(I::MPolyIdeal, J::MPolyIdeal)

Returns true if I=J, false otherwise.

Example

ulia> R, (x, y) = PolynomialRing(QQ, ["x", "y"])
(Multivariate Polynomial Ring in x, y over Rational Field, fmpq_mpoly[x, y])

julia> I = ideal(R, [x^2])
ideal generated by: x^2

julia> J = ideal(R, [x, y])^2
ideal generated by: x^2, x*y, y^2

julia> I == J
false

Containment of Ideals

Base.issubset — Method

issubset(I::MPolyIdeal, J::MPolyIdeal)

Returns true if I is contained in J, false otherwise.

Example

julia> R, (x, y) = PolynomialRing(QQ, ["x", "y"])
(Multivariate Polynomial Ring in x, y over Rational Field, fmpq_mpoly[x, y])

julia> I = ideal(R, [x^2])
ideal generated by: x^2

julia> J = ideal(R, [x, y])^2
ideal generated by: x^2, x*y, y^2

julia> issubset(I, J)
true

Ideal Membership

Oscar.ideal_membership — Method

ideal_membership(f::MPolyElem, I::MPolyIdeal)

Returns true if f is contained in I, false otherwise. Alternatively, use f in I.

Example

julia> R, (x, y) = PolynomialRing(QQ, ["x", "y"])
(Multivariate Polynomial Ring in x, y over Rational Field, fmpq_mpoly[x, y])

julia> f = x^2
x^2

julia> I = ideal(R, [x, y])^2
ideal generated by: x^2, x*y, y^2

julia> ideal_membership(f, I)
true

julia> g = x
x

julia> g in I
false

Radical Membership

Oscar.radical_membership — Method

radical_membership(f::MPolyElem, I::MPolyIdeal)

Returns true if f is contained in the radical of I, false otherwise.

Example

julia> R, (x,) = PolynomialRing(QQ, ["x"])
(Multivariate Polynomial Ring in x over Rational Field, fmpq_mpoly[x])

julia> f = x
x

julia> I = ideal(R,  [x^2])
ideal generated by: x^2

julia> radical_membership(f, I)
true

Primality Test

Nemo.isprime — Method

isprime(I::MPolyIdeal)

Returns true if the ideal I is prime, false otherwise. Proceeds by computing a primary decomposition and may take some time.

Example

julia> R, (x, y) = PolynomialRing(QQ, ["x", "y"])
(Multivariate Polynomial Ring in x, y over Rational Field, fmpq_mpoly[x, y])

julia> I = ideal(R, [x, y])^2
ideal generated by: x^2, x*y, y^2

julia> isprime(I)
false

Primary Test

Oscar.isprimary — Method

isprimary(I::MPolyIdeal)

Return true if the ideal I is primary, false otherwise. Proceeds by computing a primary decomposition and may take some time.

Example

julia> R, (x, y) = PolynomialRing(QQ, ["x", "y"])
(Multivariate Polynomial Ring in x, y over Rational Field, fmpq_mpoly[x, y])

julia> I = ideal(R, [x, y])^2
ideal generated by: x^2, x*y, y^2

julia> isprimary(I)
true

homogeneity test:        ishomogeneous(I)

.....

Decomposition

Radical

Hecke.radical — Method

radical(I::MPolyIdeal)

Returns the radical of I. If the base ring of I is a polynomial ring over a field, a combination of the algorithms of Krick and Logar (with modifications by Laplagne) and Kemper is used. For polynomial rings over the integers, the algorithm proceeds as suggested by Pfister, Sadiq, and Steidel.

Examples

julia> R, (x, y) = PolynomialRing(QQ, ["x", "y"])
(Multivariate Polynomial Ring in x, y over Rational Field, fmpq_mpoly[x, y])

julia> I = intersect(ideal(R, [x, y])^2, ideal(R, [y^2-x^3+x]))
ideal generated by: x^3*y - x*y - y^3, x^4 - x^2 - x*y^2

julia> I = intersect(I, ideal(R, [x-y-1])^2)
ideal generated by: x^5*y - 2*x^4*y^2 - 2*x^4*y + x^3*y^3 + 2*x^3*y^2 - x^2*y^3 + 2*x^2*y^2 + 2*x^2*y + 2*x*y^4 + x*y^3 - 2*x*y^2 - x*y - y^5 - 2*y^4 - y^3, x^6 - 2*x^5 - 3*x^4*y^2 - 2*x^4*y + 2*x^3*y^3 + 3*x^3*y^2 + 2*x^3*y + 2*x^3 + 5*x^2*y^2 + 2*x^2*y - x^2 + 3*x*y^4 - 5*x*y^2 - 2*x*y - 2*y^5 - 4*y^4 - 2*y^3

julia> RI = radical(I)
ideal generated by: x^4 - x^3*y - x^3 - x^2 - x*y^2 + x*y + x + y^3 + y^2

julia> R, (a, b, c, d) = PolynomialRing(ZZ, ["a", "b", "c", "d"])
(Multivariate Polynomial Ring in a, b, c, d over Integer Ring, fmpz_mpoly[a, b, c, d])

julia> I = intersect(ideal(R, [9,a,b]), ideal(R, [3,c]))
ideal generated by: 9, 3*b, 3*a, b*c, a*c

julia> I = intersect(I, ideal(R, [11,2a,7b]))
ideal generated by: 99, 3*b, 3*a, b*c, a*c

julia> I = intersect(I, ideal(R, [13a^2,17b^4]))
ideal generated by: 39*a^2, 13*a^2*c, 51*b^4, 17*b^4*c, 3*a^2*b^4, a^2*b^4*c

julia> I = intersect(I, ideal(R, [9c^5,6d^5]))
ideal generated by: 78*a^2*d^5, 117*a^2*c^5, 102*b^4*d^5, 153*b^4*c^5, 6*a^2*b^4*d^5, 9*a^2*b^4*c^5, 39*a^2*c^5*d^5, 51*b^4*c^5*d^5, 3*a^2*b^4*c^5*d^5

julia> I = intersect(I, ideal(R, [17,a^15,b^15,c^15,d^15]))
ideal generated by: 1326*a^2*d^5, 1989*a^2*c^5, 102*b^4*d^5, 153*b^4*c^5, 663*a^2*c^5*d^5, 51*b^4*c^5*d^5, 78*a^2*d^15, 117*a^2*c^15, 78*a^15*d^5, 117*a^15*c^5, 6*a^2*b^4*d^15, 9*a^2*b^4*c^15, 39*a^2*c^5*d^15, 39*a^2*c^15*d^5, 6*a^2*b^15*d^5, 9*a^2*b^15*c^5, 6*a^15*b^4*d^5, 9*a^15*b^4*c^5, 39*a^15*c^5*d^5, 3*a^2*b^4*c^5*d^15, 3*a^2*b^4*c^15*d^5, 3*a^2*b^15*c^5*d^5, 3*a^15*b^4*c^5*d^5

julia> RI = radical(I)
ideal generated by: 102*b*d, 78*a*d, 51*b*c, 39*a*c, 6*a*b*d, 3*a*b*c

Primary Decomposition

Oscar.primary_decomposition — Method

primary_decomposition(I::MPolyIdeal)

Returns a primary decomposition of the ideal I. If I is the unit ideal, [ideal(1)] is returned. If the base ring of I is a polynomial ring over a field, the algorithm of Gianni, Trager and Zacharias is used by default. Alternatively, the algorithm by Shimoyama and Yokoyama can be used by specifying alg=:SY. For polynomial rings over the integers, the algorithm proceeds as suggested by Pfister, Sadiq, and Steidel.

Examples

julia> R, (x, y) = PolynomialRing(QQ, ["x", "y"])
(Multivariate Polynomial Ring in x, y over Rational Field, fmpq_mpoly[x, y])

julia> I = intersect(ideal(R, [x, y])^2, ideal(R, [y^2-x^3+x]))
ideal generated by: x^3*y - x*y - y^3, x^4 - x^2 - x*y^2

julia> I = intersect(I, ideal(R, [x-y-1])^2)
ideal generated by: x^5*y - 2*x^4*y^2 - 2*x^4*y + x^3*y^3 + 2*x^3*y^2 - x^2*y^3 + 2*x^2*y^2 + 2*x^2*y + 2*x*y^4 + x*y^3 - 2*x*y^2 - x*y - y^5 - 2*y^4 - y^3, x^6 - 2*x^5 - 3*x^4*y^2 - 2*x^4*y + 2*x^3*y^3 + 3*x^3*y^2 + 2*x^3*y + 2*x^3 + 5*x^2*y^2 + 2*x^2*y - x^2 + 3*x*y^4 - 5*x*y^2 - 2*x*y - 2*y^5 - 4*y^4 - 2*y^3

julia> L = primary_decomposition(I)
3-element Array{Tuple{MPolyIdeal{fmpq_mpoly},MPolyIdeal{fmpq_mpoly}},1}:
 (ideal generated by: x^3 - x - y^2, ideal generated by: x^3 - x - y^2)
 (ideal generated by: x^2 - 2*x*y - 2*x + y^2 + 2*y + 1, ideal generated by: x - y - 1)
 (ideal generated by: y, x^2, ideal generated by: x, y)

julia> L = primary_decomposition(I, alg=:SY)
3-element Array{Tuple{MPolyIdeal{fmpq_mpoly},MPolyIdeal{fmpq_mpoly}},1}:
 (ideal generated by: x^3 - x - y^2, ideal generated by: x^3 - x - y^2)
 (ideal generated by: x^2 - 2*x*y - 2*x + y^2 + 2*y + 1, ideal generated by: x - y - 1)
 (ideal generated by: y, x^2, ideal generated by: y, x)

julia> R, (a, b, c, d) = PolynomialRing(ZZ, ["a", "b", "c", "d"])
(Multivariate Polynomial Ring in a, b, c, d over Integer Ring, fmpz_mpoly[a, b, c, d])

julia> I = ideal(R, [1326*a^2*d^5, 1989*a^2*c^5, 102*b^4*d^5, 153*b^4*c^5,
       663*a^2*c^5*d^5, 51*b^4*c^5*d^5, 78*a^2*d^15, 117*a^2*c^15,
       78*a^15*d^5, 117*a^15*c^5, 6*a^2*b^4*d^15, 9*a^2*b^4*c^15,
       39*a^2*c^5*d^15, 39*a^2*c^15*d^5, 6*a^2*b^15*d^5, 9*a^2*b^15*c^5,
       6*a^15*b^4*d^5, 9*a^15*b^4*c^5, 39*a^15*c^5*d^5, 3*a^2*b^4*c^5*d^15,
       3*a^2*b^4*c^15*d^5, 3*a^2*b^15*c^5*d^5, 3*a^15*b^4*c^5*d^5])
ideal generated by: 1326*a^2*d^5, 1989*a^2*c^5, 102*b^4*d^5, 153*b^4*c^5, 663*a^2*c^5*d^5, 51*b^4*c^5*d^5, 78*a^2*d^15, 117*a^2*c^15, 78*a^15*d^5, 117*a^15*c^5, 6*a^2*b^4*d^15, 9*a^2*b^4*c^15, 39*a^2*c^5*d^15, 39*a^2*c^15*d^5, 6*a^2*b^15*d^5, 9*a^2*b^15*c^5, 6*a^15*b^4*d^5, 9*a^15*b^4*c^5, 39*a^15*c^5*d^5, 3*a^2*b^4*c^5*d^15, 3*a^2*b^4*c^15*d^5, 3*a^2*b^15*c^5*d^5, 3*a^15*b^4*c^5*d^5

julia> L = primary_decomposition(I)
8-element Array{Tuple{MPolyIdeal{fmpz_mpoly},MPolyIdeal{fmpz_mpoly}},1}:
 (ideal generated by: d^5, c^5, ideal generated by: d, c)
 (ideal generated by: a^2, b^4, ideal generated by: b, a)
 (ideal generated by: 2, c^5, ideal generated by: 2, c)
 (ideal generated by: 3, ideal generated by: 3)
 (ideal generated by: 13, b^4, ideal generated by: 13, b)
 (ideal generated by: 17, a^2, ideal generated by: 17, a)
 (ideal generated by: 17, d^15, c^15, b^15, a^15, ideal generated by: 17, d, c, b, a)
 (ideal generated by: 9, 3*d^5, d^10, ideal generated by: 3, d)

Minimal Associated Primes

Oscar.minimal_primes — Method

minimal_primes(I::MPolyIdeal; alg=:GTZ)

Returns an array of the minimal associated prime ideals of I. If I is the unit ideal, [ideal(1)] is returned. If the base ring of I is a polynomial ring over a field, the algorithm of Gianni-Trager-Zacharias is used by default and characteristic sets may be used by specifying alg=:charSets. For polynomial rings over the integers, the algorithm proceeds as suggested by Pfister, Sadiq, and Steidel.

Examples

julia> R, (x, y) = PolynomialRing(QQ, ["x", "y"])
(Multivariate Polynomial Ring in x, y over Rational Field, fmpq_mpoly[x, y])

julia> I = intersect(ideal(R, [x, y])^2, ideal(R, [y^2-x^3+x]))
ideal generated by: x^3*y - x*y - y^3, x^4 - x^2 - x*y^2

julia> I = intersect(I, ideal(R, [x-y-1])^2)
ideal generated by: x^5*y - 2*x^4*y^2 - 2*x^4*y + x^3*y^3 + 2*x^3*y^2 - x^2*y^3 + 2*x^2*y^2 + 2*x^2*y + 2*x*y^4 + x*y^3 - 2*x*y^2 - x*y - y^5 - 2*y^4 - y^3, x^6 - 2*x^5 - 3*x^4*y^2 - 2*x^4*y + 2*x^3*y^3 + 3*x^3*y^2 + 2*x^3*y + 2*x^3 + 5*x^2*y^2 + 2*x^2*y - x^2 + 3*x*y^4 - 5*x*y^2 - 2*x*y - 2*y^5 - 4*y^4 - 2*y^3

julia> L = minimal_primes(I)
2-element Array{MPolyIdeal{fmpq_mpoly},1}:
 ideal generated by: x - y - 1
 ideal generated by: x^3 - x - y^2

julia> L = minimal_primes(I, alg=:charSets)
2-element Array{MPolyIdeal{fmpq_mpoly},1}:
 ideal generated by: x - y - 1
 ideal generated by: x^3 - x - y^2

julia> R, (a, b, c, d) = PolynomialRing(ZZ, ["a", "b", "c", "d"])
(Multivariate Polynomial Ring in a, b, c, d over Integer Ring, fmpz_mpoly[a, b, c, d])

julia> I = ideal(R, [1326*a^2*d^5, 1989*a^2*c^5, 102*b^4*d^5, 153*b^4*c^5,
       663*a^2*c^5*d^5, 51*b^4*c^5*d^5, 78*a^2*d^15, 117*a^2*c^15,
       78*a^15*d^5, 117*a^15*c^5, 6*a^2*b^4*d^15, 9*a^2*b^4*c^15,
       39*a^2*c^5*d^15, 39*a^2*c^15*d^5, 6*a^2*b^15*d^5, 9*a^2*b^15*c^5,
       6*a^15*b^4*d^5, 9*a^15*b^4*c^5, 39*a^15*c^5*d^5, 3*a^2*b^4*c^5*d^15,
       3*a^2*b^4*c^15*d^5, 3*a^2*b^15*c^5*d^5, 3*a^15*b^4*c^5*d^5])
ideal generated by: 1326*a^2*d^5, 1989*a^2*c^5, 102*b^4*d^5, 153*b^4*c^5, 663*a^2*c^5*d^5, 51*b^4*c^5*d^5, 78*a^2*d^15, 117*a^2*c^15, 78*a^15*d^5, 117*a^15*c^5, 6*a^2*b^4*d^15, 9*a^2*b^4*c^15, 39*a^2*c^5*d^15, 39*a^2*c^15*d^5, 6*a^2*b^15*d^5, 9*a^2*b^15*c^5, 6*a^15*b^4*d^5, 9*a^15*b^4*c^5, 39*a^15*c^5*d^5, 3*a^2*b^4*c^5*d^15, 3*a^2*b^4*c^15*d^5, 3*a^2*b^15*c^5*d^5, 3*a^15*b^4*c^5*d^5

julia> L = minimal_primes(I)
6-element Array{MPolyIdeal{fmpz_mpoly},1}:
 ideal generated by: d, c
 ideal generated by: b, a
 ideal generated by: 2, c
 ideal generated by: 3
 ideal generated by: 13, b
 ideal generated by: 17, a

Weak Equidimensional Decomposition

Oscar.equidimensional_decomposition_weak — Method

equidimensional_decomposition_weak(I::MPolyIdeal)

Returns an array of equidimensional ideals where the last element is the equidimensional hull of I, that is, the intersection of the primary components of I of maximal dimension. Each of the previous elements is an ideal of lower dimension whose associated primes are exactly the associated primes of I of that dimension. If I is the unit ideal, [ideal(1)] is returned. Uses ideas of Eisenbud, Huneke, and Vasconcelos.

Example

julia> R, (x, y) = PolynomialRing(QQ, ["x", "y"])
(Multivariate Polynomial Ring in x, y over Rational Field, fmpq_mpoly[x, y])

julia> I = intersect(ideal(R, [x, y])^2, ideal(R, [y^2-x^3+x]))
ideal generated by: x^3*y - x*y - y^3, x^4 - x^2 - x*y^2

julia> I = intersect(I, ideal(R, [x-y-1])^2)
ideal generated by: x^5*y - 2*x^4*y^2 - 2*x^4*y + x^3*y^3 + 2*x^3*y^2 - x^2*y^3 + 2*x^2*y^2 + 2*x^2*y + 2*x*y^4 + x*y^3 - 2*x*y^2 - x*y - y^5 - 2*y^4 - y^3, x^6 - 2*x^5 - 3*x^4*y^2 - 2*x^4*y + 2*x^3*y^3 + 3*x^3*y^2 + 2*x^3*y + 2*x^3 + 5*x^2*y^2 + 2*x^2*y - x^2 + 3*x*y^4 - 5*x*y^2 - 2*x*y - 2*y^5 - 4*y^4 - 2*y^3

julia> L = equidimensional_decomposition_weak(I)
2-element Array{MPolyIdeal{fmpq_mpoly},1}:
 ideal generated by: y, x
 ideal generated by: x^5 - 2*x^4*y - 2*x^4 + x^3*y^2 + 2*x^3*y - x^2*y^2 + 2*x^2*y + 2*x^2 + 2*x*y^3 + x*y^2 - 2*x*y - x - y^4 - 2*y^3 - y^

Equidimensional Decomposition of radical

Oscar.equidimensional_decomposition_radical — Method

equidimensional_decomposition_radical(I::MPolyIdeal)

Returns an array of equidimensional radical ideals increasingly ordered by dimension. For each dimension, the returned radical ideal is the intersection of the associated primes of I of that dimension. If I is the unit ideal, [ideal(1)] is returned. Uses a combination of the algorithms of Krick and Logar (with modifications by Laplagne) and Kemper.

Example

julia> R, (x, y) = PolynomialRing(QQ, ["x", "y"])
(Multivariate Polynomial Ring in x, y over Rational Field, fmpq_mpoly[x, y])

julia> I = intersect(ideal(R, [x, y])^2, ideal(R, [y^2-x^3+x]))
ideal generated by: x^3*y - x*y - y^3, x^4 - x^2 - x*y^2

julia> I = intersect(I, ideal(R, [x-y-1])^2)
ideal generated by: x^5*y - 2*x^4*y^2 - 2*x^4*y + x^3*y^3 + 2*x^3*y^2 - x^2*y^3 + 2*x^2*y^2 + 2*x^2*y + 2*x*y^4 + x*y^3 - 2*x*y^2 - x*y - y^5 - 2*y^4 - y^3, x^6 - 2*x^5 - 3*x^4*y^2 - 2*x^4*y + 2*x^3*y^3 + 3*x^3*y^2 + 2*x^3*y + 2*x^3 + 5*x^2*y^2 + 2*x^2*y - x^2 + 3*x*y^4 - 5*x*y^2 - 2*x*y - 2*y^5 - 4*y^4 - 2*y^3

julia> L = equidimensional_decomposition_radical(I)
2-element Array{MPolyIdeal{fmpq_mpoly},1}:
 ideal generated by: y, x
 ideal generated by: x^4 - x^3*y - x^3 - x^2 - x*y^2 + x*y + x + y^3 + y^2

Equidimensional Hull

Oscar.equidimensional_hull — Method

equidimensional_hull(I::MPolyIdeal)

If the base ring of I is a polynomial ring over a field, return the intersection of the primary components of I of maximal dimension. In the case of polynomials over the integers, return the intersection of the primary components of I of minimal height. If I is the unit ideal, [ideal(1)] is returned. For polynomial rings over a field, the algorithm relies on ideas as used by Gianni, Trager, and Zacharias or Krick and Logar. For polynomial rings over the integers, the algorithm proceeds as suggested by Pfister, Sadiq, and Steidel.

Examples

julia> R, (x, y) = PolynomialRing(QQ, ["x", "y"])
(Multivariate Polynomial Ring in x, y over Rational Field, fmpq_mpoly[x, y])

julia> I = intersect(ideal(R, [x, y])^2, ideal(R, [y^2-x^3+x]))
ideal generated by: x^3*y - x*y - y^3, x^4 - x^2 - x*y^2

julia> I = intersect(I, ideal(R, [x-y-1])^2)
ideal generated by: x^5*y - 2*x^4*y^2 - 2*x^4*y + x^3*y^3 + 2*x^3*y^2 - x^2*y^3 + 2*x^2*y^2 + 2*x^2*y + 2*x*y^4 + x*y^3 - 2*x*y^2 - x*y - y^5 - 2*y^4 - y^3, x^6 - 2*x^5 - 3*x^4*y^2 - 2*x^4*y + 2*x^3*y^3 + 3*x^3*y^2 + 2*x^3*y + 2*x^3 + 5*x^2*y^2 + 2*x^2*y - x^2 + 3*x*y^4 - 5*x*y^2 - 2*x*y - 2*y^5 - 4*y^4 - 2*y^3

julia> L = equidimensional_hull(I)
ideal generated by: x^5 - 2*x^4*y - 2*x^4 + x^3*y^2 + 2*x^3*y - x^2*y^2 + 2*x^2*y + 2*x^2 + 2*x*y^3 + x*y^2 - 2*x*y - x - y^4 - 2*y^3 - y^2

julia> R, (a, b, c, d) = PolynomialRing(ZZ, ["a", "b", "c", "d"])
(Multivariate Polynomial Ring in a, b, c, d over Integer Ring, fmpz_mpoly[a, b, c, d])

julia> I = ideal(R, [1326*a^2*d^5, 1989*a^2*c^5, 102*b^4*d^5, 153*b^4*c^5,
       663*a^2*c^5*d^5, 51*b^4*c^5*d^5, 78*a^2*d^15, 117*a^2*c^15,
       78*a^15*d^5, 117*a^15*c^5, 6*a^2*b^4*d^15, 9*a^2*b^4*c^15,
       39*a^2*c^5*d^15, 39*a^2*c^15*d^5, 6*a^2*b^15*d^5, 9*a^2*b^15*c^5,
       6*a^15*b^4*d^5, 9*a^15*b^4*c^5, 39*a^15*c^5*d^5, 3*a^2*b^4*c^5*d^15,
       3*a^2*b^4*c^15*d^5, 3*a^2*b^15*c^5*d^5, 3*a^15*b^4*c^5*d^5])
ideal generated by: 1326*a^2*d^5, 1989*a^2*c^5, 102*b^4*d^5, 153*b^4*c^5, 663*a^2*c^5*d^5, 51*b^4*c^5*d^5, 78*a^2*d^15, 117*a^2*c^15, 78*a^15*d^5, 117*a^15*c^5, 6*a^2*b^4*d^15, 9*a^2*b^4*c^15, 39*a^2*c^5*d^15, 39*a^2*c^15*d^5, 6*a^2*b^15*d^5, 9*a^2*b^15*c^5, 6*a^15*b^4*d^5, 9*a^15*b^4*c^5, 39*a^15*c^5*d^5, 3*a^2*b^4*c^5*d^15, 3*a^2*b^4*c^15*d^5, 3*a^2*b^15*c^5*d^5, 3*a^15*b^4*c^5*d^5

julia> L = equidimensional_hull(I)
ideal generated by: 3

Radical of the Equidimensional Hull

Oscar.equidimensional_hull_radical — Method

equidimensional_hull_radical(I::MPolyIdeal)

Returns the intersection of the associated primes of I of maximal dimension. If I is the unit ideal, [ideal(1)] is returned. Uses a combination of the algorithms of Krick and Logar (with modifications by Laplagne) and Kemper.

Example

julia> R, (x, y) = PolynomialRing(QQ, ["x", "y"])
(Multivariate Polynomial Ring in x, y over Rational Field, fmpq_mpoly[x, y])

julia> I = intersect(ideal(R, [x, y])^2, ideal(R, [y^2-x^3+x]))
ideal generated by: x^3*y - x*y - y^3, x^4 - x^2 - x*y^2

julia> I = intersect(I, ideal(R, [x-y-1])^2)
ideal generated by: x^5*y - 2*x^4*y^2 - 2*x^4*y + x^3*y^3 + 2*x^3*y^2 - x^2*y^3 + 2*x^2*y^2 + 2*x^2*y + 2*x*y^4 + x*y^3 - 2*x*y^2 - x*y - y^5 - 2*y^4 - y^3, x^6 - 2*x^5 - 3*x^4*y^2 - 2*x^4*y + 2*x^3*y^3 + 3*x^3*y^2 + 2*x^3*y + 2*x^3 + 5*x^2*y^2 + 2*x^2*y - x^2 + 3*x*y^4 - 5*x*y^2 - 2*x*y - 2*y^5 - 4*y^4 - 2*y^3

julia> L = equidimensional_hull_radical(I)
ideal generated by: x^4 - x^3*y - x^3 - x^2 - x*y^2 + x*y + x + y^3 + y^2

Absolute Primary Decomposition

absolute_primary_decomposition(I)               --->  absPrimdecGTZ

Modules over Multivariate Polynomial Rings

Module Constructors

Simple Module Operations

Sum of Submodules

N_1+N_2

Products of Ideal and Submodule

I*N

Flatness

Depth and Codimension

Affine Algebras

An affine algebra over a field $K$, or an affine $K$-algebra, or simply an affine ring, is the quotient $A=R/I$ of a multivariate polynomial ring $R$ over $K$ modulo an ideal $I$ of $A$.

Constructor

AbstractAlgebra.quo — Method

quo(R::MPolyRing, I::MPolyIdeal) -> MPolyQuoRing, Map

Creates the affine ring $R/I$ and returns the new ring as well as the projection map $R\rightarrow R/I$.

Example

julia> R, (x, y) = PolynomialRing(QQ, ["x", "y"])
(Multivariate Polynomial Ring in x, y over Rational Field, fmpq_mpoly[x, y])

julia> A, _ = quo(R, ideal(R, [x^2-y^3, x-y]))
(Quotient of Multivariate Polynomial Ring in x, y over Rational Field by ideal generated by: x^2 - y^3, x - y, Map from
Multivariate Polynomial Ring in x, y over Rational Field to Quotient of Multivariate Polynomial Ring in x, y over Rational Field by ideal generated by: x^2 - y^3, x - y defined by a julia-function with inverse
)

julia> A, f = quo(R, ideal(R, [x^2-y^3, x-y]))
(Quotient of Multivariate Polynomial Ring in x, y over Rational Field by ideal generated by: x^2 - y^3, x - y, Map from
Multivariate Polynomial Ring in x, y over Rational Field to Quotient of Multivariate Polynomial Ring in x, y over Rational Field by ideal generated by: x^2 - y^3, x - y defined by a julia-function with inverse
)

julia> f
Map from
Multivariate Polynomial Ring in x, y over Rational Field to Quotient of Multivariate Polynomial Ring in x, y over Rational Field by ideal generated by: x^2 - y^3, x - y defined by a julia-function with inverse

Data Associated to Affine Algebras

Basic Data

If A = R/I is an affine algebra, then

base_ring(A) refers to R,

modulus(A) to I,

gens(A) to the generators of A, and

ngens(A) to the number of these generators.

julia> R, (x, y, z) = PolynomialRing(QQ, ["x", "y", "z"])
(Multivariate Polynomial Ring in x, y, z over Rational Field, fmpq_mpoly[x, y, z])

julia> A, _ = quo(R, ideal(R, [y-x^2, x-z^3]))
(Quotient of Multivariate Polynomial Ring in x, y, z over Rational Field by ideal generated by: -x^2 + y, x - z^3, Map from
Multivariate Polynomial Ring in x, y, z over Rational Field to Quotient of Multivariate Polynomial Ring in x, y, z over Rational Field by ideal generated by: -x^2 + y, x - z^3 defined by a julia-function with inverse
)

julia> base_ring(A)
Multivariate Polynomial Ring in x, y, z over Rational Field

julia> modulus(A)
ideal generated by: -x^2 + y, x - z^3

julia> gens(A)
3-element Array{Oscar.MPolyQuoElem{fmpq_mpoly},1}:
 x
 y
 z

julia> ngens(A)
3

Dimension

AbstractAlgebra.Generic.dim — Method

dim(A::MPolyQuo)

Returns the dimension of A.

Example

julia> R, (x, y, z) = PolynomialRing(QQ, ["x", "y", "z"])
(Multivariate Polynomial Ring in x, y, z over Rational Field, fmpq_mpoly[x, y, z])

julia> A, _ = quo(R, ideal(R, [y-x^2, x-z^3]))
(Quotient of Multivariate Polynomial Ring in x, y, z over Rational Field by ideal generated by: -x^2 + y, x - z^3, Map from
Multivariate Polynomial Ring in x, y, z over Rational Field to Quotient of Multivariate Polynomial Ring in x, y, z over Rational Field by ideal generated by: -x^2 + y, x - z^3 defined by a julia-function with inverse
)

julia> dim(A)
1

Data Associated to Graded Affine Algebras: Hilbert Series and Hilbert Polynomial

Let $R=K[x_1, \dots x_n]$ be a polynomial ring in $n$ variables over a field $K$. Assign positive integer weights $w_i$ to the variables $x_i$, and grade $R=\bigoplus_{d\geq 0} R_d$ according to the corresponding weighted degree. Let $I$ be an ideal of $R$ which is homogeneous with respect to this grading. Then the affine $K$-algebra $A=R/I$ inherits the grading: $A = \bigoplus_{d\geq 0} A_d$, where each graded piece $A_d$ is a finite dimensional $K$-vector space. In this situation, the Hilbert function of $A$ is the function

\[H(A, \underline{\phantom{d}}): \N \rightarrow \N, d \mapsto \dim_K(d).\]

The Hilbert series of $A$ is the generating function

\[H_A(t)=\sum_{d\geq 0} H(A, d) t^d.\]

It can be written as a rational function in $t$ with denominator

\[(1-t^{w_1})\cdots (1-t^{w_n}).\]

Now suppose that the weights on the variables are all 1. Then we also have the Hilbert polynomial $P_A(t)\in\mathbb{Q}[t]$ which satisfies $H(M,d)=P_M(d)$ for all $d \gg 0$. Furthermore, the degree of $A$ is defined as the dimension of $A$ over $K$ if this dimension is finite, and as the integer $d$ such that the leading term of the Hilbert polynomial has the form $d t^e/e!$, otherwise.

CAVEAT: Currently only implemented in the case where the weights on the variables are all 1.

Oscar.hilbert_series — Method

hilbert_series(A::MPolyQuo)

Oscar.hilbert_series_reduced — Method

hilbert_series_reduced(A::MPolyQuo)

Oscar.hilbert_polynomial — Method

hilbert_polynomial(A::MPolyQuo)

AbstractAlgebra.Generic.degree — Method

degree(A::MPolyQuo)

Example

julia> R, (w, x, y, z) = PolynomialRing(QQ, ["w", "x", "y", "z"])
(Multivariate Polynomial Ring in w, x, y, z over Rational Field, fmpq_mpoly[w, x, y, z])

julia> A, _ = quo(R, ideal(R, [w*y-x^2, w*z-x*y, x*z-y^2]))
(Quotient of Multivariate Polynomial Ring in w, x, y, z over Rational Field by ideal generated by: w*y - x^2, w*z - x*y, x*z - y^2, Map from
Multivariate Polynomial Ring in w, x, y, z over Rational Field to Quotient of Multivariate Polynomial Ring in w, x, y, z over Rational Field by ideal generated by: w*y - x^2, w*z - x*y, x*z - y^2 defined by a julia-function with inverse
)

julia> hilbert_series(A)
(2*x^3 - 3*x^2 + 1, x^4 - 4*x^3 + 6*x^2 - 4*x + 1)

julia> hilbert_series_reduced(A)
(2*x + 1, x^2 - 2*x + 1)

julia> hilbert_polynomial(A)
3*x + 1

julia> degree(A)
3

Ideals in Affine Algebras

Tests on Affine Algebras

Subalgebra Membership

Oscar.subalgebra_membership — Method

subalgebra_membership(A::MPolyQuo, f::MPolyQuoElem, v::Vector{T}) where T <: MPolyQuoElem

Returns true if f is contained in the subalgebra of A generated by the elements contained in v, false otherwise.

Reducedness Test

Hecke.isreduced — Method

isreduced(A::MPolyQuo)

Returns true if A is reduced, false otherwise.

Example

julia> R, (x,) = PolynomialRing(QQ, ["x"])
(Multivariate Polynomial Ring in x over Rational Field, fmpq_mpoly[x])

julia> A, _ = quo(R, ideal(R, [x^4]))
(Quotient of Multivariate Polynomial Ring in x over Rational Field by ideal generated by: x^4, Map from
Multivariate Polynomial Ring in x over Rational Field to Quotient of Multivariate Polynomial Ring in x over Rational Field by ideal generated by: x^4 defined by a julia-function with inverse
)

julia> isreduced(A)
// ** loaded /home/runner/.julia/artifacts/8fca43e8f7685095940dea5093e165f5298f7f84/bin/../share/singular/LIB/primdec.lib (4.2.0.0,Mar_2021)
// ** loaded /home/runner/.julia/artifacts/8fca43e8f7685095940dea5093e165f5298f7f84/bin/../share/singular/LIB/ring.lib (4.1.2.0,Feb_2019)
// ** loaded /home/runner/.julia/artifacts/8fca43e8f7685095940dea5093e165f5298f7f84/bin/../share/singular/LIB/absfact.lib (4.1.2.0,Feb_2019)
// ** loaded /home/runner/.julia/artifacts/8fca43e8f7685095940dea5093e165f5298f7f84/bin/../share/singular/LIB/triang.lib (4.1.2.0,Feb_2019)
// ** loaded /home/runner/.julia/artifacts/8fca43e8f7685095940dea5093e165f5298f7f84/bin/../share/singular/LIB/matrix.lib (4.1.2.0,Feb_2019)
// ** loaded /home/runner/.julia/artifacts/8fca43e8f7685095940dea5093e165f5298f7f84/bin/../share/singular/LIB/nctools.lib (4.1.2.0,Feb_2019)
// ** loaded /home/runner/.julia/artifacts/8fca43e8f7685095940dea5093e165f5298f7f84/bin/../share/singular/LIB/inout.lib (4.1.2.0,Feb_2019)
// ** loaded /home/runner/.julia/artifacts/8fca43e8f7685095940dea5093e165f5298f7f84/bin/../share/singular/LIB/random.lib (4.1.2.0,Feb_2019)
// ** loaded /home/runner/.julia/artifacts/8fca43e8f7685095940dea5093e165f5298f7f84/bin/../share/singular/LIB/polylib.lib (4.2.0.0,Dec_2020)
// ** loaded /home/runner/.julia/artifacts/8fca43e8f7685095940dea5093e165f5298f7f84/bin/../share/singular/LIB/elim.lib (4.1.2.0,Feb_2019)
// ** loaded /home/runner/.julia/artifacts/8fca43e8f7685095940dea5093e165f5298f7f84/bin/../share/singular/LIB/general.lib (4.1.2.0,Feb_2019)
false

Normality Test

Hecke.isnormal — Method

isnormal(A::MPolyQuo)

Example

julia> R, (x, y, z) = PolynomialRing(QQ, ["x", "y", "z"])
(Multivariate Polynomial Ring in x, y, z over Rational Field, fmpq_mpoly[x, y, z])

julia> A, _ = quo(R, ideal(R, [z^2-x*y]))
(Quotient of Multivariate Polynomial Ring in x, y, z over Rational Field by ideal generated by: -x*y + z^2, Map from
Multivariate Polynomial Ring in x, y, z over Rational Field to Quotient of Multivariate Polynomial Ring in x, y, z over Rational Field by ideal generated by: -x*y + z^2 defined by a julia-function with inverse
)

julia> isnormal(A)
// ** loaded /home/runner/.julia/artifacts/8fca43e8f7685095940dea5093e165f5298f7f84/bin/../share/singular/LIB/normal.lib (4.1.2.0,Feb_2019)
// ** loaded /home/runner/.julia/artifacts/8fca43e8f7685095940dea5093e165f5298f7f84/bin/../share/singular/LIB/curveInv.lib (4.1.2.0,Feb_2019)
// ** loaded /home/runner/.julia/artifacts/8fca43e8f7685095940dea5093e165f5298f7f84/bin/../share/singular/LIB/homolog.lib (4.1.2.0,Feb_2019)
// ** loaded /home/runner/.julia/artifacts/8fca43e8f7685095940dea5093e165f5298f7f84/bin/../share/singular/LIB/deform.lib (4.1.2.0,Feb_2019)
// ** loaded /home/runner/.julia/artifacts/8fca43e8f7685095940dea5093e165f5298f7f84/bin/../share/singular/LIB/algebra.lib (4.2.0.1,Mar_2021)
// ** loaded /home/runner/.julia/artifacts/8fca43e8f7685095940dea5093e165f5298f7f84/bin/../share/singular/LIB/reesclos.lib (4.1.2.0,Feb_2019)
// ** loaded /home/runner/.julia/artifacts/8fca43e8f7685095940dea5093e165f5298f7f84/bin/../share/singular/LIB/locnormal.lib (4.1.2.0,Feb_2019)
// ** loaded /home/runner/.julia/artifacts/8fca43e8f7685095940dea5093e165f5298f7f84/bin/../share/singular/LIB/modstd.lib (4.2.0.0,Feb_2020)
// ** loaded /home/runner/.julia/artifacts/8fca43e8f7685095940dea5093e165f5298f7f84/bin/../share/singular/LIB/parallel.lib (4.1.2.0,Feb_2019)
// ** loaded /home/runner/.julia/artifacts/8fca43e8f7685095940dea5093e165f5298f7f84/bin/../share/singular/LIB/tasks.lib (4.1.2.0,Feb_2019)
// ** loaded /home/runner/.julia/artifacts/8fca43e8f7685095940dea5093e165f5298f7f84/bin/../share/singular/LIB/resources.lib (4.1.2.0,Feb_2019)
// ** loaded /home/runner/.julia/artifacts/8fca43e8f7685095940dea5093e165f5298f7f84/bin/../share/singular/LIB/modular.lib (4.2.0.0,Dec_2020)
// ** loaded /home/runner/.julia/artifacts/8fca43e8f7685095940dea5093e165f5298f7f84/bin/../share/singular/LIB/hnoether.lib (4.1.2.0,Feb_2019)
// ** loaded /home/runner/.julia/artifacts/8fca43e8f7685095940dea5093e165f5298f7f84/bin/../share/singular/LIB/primitiv.lib (4.1.2.0,Feb_2019)
// ** loaded /home/runner/.julia/artifacts/8fca43e8f7685095940dea5093e165f5298f7f84/bin/../share/singular/LIB/presolve.lib (4.1.2.0,Feb_2019)
// ** loaded /home/runner/.julia/artifacts/8fca43e8f7685095940dea5093e165f5298f7f84/bin/../share/singular/LIB/sing.lib (4.1.2.0,Feb_2019)

// 'normal' created a list, say nor, of three elements.
// To see the list type
      nor;

// * nor[1] is a list of 1 ring(s).
// To access the i-th ring nor[1][i], give it a name, say Ri, and type
     def R1 = nor[1][1]; setring R1; norid; normap;
// For the other rings type first (if R is the name of your base ring)
     setring R;
// and then continue as for R1.
// Ri/norid is the affine algebra of the normalization of R/P_i where
// P_i is the i-th component of a decomposition of the input ideal id
// and normap the normalization map from R to Ri/norid.

// * nor[2] is a list of 1 ideal(s). Let ci be the last generator
// of the ideal nor[2][i]. Then the integral closure of R/P_i is
// generated as R-submodule of the total ring of fractions by
// 1/ci * nor[2][i].

// * nor[3] is a list of an intvec of size 1 the delta invariants 
// of the components, and an integer, the total delta invariant 
// of R/id (-1 means infinite, and 0 that R/P_i resp. R/id is normal).
false

Cohen-Macaulayness Test

iscohenmacaulay(R)

Example

Homomorphisms of Affine Algebras

Constructors

Hecke.hom — Method

hom(D::MPolyQuo, C::MPolyQuo, V::Vector)

Creates the algebra homomorphism $D \rightarrow C$ defined by sending the generators of $D$ to the elements given by $V$.

Example

Data Associated to Homomorphisms of Affine Algebras

Missing docstring.

Missing docstring for domain(F::MPolyQuoHom). Check Documenter's build log for details.

Missing docstring.

Missing docstring for codomain(F::MPolyQuoHom). Check Documenter's build log for details.

Tests on Homomorphisms of Affine Algebras

Hecke.isinjective — Method

isinjective(F::MPolyQuoHom)

Return true if F is injective, false otherwise.

Hecke.issurjective — Method

issurjective(F::MPolyQuoHom)

Return true if F is surjective, false otherwise.

Hecke.isbijective — Method

isbijective(F::MPolyQuoHom)

Return true if F is bijective, false otherwise.

Base.isfinite — Method

isfinite(F::MPolyQuoHom)

Return true if F is finite, false otherwise.

Noether Normalization

Oscar.noether_normalization — Method

noether_normalization(A::MPolyQuo)

Given an affine algebra $A=R/I$ over a field $K$, the function returns a triple $(L,f,g)$, such that: $L$ is an array of $d=\dim A$ elements of $A$, all represented by linear forms in $R$, and such that $K[L]\rightarrow A$ is a Noether normalization for $A$; $f: A \rightarrow A$ is an automorphism of $A$, induced by a linear change of coordinates of $R$, and mapping the $f_i$ to the last $d$ variables of $A$; and $g = f^{-1}$.

CAVEAT: The algorithm may not terminate over a small finite field. If it terminates, the result is correct.

Example

julia> R, (x, y, z) = PolynomialRing(QQ, ["x", "y", "z"])
(Multivariate Polynomial Ring in x, y, z over Rational Field, fmpq_mpoly[x, y, z])

julia> A, _ = quo(R, ideal(R, [x*y, x*z]))
(Quotient of Multivariate Polynomial Ring in x, y, z over Rational Field by ideal generated by: x*y, x*z, Map from
Multivariate Polynomial Ring in x, y, z over Rational Field to Quotient of Multivariate Polynomial Ring in x, y, z over Rational Field by ideal generated by: x*y, x*z defined by a julia-function with inverse
)

julia> L = noether_normalization(A);

julia> L[1]
2-element Array{Oscar.MPolyQuoElem{fmpq_mpoly},1}:
 -6*x + y
 9*x - 2*y + z

julia> L[2]
AlgebraHomomorphism(Quotient of Multivariate Polynomial Ring in x, y, z over Rational Field by ideal generated by: x*y, x*z => Quotient of Multivariate Polynomial Ring in x, y, z over Rational Field by ideal generated by: x*y, x*z, Singular.spoly{Singular.n_Q}[x, 6*x + y, 3*x + 2*y + z])

julia> L[3]
AlgebraHomomorphism(Quotient of Multivariate Polynomial Ring in x, y, z over Rational Field by ideal generated by: x*y, x*z => Quotient of Multivariate Polynomial Ring in x, y, z over Rational Field by ideal generated by: x*y, x*z, Singular.spoly{Singular.n_Q}[x, -6*x + y, 9*x - 2*y + z])

Normalization of Rings

Oscar.normalize — Method

normalize(A::MPolyQuo; alg=:equidimDec)

Finds the normalization of a reduced affine algebra over a perfect field $K$: Given the quotient $A=R/I$ of a multivariate polynomial ring $R$ over $K$ modulo a radical ideal $I$, compute the integral closure $\overline{A}$ of $A$ in its total ring of fractions $Q(A)$, together with the embedding $f: A \rightarrow \overline{A}$. The function relies on the algorithm of Greuel, Laplagne, and Seelisch which proceeds by finding a suitable decomposition $I=I_1\cap\dots\cap I_r$ into radical ideals $I_k$, together with the normalization maps $f_k: R/I_k \rightarrow A_k=\overline{R/I_k}$, such that

\[f=f_1\times \dots\times f_r: A \rightarrow A_1\times \dots\times A_r=\overline{A}\]

is the normalization map of $A$. For each $k$, the function specifies two representations of $A_k$: It returns an array of triples $(A_k, f_k, \mathfrak a_k)$, where $A_k$ is represented as an affine $K$-algebra, and $f_k$ as a map of affine $K$-algebras. The third entry $\mathfrak a_k$ is a tuple $(d_k, J_k)$, consisting of an element $d_k\in A$ and an ideal $J_k\subset A$, such that $\frac{1}{d_k}J_k = A_k$ as $A$-submodules of the total ring of fractions of $A$.

By default, as a first step on its way to find the decomposition $I=I_1\cap\dots\cap I_r$, the algorithm computes an equidimensional decomposition of the radical ideal $I$. Alternatively, if specified by alg=:primeDec, the algorithm computes $I=I_1\cap\dots\cap I_r$ as the prime decomposition of the radical ideal $I$.

CAVEAT: The function does not check whether $A$ is reduced. Use isreduced(A) in case you are unsure (this may take some time).

Oscar.normalize_with_delta — Method

normalize_with_delta(A::MPolyQuo; alg=:equidimDec)

Finds the normalization

\[f=f_1\times \dots\times f_r: A \rightarrow A_1\times \dots\times A_r=\overline{A}\]

of $A$ as does normalize(A), but computes additionally the delta invariant of $A$, that is, the dimension $\dim_K(\overline{A}/A)$. More precisely, it returns a tuple whose first element is normalize(A), whose second element is an array containing the delta invariants of the $A_k$, and whose third element is the (total) delta invariant of $A$. The return value -1 in the third element indicates that the delta invariant is infinite.

Examples

julia> R, (x, y) = PolynomialRing(QQ, ["x", "y"])
(Multivariate Polynomial Ring in x, y over Rational Field, fmpq_mpoly[x, y])

julia> A, _ = quo(R, ideal(R, [(x^2-y^3)*(x^2+y^2)*x]))
(Quotient of Multivariate Polynomial Ring in x, y over Rational Field by ideal generated by: x^5 - x^3*y^3 + x^3*y^2 - x*y^5, Map from
Multivariate Polynomial Ring in x, y over Rational Field to Quotient of Multivariate Polynomial Ring in x, y over Rational Field by ideal generated by: x^5 - x^3*y^3 + x^3*y^2 - x*y^5 defined by a julia-function with inverse
)

julia> L = normalize(A)

// 'normal' created a list, say nor, of two elements.
// To see the list type
      nor;

// * nor[1] is a list of 2 ring(s).
// To access the i-th ring nor[1][i], give it a name, say Ri, and type
     def R1 = nor[1][1]; setring R1; norid; normap;
// For the other rings type first (if R is the name of your base ring)
     setring R;
// and then continue as for R1.
// Ri/norid is the affine algebra of the normalization of R/P_i where
// P_i is the i-th component of a decomposition of the input ideal id
// and normap the normalization map from R to Ri/norid.

// * nor[2] is a list of 2 ideal(s). Let ci be the last generator
// of the ideal nor[2][i]. Then the integral closure of R/P_i is
// generated as R-submodule of the total ring of fractions by
// 1/ci * nor[2][i].
2-element Array{Tuple{Oscar.MPolyQuo{Singular.spoly{Singular.n_Q}},Oscar.MPolyQuoHom,Tuple{Oscar.MPolyQuoElem{fmpq_mpoly},Oscar.MPolyQuoIdeal{fmpq_mpoly}}},1}:
 (Quotient of Singular Polynomial Ring (QQ),(T(1),T(2),T(3),x,y),(dp(3),dp(2),C) by ideal generated by: 2*T(2)*y + T(3)*y + x, T(1)*x - T(2)*y - T(3)*y, -T(1)*y + T(2)*x - y, T(1)*y^2 - T(1)*y - T(3)*x - y, T(2)*y^3 + T(2)*y^2 + T(3)*y^3 - T(3)*x^2, T(1)^2 + T(1), T(1)*T(2), T(2)^2 + T(1) + 1, T(1)*T(3) + T(2) + T(3), T(2)*T(3) - T(1) - 1, T(3)^2 + T(1)*y + T(1) + 1, x^2*y^3 + y^5 - x^4 - x^2*y^2, AlgebraHomomorphism(Quotient of Multivariate Polynomial Ring in x, y over Rational Field by ideal generated by: x^5 - x^3*y^3 + x^3*y^2 - x*y^5 => Quotient of Singular Polynomial Ring (QQ),(T(1),T(2),T(3),x,y),(dp(3),dp(2),C) by ideal generated by: 2*T(2)*y + T(3)*y + x, T(1)*x - T(2)*y - T(3)*y, -T(1)*y + T(2)*x - y, T(1)*y^2 - T(1)*y - T(3)*x - y, T(2)*y^3 + T(2)*y^2 + T(3)*y^3 - T(3)*x^2, T(1)^2 + T(1), T(1)*T(2), T(2)^2 + T(1) + 1, T(1)*T(3) + T(2) + T(3), T(2)*T(3) - T(1) - 1, T(3)^2 + T(1)*y + T(1) + 1, x^2*y^3 + y^5 - x^4 - x^2*y^2, Singular.spoly{Singular.n_Q}[x, y]), (-2*x^3 + x*y^3 - x*y^2, MPolyQuoIdeal(x^3 + x*y^2, -x^2*y + y^4, x^2*y^2 + x^2*y, -2*x^3 + x*y^3 - x*y^2)))
 (Quotient of Singular Polynomial Ring (QQ),(x,y),(dp(2),C) by ideal generated by: x, AlgebraHomomorphism(Quotient of Multivariate Polynomial Ring in x, y over Rational Field by ideal generated by: x^5 - x^3*y^3 + x^3*y^2 - x*y^5 => Quotient of Singular Polynomial Ring (QQ),(x,y),(dp(2),C) by ideal generated by: x, Singular.spoly{Singular.n_Q}[x, y]), (1, MPolyQuoIdeal(1)))

julia> R, (x, y, z) = PolynomialRing(QQ, ["x", "y", "z"])
(Multivariate Polynomial Ring in x, y, z over Rational Field, fmpq_mpoly[x, y, z])

julia> A, _ = quo(R, ideal(R, [z^3-x*y^4]))
(Quotient of Multivariate Polynomial Ring in x, y, z over Rational Field by ideal generated by: -x*y^4 + z^3, Map from
Multivariate Polynomial Ring in x, y, z over Rational Field to Quotient of Multivariate Polynomial Ring in x, y, z over Rational Field by ideal generated by: -x*y^4 + z^3 defined by a julia-function with inverse
)

julia> L = normalize(A)

// 'normal' created a list, say nor, of two elements.
// To see the list type
      nor;

// * nor[1] is a list of 1 ring(s).
// To access the i-th ring nor[1][i], give it a name, say Ri, and type
     def R1 = nor[1][1]; setring R1; norid; normap;
// For the other rings type first (if R is the name of your base ring)
     setring R;
// and then continue as for R1.
// Ri/norid is the affine algebra of the normalization of R/P_i where
// P_i is the i-th component of a decomposition of the input ideal id
// and normap the normalization map from R to Ri/norid.

// * nor[2] is a list of 1 ideal(s). Let ci be the last generator
// of the ideal nor[2][i]. Then the integral closure of R/P_i is
// generated as R-submodule of the total ring of fractions by
// 1/ci * nor[2][i].
1-element Array{Tuple{Oscar.MPolyQuo{Singular.spoly{Singular.n_Q}},Oscar.MPolyQuoHom,Tuple{Oscar.MPolyQuoElem{fmpq_mpoly},Oscar.MPolyQuoIdeal{fmpq_mpoly}}},1}:
 (Quotient of Singular Polynomial Ring (QQ),(T(1),T(2),x,y,z),(dp(2),dp(3),C) by ideal generated by: T(1)*y - T(2)*z, T(2)*y - z, -T(1)*z + x*y^2, T(1)^2 - x*z, T(1)*T(2) - x*y, T(2)^2 - T(1), x*y^4 - z^3, AlgebraHomomorphism(Quotient of Multivariate Polynomial Ring in x, y, z over Rational Field by ideal generated by: -x*y^4 + z^3 => Quotient of Singular Polynomial Ring (QQ),(T(1),T(2),x,y,z),(dp(2),dp(3),C) by ideal generated by: T(1)*y - T(2)*z, T(2)*y - z, -T(1)*z + x*y^2, T(1)^2 - x*z, T(1)*T(2) - x*y, T(2)^2 - T(1), x*y^4 - z^3, Singular.spoly{Singular.n_Q}[x, y, z]), (z^2, MPolyQuoIdeal(x*y^2*z, x*y^3, z^2)))

julia> R, (x, y) = PolynomialRing(QQ, ["x", "y"])
(Multivariate Polynomial Ring in x, y over Rational Field, fmpq_mpoly[x, y])

julia> A, _ = quo(R, ideal(R, [(x^2-y^3)*(x^2+y^2)*x]))
(Quotient of Multivariate Polynomial Ring in x, y over Rational Field by ideal generated by: x^5 - x^3*y^3 + x^3*y^2 - x*y^5, Map from
Multivariate Polynomial Ring in x, y over Rational Field to Quotient of Multivariate Polynomial Ring in x, y over Rational Field by ideal generated by: x^5 - x^3*y^3 + x^3*y^2 - x*y^5 defined by a julia-function with inverse
)

julia> L = normalize_with_delta(A, alg=:primeDec)

// 'normal' created a list, say nor, of three elements.
// To see the list type
      nor;

// * nor[1] is a list of 3 ring(s).
// To access the i-th ring nor[1][i], give it a name, say Ri, and type
     def R1 = nor[1][1]; setring R1; norid; normap;
// For the other rings type first (if R is the name of your base ring)
     setring R;
// and then continue as for R1.
// Ri/norid is the affine algebra of the normalization of R/P_i where
// P_i is the i-th component of a decomposition of the input ideal id
// and normap the normalization map from R to Ri/norid.

// * nor[2] is a list of 3 ideal(s). Let ci be the last generator
// of the ideal nor[2][i]. Then the integral closure of R/P_i is
// generated as R-submodule of the total ring of fractions by
// 1/ci * nor[2][i].

// * nor[3] is a list of an intvec of size 3 the delta invariants 
// of the components, and an integer, the total delta invariant 
// of R/id (-1 means infinite, and 0 that R/P_i resp. R/id is normal).
(Tuple{Oscar.MPolyQuo{Singular.spoly{Singular.n_Q}},Oscar.MPolyQuoHom,Tuple{Oscar.MPolyQuoElem{fmpq_mpoly},Oscar.MPolyQuoIdeal{fmpq_mpoly}}}[(Quotient of Singular Polynomial Ring (QQ),(T(1),x,y),(dp(1),dp(2),C) by ideal generated by: -T(1)*y + x, -T(1)*x + y^2, T(1)^2 - y, y^3 - x^2, AlgebraHomomorphism(Quotient of Multivariate Polynomial Ring in x, y over Rational Field by ideal generated by: x^5 - x^3*y^3 + x^3*y^2 - x*y^5 => Quotient of Singular Polynomial Ring (QQ),(T(1),x,y),(dp(1),dp(2),C) by ideal generated by: -T(1)*y + x, -T(1)*x + y^2, T(1)^2 - y, y^3 - x^2, Singular.spoly{Singular.n_Q}[x, y]), (y, MPolyQuoIdeal(x, y))), (Quotient of Singular Polynomial Ring (QQ),(T(1),x,y),(dp(1),dp(2),C) by ideal generated by: -T(1)*y + x, T(1)*x + y, T(1)^2 + 1, x^2 + y^2, AlgebraHomomorphism(Quotient of Multivariate Polynomial Ring in x, y over Rational Field by ideal generated by: x^5 - x^3*y^3 + x^3*y^2 - x*y^5 => Quotient of Singular Polynomial Ring (QQ),(T(1),x,y),(dp(1),dp(2),C) by ideal generated by: -T(1)*y + x, T(1)*x + y, T(1)^2 + 1, x^2 + y^2, Singular.spoly{Singular.n_Q}[x, y]), (y, MPolyQuoIdeal(x, y))), (Quotient of Singular Polynomial Ring (QQ),(x,y),(dp(2),C) by ideal generated by: x, AlgebraHomomorphism(Quotient of Multivariate Polynomial Ring in x, y over Rational Field by ideal generated by: x^5 - x^3*y^3 + x^3*y^2 - x*y^5 => Quotient of Singular Polynomial Ring (QQ),(x,y),(dp(2),C) by ideal generated by: x, Singular.spoly{Singular.n_Q}[x, y]), (1, MPolyQuoIdeal(1)))], [1, 1, 0], 13)

Invariant Theory

Invariant Theory of Finite Groups

...