Commutative Algebra

Introduction

This part of OSCAR is under development with regard to providing both the functionality and the documentation. At current state, the documentation serves as a guide on what to do and will be constantly updated.

Ideal Constructors for Multivariate Polynomial Rings

Hecke.ideal — Method

ideal(g::Array{T, 1}) where {T <: MPolyElem}

Creates the ideal generated by the polynomials in g.

Hecke.ideal — Method

ideal(Rx::MPolyRing, g::Array{<:Any, 1})

Creates the ideal generated by the polynomials in g, specifying the ambient ring of the ideal.

Example

julia> R, (x, y) = PolynomialRing(QQ, ["x", "y"])
(Multivariate Polynomial Ring in x, y over Rational Field, fmpq_mpoly[x, y])

julia> I = ideal([x*y-3*x,y^3-2*x^2*y])
ideal generated by: x*y - 3*x, -2*x^2*y + y^3

julia> J = ideal(R, [(x+y)^2])
ideal generated by: x^2 + 2*x*y + y^2

Gröbner Bases

Monomial Orders

Normal Forms

Computing Gröbner Bases

Oscar.groebner_assure — Method

groebner_assure(I::MPolyIdeal)

Given an ideal I in a multivariate polynomial ring this function assures that a Groebner basis w.r.t. the given monomial ordering is attached to I in I.gb.

Example

julia> R,(x,y) = PolynomialRing(QQ, ["x","y"], ordering=:degrevlex)
(Multivariate Polynomial Ring in x, y over Rational Field, fmpq_mpoly[x, y])

julia> I = ideal([x*y-3*x,y^3-2*x^2*y])
ideal generated by: x*y - 3*x, -2*x^2*y + y^3

julia> groebner_assure(I)
3-element Array{fmpq_mpoly,1}:
 x*y - 3*x
 y^3 - 6*x^2
 x^3 - 9//2*x

julia> I.gb
Oscar.BiPolyArray{fmpq_mpoly}(fmpq_mpoly[x*y - 3*x, y^3 - 6*x^2, x^3 - 9//2*x], Singular Ideal over Singular Polynomial Ring (QQ),(x,y),(dp(2),C) with generators (x*y-3*x, y^3-6*x^2, x^3-9/2*x), Multivariate Polynomial Ring in x, y over Rational Field, Singular Polynomial Ring (QQ),(x,y),(dp(2),C), false)

Example

julia> R, (x, y) = PolynomialRing(QQ, ["x", "y"], ordering=:degrevlex)
(Multivariate Polynomial Ring in x, y over Rational Field, fmpq_mpoly[x, y])

julia> I = ideal([x*y-3*x,y^3-2*x^2*y])
ideal generated by: x*y - 3*x, -2*x^2*y + y^3

julia> groebner_assure(I)
3-element Array{fmpq_mpoly,1}:
 x*y - 3*x
 y^3 - 6*x^2
 x^3 - 9//2*x

julia> I.gb
Oscar.BiPolyArray{fmpq_mpoly}(fmpq_mpoly[x*y - 3*x, y^3 - 6*x^2, x^3 - 9//2*x], Singular Ideal over Singular Polynomial Ring (QQ),(x,y),(dp(2),C) with generators (x*y-3*x, y^3-6*x^2, x^3-9/2*x), Multivariate Polynomial Ring in x, y over Rational Field, Singular Polynomial Ring (QQ),(x,y),(dp(2),C), false)

Gröbner Bases with transformation matrix

Oscar.groebner_basis_with_transform — Method

groebner_basis_with_transform(B::BiPolyArray; ord::Symbol = :degrevlex, complete_reduction::Bool = false)

Given an BiPolyArray B and optional parameters ord for a monomial ordering and complete_reduction this function computes a Groebner basis (if complete_reduction = true the reduced Groebner basis) of the ideal spanned by the elements in B w.r.t. the given monomial ordering ord and the transformation matrix from the ideal to the Groebner basis. Return value is a BiPolyArray together with a map.

Example

julia> R, (x, y) = PolynomialRing(QQ, ["x", "y"], ordering=:degrevlex)
(Multivariate Polynomial Ring in x, y over Rational Field, fmpq_mpoly[x, y])

julia> A = Oscar.BiPolyArray([x*y-3*x,y^3-2*x^2*y])
Oscar.BiPolyArray{fmpq_mpoly}(fmpq_mpoly[x*y - 3*x, -2*x^2*y + y^3], #undef, Multivariate Polynomial Ring in x, y over Rational Field, Singular Polynomial Ring (QQ),(x,y),(dp(2),C), false)

julia> B,m = groebner_basis_with_transform(A)
(Oscar.BiPolyArray{fmpq_mpoly}(fmpq_mpoly[#undef, #undef, #undef], Singular Ideal over Singular Polynomial Ring (QQ),(x,y),(dp(2),C) with generators (x*y-3*x, y^3-6*x^2, 6*x^3-27*x), Multivariate Polynomial Ring in x, y over Rational Field, Singular Polynomial Ring (QQ),(x,y),(dp(2),C), false), [1 2*x -2*x^2 + y^2 + 3*y + 9; 0 1 -x])

fglm

Gröbner walks

Hilbert-driven

Leading Ideals

Oscar.leading_ideal — Method

leading_ideal(g::Array{T, 1}, args...) where { T <: MPolyElem }

Given an array of multivariate polynomials this function returns the ideal generated by the leading monomials of the given array elements. If not otherwise given as a further argument this is done w.r.t. the degree reverse lexicographical monomial ordering.

Example

julia> R, (x, y) = PolynomialRing(QQ, ["x", "y"])
(Multivariate Polynomial Ring in x, y over Rational Field, fmpq_mpoly[x, y])

julia> L = leading_ideal([x*y^2-3*x, x^3-14*y^5])
ideal generated by: x*y^2, x^3

Oscar.leading_ideal — Method

leading_ideal(I::MPolyIdeal)

Given a multivariate polynomial ideal Ì this function returns the leading ideal for I. This is done w.r.t. the given monomial ordering in the polynomial ring of I.

Example

julia> R, (x, y) = PolynomialRing(QQ, ["x", "y"])
(Multivariate Polynomial Ring in x, y over Rational Field, fmpq_mpoly[x, y])

julia> I = ideal(R,[x*y^2-3*x, x^3-14*y^5])
ideal generated by: x*y^2 - 3*x, x^3 - 14*y^5

julia> L = leading_ideal(I)
ideal generated by: x*y^2, x^4, x^3

Gröbner Bases over the integers

....

Syzygies

Generators of syzygies

Oscar.syzygy_generators — Method

syzygy_generators(a::Array{<:MPolyElem, 1})

Given an array of multivariate polynomials this function returns the generators of the syzygy module of the ideal generated by the given array elements.

Example

julia> R, (x, y) = PolynomialRing(QQ, ["x", "y"])
(Multivariate Polynomial Ring in x, y over Rational Field, fmpq_mpoly[x, y])

julia> S = syzygy_generators([x^3+y+2,x*y^2-13*x^2,y-14])
3-element Array{AbstractAlgebra.Generic.FreeModuleElem{fmpq_mpoly},1}:
 (0, -y + 14, -13*x^2 + x*y^2)
 (-169*y + 2366, -13*x*y + 182*x - 196*y + 2744, 13*x^2*y^2 - 2548*x^2 + 196*x*y^2 + 169*y + 338)
 (-13*x^2 + 196*x, -x^3 - 16, x^4*y + 14*x^4 + 13*x^2 + 16*x*y + 28*x)

Ideals in Multivariate Polynomial Rings

Introduction

Distinction: ideals <–> homogeneous ideals. To be discussed.

Ideal Operations for Multivariate Polynomial Rings

Simple ideal Operations

Powers of Ideal

Base.:^ — Method

:^(I::MPolyIdeal, m::Int)

Returns the m-th power of I.

Sum of Ideals

Base.:+ — Method

:+(I::MPolyIdeal, J::MPolyIdeal)

Returns the sum of I and J.

Product of Ideals

Base.:* — Method

:*(I::MPolyIdeal, J::MPolyIdeal)

Returns the product of I and J.

Example

julia> R, (x, y, z) = PolynomialRing(QQ, ["x", "y", "z"])
(Multivariate Polynomial Ring in x, y, z over Rational Field, fmpq_mpoly[x, y, z])

julia> I = ideal(R, [x, y])
ideal generated by: x, y

julia> J = ideal(R, [z^2])
ideal generated by: z^2

julia> I^3
ideal generated by: x^3, x^2*y, x*y^2, y^3

julia> I+J
ideal generated by: x, y, z^2

julia> I*J
ideal generated by: x*z^2, y*z^2

Intersection of Ideals

Base.intersect — Method

intersect(I::MPolyIdeal, J::MPolyIdeal)

Returns the intersection of I and J.

Example

julia> R, (x, y) = PolynomialRing(QQ, ["x", "y"])
(Multivariate Polynomial Ring in x, y over Rational Field, fmpq_mpoly[x, y])

julia> I = intersect(ideal(R, [x, y])^2, ideal(R, [y^2-x^3+x]))
ideal generated by: x^3*y - x*y - y^3, x^4 - x^2 - x*y^2

Ideal Quotients

Hecke.quotient — Method

quotient(I::MPolyIdeal, J::MPolyIdeal)

Returns the ideal quotient of I by J. Alternatively, use I:J.

Example

julia> R, (x, y, z) = PolynomialRing(QQ, ["x", "y", "z"])
(Multivariate Polynomial Ring in x, y, z over Rational Field, fmpq_mpoly[x, y, z])

julia> I = ideal(R, [x^4+x^2*y*z+y^3*z, y^4+x^3*z+x*y^2*z, x^3*y+x*y^3])
ideal generated by: x^4 + x^2*y*z + y^3*z, x^3*z + x*y^2*z + y^4, x^3*y + x*y^3

julia> J = ideal(R,[x,y,z])^2
ideal generated by: x^2, x*y, x*z, y^2, y*z, z^2

julia> L = quotient(I,J)
ideal generated by: x^3*z + x*y^2*z + y^4, x^3*y + x*y^3, x^4 + x^2*y*z + y^3*z, x^3*z^2 - x^2*y*z^2 + x*y^2*z^2 - y^3*z^2, x^2*y^2*z - x^2*y*z^2 - y^3*z^2, x^3*z^2 + x^2*y^3 - x^2*y^2*z + x*y^2*z^2

julia> I:J
ideal generated by: x^3*z + x*y^2*z + y^4, x^3*y + x*y^3, x^4 + x^2*y*z + y^3*z, x^3*z^2 - x^2*y*z^2 + x*y^2*z^2 - y^3*z^2, x^2*y^2*z - x^2*y*z^2 - y^3*z^2, x^3*z^2 + x^2*y^3 - x^2*y^2*z + x*y^2*z^2

Saturation

Oscar.saturation — Method

saturation(I::MPolyIdeal, J::MPolyIdeal)

Returns the saturation of I with respect to J.

Example

julia> R, (x, y, z) = PolynomialRing(QQ, ["x", "y", "z"])
(Multivariate Polynomial Ring in x, y, z over Rational Field, fmpq_mpoly[x, y, z])

julia> I = ideal(R, [x^4+x^2*y*z+y^3*z, y^4+x^3*z+x*y^2*z, x^3*y+x*y^3])
ideal generated by: x^4 + x^2*y*z + y^3*z, x^3*z + x*y^2*z + y^4, x^3*y + x*y^3

julia> J = ideal(R, [x,y,z])^2
ideal generated by: x^2, x*y, x*z, y^2, y*z, z^2

julia> L = saturation(I,J)
ideal generated by: x^3 - x^2*y + x*y^2 - y^3, x^2*y*z + y^4 + y^3*z, x^2*y^2 + y^4

Elimination

Oscar.eliminate — Method

eliminate(I::MPolyIdeal, polys::Array{MPolyElem, 1})

Given a list of polynomials which are variables, these variables are eliminated from I. That is, the function returns the ideal of all polynomials in I which only depend on the remaining variables.

Oscar.eliminate — Method

eliminate(I::MPolyIdeal, polys::AbstractArray{Int, 1})

Given a list of indices which specify variables, these variables are eliminated from I. That is, the function returns the ideal of all polynomials in I which only depend on the remaining variables.

Example

julia> R, (t, x, y, z) = PolynomialRing(QQ, ["t", "x", "y", "z"])
(Multivariate Polynomial Ring in t, x, y, z over Rational Field, fmpq_mpoly[t, x, y, z])

julia> I = ideal(R, [t-x, t^2-y, t-z^3])
ideal generated by: t - x, t^2 - y, t - z^3

julia> A = [t]
1-element Array{fmpq_mpoly,1}:
 t

julia> AA = [1]
1-element Array{Int64,1}:
 1

julia> TC = eliminate(I,A)
ideal generated by: x^2 - y, -x + z^3

julia> TC = eliminate(I,AA)
ideal generated by: x^2 - y, -x + z^3

Homogenization and Dehomogenization

homogenize(I,t)   CAVEAT: Als Ideal! Auch poly, vector, etc. Siehe M2.

dehomogenize(I,t)

Decomposition

Radical

Hecke.radical — Method

radical(I::MPolyIdeal)

Returns the radical of I. If the base ring of I is a polynomial ring over a field, a combination of the algorithms of Krick and Logar (with modifications by Laplagne) and Kemper is used. For polynomial rings over the integers, the algorithm proceeds as suggested by Pfister, Sadiq, and Steidel.

Example

julia> R, (x, y) = PolynomialRing(QQ, ["x", "y"])
(Multivariate Polynomial Ring in x, y over Rational Field, fmpq_mpoly[x, y])

julia> I = intersect(ideal(R, [x, y])^2, ideal(R, [y^2-x^3+x]))
ideal generated by: x^3*y - x*y - y^3, x^4 - x^2 - x*y^2

julia> I = intersect(I, ideal(R, [x-y-1])^2)
ideal generated by: x^5*y - 2*x^4*y^2 - 2*x^4*y + x^3*y^3 + 2*x^3*y^2 - x^2*y^3 + 2*x^2*y^2 + 2*x^2*y + 2*x*y^4 + x*y^3 - 2*x*y^2 - x*y - y^5 - 2*y^4 - y^3, x^6 - 2*x^5 - 3*x^4*y^2 - 2*x^4*y + 2*x^3*y^3 + 3*x^3*y^2 + 2*x^3*y + 2*x^3 + 5*x^2*y^2 + 2*x^2*y - x^2 + 3*x*y^4 - 5*x*y^2 - 2*x*y - 2*y^5 - 4*y^4 - 2*y^3

julia> RI = radical(I)
// ** loaded /home/fieker/.julia/artifacts/49e6c5baf692a764a6bb5d397d0fdb8e728eb669/bin/../share/singular/LIB/primdec.lib (4.1.2.0,Feb_2019)
// ** loaded /home/fieker/.julia/artifacts/49e6c5baf692a764a6bb5d397d0fdb8e728eb669/bin/../share/singular/LIB/ring.lib (4.1.2.0,Feb_2019)
// ** loaded /home/fieker/.julia/artifacts/49e6c5baf692a764a6bb5d397d0fdb8e728eb669/bin/../share/singular/LIB/absfact.lib (4.1.2.0,Feb_2019)
// ** loaded /home/fieker/.julia/artifacts/49e6c5baf692a764a6bb5d397d0fdb8e728eb669/bin/../share/singular/LIB/triang.lib (4.1.2.0,Feb_2019)
// ** loaded /home/fieker/.julia/artifacts/49e6c5baf692a764a6bb5d397d0fdb8e728eb669/bin/../share/singular/LIB/matrix.lib (4.1.2.0,Feb_2019)
// ** loaded /home/fieker/.julia/artifacts/49e6c5baf692a764a6bb5d397d0fdb8e728eb669/bin/../share/singular/LIB/nctools.lib (4.1.2.0,Feb_2019)
// ** loaded /home/fieker/.julia/artifacts/49e6c5baf692a764a6bb5d397d0fdb8e728eb669/bin/../share/singular/LIB/inout.lib (4.1.2.0,Feb_2019)
// ** loaded /home/fieker/.julia/artifacts/49e6c5baf692a764a6bb5d397d0fdb8e728eb669/bin/../share/singular/LIB/random.lib (4.1.2.0,Feb_2019)
// ** loaded /home/fieker/.julia/artifacts/49e6c5baf692a764a6bb5d397d0fdb8e728eb669/bin/../share/singular/LIB/polylib.lib (4.2.0.0,Dec_2020)
// ** loaded /home/fieker/.julia/artifacts/49e6c5baf692a764a6bb5d397d0fdb8e728eb669/bin/../share/singular/LIB/elim.lib (4.1.2.0,Feb_2019)
// ** loaded /home/fieker/.julia/artifacts/49e6c5baf692a764a6bb5d397d0fdb8e728eb669/bin/../share/singular/LIB/general.lib (4.1.2.0,Feb_2019)
ideal generated by: x^4 - x^3*y - x^3 - x^2 - x*y^2 + x*y + x + y^3 + y^2

julia> R, (a, b, c, d) = PolynomialRing(ZZ, ["a", "b", "c", "d"])
(Multivariate Polynomial Ring in a, b, c, d over Integer Ring, fmpz_mpoly[a, b, c, d])

julia> I = intersect(ideal(R, [9,a,b]), ideal(R, [3,c]))
ideal generated by: 9, 3*b, 3*a, b*c, a*c

julia> I = intersect(I, ideal(R, [11,2a,7b]))
ideal generated by: 99, 3*b, 3*a, b*c, a*c

julia> I = intersect(I, ideal(R, [13a^2,17b^4]))
ideal generated by: 39*a^2, 13*a^2*c, 51*b^4, 17*b^4*c, 3*a^2*b^4, a^2*b^4*c

julia> I = intersect(I, ideal(R, [9c^5,6d^5]))
ideal generated by: 78*a^2*d^5, 117*a^2*c^5, 102*b^4*d^5, 153*b^4*c^5, 6*a^2*b^4*d^5, 9*a^2*b^4*c^5, 39*a^2*c^5*d^5, 51*b^4*c^5*d^5, 3*a^2*b^4*c^5*d^5

julia> I = intersect(I, ideal(R, [17,a^15,b^15,c^15,d^15]))
ideal generated by: 1326*a^2*d^5, 1989*a^2*c^5, 102*b^4*d^5, 153*b^4*c^5, 663*a^2*c^5*d^5, 51*b^4*c^5*d^5, 78*a^2*d^15, 117*a^2*c^15, 78*a^15*d^5, 117*a^15*c^5, 6*a^2*b^4*d^15, 9*a^2*b^4*c^15, 39*a^2*c^5*d^15, 39*a^2*c^15*d^5, 6*a^2*b^15*d^5, 9*a^2*b^15*c^5, 6*a^15*b^4*d^5, 9*a^15*b^4*c^5, 39*a^15*c^5*d^5, 3*a^2*b^4*c^5*d^15, 3*a^2*b^4*c^15*d^5, 3*a^2*b^15*c^5*d^5, 3*a^15*b^4*c^5*d^5

julia> RI = radical(I)
// ** loaded /home/fieker/.julia/artifacts/49e6c5baf692a764a6bb5d397d0fdb8e728eb669/bin/../share/singular/LIB/primdecint.lib (4.1.2.0,Feb_2019)
ideal generated by: 102*b*d, 78*a*d, 51*b*c, 39*a*c, 6*a*b*d, 3*a*b*c

Primary Decomposition

Oscar.primary_decomposition — Method

primary_decomposition(I::MPolyIdeal)

Returns a primary decomposition of the ideal I. If I is the unit ideal, [ideal(1)] is returned. If the base ring of I is a polynomial ring over a field, the algorithm of Gianni, Trager and Zacharias is used by default. Alternatively, the algorithm by Shimoyama and Yokoyama can be used by specifying alg=:SY. For polynomial rings over the integers, the algorithm proceeds as suggested by Pfister, Sadiq, and Steidel.

Example

julia> R, (x, y) = PolynomialRing(QQ, ["x", "y"])
(Multivariate Polynomial Ring in x, y over Rational Field, fmpq_mpoly[x, y])

julia> I = intersect(ideal(R, [x, y])^2, ideal(R, [y^2-x^3+x]))
ideal generated by: x^3*y - x*y - y^3, x^4 - x^2 - x*y^2

julia> I = intersect(I, ideal(R, [x-y-1])^2)
ideal generated by: x^5*y - 2*x^4*y^2 - 2*x^4*y + x^3*y^3 + 2*x^3*y^2 - x^2*y^3 + 2*x^2*y^2 + 2*x^2*y + 2*x*y^4 + x*y^3 - 2*x*y^2 - x*y - y^5 - 2*y^4 - y^3, x^6 - 2*x^5 - 3*x^4*y^2 - 2*x^4*y + 2*x^3*y^3 + 3*x^3*y^2 + 2*x^3*y + 2*x^3 + 5*x^2*y^2 + 2*x^2*y - x^2 + 3*x*y^4 - 5*x*y^2 - 2*x*y - 2*y^5 - 4*y^4 - 2*y^3

julia> L = primary_decomposition(I)
3-element Array{Tuple{Oscar.MPolyIdeal{fmpq_mpoly},Oscar.MPolyIdeal{fmpq_mpoly}},1}:
 (ideal generated by: x^3 - x - y^2, ideal generated by: x^3 - x - y^2)
 (ideal generated by: x^2 - 2*x*y - 2*x + y^2 + 2*y + 1, ideal generated by: x - y - 1)
 (ideal generated by: y, x^2, ideal generated by: x, y)

julia> L = primary_decomposition(I, alg=:SY)
3-element Array{Tuple{Oscar.MPolyIdeal{fmpq_mpoly},Oscar.MPolyIdeal{fmpq_mpoly}},1}:
 (ideal generated by: x^3 - x - y^2, ideal generated by: x^3 - x - y^2)
 (ideal generated by: x^2 - 2*x*y - 2*x + y^2 + 2*y + 1, ideal generated by: x - y - 1)
 (ideal generated by: y, x^2, ideal generated by: y, x)

julia> R, (a, b, c, d) = PolynomialRing(ZZ, ["a", "b", "c", "d"])
(Multivariate Polynomial Ring in a, b, c, d over Integer Ring, fmpz_mpoly[a, b, c, d])

julia> I = ideal(R, [1326*a^2*d^5, 1989*a^2*c^5, 102*b^4*d^5, 153*b^4*c^5,
       663*a^2*c^5*d^5, 51*b^4*c^5*d^5, 78*a^2*d^15, 117*a^2*c^15,
       78*a^15*d^5, 117*a^15*c^5, 6*a^2*b^4*d^15, 9*a^2*b^4*c^15,
       39*a^2*c^5*d^15, 39*a^2*c^15*d^5, 6*a^2*b^15*d^5, 9*a^2*b^15*c^5,
       6*a^15*b^4*d^5, 9*a^15*b^4*c^5, 39*a^15*c^5*d^5, 3*a^2*b^4*c^5*d^15,
       3*a^2*b^4*c^15*d^5, 3*a^2*b^15*c^5*d^5, 3*a^15*b^4*c^5*d^5])
ideal generated by: 1326*a^2*d^5, 1989*a^2*c^5, 102*b^4*d^5, 153*b^4*c^5, 663*a^2*c^5*d^5, 51*b^4*c^5*d^5, 78*a^2*d^15, 117*a^2*c^15, 78*a^15*d^5, 117*a^15*c^5, 6*a^2*b^4*d^15, 9*a^2*b^4*c^15, 39*a^2*c^5*d^15, 39*a^2*c^15*d^5, 6*a^2*b^15*d^5, 9*a^2*b^15*c^5, 6*a^15*b^4*d^5, 9*a^15*b^4*c^5, 39*a^15*c^5*d^5, 3*a^2*b^4*c^5*d^15, 3*a^2*b^4*c^15*d^5, 3*a^2*b^15*c^5*d^5, 3*a^15*b^4*c^5*d^5

julia> L = primary_decomposition(I)
8-element Array{Tuple{Oscar.MPolyIdeal{fmpz_mpoly},Oscar.MPolyIdeal{fmpz_mpoly}},1}:
 (ideal generated by: d^5, c^5, ideal generated by: d, c)
 (ideal generated by: a^2, b^4, ideal generated by: b, a)
 (ideal generated by: 2, c^5, ideal generated by: 2, c)
 (ideal generated by: 3, ideal generated by: 3)
 (ideal generated by: 13, b^4, ideal generated by: 13, b)
 (ideal generated by: 17, a^2, ideal generated by: 17, a)
 (ideal generated by: 17, d^15, c^15, b^15, a^15, ideal generated by: 17, d, c, b, a)
 (ideal generated by: 9, 3*d^5, d^10, ideal generated by: 3, d)

Minimal Associated Primes

Oscar.minimal_primes — Method

minimal_primes(I::MPolyIdeal; alg=:GTZ)

Returns an array of the minimal associated prime ideals of I. If I is the unit ideal, [ideal(1)] is returned. If the base ring of I is a polynomial ring over a field, the algorithm of Gianni-Trager-Zacharias is used by default and characteristic sets may be used by specifying alg=:charSets. For polynomial rings over the integers, the algorithm proceeds as suggested by Pfister, Sadiq, and Steidel.

Example

julia> R, (x, y) = PolynomialRing(QQ, ["x", "y"])
(Multivariate Polynomial Ring in x, y over Rational Field, fmpq_mpoly[x, y])

julia> I = intersect(ideal(R, [x, y])^2, ideal(R, [y^2-x^3+x]))
ideal generated by: x^3*y - x*y - y^3, x^4 - x^2 - x*y^2

julia> I = intersect(I, ideal(R, [x-y-1])^2)
ideal generated by: x^5*y - 2*x^4*y^2 - 2*x^4*y + x^3*y^3 + 2*x^3*y^2 - x^2*y^3 + 2*x^2*y^2 + 2*x^2*y + 2*x*y^4 + x*y^3 - 2*x*y^2 - x*y - y^5 - 2*y^4 - y^3, x^6 - 2*x^5 - 3*x^4*y^2 - 2*x^4*y + 2*x^3*y^3 + 3*x^3*y^2 + 2*x^3*y + 2*x^3 + 5*x^2*y^2 + 2*x^2*y - x^2 + 3*x*y^4 - 5*x*y^2 - 2*x*y - 2*y^5 - 4*y^4 - 2*y^3

julia> L = minimal_primes(I)
2-element Array{Oscar.MPolyIdeal{fmpq_mpoly},1}:
 ideal generated by: x - y - 1
 ideal generated by: x^3 - x - y^2

julia> L = minimal_primes(I, alg=:charSets)
2-element Array{Oscar.MPolyIdeal{fmpq_mpoly},1}:
 ideal generated by: x - y - 1
 ideal generated by: x^3 - x - y^2

julia> R, (a, b, c, d) = PolynomialRing(ZZ, ["a", "b", "c", "d"])
(Multivariate Polynomial Ring in a, b, c, d over Integer Ring, fmpz_mpoly[a, b, c, d])

julia> I = ideal(R, [1326*a^2*d^5, 1989*a^2*c^5, 102*b^4*d^5, 153*b^4*c^5,
       663*a^2*c^5*d^5, 51*b^4*c^5*d^5, 78*a^2*d^15, 117*a^2*c^15,
       78*a^15*d^5, 117*a^15*c^5, 6*a^2*b^4*d^15, 9*a^2*b^4*c^15,
       39*a^2*c^5*d^15, 39*a^2*c^15*d^5, 6*a^2*b^15*d^5, 9*a^2*b^15*c^5,
       6*a^15*b^4*d^5, 9*a^15*b^4*c^5, 39*a^15*c^5*d^5, 3*a^2*b^4*c^5*d^15,
       3*a^2*b^4*c^15*d^5, 3*a^2*b^15*c^5*d^5, 3*a^15*b^4*c^5*d^5])
ideal generated by: 1326*a^2*d^5, 1989*a^2*c^5, 102*b^4*d^5, 153*b^4*c^5, 663*a^2*c^5*d^5, 51*b^4*c^5*d^5, 78*a^2*d^15, 117*a^2*c^15, 78*a^15*d^5, 117*a^15*c^5, 6*a^2*b^4*d^15, 9*a^2*b^4*c^15, 39*a^2*c^5*d^15, 39*a^2*c^15*d^5, 6*a^2*b^15*d^5, 9*a^2*b^15*c^5, 6*a^15*b^4*d^5, 9*a^15*b^4*c^5, 39*a^15*c^5*d^5, 3*a^2*b^4*c^5*d^15, 3*a^2*b^4*c^15*d^5, 3*a^2*b^15*c^5*d^5, 3*a^15*b^4*c^5*d^5

julia> L = minimal_primes(I)
6-element Array{Oscar.MPolyIdeal{fmpz_mpoly},1}:
 ideal generated by: d, c
 ideal generated by: b, a
 ideal generated by: 2, c
 ideal generated by: 3
 ideal generated by: 13, b
 ideal generated by: 17, a

Weak Equidimensional Decomposition

Oscar.equidimensional_decomposition_weak — Method

equidimensional_decomposition_weak(I::MPolyIdeal)

Returns an array of equidimensional ideals where the last element is the equidimensional hull of I, that is, the intersection of the primary components of I of maximal dimension. Each of the previous elements is an ideal of lower dimension whose associated primes are exactly the associated primes of I of that dimension. If I is the unit ideal, [ideal(1)] is returned. Uses ideas of Eisenbud, Huneke, and Vasconcelos.

Example

julia> R, (x, y) = PolynomialRing(QQ, ["x", "y"])
(Multivariate Polynomial Ring in x, y over Rational Field, fmpq_mpoly[x, y])

julia> I = intersect(ideal(R, [x, y])^2, ideal(R, [y^2-x^3+x]))
ideal generated by: x^3*y - x*y - y^3, x^4 - x^2 - x*y^2

julia> I = intersect(I, ideal(R, [x-y-1])^2)
ideal generated by: x^5*y - 2*x^4*y^2 - 2*x^4*y + x^3*y^3 + 2*x^3*y^2 - x^2*y^3 + 2*x^2*y^2 + 2*x^2*y + 2*x*y^4 + x*y^3 - 2*x*y^2 - x*y - y^5 - 2*y^4 - y^3, x^6 - 2*x^5 - 3*x^4*y^2 - 2*x^4*y + 2*x^3*y^3 + 3*x^3*y^2 + 2*x^3*y + 2*x^3 + 5*x^2*y^2 + 2*x^2*y - x^2 + 3*x*y^4 - 5*x*y^2 - 2*x*y - 2*y^5 - 4*y^4 - 2*y^3

julia> L = equidimensional_decomposition_weak(I)
2-element Array{Oscar.MPolyIdeal{fmpq_mpoly},1}:
 ideal generated by: y, x
 ideal generated by: x^5 - 2*x^4*y - 2*x^4 + x^3*y^2 + 2*x^3*y - x^2*y^2 + 2*x^2*y + 2*x^2 + 2*x*y^3 + x*y^2 - 2*x*y - x - y^4 - 2*y^3 - y^2

Equidimensional Decomposition of radical

Oscar.equidimensional_decomposition_radical — Method

equidimensional_decomposition_radical(I::MPolyIdeal)

Returns an array of equidimensional radical ideals increasingly ordered by dimension. For each dimension, the returned radical ideal is the intersection of the associated primes of I of that dimension. If I is the unit ideal, [ideal(1)] is returned. Uses a combination of the algorithms of Krick and Logar (with modifications by Laplagne) and Kemper.

Example

julia> R, (x, y) = PolynomialRing(QQ, ["x", "y"])
(Multivariate Polynomial Ring in x, y over Rational Field, fmpq_mpoly[x, y])

julia> I = intersect(ideal(R, [x, y])^2, ideal(R, [y^2-x^3+x]))
ideal generated by: x^3*y - x*y - y^3, x^4 - x^2 - x*y^2

julia> I = intersect(I, ideal(R, [x-y-1])^2)
ideal generated by: x^5*y - 2*x^4*y^2 - 2*x^4*y + x^3*y^3 + 2*x^3*y^2 - x^2*y^3 + 2*x^2*y^2 + 2*x^2*y + 2*x*y^4 + x*y^3 - 2*x*y^2 - x*y - y^5 - 2*y^4 - y^3, x^6 - 2*x^5 - 3*x^4*y^2 - 2*x^4*y + 2*x^3*y^3 + 3*x^3*y^2 + 2*x^3*y + 2*x^3 + 5*x^2*y^2 + 2*x^2*y - x^2 + 3*x*y^4 - 5*x*y^2 - 2*x*y - 2*y^5 - 4*y^4 - 2*y^3

julia> L = equidimensional_decomposition_radical(I)
2-element Array{Oscar.MPolyIdeal{fmpq_mpoly},1}:
 ideal generated by: x^4 - x^3*y - x^3 - x^2 - x*y^2 + x*y + x + y^3 + y^2
 ideal generated by: y, x

Equidimensional Hull

Oscar.equidimensional_hull — Method

equidimensional_hull(I::MPolyIdeal)

If the base ring of I is a polynomial ring over a field, return the intersection of the primary components of I of maximal dimension. In the case of polynomials over the integers, return the intersection of the primary components of I of minimal height. If I is the unit ideal, [ideal(1)] is returned. For polynomial rings over a field, the algorithm relies on ideas as used by Gianni, Trager, and Zacharias or Krick and Logar. For polynomial rings over the integers, the algorithm proceeds as suggested by Pfister, Sadiq, and Steidel.

Example

julia> R, (x, y) = PolynomialRing(QQ, ["x", "y"])
(Multivariate Polynomial Ring in x, y over Rational Field, fmpq_mpoly[x, y])

julia> I = intersect(ideal(R, [x, y])^2, ideal(R, [y^2-x^3+x]))
ideal generated by: x^3*y - x*y - y^3, x^4 - x^2 - x*y^2

julia> I = intersect(I, ideal(R, [x-y-1])^2)
ideal generated by: x^5*y - 2*x^4*y^2 - 2*x^4*y + x^3*y^3 + 2*x^3*y^2 - x^2*y^3 + 2*x^2*y^2 + 2*x^2*y + 2*x*y^4 + x*y^3 - 2*x*y^2 - x*y - y^5 - 2*y^4 - y^3, x^6 - 2*x^5 - 3*x^4*y^2 - 2*x^4*y + 2*x^3*y^3 + 3*x^3*y^2 + 2*x^3*y + 2*x^3 + 5*x^2*y^2 + 2*x^2*y - x^2 + 3*x*y^4 - 5*x*y^2 - 2*x*y - 2*y^5 - 4*y^4 - 2*y^3

julia> L = equidimensional_hull(I)
ideal generated by: x^5 - 2*x^4*y - 2*x^4 + x^3*y^2 + 2*x^3*y - x^2*y^2 + 2*x^2*y + 2*x^2 + 2*x*y^3 + x*y^2 - 2*x*y - x - y^4 - 2*y^3 - y^2

julia> R, (a, b, c, d) = PolynomialRing(ZZ, ["a", "b", "c", "d"])
(Multivariate Polynomial Ring in a, b, c, d over Integer Ring, fmpz_mpoly[a, b, c, d])

julia> I = ideal(R, [1326*a^2*d^5, 1989*a^2*c^5, 102*b^4*d^5, 153*b^4*c^5,
       663*a^2*c^5*d^5, 51*b^4*c^5*d^5, 78*a^2*d^15, 117*a^2*c^15,
       78*a^15*d^5, 117*a^15*c^5, 6*a^2*b^4*d^15, 9*a^2*b^4*c^15,
       39*a^2*c^5*d^15, 39*a^2*c^15*d^5, 6*a^2*b^15*d^5, 9*a^2*b^15*c^5,
       6*a^15*b^4*d^5, 9*a^15*b^4*c^5, 39*a^15*c^5*d^5, 3*a^2*b^4*c^5*d^15,
       3*a^2*b^4*c^15*d^5, 3*a^2*b^15*c^5*d^5, 3*a^15*b^4*c^5*d^5])
ideal generated by: 1326*a^2*d^5, 1989*a^2*c^5, 102*b^4*d^5, 153*b^4*c^5, 663*a^2*c^5*d^5, 51*b^4*c^5*d^5, 78*a^2*d^15, 117*a^2*c^15, 78*a^15*d^5, 117*a^15*c^5, 6*a^2*b^4*d^15, 9*a^2*b^4*c^15, 39*a^2*c^5*d^15, 39*a^2*c^15*d^5, 6*a^2*b^15*d^5, 9*a^2*b^15*c^5, 6*a^15*b^4*d^5, 9*a^15*b^4*c^5, 39*a^15*c^5*d^5, 3*a^2*b^4*c^5*d^15, 3*a^2*b^4*c^15*d^5, 3*a^2*b^15*c^5*d^5, 3*a^15*b^4*c^5*d^5

julia> L = equidimensional_hull(I)
ideal generated by: 3

Radical of the Equidimensional Hull

Oscar.equidimensional_hull_radical — Method

equidimensional_hull_radical(I::MPolyIdeal)

Returns the intersection of the associated primes of I of maximal dimension. If I is the unit ideal, [ideal(1)] is returned. Uses a combination of the algorithms of Krick and Logar (with modifications by Laplagne) and Kemper.

Example

julia> R, (x, y) = PolynomialRing(QQ, ["x", "y"])
(Multivariate Polynomial Ring in x, y over Rational Field, fmpq_mpoly[x, y])

julia> I = intersect(ideal(R, [x, y])^2, ideal(R, [y^2-x^3+x]))
ideal generated by: x^3*y - x*y - y^3, x^4 - x^2 - x*y^2

julia> I = intersect(I, ideal(R, [x-y-1])^2)
ideal generated by: x^5*y - 2*x^4*y^2 - 2*x^4*y + x^3*y^3 + 2*x^3*y^2 - x^2*y^3 + 2*x^2*y^2 + 2*x^2*y + 2*x*y^4 + x*y^3 - 2*x*y^2 - x*y - y^5 - 2*y^4 - y^3, x^6 - 2*x^5 - 3*x^4*y^2 - 2*x^4*y + 2*x^3*y^3 + 3*x^3*y^2 + 2*x^3*y + 2*x^3 + 5*x^2*y^2 + 2*x^2*y - x^2 + 3*x*y^4 - 5*x*y^2 - 2*x*y - 2*y^5 - 4*y^4 - 2*y^3

julia> L = equidimensional_hull_radical(I)
ideal generated by: x^4 - x^3*y - x^3 - x^2 - x*y^2 + x*y + x + y^3 + y^2

Absolute Primary Decomposition

absolute_primary_decomposition(I)               --->  absPrimdecGTZ

Tests on Ideals

Equality of Ideals

Base.:== — Method

:(==)(I::MPolyIdeal, J::MPolyIdeal)

Returns true if I=J, false otherwise.

Example

julia> R, (x, y) = PolynomialRing(QQ, ["x", "y"])
(Multivariate Polynomial Ring in x, y over Rational Field, fmpq_mpoly[x, y])

julia> I = ideal(R, [x^2])
ideal generated by: x^2

julia> J = ideal(R, [x, y])^2
ideal generated by: x^2, x*y, y^2

julia> I == J
false

Containment of Ideals

Base.issubset — Method

issubset(I::MPolyIdeal, J::MPolyIdeal)

Returns true if I is contained in J, false otherwise.

Example

julia> R, (x, y) = PolynomialRing(QQ, ["x", "y"])
(Multivariate Polynomial Ring in x, y over Rational Field, fmpq_mpoly[x, y])

julia> I = ideal(R, [x^2])
ideal generated by: x^2

julia> J = ideal(R, [x, y])^2
ideal generated by: x^2, x*y, y^2

julia> issubset(I, J)
true

Ideal Membership

Base.in — Method

in(f::MPolyElem, I::MPolyIdeal)

Returns true if f is contained in J and false, otherwise.

Example

julia> R, (x, y) = PolynomialRing(QQ, ["x", "y"])
(Multivariate Polynomial Ring in x, y over Rational Field, fmpq_mpoly[x, y])

julia> f = x^2
x^2

julia> J = ideal(R, [x, y])^2
ideal generated by: x^2, x*y, y^2

julia> in(f, J)
true

Primality Test

Nemo.isprime — Method

isprime(I::MPolyIdeal)

Returns true if the ideal I is prime, false otherwise. Proceeds by computing a primary decomposition.

Example

julia> R, (x, y) = PolynomialRing(QQ, ["x", "y"])
(Multivariate Polynomial Ring in x, y over Rational Field, fmpq_mpoly[x, y])

julia> I = ideal(R, [x, y])^2
ideal generated by: x^2, x*y, y^2

julia> isprime(I)
false

Primary Test

Oscar.isprimary — Method

isprimary(I::MPolyIdeal)

Return true if the ideal I is primary, false otherwise. Proceeds by computing a primary decomposition.

Example

julia> R, (x, y) = PolynomialRing(QQ, ["x", "y"])
(Multivariate Polynomial Ring in x, y over Rational Field, fmpq_mpoly[x, y])

julia> I = ideal(R, [x, y])^2
ideal generated by: x^2, x*y, y^2

julia> isprimary(I)
true

homogeneity test:        ishomogeneous(I)

.....

Data Associated to Ideals

AbstractAlgebra.base_ring — Method

base_ring(I::MPolyIdeal{S}) where {S}

Returns the ambient ring of I.

Example

julia> R, (x, y) = PolynomialRing(QQ, ["x", "y"])
(Multivariate Polynomial Ring in x, y over Rational Field, fmpq_mpoly[x, y])

julia> I = ideal(R, [x, y])^2
ideal generated by: x^2, x*y, y^2

julia> base_ring(I)
Multivariate Polynomial Ring in x, y over Rational Field

Number of Generators

AbstractAlgebra.Generic.ngens — Method

ngens(I::MPolyIdeal)

Returns the number of generators of I.

Example

julia> R, (x, y) = PolynomialRing(QQ, ["x", "y"])
(Multivariate Polynomial Ring in x, y over Rational Field, fmpq_mpoly[x, y])

julia> I = ideal(R, [x, y])^2
ideal generated by: x^2, x*y, y^2

julia> ngens(I)
3

Generators

AbstractAlgebra.gens — Method

gens(I::MPolyIdeal)

Returns the generators of I as an array of multivariate polynomials.

Example

julia> R, (x, y) = PolynomialRing(QQ, ["x", "y"])
(Multivariate Polynomial Ring in x, y over Rational Field, fmpq_mpoly[x, y])

julia> I = ideal(R, [x, y])^2
ideal generated by: x^2, x*y, y^2

julia> gens(I)
3-element Array{fmpq_mpoly,1}:
 x^2
 x*y
 y^2

Dimension

AbstractAlgebra.Generic.dim — Method

dim(I::MPolyIdeal)

Returns the dimension of I.

Example

julia> R, (x, y, z) = PolynomialRing(QQ, ["x", "y", "z"])
(Multivariate Polynomial Ring in x, y, z over Rational Field, fmpq_mpoly[x, y, z])

julia> I = ideal(R, [y-x^2, x-z^3])
ideal generated by: -x^2 + y, x - z^3

julia> dim(I)
1

Codimension

Oscar.codim — Method

codim(I::MPolyIdeal)

Returns the codimension of I.

Example

julia> R, (x, y, z) = PolynomialRing(QQ, ["x", "y", "z"])
(Multivariate Polynomial Ring in x, y, z over Rational Field, fmpq_mpoly[x, y, z])

julia> I = ideal(R, [y-x^2, x-z^3])
ideal generated by: -x^2 + y, x - z^3

julia> codim(I)
2

Data Associated to Homogeneous Ideals

                         min_base(I)   (or so)
 
degree:                  degree(I)                             (type integer)
Hilbert function:        hilbert_function(d,I)                 (type integer)
Hilbert series:          hilbert_series(I)                     (type univariate rational function)  numerator, denominator
reduced Hilbert series:  reduced_Hilbert_series(I)             (type univariate rational function)
Hilbert polynomial:      hilbert_polynomial(I)                 (type univariate polynomial, direkt in Julia von hilbert_series(I))

Modules over Multivariate Polynomial Rings

Module Constructors

Simple Module Operations

Sum of Submodules

N_1+N_2

Products of Ideal and Submodule

I*N

Flatness

Depth and Codimension

Noether Normalization

Normalization of Rings

Oscar.normalize — Method

normalize(Q::MPolyQuo)

Given $Q$, a quotient by a radical ideal, return an array of size, say $r$, where each entry of the array is a triple $(S, M, f)$ corresponding to one of the $r$ components of $Q$ as found by Singular's algorithms. For each component, $S$ is a quotient ring giving the normalization, $M$ is a map from $Q$ to $S$, and $f$ is a fractional ideal of $Q$ giving the normalization. The fractional ideal $f$ is returned as a tuple $(d, i)$ where $d \in Q$ is the denominator, and $i$ is an ideal of $Q$.

Example

julia> R, (x, y, z) = PolynomialRing(QQ, ["x", "y", "z"])
(Multivariate Polynomial Ring in x, y, z over Rational Field, fmpq_mpoly[x, y, z])

julia> Q, _ = quo(R, ideal(R, [z - x^4, z - y^6]))
(Quotient of Multivariate Polynomial Ring in x, y, z over Rational Field by ideal generated by: -x^4 + z, -y^6 + z, Map from
Multivariate Polynomial Ring in x, y, z over Rational Field to Quotient of Multivariate Polynomial Ring in x, y, z over Rational Field by ideal generated by: -x^4 + z, -y^6 + z defined by a julia-function with inverse
)

julia> L = normalize(Q)
// ** loaded /home/fieker/.julia/artifacts/49e6c5baf692a764a6bb5d397d0fdb8e728eb669/bin/../share/singular/LIB/normal.lib (4.1.2.0,Feb_2019)
// ** loaded /home/fieker/.julia/artifacts/49e6c5baf692a764a6bb5d397d0fdb8e728eb669/bin/../share/singular/LIB/curveInv.lib (4.1.2.0,Feb_2019)
// ** loaded /home/fieker/.julia/artifacts/49e6c5baf692a764a6bb5d397d0fdb8e728eb669/bin/../share/singular/LIB/homolog.lib (4.1.2.0,Feb_2019)
// ** loaded /home/fieker/.julia/artifacts/49e6c5baf692a764a6bb5d397d0fdb8e728eb669/bin/../share/singular/LIB/deform.lib (4.1.2.0,Feb_2019)
// ** loaded /home/fieker/.julia/artifacts/49e6c5baf692a764a6bb5d397d0fdb8e728eb669/bin/../share/singular/LIB/algebra.lib (4.1.2.0,Feb_2019)
// ** loaded /home/fieker/.julia/artifacts/49e6c5baf692a764a6bb5d397d0fdb8e728eb669/bin/../share/singular/LIB/reesclos.lib (4.1.2.0,Feb_2019)
// ** loaded /home/fieker/.julia/artifacts/49e6c5baf692a764a6bb5d397d0fdb8e728eb669/bin/../share/singular/LIB/locnormal.lib (4.1.2.0,Feb_2019)
// ** loaded /home/fieker/.julia/artifacts/49e6c5baf692a764a6bb5d397d0fdb8e728eb669/bin/../share/singular/LIB/modstd.lib (4.1.2.0,Feb_2019)
// ** loaded /home/fieker/.julia/artifacts/49e6c5baf692a764a6bb5d397d0fdb8e728eb669/bin/../share/singular/LIB/parallel.lib (4.1.2.0,Feb_2019)
// ** loaded /home/fieker/.julia/artifacts/49e6c5baf692a764a6bb5d397d0fdb8e728eb669/bin/../share/singular/LIB/tasks.lib (4.1.2.0,Feb_2019)
// ** loaded /home/fieker/.julia/artifacts/49e6c5baf692a764a6bb5d397d0fdb8e728eb669/bin/../share/singular/LIB/resources.lib (4.1.2.0,Feb_2019)
// ** loaded /home/fieker/.julia/artifacts/49e6c5baf692a764a6bb5d397d0fdb8e728eb669/bin/../share/singular/LIB/modular.lib (4.2.0.0,Dec_2020)
// ** loaded /home/fieker/.julia/artifacts/49e6c5baf692a764a6bb5d397d0fdb8e728eb669/bin/../share/singular/LIB/hnoether.lib (4.1.2.0,Feb_2019)
// ** loaded /home/fieker/.julia/artifacts/49e6c5baf692a764a6bb5d397d0fdb8e728eb669/bin/../share/singular/LIB/primitiv.lib (4.1.2.0,Feb_2019)
// ** loaded /home/fieker/.julia/artifacts/49e6c5baf692a764a6bb5d397d0fdb8e728eb669/bin/../share/singular/LIB/presolve.lib (4.1.2.0,Feb_2019)
// ** loaded /home/fieker/.julia/artifacts/49e6c5baf692a764a6bb5d397d0fdb8e728eb669/bin/../share/singular/LIB/sing.lib (4.1.2.0,Feb_2019)

// 'normal' created a list, say nor, of two elements.
// To see the list type
      nor;

// * nor[1] is a list of 1 ring(s).
// To access the i-th ring nor[1][i], give it a name, say Ri, and type
     def R1 = nor[1][1]; setring R1; norid; normap;
// For the other rings type first (if R is the name of your base ring)
     setring R;
// and then continue as for R1.
// Ri/norid is the affine algebra of the normalization of R/P_i where
// P_i is the i-th component of a decomposition of the input ideal id
// and normap the normalization map from R to Ri/norid.

// * nor[2] is a list of 1 ideal(s). Let ci be the last generator
// of the ideal nor[2][i]. Then the integral closure of R/P_i is
// generated as R-submodule of the total ring of fractions by
// 1/ci * nor[2][i].
1-element Array{Tuple{Oscar.MPolyQuo{Singular.spoly{Singular.n_Q}},Oscar.MPolyQuoHom,Tuple{Oscar.MPolyQuoElem{fmpq_mpoly},Oscar.MPolyQuoIdeal{fmpq_mpoly}}},1}:
 (Quotient of Singular Polynomial Ring (QQ),(T(1),T(2),T(3),x,y,z),(dp(3),dp(3),C) by ideal generated by: -T(3)*y+x, T(1)*x-T(2)*y, -T(2)*x+y^2, T(1)*y^2-T(3)*x, T(2)*y^3-T(3)*x^2, -T(2)*x^2+T(3)*y^3, -T(1)*z+T(2)*x^3*y, T(3)*x^3*y-z, T(1)^2-1, T(1)*T(2)-T(3), T(2)^2-T(1)*y, T(1)*T(3)-T(2), T(2)*T(3)-y, T(3)^2-T(1)*y, x^4-z, y^6-z, AlgebraHomomorphism(Quotient of Multivariate Polynomial Ring in x, y, z over Rational Field by ideal generated by: -x^4 + z, -y^6 + z => Quotient of Singular Polynomial Ring (QQ),(T(1),T(2),T(3),x,y,z),(dp(3),dp(3),C) by ideal generated by: -T(3)*y+x, T(1)*x-T(2)*y, -T(2)*x+y^2, T(1)*y^2-T(3)*x, T(2)*y^3-T(3)*x^2, -T(2)*x^2+T(3)*y^3, -T(1)*z+T(2)*x^3*y, T(3)*x^3*y-z, T(1)^2-1, T(1)*T(2)-T(3), T(2)^2-T(1)*y, T(1)*T(3)-T(2), T(2)*T(3)-y, T(3)^2-T(1)*y, x^4-z, y^6-z, Singular.spoly{Singular.n_Q}[x, y, z]), (z, MPolyQuoIdeal(x^2*y^3, x^3*y^2, x*y^5, z)))

Properties of Rings

Normality

isnormal(R)

Cohen-Macaulayness

iscohenmacaulay(R)

Invariant Theory

Invariant Theory of Finite Groups

...