Polynomial Rings

Coefficients

Constructions

Gradings

Polynomial Rings

Referring to the sections on rings and fields for more details, we summarize how to construct multivariate polynomial rings. Also, we show how to decorate multivariate polynomial rings with gradings.

Coefficients

Multivariate polynomial rings over coefficient rings from the list below allow for Gröbner basis computations :

The field of rational numbers $\mathbb{Q}$

julia> QQ
Rational Field

Finite fields $\mathbb{F_p}$, $p$ a prime

julia> GF(3)
Galois field with characteristic 3

julia> GF(ZZ(2)^127 - 1)
Galois field with characteristic 170141183460469231731687303715884105727

Finite fields $\mathbb{F}_{p^n}$ with $p^n$ elements, $p$ a prime

julia> FiniteField(2, 70, "a")
(Finite field of degree 70 over F_2, a)

Simple algebraic extensions of $\mathbb{Q}$ or $\mathbb{F}_p$

julia> Qx, x = QQ["x"]
(Univariate Polynomial Ring in x over Rational Field, x)

julia> K, a = NumberField(x^2 + 1, "a")
(Number field over Rational Field with defining polynomial x^2 + 1, a)

julia> F = GF(3)
Galois field with characteristic 3

julia> Fx, x = F["x"]
(Univariate Polynomial Ring in x over Galois field with characteristic 3, x)

julia> K, a = FiniteField(x^2 + 1, "a")
(Finite field of degree 2 over F_3, a)

Purely transcendental extensions of $\mathbb{Q}$ or $\mathbb{F}_p$

julia> T, (s, t) = PolynomialRing(QQ, ["s", "t"]);

julia> QT = FractionField(T)
Fraction field of Multivariate Polynomial Ring in s, t over Rational Field

julia> parent(t)
Multivariate Polynomial Ring in s, t over Rational Field

julia> parent(1//t)
Fraction field of Multivariate Polynomial Ring in s, t over Rational Field

julia> T, t = PolynomialRing(GF(3), "t");

julia> QT = FractionField(T)
Fraction field of Univariate Polynomial Ring in t over Galois field with characteristic 3

The ring of integers $\mathbb{Z}$

julia> ZZ
Integer Ring

Constructions

The basic constructor for multivariate polynomial rings reads as follows:

PolynomialRing(C::Ring, v::Vector{String}; ordering=:lex)

Its use, together with that of a convenient constructor for handling variables with multi-indices, is indicated below:

Examples

julia> R, (x, y, z) = PolynomialRing(QQ, ["x", "y", "z"])
(Multivariate Polynomial Ring in x, y, z over Rational Field, fmpq_mpoly[x, y, z])

julia> base_ring(R)
Rational Field

julia> gens(R)
3-element Array{fmpq_mpoly,1}:
 x
 y
 z

julia> gen(R, 2)
y

julia> R[3]
z

julia> ngens(R)
3

julia> f = x^2+y*z
x^2 + y*z

julia> parent(f)
Multivariate Polynomial Ring in x, y, z over Rational Field

julia> typeof(f)
fmpq_mpoly

julia> S, (a,b) = PolynomialRing(ZZ, ["a", "b"])
(Multivariate Polynomial Ring in a, b over Integer Ring, fmpz_mpoly[a, b])

julia> g = a*b
a*b

julia> typeof(g)
fmpz_mpoly

julia> V = ["x[1]", "x[2]"]
2-element Array{String,1}:
 "x[1]"
 "x[2]"

julia> T, x = PolynomialRing(GF(3), V)
(Multivariate Polynomial Ring in x[1], x[2] over Galois field with characteristic 3, gfp_mpoly[x[1], x[2]])

julia> x
2-element Array{gfp_mpoly,1}:
 x[1]
 x[2]

julia> R, x, y, z = PolynomialRing(QQ, "x" => (1:3, 1:4), "y" => 1:2, "z" => (1:1, 1:1, 1:1))
(Multivariate Polynomial Ring in 15 variables x[1, 1], x[2, 1], x[3, 1], x[1, 2], ..., z[1, 1, 1] over Rational Field, fmpq_mpoly[x[1, 1] x[1, 2] x[1, 3] x[1, 4]; x[2, 1] x[2, 2] x[2, 3] x[2, 4]; x[3, 1] x[3, 2] x[3, 3] x[3, 4]], fmpq_mpoly[y[1], y[2]], fmpq_mpoly[z[1, 1, 1]])

julia> x
3×4 Array{fmpq_mpoly,2}:
 x[1, 1]  x[1, 2]  x[1, 3]  x[1, 4]
 x[2, 1]  x[2, 2]  x[2, 3]  x[2, 4]
 x[3, 1]  x[3, 2]  x[3, 3]  x[3, 4]

Gradings

The following functions allow one to assign gradings to multivariate polynomial rings:

grade — Method

grade(R::MPolyRing, W::Vector{Int})

Grade R by assigning weights to the variables according to the entries of W.

grade(R::MPolyRing)

Grade R by assigning weight 1 to each variable.

Examples

julia> R, (x, y, z) = PolynomialRing(QQ, ["x", "y", "z"])
(Multivariate Polynomial Ring in x, y, z over Rational Field, fmpq_mpoly[x, y, z])

julia> T, (x, y, z) = grade(R)
(Multivariate Polynomial Ring in x, y, z over Rational Field graded by 
  x -> [1]
  y -> [1]
  z -> [1], MPolyElem_dec{fmpq,fmpq_mpoly}[x, y, z])

julia> W = [1, 2, 3];

julia> S, (x, y, z) = grade(R, W)
(Multivariate Polynomial Ring in x, y, z over Rational Field graded by 
  x -> [1]
  y -> [2]
  z -> [3], MPolyElem_dec{fmpq,fmpq_mpoly}[x, y, z])

More directly, graded polynomial rings can be constructed as follows:

GradedPolynomialRing — Method

GradedPolynomialRing(C::Ring, V::Vector{String}, W::Vector{Int}; ordering=:lex)

Return a multivariate polynomial ring with weights assigned to the variables according to the entries of W.

GradedPolynomialRing(C::Ring, V::Vector{String}; ordering=:lex)

Return a multivariate polynomial ring with weight 1 assigned to each variable.

Examples

julia> R, (x, y, z) = GradedPolynomialRing(QQ, ["x", "y", "z"], [1, 2, 3])
(Multivariate Polynomial Ring in x, y, z over Rational Field graded by 
  x -> [1]
  y -> [2]
  z -> [3], MPolyElem_dec{fmpq,fmpq_mpoly}[x, y, z])

Note

The return type of PolynomialRing as well as that of the constructors for graded rings is a subtype of MPolyRing.