Gröbner and Standard Bases Over Fields

Division With Remainder

Gröbner and Standard Bases

Leading Ideals

Normal Forms

Syzygies

Gröbner and Standard Bases Over Fields

Let $K[x] = K[x_1, \dots, x_n]$ be a polynomial ring over a field $K$, and let $>$ be a monomial ordering on $\text{Mon}_n(x)$.

Given a polynomial $f\in K[x]$, write $f$ as the sum of its nonzero terms as follows:

\[f = a_\alpha x^\alpha + a_{\beta_1} x^{\beta_1} + \dots + a_{\beta_s} x^{\beta_s},\quad x^\alpha > x^{\beta_1} > \dots > x^{\beta_s} .\]

Then, with respect to $>$, we refer to $\text{LT}_>(f) := a_\alpha x^\alpha$, $\text{LM}_>(f) := x^\alpha$, $\text{LE}_>(f) := \alpha$, $\text{LC}_>(f) := a_\alpha$, and $\text{tail}_>(f) := f - \text{LT}_>(f)$ as the leading term, the leading monomial, the leading exponent, the leading coefficient, and the tail of $f$, respectively.

Next note that the set

\[U_>:= \{u\in K[x]\setminus \{0\} \mid {\text{LM}}_>(u)=1 \}\]

is a multiplicatively closed subset of $K[x]$. Consider the localization

\[K[x]_>:= K[x][U^{-1}] = \left\{ \frac{f}{u} \:\bigg|\: f \in K[x], \, u\in U_>\right\}.\]

Then $K[x]\subseteq K[x]_>\subseteq K[x]_{\langle x \rangle}.$ Moreover,

$K[x] = K[x]_>$ iff $>$ is global, and

$K[x]_> = K[x]_{\langle x \rangle}$ iff $>$ is local.

Extending the notation introduced for polynomials, let now $f\in K[x]_>$. Choose $u\in U_>$ such that $uf\in K[x]$. Then, with respect to $>$, the leading term of $f$ is defined to be $\text{LT}_>(f) := \text{LT}_>(uf)$ (this definition is independent of the choice of $u$). The leading monomial $\text{LM}_>(f)$, the leading exponent $\text{LE}_>(f)$, the leading coefficient $\text{LK}_>(f)$, and the tail $\text{tail}_>(f)$ of $f$ are defined similarly.

Note

We use the analogous notation for elements of free modules with a given monomial ordering.

Note

The OSCAR functions discussed in this section depend on a monomial ordering which may be entered as a keyword argument. For a polynomial ring $R$, the respective default ordering is degrevlex except if $R$ is $\mathbb Z$-graded with positive weights. Then the corresponding wdegrevlex ordering is used. For a free $R$-module $F$, the default ordering is default_ordering (R)*lex(gens(F)).

Here are some illustrating OSCAR examples:

Examples

julia> R, (x, y, z) = PolynomialRing(QQ, ["x", "y", "z"])(Multivariate Polynomial Ring in x, y, z over Rational Field, fmpq_mpoly[x, y, z])
julia> default_ordering(R)degrevlex([x, y, z])
julia> F = free_module(R, 2)Free module of rank 2 over Multivariate Polynomial Ring in x, y, z over Rational Field
julia> default_ordering(F)degrevlex([x, y, z])*lex([gen(1), gen(2)])
julia> S, _ = grade(R, [1, 2, 3])(Multivariate Polynomial Ring in x, y, z over Rational Field graded by
  x -> [1]
  y -> [2]
  z -> [3], MPolyElem_dec{fmpq, fmpq_mpoly}[x, y, z])
julia> default_ordering(S)wdegrevlex([x, y, z], [1, 2, 3])

julia> R, (x, y, z) = PolynomialRing(QQ, ["x", "y", "z"])(Multivariate Polynomial Ring in x, y, z over Rational Field, fmpq_mpoly[x, y, z])
julia> f = 3*z^3+2*x*y+12*x*y + 3*z^3 + 1

Division With Remainder

Buchberger's algorithm for computing Gröbner bases makes use of reduction steps which rely on multivariate division with remainder. For the computation of standard bases, a variant of the division algorithm due to Mora comes into play. In that context, with notation as above, we use the following terminology.

First suppose that $>$ is global and let polynomials $g\in K[x]$ and $f_1, \dots, f_r\in K[x]\setminus \{0\}$ be given. In this situation, multivariate division with remainder allows us to compute expressions

\[g = q_1f_1+\dots q_rf_r + h, \; h\in K[x], \;\text{ all }\; q_i \in K[x]\]

such that:

If both sides are nonzero, then $\text{LM}_>(g) \ge \text{LM}_>(q_if_i)$.

If $h$ is nonzero, then $\text{LM}_>(h)$ is not divisible by any $\text{LM}_>(f_i)$.

Each such expression is called a standard representation for $g$ with quotients $q_i$ and remainder $h$ (in terms of the $f_i$, with respect to $>$). We say that the remainder $h$ in a standard representation is fully reduced if it satisfies the stronger condition below:

If $h$ is nonzero, then no term of $h$ is divisible by any $\text{LM}_>(f_i)$.

Without restrictions on $>$, let elements $g\in K[x]_>$ and $f_1, \dots, f_r\in K[x]\setminus \{0\}$ be given. In this more general situation, Mora division with remainder allows us to compute expressions

\[ug = q_1f_1+\dots q_rf_r + h, \; h\in K[x]_>, \;\text{ all }\; q_i \in K[x]_>\]

such that:

$u$ is a unit of $ K[x]_>$.

If both sides are nonzero, then $\text{LM}_>(g) \ge \text{LM}_>(q_if_i)$.

If $h$ is nonzero, then $\text{LM}_>(h)$ is not divisible by any $\text{LM}_>(f_i)$.

Each such expression is called a weak standard representation for $g$ with quotients $q_i$ and remainder $h$ (in terms of the $f_i$, with respect to $>$). If $g\in K[x])$, we speak of a polynomial weak standard representation if $u$ and the $q_i$ are elements of $K[x].$ While from a theoretical point of view it makes still sense to speak of fully reduced remainders, such remainders may not be computable (in finitely many steps).

The OSCAR functions discussed below compute standard representations and polynomial weak standard representations, respectively.

reduce(I::IdealGens, J::IdealGens; ordering::MonomialOrdering = default_ordering(base_ring(J)))

reduce(f::T, F::Vector{T}; ordering::MonomialOrdering = default_ordering(parent(f))) where {T <: MPolyElem}

reduce(F::Vector{T}, G::Vector{T}; ordering::MonomialOrdering = default_ordering(parent(F[1]))) where {T <: MPolyElem}

reduce_with_quotients(I::IdealGens, J::IdealGens; ordering::MonomialOrdering = default_ordering(base_ring(J)))

reduce_with_quotients_and_units(I::IdealGens, J::IdealGens; ordering::MonomialOrdering = default_ordering(base_ring(J)))

Gröbner and Standard Bases

Still keeping the notation introduced at the beginning of this section, let $I$ be an ideal of $K[x]_>$. Then the leading ideal of $I$ is the ideal of $K[x]$ defined by

\[\text{L}_>(I):=\langle \text{LT}_>(f) \mid f\in I\rangle\subset K[x].\]

A standard basis of $I$ is a finite set of elements of $I$ such that the leading terms of these elements generate $\text{L}_>(I)$. A finite subset of $K[x]_>$ is a standard basis if it is a standard basis for the ideal it generates.

We call a standard basis $\{f_1,\dots, f_r\}\subset K[x]_>$ minimal if $\text{LM}_>(f_i)\neq \text{LM}_>(f_j)$, for $i\neq j$. We call it reduced if it is minimal and no term of $f_i$ is divisible by $\text{LM}_>(f_j)$, for $i\neq j$.

Note

The above definitions deviate from the usual textbook definitions as we do not ask that the leading coefficients of the standard basis elements are 1.

Note

The standard bases returned by OSCAR are always minimal in the sense above.

If $>$ is global, that is, $K[x] = K[x]_>$, then we use the expression Gröbner basis instead of standard basis. Note that reduced Gröbner bases can always be computed, but reduced standard bases may not be computatable (in finitely many steps).

Note

We use the analogous notation for submodules of free modules with a given monomial ordering.

groebner_basis — Method

groebner_basis(I::MPolyIdeal;
  ordering::MonomialOrdering = default_ordering(base_ring(I)),
  complete_reduction::Bool = false)

Given an ideal I and optional parameters monomial ordering ordering and complete_reduction, compute a Groebner basis (if complete_reduction = true the reduced Groebner basis) of I w.r.t. the given monomial ordering ordering (as default degree reverse lexicographical).

Examples

julia> R, (x, y) = PolynomialRing(QQ, ["x", "y"])
(Multivariate Polynomial Ring in x, y over Rational Field, fmpq_mpoly[x, y])

julia> I = ideal([x*y-3*x,y^3-2*x^2*y])
ideal(x*y - 3*x, -2*x^2*y + y^3)

julia> H = groebner_basis(I, ordering=lex(R))
Gröbner basis with elements
1 -> y^4 - 3*y^3
2 -> x*y - 3*x
3 -> 6*x^2 - y^3
with respect to the ordering
lex([x, y])

standard_basis — Method

standard_basis(I::MPolyIdeal;
  ordering::MonomialOrdering = default_ordering(base_ring(I)),
  complete_reduction::Bool = false)

Compute a standard basis of I for the monomial ordering ordering.

Note: Since there is no general notion of complete reduction for non-global orderings, this option is not available for this command; instead, use groebner_basis directly.

Examples

julia> R,(x,y) = PolynomialRing(QQ, ["x","y"])
(Multivariate Polynomial Ring in x, y over Rational Field, fmpq_mpoly[x, y])

julia> I = ideal([x*(x+1), x^2-y^2+(x-2)*y])
ideal(x^2 + x, x^2 + x*y - y^2 - 2*y)

julia> standard_basis(I, ordering=negdegrevlex(R))
Standard basis with elements
1 -> x
2 -> y
with respect to the ordering
negdegrevlex([x, y])

Gröbner Bases with transformation matrix

groebner_basis_with_transformation_matrix — Method

groebner_basis_with_transformation_matrix(I::MPolyIdeal;
  ordering::MonomialOrdering = default_ordering(base_ring(I)),
  complete_reduction::Bool=false)

Return a pair G, m where G is a Groebner basis of the ideal I with respect to the monomial ordering ordering (as default degree reverse lexicographical), and m is a transformation matrix from gens(I)toG. If complete_reductionis set to true (as default 'false') then G will be the reduced Groebner basis.

Examples

julia> R,(x,y) = PolynomialRing(QQ,["x","y"])
(Multivariate Polynomial Ring in x, y over Rational Field, fmpq_mpoly[x, y])

julia> I = ideal([x*y^2-1,x^3+y^2+x*y])
ideal(x*y^2 - 1, x^3 + x*y + y^2)

julia> G,m = groebner_basis_with_transformation_matrix(I)
(fmpq_mpoly[x*y^2 - 1, x^3 + x*y + y^2, x^2 + y^4 + y], [1 0 -x^2-y; 0 1 y^2])

julia> gens(I)*m == G
true

fglm

Gröbner walks

Hilbert-driven

Expert functions for Gröbner bases

The following functions are low-level implementations of various Gröbner basis algorithms with many adjustable arguments. Only use these functions directly if you know what you are doing.

f4 — Method

f4(I::MPolyIdeal, <keyword arguments>)

Compute a Groebner basis of the given ideal I w.r.t. to the degree reverse lexicographical monomial ordering using Faugère's F4 algorithm. See Jean-Charles Faugère (1999) for more information.

Note: At the moment only ground fields of characteristic p, p prime, p < 2^{31} are supported.

Arguments

Ì::MPolyIdeal: input ideal.

initial_hts::Int=17: initial hash table size log_2.

nr_thrds::Int=1: number of threads for parallel linear algebra.

max_nr_pairs::Int=0: maximal number of pairs per matrix, only bounded by minimal degree if 0.

la_option::Int=2: linear algebra option: exact sparse-dense (1), exact sparse (2, default), probabilistic sparse-dense (42), probabilistic sparse(44).

eliminate::Int=0: size of first block of variables to be eliminated.

complete_reduction::Bool=true: compute a reduced Gröbner basis for I

info_level::Int=0: info level printout: off (0, default), summary (1), detailed (2).

Examples

julia> R,(x,y,z) = PolynomialRing(GF(101), ["x","y","z"], ordering=:degrevlex)
(Multivariate Polynomial Ring in x, y, z over Galois field with characteristic 101, gfp_mpoly[x, y, z])

julia> I = ideal(R, [x+2*y+2*z-1, x^2+2*y^2+2*z^2-x, 2*x*y+2*y*z-y])
ideal(x + 2*y + 2*z + 100, x^2 + 2*y^2 + 2*z^2 + 100*x, 2*x*y + 2*y*z + 100*y)

 
julia> f4(I)
Gröbner basis with elements
1 -> x + 2*y + 2*z + 100
2 -> y*z + 82*z^2 + 10*y + 40*z
3 -> y^2 + 60*z^2 + 20*y + 81*z
4 -> z^3 + 28*z^2 + 64*y + 13*z
with respect to the ordering
degrevlex([x, y, z])

julia> isdefined(I, :gb)
true

Leading Ideals

leading_ideal — Method

leading_ideal(g::Vector{T}; ordering::MonomialOrdering) where { T <: MPolyElem }

Return the ideal generated by the leading monomials of the given polynomials w.r.t. the given monomial ordering.

Examples

julia> R, (x, y) = PolynomialRing(QQ, ["x", "y"])
(Multivariate Polynomial Ring in x, y over Rational Field, fmpq_mpoly[x, y])

julia> L = leading_ideal([x*y^2-3*x, x^3-14*y^5], ordering=degrevlex(R))
ideal(x*y^2, y^5)

julia> L = leading_ideal([x*y^2-3*x, x^3-14*y^5], ordering=lex(R))
ideal(x*y^2, x^3)

leading_ideal — Method

leading_ideal(I::MPolyIdeal; ordering::MonomialOrdering)

Given a multivariate polynomial ideal Ì this function returns the leading ideal for I. This is done w.r.t. the given monomial ordering.

Examples

julia> R, (x, y) = PolynomialRing(QQ, ["x", "y"])
(Multivariate Polynomial Ring in x, y over Rational Field, fmpq_mpoly[x, y])

julia> I = ideal(R,[x*y^2-3*x, x^3-14*y^5])
ideal(x*y^2 - 3*x, x^3 - 14*y^5)

julia> L = leading_ideal(I, ordering=degrevlex(R))
ideal(x*y^2, x^4, y^5)

julia> L = leading_ideal(I, ordering=lex(R))
ideal(y^7, x*y^2, x^3)

Normal Forms

normal_form — Method

normal_form(f::T, J::MPolyIdeal) where { T <: MPolyElem }

Compute the normal form of the polynomial f w.r.t. a Groebner basis of J.

CAVEAT: This computation needs a Groebner basis of J. If this Groebner basis is not available, one is computed automatically. This may take some time.

Examples

julia> R,(a,b,c) = PolynomialRing(QQ,["a","b","c"])
(Multivariate Polynomial Ring in a, b, c over Rational Field, fmpq_mpoly[a, b, c])

julia> J = ideal(R,[-1+c+b,-1+b+c*a+2*a*b])
ideal(b + c - 1, 2*a*b + a*c + b - 1)

julia> gens(groebner_basis(J))
2-element Vector{fmpq_mpoly}:
 b + c - 1
 a*c - 2*a + c

julia> normal_form(-1+c+b+a^3, J)
a^3

normal_form — Method

normal_form(A::Vector{T}, J::MPolyIdeal) where { T <: MPolyElem }

Compute the normal form of the elements of the array A w.r.t. a Groebner basis of J.

CAVEAT: This computation needs a Groebner basis of J. If this Groebner basis is not available, one is computed automatically. This may take some time.

Examples

julia> R,(a,b,c) = PolynomialRing(QQ,["a","b","c"])
(Multivariate Polynomial Ring in a, b, c over Rational Field, fmpq_mpoly[a, b, c])

julia> A = [-1+c+b+a^3,-1+b+c*a+2*a^3,5+c*b+c^2*a]
3-element Vector{fmpq_mpoly}:
 a^3 + b + c - 1
 2*a^3 + a*c + b - 1
 a*c^2 + b*c + 5

julia> J = ideal(R,[-1+c+b,-1+b+c*a+2*a*b])
ideal(b + c - 1, 2*a*b + a*c + b - 1)

julia> gens(groebner_basis(J))
2-element Vector{fmpq_mpoly}:
 b + c - 1
 a*c - 2*a + c

julia> normal_form(A, J)
3-element Vector{fmpq_mpoly}:
 a^3
 2*a^3 + 2*a - 2*c
 4*a - 2*c^2 - c + 5

Syzygies

syzygy_generators — Method

syzygy_generators(a::Vector{<:MPolyElem})

Return generators for the syzygies on the given polynomials.

Examples

julia> R, (x, y) = PolynomialRing(QQ, ["x", "y"])
(Multivariate Polynomial Ring in x, y over Rational Field, fmpq_mpoly[x, y])

julia> S = syzygy_generators([x^3+y+2,x*y^2-13*x^2,y-14])
3-element Vector{FreeModElem{fmpq_mpoly}}:
 (-y + 14)*e[2] + (-13*x^2 + x*y^2)*e[3]
 (-169*y + 2366)*e[1] + (-13*x*y + 182*x - 196*y + 2744)*e[2] + (13*x^2*y^2 - 2548*x^2 + 196*x*y^2 + 169*y + 338)*e[3]
 (-13*x^2 + 196*x)*e[1] + (-x^3 - 16)*e[2] + (x^4*y + 14*x^4 + 13*x^2 + 16*x*y + 28*x)*e[3]