Localizations of commutative rings

Suppose $R$ is a commutative ring with unit and $S \subset R$ is a multiplicatively closed set containing $1 \in R$. Then we can form the localization of $R$ at $S$

\[ R[S^{-1}] = \left\{ \frac{p}{q} : p,q \in R, \, q \in S \right\},\]

with its standard arithmetic for fractions. See, for instance, [Eis95] for an account on localizations.

Oscar provides a general framework for such localizations, originally intended to be used with multivariate polynomial rings $R$ over some base field $\mathbb k$, but also applicable to more general commutative rings.

In the case of polynomials, the localization framework provides the structure for certain algorithms using standard bases. Note that, in general, localizations of polynomial algebras are not finitely generated as algebras over $\mathbb k$; for instance when localizing at some maximal ideal $\mathfrak m \subset R$. However, many ideal- and module-theoretic questions in the localization $R[S^{-1}]$, such as e.g. the ideal membership, can be transformed to questions on ideals and modules over the base ring $R$ and then solved using Groebner- or standard-basis techniques. This makes it important to regard localizations $R[S^{-1}]$ as rings with a history of creation from the original pair $S \subset R$.

The localization interface

Localized rings

The interface that needs to be implemented for any concrete instance of localized rings is the following. Multiplicatively closed sets are derived from the abstract type

AbsMultSet — Type

AbsMultSet{RingType, RingElemType}

Abstract type for a multiplicatively closed set in a commutative (Noetherian) ring R of type RingType with elements of type RingElemType.

The basic functionality that has to be implemented for any concrete type derived from this is to be able to check containment of elements via

in — Method

in(f::RingElemType, S::AbsMultSet{RingType, RingElemType}) where {RingType, RingElemType}

Returns true if f belongs to S; false otherwise.

Note: If this routine is not implemented, the function call will default to the execution of an error message.

This is supposed to be an extension of the methods of the function Base.in.

A localized ring should then be derived from

AbsLocalizedRing — Type

AbsLocalizedRing{RingType, RingElemType, MultSetType}

The localization R[S⁻¹] of a ring R of type RingType with elements of type RingElemType at a multiplicatively closed set S of type MultSetType.

In general, the arithmetic of such a localized ring R[S⁻¹] should be implemented using fractions of elements in the original ring R. The methods provided for the multiplicatively closed set S can be used to check whether a given denominator is admissible for the specific localization.

Depending on the actual type of R and S, further functionality can then be provided using different Groebner-basis driven backends.

The basic way to construct localized rings is to first specify a multiplicative set S and then call

Localization — Method

Localization(S::AbsMultSet)

Returns the localization of the ambient_ring of S at S itself.

This method must be implemented with a dispatch depending on the concrete type of S.

For any concrete instance of type AbsLocalizedRing the following methods must be implemented:

base_ring — Method

base_ring(W::AbsLocalizedRing)

Returns the base ring R for a localized ring of the form W = R[S⁻¹].

inverted_set — Method

inverted_set(W::AbsLocalizedRing)

Returns the set S of at which has been localized for a localized ring W = R[S⁻¹].

Also, conversion of fractions to elements of localized rings must be implemented in the form (W::AbsLocalizedRing{RingType, RingElemType, MultSetType})(a::RingElemType) where {RingType, RingElemType, MultSetType}, taking $a$ to the element $\frac{a}{1}$. For more general fractions one needs (W::AbsLocalizedRing{RingType, RingElemType, MultSetType})(a::RingElemType, b::RingElemType) where {RingType, RingElemType, MultSetType}, mapping a pair $(a, b)$ to the fraction $\frac{a}{b}$.

The elements of localized rings must be derived from

AbsLocalizedRingElem — Type

AbsLocalizedRingElem{RingType, RingElemType, MultSetType}

The abstract type of an element of the localization R[S⁻¹] of a commutative ring R of type RingType with elements of type RingElemType at a multiplicatively closed set S of type MultSetType.

For any concrete instance F of AbsLocalizedRingElem there must be the following methods:

numerator — Method

numerator(f::AbsLocalizedRingElem)

Returns the numerator of f.

denominator — Method

denominator(f::AbsLocalizedRingElem)

Returns the denominator of f.

parent — Method

parent(f::AbsLocalizedRingElem)

Returns the parent ring R[S⁻¹] of f.

A default version of the arithmetic is implemented on the generic level using the above functionality and the arithmetic for the original ring. Depending on the actual concrete instance, one might wish to provide more fine-tuned methods, starting e.g. by implementing

reduce_fraction — Method

reduce_fraction(a::AbsLocalizedRingElem)

Reduce the fraction a = p/q. Warning: The catchall-implementation does nothing!

Note that this is called after every arithmetic operation (addition, multiplication,...), so the computations carried out here should be computationally cheap.

Two toy examples for implementations of this interface for localizations of the integers $\mathbb Z$ and rings of the form $\mathbb Z/n\mathbb Z$ can be found in the test files test/Rings/integer-localizations.jl and test/Rings/nmod-localizations.jl.

Note: Any concrete type for localized rings is also required to implement the general Ring Interface of Oscar! This has not been done to a full extent for the previous two examples, but for MPolyLocalizedRing; see below.

Ideals in localized rings

One of the main reasons to implement localizations in the first place is that this process preserves the property of a ring to be Noetherian; which is crucial for computer algebra. In this regard, we have

AbsLocalizedIdeal — Type

AbsLocalizedIdeal{RingType, RingElemType, MultSetType}

Abstract type for finitely generated ideals $I ⊂ R[S⁻¹]$ in localized rings.

The required getter methods are

gens — Method

gens(A::AbsAlgAss, return_full_basis::Typel{Val{T}} = Val{false};
     thorough_search::Bool = false) where T
  -> Vector{AbsAlgAssElem}

Returns a subset of basis(A), which generates $A$ as an algebra over base_ring(A). If return_full_basis is set to Val{true}, the function also returns a Vector{AbsAlgAssElem} containing a full basis consisting of monomials in the generators and a Vector{Vector{Tuple{Int, Int}}} containing the information on how these monomials are built. E. g.: If the function returns g, full_basis and v, then we have full_basis[i] = prod( g[j]^k for (j, k) in v[i] ). If thorough_search is true, the number of returned generators is possibly smaller. This will in general increase the runtime. It is not guaranteed that the number of generators is minimal in any case.

gens(A::AlgGrp, return_full_basis::Type{Val{T}} = Val{false})
  -> Vector{AlgGrpElem}

Returns a subset of basis(A), which generates $A$ as an algebra over base_ring(A). If return_full_basis is set to Val{true}, the function also returns a Vector{AbsAlgAssElem} containing a full basis consisting of monomials in the generators and a Vector{Vector{Tuple{Int, Int}}} containing the information on how these monomials are built. E. g.: If the function returns g, full_basis and v, then we have full_basis[i] = prod( g[j]^k for (j, k) in v[i] ).

gens(M::SubQuo{T}) where T

Return the generators of M.

gens(I::sideal)

Return the generators in the internal representation of the ideal $I$ as an array.

base_ring — Method

base_ring(I::MPolyIdeal{S}) where {S}

Return the ambient ring of I.

Examples

julia> R, (x, y) = PolynomialRing(QQ, ["x", "y"])
(Multivariate Polynomial Ring in x, y over Rational Field, fmpq_mpoly[x, y])

julia> I = ideal(R, [x, y])^2
ideal(x^2, x*y, y^2)

julia> base_ring(I)
Multivariate Polynomial Ring in x, y over Rational Field

The constructors to be implemented are

   ideal(W::AbsLocalizedRing{RingType, RingElemType, MultSetType}, f::AbsLocalizedRingElem{RingType, RingElemType, MultSetType})
   ideal(W::AbsLocalizedRing{RingType, RingElemType, MultSetType}, f::RingElemType)
   ideal(W::AbsLocalizedRing{RingType, RingElemType, MultSetType}, v::Vector{AbsLocalizedRingElem{RingType, RingElemType, MultSetType}})
   ideal(W::AbsLocalizedRing{RingType, RingElemType, MultSetType}, v::Vector{RingElemType})

for a single and a list of generators from both the base_ring of W and from W itself.

The minimal functionality which should be implemented for ideals is the test for ideal membership

Base.in(
    f::AbsLocalizedRingElem{RingType, RingElemType, MultSetType}, 
    I::AbsLocalizedIdeal{RingType, RingElemType, MultSetType}
  ) where {RingType, RingElemType, MultSetType}

and again the same for elements f of type RingElemType.

Basic operations on ideals which are already implemented on the generic level are

Base.:*(I::T, J::T) where {T<:AbsLocalizedIdeal}
Base.:+(I::T, J::T) where {T<:AbsLocalizedIdeal}

Everything else, such as e.g. intersections of ideals, has to be implemented for the specific types by the user.

Localizations of multivariate polynomial rings

The available types of multiplicative sets are

MPolyComplementOfPrimeIdeal — Type

MPolyComplementOfPrimeIdeal{ BaseRingType, BaseRingElemType, RingType, RingElemType } <: AbsMultSet{ RingType, RingElemType }

The complement of a prime ideal P ⊂ 𝕜[x₁,…,xₙ] in a multivariate polynomial ring with elements of type RingElemType over a base ring 𝕜 of type BaseRingType.

MPolyComplementOfKPointIdeal — Type

MPolyComplementOfKPointIdeal{ BaseRingType, BaseRingElemType, RingType, RingElemType } <: AbsMultSet{ RingType, RingElemType }

Complement of a maximal ideal $𝔪 = ⟨x₁-a₁,…,xₙ-aₙ⟩⊂ 𝕜[x₁,…xₙ]$ with $aᵢ∈ 𝕜$.

MPolyPowersOfElement — Type

MPolyPowersOfElement{ BaseRingType, BaseRingElemType, RingType, RingElemType } <: AbsMultSet{ RingType, RingElemType }

The set S = { aᵏ : k ∈ ℕ₀ } for some $a ∈ R$ with $R$ of type BaseRingType.

Localizations of multivariate polynomial rings are of type

MPolyLocalizedRing — Type

MPolyLocalizedRing{ BaseRingType, BaseRingElemType, RingType, RingElemType, MultSetType } <: AbsLocalizedRing{ RingType, RingType, MultSetType }

The localization of a multivariate polynomial ring $R = 𝕜[x₁,…,xₙ]$ over a base field $𝕜$ of type BaseRingType and with elements of type RingElemType at a multiplicative set $S ⊂ R$ of type MultSetType.

with elements of type

MPolyLocalizedRingElem — Type

MPolyLocalizedRingElem{ BaseRingType, BaseRingElemType, RingType, RingElemType, MultSetType } <: AbsLocalizedRingElem{ RingType, RingElemType, MultSetType }

Elements of localizations of polynomial rings.

In order to make the Singular functionality available there is

LocalizedBiPolyArray — Type

LocalizedBiPolyArray{BRT, BRET, RT, RET, MST}

Main workhorse for binding of ideals in localizations $R[S⁻¹]$ of multivariate polynomial rings $R = 𝕜[x₁,…,xₙ]$ to Singular.

which has a monomial ordering and a Singular ring associated to it. This is also the type returned by any Groebner- or standard basis computation.

Ideals in localizations of multivariate polynomial rings are of the type

MPolyLocalizedIdeal — Type

MPolyLocalizedIdeal{BRT, BRET, RT, RET, MST} <: AbsLocalizedIdeal{RT, RET, MST}

Ideals in localizations of polynomial rings.

and can be constructed using ideal(W::MPolyLocalizedRing, ...), either from single elements or a list of elements from both the localized ring or its base ring before localization.

Groebner and standard bases of ideals can be computed for explicit orderings using

    groebner_basis(I::MPolyLocalizedIdeal, ord::Symbol)

Note that depending on the type parameters of I, this method is dispatched differently, which will also lead to different interpretations of the ordering. For instance, for multiplicative sets of type MPolyComplementOfKPointIdeal, a shift of variables taking the geometric point to the origin is applied to all polynomials when passing to the singular side.

If the second argument is omitted, a default ordering will be chosen, depending on the type of the multiplicative set.