Modules Over Polynomial Rings

Types

Constructors

Data Associated to Subqotients

Elements of Subqotients

Tests on Subqotients

Basic Operations on Subqotients

Submodules and Quotients

Homomorphisms of Subqotients

Operations on Homomorphisms of Subqotients

Subquotients Related to Homomorphisms

Presentations

Syzygies and Free Resolutions

Hom and Ext

Tensorproduct and Tor

Modules Over Polynomial Rings

In this section, the term module will refer to a finitely presented module over a multivariate polynomial ring. In OSCAR, the most general way of implementing such a module is that of a subquotient, that is, as a submodule of a quotient of a free module. Explicitly, a subquotient $M$ over the ring $R$ is a module of type

\[M = (\text{im } a + \text{im } b)/\text{im } b,\]

where

\[a:R^s ⟶R^p \;\text{ and }\; b:R^t ⟶R^p\]

are two homomorphisms of free $R$-modules with the same codomain. We then refer to

the module $M$ as the subquotient defined by $a$ and $b$,

the codomain $R^p$ as the ambient free module of $M$,

the images of the canonical basis vectors of $R^s$ in $M$ as the generators of $M$, and

the images of the canonical basis vectors of $R^t$ in $R^p$ as the relations of $M$.

Alternatively, we speak of the subquotient of $\text{im } a$ by $\text{im } b$ or the subquotient defined by $A$ and $B$, where $A$ and $B$ are the matrices representing $a$ and $b$, respectively.

Note

Recall from the section on free modules that by a free $R$-module we mean a free module of type $R^p$ , where we think of $R^p$ as a free module with a given basis, namely the basis of standard unit vectors. Accordingly, elements of free modules are represented by coordinate vectors, and homomorphisms between free modules by matrices. Here, by convention, vectors are row vectors, and matrices operate by multiplication on the right.

Types

The abstract supertype for all finitely presented modules over commutative rings in OSCAR is ModuleFP{T}. For subquotients, OSCAR provides the abstract type AbstractSubQuo{T} <: ModuleFP{T} and its concrete descendant SubQuo{T}.

Constructors

subquotient — Method

subquotient(a::FreeModuleHom{T}, b::FreeModuleHom{T}) where T

Given free module homomorphisms a and b with the same codomain, return $(\text{im } a + \text{im } b)/\text{im } b$.

subquotient(F::FreeMod{T}, A::MatElem{T}, B::MatElem{T}) where T

Given matrices A and B with rank F columns, return $(\text{im } a + \text{im } b)/\text{im } b$, where a and b are the free module homomorphisms with codomain F represented by A and B.

subquotient(A::MatElem{T}, B::MatElem{T}) where T

Given matrices A and B with the same number of columns, create a free module, F, whose rank is that number, and return subquotient(F, A, B).

Examples

julia> R, (x, y, z) = PolynomialRing(QQ, ["x", "y", "z"])
(Multivariate Polynomial Ring in x, y, z over Rational Field, fmpq_mpoly[x, y, z])

julia> A = R[x; y]
[x]
[y]

julia> B = R[x^2; x*y; y^2; z^4]
[x^2]
[x*y]
[y^2]
[z^4]

julia> M = subquotient(A, B)
Subquotient of Submodule with 2 generators
1 -> x*e[1]
2 -> y*e[1]
by Submodule with 4 generators
1 -> x^2*e[1]
2 -> x*y*e[1]
3 -> y^2*e[1]
4 -> z^4*e[1]

Data Associated to Subqotients

If M is a subquotient with ambient free R-module F, then

base_ring(M) refers to R,

ambient_free_module(M) to F,

gens(M) to the generators of M,

ngens(M) to the number of these generators,

M[i], gen(M, i) to the ith such generator, and

rels(M) to the relations of M.

Examples

julia> R, (x, y, z) = PolynomialRing(QQ, ["x", "y", "z"])(Multivariate Polynomial Ring in x, y, z over Rational Field, fmpq_mpoly[x, y, z])
julia> A = R[x; y][x]
[y]
julia> B = R[x^2; x*y; y^2; z^4][x^2]
[x*y]
[y^2]
[z^4]
julia> M = SubQuo(A, B)Subquotient of Submodule with 2 generators
1 -> x*e[1]
2 -> y*e[1]
by Submodule with 4 generators
1 -> x^2*e[1]
2 -> x*y*e[1]
3 -> y^2*e[1]
4 -> z^4*e[1]
julia> base_ring(M)Multivariate Polynomial Ring in x, y, z over Rational Field
julia> ambient_free_module(M)Free module of rank 1 over Multivariate Polynomial Ring in x, y, z over Rational Field
julia> gens(M)2-element Vector{SubQuoElem{fmpq_mpoly}}:
 x*e[1]
 y*e[1]
julia> ngens(M)2
julia> gen(M, 2)y*e[1]
julia> rels(M)4-element Vector{FreeModElem{fmpq_mpoly}}:
 x^2*e[1]
 x*y*e[1]
 y^2*e[1]
 z^4*e[1]

Elements of Subqotients

The abstract supertype for all elements of finitely presented modules over commutative rings in OSCAR is ModuleFPElem{T}. The abstract type for elements of subquotients is AbstractSubQuoElem{T} <: ModuleFPElem{T}. Its concrete descendant SubQuoElem{T} implements an element $m$ of a subquotient $M$ over the ring $R$ as a sparse row, that is, as an object of type SRow{T}. This object specifies the coefficients of an $R$-linear combination of the generators of $M$ which gives $m$. To create an element, enter the coefficients as an object of type SRow{T} or Vector{T}:

(M::SubQuo{T})(c::SRow{T}) where T

(M::SubQuo{T})(c::Vector{T}) where T

Alternatively, directly write the element as an $R$-linear combination of generators of $M$.

Examples

julia> R, (x, y, z) = PolynomialRing(QQ, ["x", "y", "z"])(Multivariate Polynomial Ring in x, y, z over Rational Field, fmpq_mpoly[x, y, z])
julia> A = R[x; y][x]
[y]
julia> B = R[x^2; x*y; y^2; z^4][x^2]
[x*y]
[y^2]
[z^4]
julia> M = SubQuo(A, B)Subquotient of Submodule with 2 generators
1 -> x*e[1]
2 -> y*e[1]
by Submodule with 4 generators
1 -> x^2*e[1]
2 -> x*y*e[1]
3 -> y^2*e[1]
4 -> z^4*e[1]
julia> m = M(sparse_row(R, [(1,z),(2,one(R))]))(x*z + y)*e[1]
julia> n = M([z, one(R)])(x*z + y)*e[1]
julia> o = z*M[1] + M[2](x*z + y)*e[1]
julia> m == n == otrue

Given an element m of a subquotient M,

parent(m) refers to M,

coefficients(m) to an object of type SRow{T} specifying the coefficients of an $R$-linear combination of the generators of $M$ which gives $m$, and

ambient_representative(m) to an element of the ambient free module of M which represents m.

Given an element f of the ambient free module of a subquotient M such that f represents an element of M, the function below creates the represented element:

(M::SubQuo{T})(f::FreeModElem{T}; check::Bool = true) where T

By default (check = true), it is tested whether f indeed represents an element of M. If this is already clear, it may be convenient to omit the test (check = false).

Examples

julia> R, (x, y, z) = PolynomialRing(QQ, ["x", "y", "z"]);
julia> A = R[x; y][x]
[y]
julia> B = R[x^2; x*y; y^2; z^4][x^2]
[x*y]
[y^2]
[z^4]
julia> M = SubQuo(A, B);
julia> m = z*M[1] + M[2](x*z + y)*e[1]
julia> parent(m)Subquotient of Submodule with 2 generators
1 -> x*e[1]
2 -> y*e[1]
by Submodule with 4 generators
1 -> x^2*e[1]
2 -> x*y*e[1]
3 -> y^2*e[1]
4 -> z^4*e[1]
julia> coefficients(m)Sparse row with positions [1, 2] and values fmpq_mpoly[z, 1]
julia> fm = ambient_representative(m)(x*z + y)*e[1]
julia> typeof(m)SubQuoElem{fmpq_mpoly}
julia> typeof(fm)FreeModElem{fmpq_mpoly}
julia> parent(fm) == ambient_free_module(M)true
julia> F = ambient_free_module(M)Free module of rank 1 over Multivariate Polynomial Ring in x, y, z over Rational Field
julia> f = x*F[1]x*e[1]
julia> M(f)x*e[1]
julia> typeof(f)FreeModElem{fmpq_mpoly}
julia> typeof(M(f))SubQuoElem{fmpq_mpoly}

The zero element of a subquotient is obtained as follows:

zero — Method

zero(M::SubQuo)

Return the zero element of M.

Whether a given element of a subquotient is zero can be tested as follows:

iszero — Method

iszero(m::SubQuoElem)

Return true if m is zero, false otherwise.

Examples

julia> R, (x, y, z) = PolynomialRing(QQ, ["x", "y", "z"]);

julia> A = R[x; y]
[x]
[y]

julia> B = R[x^2; x*y; y^2; z^4]
[x^2]
[x*y]
[y^2]
[z^4]

julia> M = SubQuo(A, B)
Subquotient of Submodule with 2 generators
1 -> x*e[1]
2 -> y*e[1]
by Submodule with 4 generators
1 -> x^2*e[1]
2 -> x*y*e[1]
3 -> y^2*e[1]
4 -> z^4*e[1]

julia> iszero(M[1])
false

julia> iszero(x*M[1])
true

Tests on Subqotients

Basic Operations on Subqotients

Submodules and Quotients

Homomorphisms of Subqotients

Operations on Homomorphisms of Subqotients

Subquotients Related to Homomorphisms

Kernel

Image

Cokernel

Homology

Presentations

Syzygies and Free Resolutions

Hom and Ext

Tensorproduct and Tor