Creating Polynomial Rings

Types

Constructors

Coefficient Rings

Data Associated to Polynomial Rings

Elements of Polynomial Rings

Gradings

Homomorphisms of Polynomial Rings

Creating Polynomial Rings

Referring to the corrresponding section of the ring chapter for details, we recall how to create multivariate polynomial rings and their elements. We also show how to assign gradings to these rings. Homomorphisms of multivariate polynomial rings are described in a more general context in the section on affine algebras.

Types

Note

All types for multivariate polynomial rings belong to the abstract type MPolyRing{T}, irrespective of whether they are graded or not. In contrast, the types for univariate polynomial rings belong to PolyRing{T}. The elements of these rings are stored in a sparse and dense format, respectively.

Constructors

The standard constructor below allows one to build multivariate polynomial rings:

PolynomialRing(C::Ring, v::Vector{String}; ordering=:lex)

Its return value is a tuple, say R, vars, consisting of a polynomial ring R with coefficient ring C and a vector vars of generators (variables) which print according to the strings in the vector v . The input ordering=:lex refers to the lexicograpical monomial ordering which specifies the default way of storing and displaying polynomials in OSCAR (terms are sorted in descending order). The other possible choices are :deglex and :degrevlex. Gröbner bases, however, can be computed with respect to any monomial ordering. See the section on Gröbner bases.

Examples

julia> R1, x = PolynomialRing(QQ, ["x"])(Multivariate Polynomial Ring in x over Rational Field, fmpq_mpoly[x])
julia> typeof(x)Vector{fmpq_mpoly} (alias for Array{fmpq_mpoly, 1})
julia> R2, (x,) = PolynomialRing(QQ, ["x"])(Multivariate Polynomial Ring in x over Rational Field, fmpq_mpoly[x])
julia> typeof(x)fmpq_mpoly
julia> R3, x = PolynomialRing(QQ, "x")(Univariate Polynomial Ring in x over Rational Field, x)
julia> typeof(x)fmpq_poly

julia> R, (x, y, z) = PolynomialRing(ZZ, ["x", "y", "z"])(Multivariate Polynomial Ring in x, y, z over Integer Ring, fmpz_mpoly[x, y, z])
julia> typeof(R)FmpzMPolyRing
julia> typeof(x)fmpz_mpoly

julia> V = ["x[1]", "x[2]"]2-element Vector{String}:
 "x[1]"
 "x[2]"
julia> T, x = PolynomialRing(GF(3), V)(Multivariate Polynomial Ring in x[1], x[2] over Galois field with characteristic 3, gfp_mpoly[x[1], x[2]])
julia> x2-element Vector{gfp_mpoly}:
 x[1]
 x[2]

The constructor illustrated below allows for the convenient handling of variables with multi-indices:

julia> R, x, y, z = PolynomialRing(QQ, "x" => (1:3, 1:4), "y" => 1:2, "z" => (1:1, 1:1, 1:1))(Multivariate Polynomial Ring in 15 variables x[1, 1], x[2, 1], x[3, 1], x[1, 2], ..., z[1, 1, 1] over Rational Field, fmpq_mpoly[x[1, 1] x[1, 2] x[1, 3] x[1, 4]; x[2, 1] x[2, 2] x[2, 3] x[2, 4]; x[3, 1] x[3, 2] x[3, 3] x[3, 4]], fmpq_mpoly[y[1], y[2]], fmpq_mpoly[z[1, 1, 1]])
julia> x3×4 Matrix{fmpq_mpoly}:
 x[1, 1]  x[1, 2]  x[1, 3]  x[1, 4]
 x[2, 1]  x[2, 2]  x[2, 3]  x[2, 4]
 x[3, 1]  x[3, 2]  x[3, 3]  x[3, 4]
julia> y2-element Vector{fmpq_mpoly}:
 y[1]
 y[2]
julia> z1×1×1 Array{fmpq_mpoly, 3}:
[:, :, 1] =
 z[1, 1, 1]

Coefficient Rings

Gröbner bases are implemented for multivariate polynomial rings over the fields and rings from this list:

The field of rational numbers $\mathbb{Q}$

julia> QQRational Field

Finite fields $\mathbb{F_p}$, $p$ a prime

julia> GF(3)Galois field with characteristic 3
julia> GF(ZZ(2)^127 - 1)Galois field with characteristic 170141183460469231731687303715884105727

Finite fields $\mathbb{F}_{p^n}$ with $p^n$ elements, $p$ a prime

julia> FiniteField(2, 70, "a")(Finite field of degree 70 over F_2, a)

Simple algebraic extensions of $\mathbb{Q}$ or $\mathbb{F}_p$

julia> T, t = PolynomialRing(QQ, "t")(Univariate Polynomial Ring in t over Rational Field, t)
julia> K, a = NumberField(t^2 + 1, "a")(Number field over Rational Field with defining polynomial t^2 + 1, a)
julia> F = GF(3)Galois field with characteristic 3
julia> T, t = PolynomialRing(F, "t")(Univariate Polynomial Ring in t over Galois field with characteristic 3, t)
julia> K, a = FiniteField(t^2 + 1, "a")(Finite field of degree 2 over F_3, a)

Purely transcendental extensions of $\mathbb{Q}$ or $\mathbb{F}_p$

julia> T, t = PolynomialRing(QQ, "t")(Univariate Polynomial Ring in t over Rational Field, t)
julia> QT = FractionField(T)Fraction field of Univariate Polynomial Ring in t over Rational Field
julia> parent(t)Univariate Polynomial Ring in t over Rational Field
julia> parent(1//t)Fraction field of Univariate Polynomial Ring in t over Rational Field
julia> T, (s, t) = PolynomialRing(GF(3), ["s", "t"]);
julia> QT = FractionField(T)Fraction field of Multivariate Polynomial Ring in s, t over Galois field with characteristic 3

The ring of integers $\mathbb{Z}$

julia> ZZInteger Ring

Data Associated to Polynomial Rings

If R is a multivariate polynomial ring R with coefficient ring C, then

base_ring(R) refers to C,

gens(R) to the generators (variables) of R,

ngens(R) to the number of these generators, and

gen(R, i) as well as R[i] to the ith generator.

Examples

julia> R, (x, y, z) = PolynomialRing(QQ, ["x", "y", "z"])(Multivariate Polynomial Ring in x, y, z over Rational Field, fmpq_mpoly[x, y, z])
julia> base_ring(R)Rational Field
julia> gens(R)3-element Vector{fmpq_mpoly}:
 x
 y
 z
julia> gen(R, 2)y
julia> R[3]z
julia> ngens(R)3

Elements of Polynomial Rings

Note

The types for multivariate polynomials belong to the abstract type MPolyElem{T}.

One way to construct elements of a multivariate polynomial ring R is to build up polynomials from the generators (variables) of R using basic arithmetic as shown below:

Examples

julia> R, (x, y, z) = PolynomialRing(QQ, ["x", "y", "z"])(Multivariate Polynomial Ring in x, y, z over Rational Field, fmpq_mpoly[x, y, z])
julia> f = 3*x^2+y*z3*x^2 + y*z
julia> typeof(f)fmpq_mpoly

Alternatively, there is the following constructor:

(R::MPolyRing{T})(c::Vector{T}, e::Vector{Vector{Int}}) where T <: RingElem

Its return value is the element of R whose nonzero coefficients are specified by the elements of c, with exponent vectors given by the elements of e.

Examples

julia> R, (x, y, z) = PolynomialRing(QQ, ["x", "y", "z"])(Multivariate Polynomial Ring in x, y, z over Rational Field, fmpq_mpoly[x, y, z])
julia> f = R(QQ.([3, 1]), [[2, 0, 0], [0, 1, 1]])3*x^2 + y*z
julia> QQ.([3, 1]) == map(QQ, [3, 1])true

If an element f of R has been constructed, then

parent(f) refers to R,

monomial(f, i) to the i-th monomial of f,

term(f, i) to the i-th term of f,

coeff(f, i) to the coefficient of the i-th term of f, and

exponent_vector(f, i) to the exponent vector of the i-th term of f.

Examples

julia> R, (x, y) = PolynomialRing(GF(5), ["x", "y"])(Multivariate Polynomial Ring in x, y over Galois field with characteristic 5, gfp_mpoly[x, y])
julia> c = map(GF(5), [1, 2, 3])3-element Vector{gfp_elem}:
 1
 2
 3
julia> e = [[3, 2], [1, 0], [0, 1]]3-element Vector{Vector{Int64}}:
 [3, 2]
 [1, 0]
 [0, 1]
julia> f = R(c, e)x^3*y^2 + 2*x + 3*y
julia> parent(f)Multivariate Polynomial Ring in x, y over Galois field with characteristic 5
julia> coeff(f, 2)2
julia> exponent_vector(f, 2)2-element Vector{Int64}:
 1
 0
julia> monomial(f, 2)x
julia> term(f, 2)2*x

Note

An often more effective way to construct polynomials is to use the MPoly build context as explained in the chapter on rings.

Gradings

The following functions allow one to assign gradings to multivariate polynomial rings:

grade — Method

grade(R::MPolyRing, W::Vector{Int})

Grade R by assigning weights to the variables according to the entries of W.

grade(R::MPolyRing)

Grade R by assigning weight 1 to each variable.

Examples

julia> R, (x, y, z) = PolynomialRing(QQ, ["x", "y", "z"])
(Multivariate Polynomial Ring in x, y, z over Rational Field, fmpq_mpoly[x, y, z])

julia> typeof(R)
FmpqMPolyRing

julia> typeof(x)
fmpq_mpoly

julia> S, (x, y, z) = grade(R)
(Multivariate Polynomial Ring in x, y, z over Rational Field graded by
  x -> [1]
  y -> [1]
  z -> [1], MPolyElem_dec{fmpq, fmpq_mpoly}[x, y, z])

julia> typeof(S)
MPolyRing_dec{fmpq, FmpqMPolyRing}

julia> typeof(S) <: MPolyRing
true

julia> typeof(x)
MPolyElem_dec{fmpq, fmpq_mpoly}

julia> W = [1, 2, 3];

julia> T, (x, y, z) = grade(R, W)
(Multivariate Polynomial Ring in x, y, z over Rational Field graded by
  x -> [1]
  y -> [2]
  z -> [3], MPolyElem_dec{fmpq, fmpq_mpoly}[x, y, z])

More directly, graded polynomial rings can be constructed as follows:

GradedPolynomialRing — Method

GradedPolynomialRing(C::Ring, V::Vector{String}, W::Vector{Int}; ordering=:lex)

Return a multivariate polynomial ring with weights assigned to the variables according to the entries of W.

GradedPolynomialRing(C::Ring, V::Vector{String}; ordering=:lex)

Return a multivariate polynomial ring with weight 1 assigned to each variable.

Examples

julia> R, (x, y, z) = GradedPolynomialRing(QQ, ["x", "y", "z"], [1, 2, 3])
(Multivariate Polynomial Ring in x, y, z over Rational Field graded by
  x -> [1]
  y -> [2]
  z -> [3], MPolyElem_dec{fmpq, fmpq_mpoly}[x, y, z])

Note

OSCAR functionality for working with gradings by arbitrary abelian groups is currently under development.

Homomorphisms of Polynomial Rings

How to handle homomorphisms of multivariate polynomial rings is described in a more general context in the section on affine algebras.