G4-Fluxes

\[G_4\]

-fluxes are at the heart of F-theory model building.

Constructors

We currently support the following constructor:

g4_flux — Method

g4_flux(model::AbstractFTheoryModel, class::CohomologyClass)

Construct a G4-flux candidate on an F-theory model. This functionality is currently limited to

Weierstrass models,

global Tate models,

hypersurface models.

Furthermore, our functionality requires a concrete geometry. That is, the base space as well as the ambient space must be toric varieties. In the toric ambient space $X_\Sigma$, the elliptically fibered space $Y$ that defines the F-theory model, is given by a hypersurface (cut out by the Weierstrass, Tate or hypersurface polynomial, respectively).

In this setting, we assume that a $G_4$-flux candidate is represented by a cohomology class $h$ in $H^{(2,2)} (X_\Sigma)$. The actual $G_4$-flux candidate is then obtained by restricting $h$ to $Y$.

It is worth recalling that the $G_4$-flux candidate is subject to the quantization condition $G_4 + \frac{1}{2} c_2(Y) \in H^{/2,2)}( Y_, \mathbb{Z})$ (see [Wit97]). This condition is very hard to verify. However, it is relatively easy to gather evidence for this condition to be satisfied/show that it is violated. To this end, let $D_1$, $D_2$ be two toric divisors in $X_\Sigma$, then the topological intersection number $\left[ h|_Y \right] \cdot \left[ P \right] \cdot \left[ D_1 \right] \cdot \left[ D_2 \right]$ must be an integer. Even this rather elementary check can be computationally expensive. Users can therefore decide to skip this check upon construction by setting the parameter check to the value false.

Another bottleneck can be the computation of the cohomology ring, which is necessary to work with cohomology classes on the toric ambient space, which in turn define the G4-flux, as explained above. The reason for this is, that by employing the theory explained in [CLS11], we can only work out the cohomology ring of simpicial and complete (i.e. compact) toric varieties. However, checking if a toric variety is complete (i.e. compact) can take a long time. If the geometry in question is involved and you already know that the variety is simplicial and complete, then we recommend to trigger the computation of the cohomology ring with check = false. This will avoid this time consuming test.

An example is in order.

Examples

julia> qsm_model = literature_model(arxiv_id = "1903.00009", model_parameters = Dict("k" => 4))
Hypersurface model over a concrete base

julia> cohomology_ring(ambient_space(qsm_model), check = false);

julia> g4_class = cohomology_class(anticanonical_divisor_class(ambient_space(qsm_model)))^2;

julia> g4f = g4_flux(qsm_model, g4_class)
G4-flux candidate
  - Elementary quantization checks: satisfied
  - Tadpole cancellation check: not executed
  - Verticality checks: not executed
  - Non-abelian gauge group: breaking pattern not analyzed

julia> g4f2 = g4_flux(qsm_model, g4_class, check = false)
G4-flux candidate
  - Elementary quantization checks: not executed
  - Tadpole cancellation check: not executed
  - Verticality checks: not executed
  - Non-abelian gauge group: breaking pattern not analyzed

Experimental

This function is part of the experimental code in Oscar. Please read here for more details.

Attributes

We currently support the following attributes:

model — Method

model(gf::G4Flux)

Return the F-theory model for which this $G_4$-flux candidate is defined.

julia> qsm_model = literature_model(arxiv_id = "1903.00009", model_parameters = Dict("k" => 4))
Hypersurface model over a concrete base

julia> cohomology_ring(ambient_space(qsm_model), check = false);

julia> g4_class = cohomology_class(anticanonical_divisor_class(ambient_space(qsm_model)))^2;

julia> g4f = g4_flux(qsm_model, g4_class, check = false)
G4-flux candidate
  - Elementary quantization checks: not executed
  - Tadpole cancellation check: not executed
  - Verticality checks: not executed
  - Non-abelian gauge group: breaking pattern not analyzed  

julia> model(g4f)
Hypersurface model over a concrete base

Experimental

This function is part of the experimental code in Oscar. Please read here for more details.

cohomology_class — Method

cohomology_class(gf::G4Flux)

Return the cohomology class which defines the $G_4$-flux candidate.

julia> qsm_model = literature_model(arxiv_id = "1903.00009", model_parameters = Dict("k" => 4))
Hypersurface model over a concrete base

julia> cohomology_ring(ambient_space(qsm_model), check = false);

julia> g4_class = cohomology_class(anticanonical_divisor_class(ambient_space(qsm_model)))^2;

julia> g4f = g4_flux(qsm_model, g4_class, check = false)
G4-flux candidate
  - Elementary quantization checks: not executed
  - Tadpole cancellation check: not executed
  - Verticality checks: not executed
  - Non-abelian gauge group: breaking pattern not analyzed

julia> cohomology_class(g4f) == g4_class
true

Experimental

This function is part of the experimental code in Oscar. Please read here for more details.

d3_tadpole_constraint — Method

d3_tadpole_constraint(gf::G4Flux; check::Bool = true)

Return the d3-tapdole constraint of a G4-flux, that is compute the quantity $- \frac{1}{2} \cdot G_4^2 + \frac{1}{24} \cdot \chi(\widehat{Y}_4) \stackrel{!}{\geq} 0$.

julia> qsm_model = literature_model(arxiv_id = "1903.00009", model_parameters = Dict("k" => 4))
Hypersurface model over a concrete base

julia> divs = torusinvariant_prime_divisors(ambient_space(qsm_model));

julia> e1 = cohomology_class(divs[35]);e2 = cohomology_class(divs[32]);e4 = cohomology_class(divs[34]);

julia> u = cohomology_class(divs[33]);v = cohomology_class(divs[30]);pb_Kbar = cohomology_class(sum([divs[k] for k in 1:29]));

julia> g4_class = (-3) // kbar3(qsm_model) * (5 * e1 * e4 + pb_Kbar * (-3 * e1 - 2 * e2 - 6 * e4 + pb_Kbar - 4 * u + v));

julia> g4 = g4_flux(qsm_model, g4_class, check = false)
G4-flux candidate
  - Elementary quantization checks: not executed
  - Tadpole cancellation check: not executed
  - Verticality checks: not executed
  - Non-abelian gauge group: breaking pattern not analyzed

julia> d3_tadpole_constraint(g4, check = false)
12

Experimental

This function is part of the experimental code in Oscar. Please read here for more details.

Methods

You can create instances of families of G4-fluxes, i.e. a flux that is contained in said family. For this, we support the following:

flux_instance — Method

flux_instance(fgs::FamilyOfG4Fluxes, int_combination::ZZMatrix, rat_combination::QQMatrix; check::Bool = true)

Create an element of a family of G4-fluxes.

Examples

julia> qsm_model = literature_model(arxiv_id = "1903.00009", model_parameters = Dict("k" => 2021))
Hypersurface model over a concrete base

julia> mat_int = zero_matrix(QQ, 37, 1);

julia> mat_int[1,1] = 1;

julia> mat_rat = zero_matrix(QQ, 37, 1);

julia> mat_rat[2,1] = 1;

julia> fgs = family_of_g4_fluxes(qsm_model, mat_int, mat_rat, check = false)
A family of G4 fluxes:
  - Elementary quantization checks: not executed
  - Verticality checks: not executed
  - Non-abelian gauge group: breaking pattern not analyzed

julia> int_combination = matrix(ZZ, [[3]])
[3]

julia> rat_combination = matrix(QQ, [[5//2]])
[5//2]

julia> flux_instance(fgs, int_combination, rat_combination, check = false)
G4-flux candidate
  - Elementary quantization checks: not executed
  - Tadpole cancellation check: not executed
  - Verticality checks: not executed
  - Non-abelian gauge group: breaking pattern not analyzed

Experimental

This function is part of the experimental code in Oscar. Please read here for more details.

random_flux_instance — Method

random_flux_instance(fgs::FamilyOfG4Fluxes; check::Bool = true)

Create a random element of a family of G4-fluxes.

Examples

julia> qsm_model = literature_model(arxiv_id = "1903.00009", model_parameters = Dict("k" => 2021))
Hypersurface model over a concrete base

julia> mat_int = zero_matrix(QQ, 37, 1);

julia> mat_int[1,1] = 1;

julia> mat_rat = zero_matrix(QQ, 37, 1);

julia> mat_rat[2,1] = 1;

julia> fgs = family_of_g4_fluxes(qsm_model, mat_int, mat_rat, check = false)
A family of G4 fluxes:
  - Elementary quantization checks: not executed
  - Verticality checks: not executed
  - Non-abelian gauge group: breaking pattern not analyzed

julia> random_flux_instance(fgs, check = false)
G4-flux candidate
  - Elementary quantization checks: not executed
  - Tadpole cancellation check: not executed
  - Verticality checks: not executed
  - Non-abelian gauge group: breaking pattern not analyzed

Experimental

This function is part of the experimental code in Oscar. Please read here for more details.

For convenience, we also support the following method to create a random $G_4$-flux with prescribed properties on a given F-theory model:

random_flux — Method

random_flux(m::AbstractFTheoryModel; vert::Bool = false, not_breaking::Bool = false, check::Bool = true)

Create a random $G_4$-flux on a given F-theory model.

Examples

julia> qsm_model = literature_model(arxiv_id = "1903.00009", model_parameters = Dict("k" => 2021))
Hypersurface model over a concrete base

julia> rf = random_flux(qsm_model, vert = true, check = false)
G4-flux candidate
  - Elementary quantization checks: satisfied
  - Tadpole cancellation check: not executed
  - Verticality checks: satisfied
  - Non-abelian gauge group: broken

Experimental

This function is part of the experimental code in Oscar. Please read here for more details.

Properties

We currently support the following properties:

passes_elementary_quantization_checks — Method

passes_elementary_quantization_checks(gf::G4Flux)

G4-fluxes are subject to the quantization condition [Wit97] $G_4 + \frac{1}{2} c_2(Y) \in H^{(2,2)}(Y, \mathbb{Z})$. It is hard to verify that this condition is met. However, we can execute a number of simple consistency checks, by verifying that $\int_{Y}{G_4 \wedge [D_1] \wedge [D_2]} \in \mathbb{Z}$ for any two toric divisors $D_1$, $D_2$. If all of these simple consistency checks are met, this method will return true and otherwise false.

It is worth mentioning that currently (August 2024), we only support this check for $G_4$-fluxes defined on Weierstrass, global Tate and hypersurface models. If this condition is not met, this method will return an error.

julia> qsm_model = literature_model(arxiv_id = "1903.00009", model_parameters = Dict("k" => 4))
Hypersurface model over a concrete base

julia> cohomology_ring(ambient_space(qsm_model), check = false);

julia> g4_class = cohomology_class(anticanonical_divisor_class(ambient_space(qsm_model)))^2;

julia> g4 = g4_flux(qsm_model, g4_class, check = false)
G4-flux candidate
  - Elementary quantization checks: not executed
  - Tadpole cancellation check: not executed
  - Verticality checks: not executed
  - Non-abelian gauge group: breaking pattern not analyzed

julia> passes_elementary_quantization_checks(g4)
true

julia> g4
G4-flux candidate
  - Elementary quantization checks: satisfied
  - Tadpole cancellation check: not executed
  - Verticality checks: not executed
  - Non-abelian gauge group: breaking pattern not analyzed

Experimental

This function is part of the experimental code in Oscar. Please read here for more details.

Ambient Space Models for G4-Fluxes

Focus on 4-dimensional F-theory models $m$, such that the resolution $\widehat{Y}_4$ of the defining singular elliptically fibered CY 4-fold $\Y_4 \twoheadrightarrow B_3$ is defined as hypersurface in a complete and simplicial toric variety $X_\Sigma$. In such a setup, it is convenient to focus on $G_4$-fluxes modelled from the restriction of elements of $H^{(2,2)}( X_\Sigma, \mathbb{Q})$ to $\widehat{Y}_4$. This method identifies a basis of $H^{(2,2)}( X_\Sigma, \mathbb{Q})$ and filters out elements, whose restricton to $\widehat{Y}_4$ is obviously trivial.

It is important to elaborate a bit more on the meaning of "obviously". To this end, fix a basis element of $H^{(2,2)}( X_\Sigma, \mathbb{Q})$. Let us denote a corresponding algebraic cycle by $A = \mathbb{V}(x_i, x_j) \subset X_\Sigma$, where $x_i$, $x_j$ are suitable homogeneous coordinates. Furthermore, let $\widehat{Y}_4 = \mathbb{V}( p ) \subset X_\Sigma$. Then of course, we can look at the set-theoretic intersection $\mathbb{V}( p, x_i, x_j)$. Provided that $p(x_i = 0, x_j = 0)$ is a non-zero constant, this set-theoretic intersection is trivial. This is exactly the check conducted by the method ambient_space_models_of_g4_fluxes below. However, for reasons of simplicity, this approach is avoid a number of sutleties.

Namely, we really have to work out the intersection in the Chow ring, that is we should consider the rational equivalence class of the algebraic cycle $A$ and intersect this class with the rational equivalence class of the algebraic cycle $\mathbb{V}( p )$. In particular, for "unlucky" choices of $i, j$, the algebraic cycles $\mathbb{V}( p )$ and $\mathbb{V}(x_i, x_j)$ may not intersect transversely. This is for instance the case if $i = j$. Such phenomena are addressed in theory by "moving the algebraic cycles into general position", but in practice this somewhat tricky. Instances include the following:

\[i = j\]
: Then apparently, a self-intersection of $\mathbb{V}(x_i)$ is involved.

\[p(x_i, x_j) \equiv 0\]
: This is unexpected for dimensional reasons, and indicates a

non-transverse intersection.

In both instances, one makes use of the linear relations of $X_\Sigma$ to replace the cycle $\mathbb{V}(x_i)$ (and/or $\mathbb{V}(x_j)$) with a rational combination of algebraic cycles $R = \sum_{k = 1}^{N}{c_k \cdot A_k}$, such that $R$ is rationally equivalent to $\mathbb{V}(x_i)$. From experience, it is then rather common that $R$ and $\mathbb{V}(x_i)$ intersect transversely. And if not, then modify the non-transverse intersections by using the linear relations again to replace an involved algebraic cycle.

ambient_space_models_of_g4_fluxes — Method

ambient_space_models_of_g4_fluxes(m::AbstractFTheoryModel; check::Bool = true)::Vector{CohomologyClass}

Given an F-theory model $m$ defined as hypersurface in a simplicial and complete toric base, we this method first computes a basis of $H^(2,2)(X, \mathbb{Q})$ (by use of the method basis_of_h22 below) and then filters out "some" basis elements whose restriction to the hypersurface in question is trivial. The exact meaning of "some" is explained above this method.

Note that it can be computationally very demanding to check if a toric variety $X$ is complete (and simplicial). The optional argument check can be set to false to skip these tests.

Examples

julia> B3 = projective_space(NormalToricVariety, 3)
Normal toric variety

julia> Kbar = anticanonical_divisor_class(B3)
Divisor class on a normal toric variety

julia> t = literature_model(arxiv_id = "1109.3454", equation = "3.1", base_space = B3, defining_classes = Dict("w"=>Kbar))
Construction over concrete base may lead to singularity enhancement. Consider computing singular_loci. However, this may take time!

Global Tate model over a concrete base -- SU(5)xU(1) restricted Tate model based on arXiv paper 1109.3454 Eq. (3.1)

julia> g4_amb_list = ambient_space_models_of_g4_fluxes(t)
2-element Vector{CohomologyClass}:
 Cohomology class on a normal toric variety given by z^2
 Cohomology class on a normal toric variety given by y^2

julia> qsm_model = literature_model(arxiv_id = "1903.00009", model_parameters = Dict("k" => 8))
Hypersurface model over a concrete base

julia> g4_amb_list = ambient_space_models_of_g4_fluxes(qsm_model, check = false);

julia> length(g4_amb_list) == 172
true

Experimental

This function is part of the experimental code in Oscar. Please read here for more details.

basis_of_h22 — Method

basis_of_h22(v::NormalToricVariety; check::Bool = true)::Vector{CohomologyClass}

By virtue of Theorem 12.4.1 in [CLS11], one can compute a monomial basis of $H^4(X, \mathbb{Q})$ for a simplicial, complete toric variety $X$ by truncating its cohomology ring to degree $2$. Inspired by this, this method identifies a basis of $H^{(2,2)}(X, \mathbb{Q})$ by multiplying pairs of cohomology classes associated with toric coordinates.

By definition, $H^{(2,2)}(X, \mathbb{Q})$ is a subset of $H^{4}(X, \mathbb{Q})$. However, by Theorem 9.3.2 in [CLS11], for complete and simplicial toric varieties and $p \neq q$ it holds $H^{(p,q)}(X, \mathbb{Q}) = 0$. It follows that for such varieties $H^{(2,2)}(X, \mathbb{Q}) = H^4(X, \mathbb{Q})$ and the vector space dimension of those spaces agrees with the Betti number $b_4(X)$.

Note that it can be computationally very demanding to check if a toric variety $X$ is complete (and simplicial). The optional argument check can be set to false to skip these tests.

Examples

julia> Y1 = hirzebruch_surface(NormalToricVariety, 2)
Normal toric variety

julia> Y2 = hirzebruch_surface(NormalToricVariety, 2)
Normal toric variety

julia> Y = Y1 * Y2
Normal toric variety

julia> h22_basis = basis_of_h22(Y, check = false)
6-element Vector{CohomologyClass}:
 Cohomology class on a normal toric variety given by xx2*yx2
 Cohomology class on a normal toric variety given by xt2*yt2
 Cohomology class on a normal toric variety given by xx2*yt2
 Cohomology class on a normal toric variety given by xt2*yx2
 Cohomology class on a normal toric variety given by yx2^2
 Cohomology class on a normal toric variety given by xx2^2

julia> betti_number(Y, 4) == length(h22_basis)
true

Experimental

This function is part of the experimental code in Oscar. Please read here for more details.

Families of G4-Fluxes

Constructor

Physics often is interested in families of G4-fluxes, that are expressed in terms of integral or rational combinations of the ambient space models of G4-fluxes, that we described above. For such families, we currently support the following constructor:

family_of_g4_fluxes — Method

family_of_g4_fluxes(m::AbstractFTheoryModel, mat_int::QQMatrix, mat_rat::QQMatrix; check::Bool = true)

Given an F-theory model with a toric ambient space, we can identify ambient space candidates of G4-fluxes. In terms of these candidates, we can define a family of G4-fluxes as:

\[\mathbb{Z}\]
-linear combinations, provided by a matrix $mat_int$,

\[\mathbb{Q}\]
-linear combinations, provided by a matrix $mat_rat$.

An example is in order.

Examples

julia> qsm_model = literature_model(arxiv_id = "1903.00009", model_parameters = Dict("k" => 2021))
Hypersurface model over a concrete base

julia> mat_int = zero_matrix(QQ, 37, 1);

julia> mat_int[1,1] = 1;

julia> mat_rat = zero_matrix(QQ, 37, 1);

julia> mat_rat[2,1] = 1;

julia> family_of_g4_fluxes(qsm_model, mat_int, mat_rat, check = false)
A family of G4 fluxes:
  - Elementary quantization checks: not executed
  - Verticality checks: not executed
  - Non-abelian gauge group: breaking pattern not analyzed

Experimental

This function is part of the experimental code in Oscar. Please read here for more details.

Attributes

Families of G4-fluxes currently support the following attributes:

model — Method

model(gf::FamilyOfG4Fluxes)

Return the F-theory model for which this family of $G_4$-flux candidates is defined.

julia> qsm_model = literature_model(arxiv_id = "1903.00009", model_parameters = Dict("k" => 2021))
Hypersurface model over a concrete base

julia> mat_int = zero_matrix(QQ, 37, 1);

julia> mat_int[1,1] = 1;

julia> mat_rat = zero_matrix(QQ, 37, 1);

julia> mat_rat[2,1] = 1;

julia> f_gs = family_of_g4_fluxes(qsm_model, mat_int, mat_rat, check = false)
A family of G4 fluxes:
  - Elementary quantization checks: not executed
  - Verticality checks: not executed
  - Non-abelian gauge group: breaking pattern not analyzed

julia> model(f_gs) == qsm_model
true

Experimental

This function is part of the experimental code in Oscar. Please read here for more details.

matrix_integral — Method

matrix_integral(gf::FamilyOfG4Fluxes)

Return the matrix whose columns specify those combinations of ambient space G4-flux candidates, of which integral linear combinations are contained in this family of G4-fluxes.

julia> qsm_model = literature_model(arxiv_id = "1903.00009", model_parameters = Dict("k" => 2021))
Hypersurface model over a concrete base

julia> mat_int = zero_matrix(QQ, 37, 1);

julia> mat_int[1,1] = 1;

julia> mat_rat = zero_matrix(QQ, 37, 1);

julia> mat_rat[2,1] = 1;

julia> f_gs = family_of_g4_fluxes(qsm_model, mat_int, mat_rat, check = false)
A family of G4 fluxes:
  - Elementary quantization checks: not executed
  - Verticality checks: not executed
  - Non-abelian gauge group: breaking pattern not analyzed

julia> matrix_integral(f_gs) == mat_int
true

Experimental

This function is part of the experimental code in Oscar. Please read here for more details.

matrix_rational — Method

matrix_rational(gf::FamilyOfG4Fluxes)

Return the matrix whose columns specify those combinations of ambient space G4-flux candidates, of which rational linear combinations are contained in this family of G4-fluxes.

julia> qsm_model = literature_model(arxiv_id = "1903.00009", model_parameters = Dict("k" => 2021))
Hypersurface model over a concrete base

julia> mat_int = zero_matrix(QQ, 37, 1);

julia> mat_int[1,1] = 1;

julia> mat_rat = zero_matrix(QQ, 37, 1);

julia> mat_rat[2,1] = 1;

julia> f_gs = family_of_g4_fluxes(qsm_model, mat_int, mat_rat, check = false)
A family of G4 fluxes:
  - Elementary quantization checks: not executed
  - Verticality checks: not executed
  - Non-abelian gauge group: breaking pattern not analyzed

julia> matrix_rational(f_gs) == mat_rat
true

Experimental

This function is part of the experimental code in Oscar. Please read here for more details.

d3_tadpole_constraint — Method

d3_tadpole_constraint(fgs::FamilyOfG4Fluxes; check::Bool = true)

Return the d3-tapdole constraint of a family of G4-fluxes. Recall that for a given $G_4$-flux, this constraint is $- \frac{1}{2} \cdot G_4^2 + \frac{1}{24} \cdot \chi(\widehat{Y}_4) \stackrel{!}{\geq} 0$.

Note that the family of fluxes is specified by linear combination of cohomology classes, some with rational and some with integral coefficients. In terms of these coefficients the d3-tadpole constraint is the demand that a quadratic polynomial in the coefficients evaluates to a non-negative number. This method returns said polynomial in the cofficients. In order to evaluate the D3-tadpole for a particular $G_4$-flux, one has to evaluate this polynomial for the numeric coefficient values that correspond to the given $G_4$-flux.

We use symbols $a_i$ to indicate integral coefficients and $r_i$ to indicate rational coefficients.

julia> qsm_model = literature_model(arxiv_id = "1903.00009", model_parameters = Dict("k" => 2021))
Hypersurface model over a concrete base

julia> fgs = special_flux_family(qsm_model, check = false)
A family of G4 fluxes:
  - Elementary quantization checks: satisfied
  - Verticality checks: failed
  - Non-abelian gauge group: broken

julia> d3_tadpole_constraint(fgs);

Experimental

This function is part of the experimental code in Oscar. Please read here for more details.

Properties

Families of G4-fluxes currently support the following properties:

is_well_quantized — Method

is_well_quantized(fgs::FamilyOfG4Fluxes)

In case it is known if the family of G4-fluxes is well-quantized, this method returns this boolean value – true if well-quantized and false if not well-quantized. In case it is not known if the family of G4-fluxes is well-quantized, an error is raised.

julia> qsm_model = literature_model(arxiv_id = "1903.00009", model_parameters = Dict("k" => 2021))
Hypersurface model over a concrete base

julia> gf = special_flux_family(qsm_model, check = false)
A family of G4 fluxes:
  - Elementary quantization checks: satisfied
  - Verticality checks: failed
  - Non-abelian gauge group: broken

julia> is_well_quantized(gf)
true

Experimental

This function is part of the experimental code in Oscar. Please read here for more details.

is_vertical — Method

is_vertical(fgs::FamilyOfG4Fluxes)

In case it is known if the family of G4-fluxes is vertical, this method returns this boolean value – true if vertical and false if not. In case it is not known if the family of G4-fluxes is vertical, an error is raised.

julia> qsm_model = literature_model(arxiv_id = "1903.00009", model_parameters = Dict("k" => 2021))
Hypersurface model over a concrete base

julia> gf = special_flux_family(qsm_model, check = false)
A family of G4 fluxes:
  - Elementary quantization checks: satisfied
  - Verticality checks: failed
  - Non-abelian gauge group: broken

julia> is_vertical(gf)
false

julia> gf2 = special_flux_family(qsm_model, vert = true, check = false)
A family of G4 fluxes:
  - Elementary quantization checks: satisfied
  - Verticality checks: satisfied
  - Non-abelian gauge group: broken

julia> is_vertical(gf2)
true

Experimental

This function is part of the experimental code in Oscar. Please read here for more details.

breaks_non_abelian_gauge_group — Method

breaks_non_abelian_gauge_group(fgs::FamilyOfG4Fluxes)

In case it is known if the family of G4-fluxes does break the non-abelian gauge group, this method returns this boolean value – true if it break the non-abelian gauge group and false if not. In case it is not known if the family of G4-fluxes does break the non-abelian gauge group, an error is raised.

julia> qsm_model = literature_model(arxiv_id = "1903.00009", model_parameters = Dict("k" => 2021))
Hypersurface model over a concrete base

julia> gf = special_flux_family(qsm_model, check = false)
A family of G4 fluxes:
  - Elementary quantization checks: satisfied
  - Verticality checks: failed
  - Non-abelian gauge group: broken

julia> breaks_non_abelian_gauge_group(gf)
true

julia> gf2 = special_flux_family(qsm_model, vert = true, check = false)
A family of G4 fluxes:
  - Elementary quantization checks: satisfied
  - Verticality checks: satisfied
  - Non-abelian gauge group: broken

julia> breaks_non_abelian_gauge_group(gf2)
true

julia> gf3 = special_flux_family(qsm_model, vert = true, not_breaking = true, check = false)
A family of G4 fluxes:
  - Elementary quantization checks: satisfied
  - Verticality checks: satisfied
  - Non-abelian gauge group: not broken

julia> breaks_non_abelian_gauge_group(gf3)
false

Experimental

This function is part of the experimental code in Oscar. Please read here for more details.

Special Families of G4-Fluxes

Among the $G_4$-flux candidates, the physics is interested in the well-quantized fluxes. That is, those cohomology classes which integrate to an integer against any other 2-cycle in the elliptic 4-fold $\widehat{Y}_4$. Even in theory, this is a hard task. In practice, one therefore focuses on consistency checks. In the case at hand, we can integrate any ambient space $G_4$-flux candidate against a pair of (algebraic cycles associated to) toric divisors. If for any two toric divisors, the result is an integer, then this $G_4$-flux candidate passes a rather non-trivial and necessary test.

Similarly, we have a method for all vertical and well-quantized ambient space $G_4$-flux candidates, as well as those vertical and well-quantized ambient space $G_4$-flux candidates, which do not break the non-abelian gauge group.

For these families, we support the following constructor. Please note that this method may take a long time to execute for involved geometries $\widehat{Y}_4$.

well_quantized_ambient_space_models_of_g4_fluxes — Method

well_quantized_ambient_space_models_of_g4_fluxes(m::AbstractFTheoryModel; check::Bool = true)

Given an F-theory model $m$ defined as hypersurface in a simplicial and complete toric base, this method computes a basis of all well-quantized ambient space $G_4$-fluxes. The result of this operation is a tuple of two matrices. The columns of the first matrix specify those (rational) combinations of ambient space $G_4$-fluxes, of which one may only take $\mathbb{Z}$-linear combinations without violating flux quantization. The columns of the second matrix specify those (rational) combinations of ambient space $G_4$-fluxes, for which any rational linear combination satisfies the flux quantization condition.

Crucially, this method assumes that $c_2( \widehat{Y}_4)$ is even. Currently, no check is conducted and no error raised. Use with care!

Recall that this is relevant in so much as the quantization condition asks to verify if the twist of the given $G_4$-flux by $1/2 \cdot c_2( \widehat{Y}_4)$ is even. Recall also that it is known that for many F-theory models, $c_2( \widehat{Y}_4)$ is an even class. For instance, this applies to all F-theory compactifications on an elliptically fibered smooth Calabi-Yau 4-fold with a globally defined Weierstrass model [CS12]. For instance, this means that all of the F-theory QSMs [CHLLT19] have an even $c_2( \widehat{Y}_4)$.

It can be computationally very demanding to check if a toric variety $X$ is complete (and simplicial). The optional argument check can be set to false to skip these tests.

Examples

julia> B3 = projective_space(NormalToricVariety, 3)
Normal toric variety

julia> Kbar = anticanonical_divisor_class(B3)
Divisor class on a normal toric variety

julia> t = literature_model(arxiv_id = "1109.3454", equation = "3.1", base_space = B3, defining_classes = Dict("w"=>Kbar))
Construction over concrete base may lead to singularity enhancement. Consider computing singular_loci. However, this may take time!

Global Tate model over a concrete base -- SU(5)xU(1) restricted Tate model based on arXiv paper 1109.3454 Eq. (3.1)

julia> ambient_space_models_of_g4_fluxes(t, check = false);

julia> fg = well_quantized_ambient_space_models_of_g4_fluxes(t, check = false);

julia> matrix_integral(fg)
[1//4   -3//16]
[   0   1//144]

julia> matrix_rational(fg)
2 by 0 empty matrix

Here is a more interesting example.

julia> qsm_model = literature_model(arxiv_id = "1903.00009", model_parameters = Dict("k" => 2021))
Hypersurface model over a concrete base

julia> g4_base = ambient_space_models_of_g4_fluxes(qsm_model, check = false);

julia> length(g4_base)
37

julia> fg = well_quantized_ambient_space_models_of_g4_fluxes(qsm_model, check = false);

julia> size(matrix_integral(fg))
(37, 37)

julia> size(matrix_rational(fg))
(37, 0)

julia> M = matrix_integral(fg);

julia> g4_class = sum(M[i,j]*g4_base[i] for i in 1:length(g4_base) for j in 1:size(M,2));

julia> g4 = g4_flux(qsm_model, g4_class, check = false)
G4-flux candidate
  - Elementary quantization checks: not executed
  - Tadpole cancellation check: not executed
  - Verticality checks: not executed
  - Non-abelian gauge group: breaking pattern not analyzed

julia> passes_elementary_quantization_checks(g4)
true

Experimental

This function is part of the experimental code in Oscar. Please read here for more details.

well_quantized_and_vertical_ambient_space_models_of_g4_fluxes — Method

well_quantized_and_vertical_ambient_space_models_of_g4_fluxes(m::AbstractFTheoryModel; check::Bool = true)

Given an F-theory model $m$ defined as hypersurface in a simplicial and complete toric base, this method computes a basis of all well-quantized and vertical ambient space $G_4$-fluxes. The result of this operation is a tuple of two matrices. The columns of the first matrix specify those (rational) combinations of ambient space $G_4$-fluxes, of which one may only take $\mathbb{Z}$-linear combinations without violating flux quantization. The columns of the second matrix specify those (rational) combinations of ambient space $G_4$-fluxes, for which any rational linear combination satisfies the flux quantization condition.

Crucially, this method assumes that $c_2( \widehat{Y}_4)$ is even. Currently, no check is conducted and no error raised. Use with care!

Recall that this is relevant in so much as the quantization condition asks to verify if the twist of the given $G_4$-flux by $1/2 \cdot c_2( \widehat{Y}_4)$ is even. Recall also that it is known that for many F-theory models, $c_2( \widehat{Y}_4)$ is an even class. For instance, this applies to all F-theory compactifications on an elliptically fibered smooth Calabi-Yau 4-fold with a globally defined Weierstrass model [CS12]. For instance, this means that all of the F-theory QSMs [CHLLT19] have an even $c_2( \widehat{Y}_4)$.

It can be computationally very demanding to check if a toric variety $X$ is complete (and simplicial). The optional argument check can be set to false to skip these tests.

Examples

julia> B3 = projective_space(NormalToricVariety, 3)
Normal toric variety

julia> Kbar = anticanonical_divisor_class(B3)
Divisor class on a normal toric variety

julia> t = literature_model(arxiv_id = "1109.3454", equation = "3.1", base_space = B3, defining_classes = Dict("w"=>Kbar))
Construction over concrete base may lead to singularity enhancement. Consider computing singular_loci. However, this may take time!

Global Tate model over a concrete base -- SU(5)xU(1) restricted Tate model based on arXiv paper 1109.3454 Eq. (3.1)

julia> fg = well_quantized_and_vertical_ambient_space_models_of_g4_fluxes(t, check = false);

julia> matrix_integral(fg)
2 by 0 empty matrix

julia> matrix_rational(fg)
2 by 0 empty matrix

Here is a more interesting example, in which we verify with our software tool for one particular F-theory QSM, that the choice of $G_4$-flux presented in [CHLLT19], is indeed vertical and satisfies the necessary conditions for being well-quantized.

julia> qsm_model = literature_model(arxiv_id = "1903.00009", model_parameters = Dict("k" => 2021))
Hypersurface model over a concrete base

julia> fg = well_quantized_and_vertical_ambient_space_models_of_g4_fluxes(qsm_model, check = false);

julia> size(matrix_integral(fg))
(37, 25)

julia> size(matrix_rational(fg))
(37, 0)

julia> M=matrix_integral(fg);

julia> g4_base = ambient_space_models_of_g4_fluxes(qsm_model, check = false);

julia> g4_classes = [sum(M[i,j]*g4_base[i] for i in 1:length(g4_base)) for j in 1:size(M,2)];

julia> length(g4_classes) == 25
true

julia> g4_classes[end]
Cohomology class on a normal toric variety given by 293//300*x4*e2 + 143//150*x4*u - 283//25*x4*e4 + 143//150*x4*e1 + 1643//300*x4*w - 599//150*x5*x8 - 7//150*x5*e2 - 7//75*x5*u - 7//50*x5*e4 - 7//75*x5*e1 - 89//300*x5*w + 896//75*x6*x7 + 1//20*x6*e2 + 1//10*x6*u + 2//5*x6*e4 + 1//10*x6*e1 - 1//5*x6*w - 599//75*x7*x8 - 7//75*x7*e2 - 14//75*x7*u - 7//25*x7*e4 - 14//75*x7*e1 - 89//150*x7*w + 208//75*x8^2 + 298//75*x8*x9 + 1//150*x8*e2 - 73//150*x8*u - 12//25*x8*e4 - 73//150*x8*e1 - 37//75*x8*w + 82//75*x9^2 - 7//150*x9*e2 - 7//75*x9*u + 9//25*x9*e4 - 7//75*x9*e1 - 41//75*x9*w + 11//30*e1*w

julia> g4_list = [g4_flux(qsm_model, cl, check = false) for cl in g4_classes];

julia> all(k -> passes_elementary_quantization_checks(k), g4_list)
true

julia> all(k -> passes_verticality_checks(k), g4_list)
true

julia> c = [60, 51, 90, 0, 24, 51, -24, 45, 30, 0, -48, 90, -57, 60, 30, 15, 120, 0, -60, 0, -720, -420, -270, -60, -2190];

julia> qsm_g4_candidate = g4_flux(qsm_model, sum(c[i]*g4_classes[i] for i in 1:length(g4_classes)), check = false)
G4-flux candidate
  - Elementary quantization checks: not executed
  - Tadpole cancellation check: not executed
  - Verticality checks: not executed
  - Non-abelian gauge group: breaking pattern not analyzed

julia> passes_elementary_quantization_checks(qsm_g4_candidate)
true

julia> passes_verticality_checks(qsm_g4_candidate)
true

Experimental

This function is part of the experimental code in Oscar. Please read here for more details.

well_quantized_and_vertical_and_no_non_abelian_gauge_group_breaking_ambient_space_models_of_g4_fluxes — Method

well_quantized_and_vertical_and_no_non_abelian_gauge_group_breaking_ambient_space_models_of_g4_fluxes(m::AbstractFTheoryModel; check::Bool = true)

Given an F-theory model $m$ defined as hypersurface in a simplicial and complete toric base, this method computes a basis of all well-quantized and vertical ambient space $G_4$-fluxes which do not break any non-abelian gauge group factor. The result of this operation is a tuple of two matrices. The columns of the first matrix specify those (rational) combinations of ambient space $G_4$-fluxes, of which one may only take $\mathbb{Z}$-linear combinations without violating flux quantization. The columns of the second matrix specify those (rational) combinations of ambient space $G_4$-fluxes, for which any rational linear combination satisfies the flux quantization condition.

Crucially, this method assumes that $c_2( \widehat{Y}_4)$ is even. Currently, no check is conducted and no error raised. Use with care!

Recall that this is relevant in so much as the quantization condition asks to verify if the twist of the given $G_4$-flux by $1/2 \cdot c_2( \widehat{Y}_4)$ is even. Recall also that it is known that for many F-theory models, $c_2( \widehat{Y}_4)$ is an even class. For instance, this applies to all F-theory compactifications on an elliptically fibered smooth Calabi-Yau 4-fold with a globally defined Weierstrass model [CS12]. For instance, this means that all of the F-theory QSMs [CHLLT19] have an even $c_2( \widehat{Y}_4)$.

It can be computationally very demanding to check if a toric variety $X$ is complete (and simplicial). The optional argument check can be set to false to skip these tests.

Examples

julia> B3 = projective_space(NormalToricVariety, 3)
Normal toric variety

julia> Kbar = anticanonical_divisor_class(B3)
Divisor class on a normal toric variety

julia> t = literature_model(arxiv_id = "1109.3454", equation = "3.1", base_space = B3, defining_classes = Dict("w"=>Kbar))
Construction over concrete base may lead to singularity enhancement. Consider computing singular_loci. However, this may take time!

Global Tate model over a concrete base -- SU(5)xU(1) restricted Tate model based on arXiv paper 1109.3454 Eq. (3.1)

julia> fg = well_quantized_and_vertical_and_no_non_abelian_gauge_group_breaking_ambient_space_models_of_g4_fluxes(t, check = false);

julia> matrix_integral(fg)
2 by 0 empty matrix

julia> matrix_rational(fg)
2 by 0 empty matrix

Here is a more interesting example, in which we verify with our software tool for one particular F-theory QSM, that the choice of $G_4$-flux presented in [CHLLT19], is indeed vertical, satisfies the necessary conditions for being well-quantized and does not break the non-abelian gauge group.

julia> qsm_model = literature_model(arxiv_id = "1903.00009", model_parameters = Dict("k" => 2021))
Hypersurface model over a concrete base

julia> fg = well_quantized_and_vertical_and_no_non_abelian_gauge_group_breaking_ambient_space_models_of_g4_fluxes(qsm_model, check = false);

julia> size(matrix_integral(fg))
(37, 1)

julia> size(matrix_rational(fg))
(37, 0)

julia> M = matrix_integral(fg);

julia> g4_base = ambient_space_models_of_g4_fluxes(qsm_model, check = false);

julia> g4_classes = [sum(M[i,j]*g4_base[i] for i in 1:length(g4_base)) for j in 1:size(M,2)];

julia> length(g4_classes) == 1
true

julia> g4_classes[end]
Cohomology class on a normal toric variety given by 1//30*x4*e2 + 1//15*x4*u + 1//10*x4*e4 + 1//15*x4*e1 + 1//30*x4*w - 11//30*x5*x8 + 1//15*x5*e2 + 2//15*x5*u + 1//5*x5*e4 + 2//15*x5*e1 + 1//15*x5*w - 1//15*x6*x7 - 11//15*x7*x8 + 2//15*x7*e2 + 4//15*x7*u + 2//5*x7*e4 + 4//15*x7*e1 + 2//15*x7*w - 8//15*x8^2 - 8//15*x8*x9 + 2//15*x8*e2 + 4//15*x8*u + 2//5*x8*e4 + 4//15*x8*e1 + 2//15*x8*w - 2//15*x9^2 + 1//15*x9*e2 + 2//15*x9*u + 1//5*x9*e4 + 2//15*x9*e1 + 1//15*x9*w - 1//6*e1*w

julia> g4_candidate = g4_flux(qsm_model, g4_classes[end], check = false);

julia> passes_elementary_quantization_checks(g4_candidate)
true

julia> passes_verticality_checks(g4_candidate)
true

julia> breaks_non_abelian_gauge_group(g4_candidate)
false

julia> c = [3];

julia> qsm_g4_candidate = g4_flux(qsm_model, sum(c[i]*g4_classes[i] for i in 1:length(g4_classes)), check = false)
G4-flux candidate
  - Elementary quantization checks: not executed
  - Tadpole cancellation check: not executed
  - Verticality checks: not executed
  - Non-abelian gauge group: breaking pattern not analyzed

julia> passes_elementary_quantization_checks(qsm_g4_candidate)
true

julia> passes_verticality_checks(qsm_g4_candidate)
true

julia> breaks_non_abelian_gauge_group(qsm_g4_candidate)
false

julia> divs = torusinvariant_prime_divisors(ambient_space(qsm_model));

julia> e1 = cohomology_class(divs[15]);

julia> e2 = cohomology_class(divs[12]);

julia> e4 = cohomology_class(divs[14]);

julia> u = cohomology_class(divs[13]);

julia> v = cohomology_class(divs[10]);

julia> pb_Kbar = cohomology_class(sum([divs[k] for k in 1:9]));

julia> g4_class = (-3) // kbar3(qsm_model) * (5 * e1 * e4 + pb_Kbar * (-3 * e1 - 2 * e2 - 6 * e4 + pb_Kbar - 4 * u + v));

julia> qsm_g4_candidate == g4_flux(qsm_model, g4_class)
true

Experimental

This function is part of the experimental code in Oscar. Please read here for more details.

We understand that these function names are very long. For this reason, we also support the following convenience method:

special_flux_family — Method

special_flux_family(m::AbstractFTheoryModel; vert::Bool = false, not_breaking::Bool = false, check::Bool = true)

Given an F-theory model $m$ defined as hypersurface in a simplicial and complete toric base, this method computes a family of G4-fluxes with given properties.

This family is modelled by the restriction of cohomology classes on the toric ambient space to the hypersurface in question. In the toric ambient space, those cohomology classes are vertical, i.e. are of the form $a \wedge b$ for $a,b \in H^(1,1)(X_\Sigma)$ with $X_\Sigma$. Note that this does NOT mean that they are vertical on the hypersurface, which defines the actual F-theory geometry. We further subject this family to the consistency conditions for being well-quantized. Unless explicitly asked differently, it is this family of ambient space G4-flux candidates, that is being returned by this function.

However, we allow for optional arguments, which refine the family further:

vert = true: The returned family of fluxes is vertical on the hypersurface in question.

not_breaking = true: The returned family of fluxes does not break the non-abelian gauge group.

Below are examples that examplify the use of these optional arguments and which families of $G_4$-fluxes are being returned.

Examples

julia> qsm_model = literature_model(arxiv_id = "1903.00009", model_parameters = Dict("k" => 2021))
Hypersurface model over a concrete base

julia> gf1 = special_flux_family(qsm_model, check = false)
A family of G4 fluxes:
  - Elementary quantization checks: satisfied
  - Verticality checks: failed
  - Non-abelian gauge group: broken

julia> gf2 = special_flux_family(qsm_model, vert = true, check = false)
A family of G4 fluxes:
  - Elementary quantization checks: satisfied
  - Verticality checks: satisfied
  - Non-abelian gauge group: broken

julia> gf3 = special_flux_family(qsm_model, vert = true, not_breaking = true, check = false)
A family of G4 fluxes:
  - Elementary quantization checks: satisfied
  - Verticality checks: satisfied
  - Non-abelian gauge group: not broken

Experimental

This function is part of the experimental code in Oscar. Please read here for more details.