Global Tate models

A global Tate model describes a particular form of an elliptic fibration. We focus on elliptic fibrations over base 3-folds $B3$. Consider the weighted projective space $\mathbb{P}^{2,3,1}$ with coordinates $x, y, z$. In addition, consider

$a_1 \in H^0( B_3, \overline{K}_{B_3} )$,

$a_2 \in H^0( B_3, \overline{K}_{B_3}^{\otimes 2} )$,

$a_3 \in H^0( B_3, \overline{K}_{B_3}^{\otimes 3} )$,

$a_4 \in H^0( B_3, \overline{K}_{B_3}^{\otimes 4} )$,

$a_6 \in H^0( B_3, \overline{K}_{B_3}^{\otimes 6} )$.

Then form a $\mathbb{P}^{2,3,1}$-bundle over $B3$ such that

$x$ transforms as a section of $2 \overline{K}_{B_3}$,

$y$ transforms as a section of $3 \overline{K}_{B_3}$,

$z$ transforms as a section of $0 \overline{K}_{B_3} = \mathcal{O}_{B_3}$.

In this 5-fold ambient space, a global Tate model is the hypersurface defined by the vanishing of the Tate polynomial $P_T = x^3 - y^2 - x y z a_1 + x^2 z^2 a_2 - y z^3 a_3 + x z^4 a_4 + z^6 a_6$.

Crucially, for non-trivial F-theory settings, the elliptic fibration in question must be singular. In fact, by construction, one usually engineers certain singularities. For this, vanishing orders of the sections $a_i$ above need to specified. The following table–-often referred to as the Tate table and taken from Timo Weigand (2010)–-summarizes the singularities introduced by certain vanishing orders:

sing. type	$\mathrm{ord}(\Delta)$	singularity	group $G$	$a_1$	$a_2$	$a_3$	$a_4$	$a_6$

$I_0$	$0$			$0$	$0$	$0$	$0$	$0$

$I_1$	$1$			$0$	$0$	$1$	$1$	$1$

$I_2$	$2$	$A_1$	$SU(2)$	$0$	$0$	$1$	$1$	$2$

$I_{2k}^{ns}$	$2k$	$C_k$	$Sp(k)$	$0$	$0$	$k$	$k$	$2k$

$I_{2k}^s$	$2k$	$A_{2k-1}$	$SU(2k)$	$0$	$1$	$k$	$k$	$2k$

$I_{2k+1}^{ns}$	$2k+1$		$Sp(k)$	$0$	$0$	$k+1$	$k+1$	$2k+1$

$I_{2k+1}^{s}$	$2k+1$	$A_{2k}$	$SU(2k+1)$	$0$	$1$	$k$	$k+1$	$2k+1$

$II$	$2$			$1$	$1$	$1$	$1$	$1$

$III$	$3$	$A_1$	$SU(2)$	$1$	$1$	$1$	$1$	$2$

$IV^{ns}$	$4$		$Sp(1)$	$1$	$1$	$1$	$2$	$2$

$IV^s$	$4$	$A_2$	$SU(3)$	$1$	$1$	$1$	$2$	$3$

$I_0^{*ns}$	$6$	$G_2$	$G_2$	$1$	$1$	$2$	$2$	$3$

$I_0^{*ss}$	$6$	$B_3$	$SO(7)$	$1$	$1$	$2$	$2$	$4$

$I_0^{*s}$	$6$	$D_4$	$SO(8)$	$1$	$1$	$2$	$2$	$4$

$I_1^{*ns}$	$7$	$B_4$	$SO(9)$	$1$	$1$	$2$	$3$	$4$

$I_1^{*s}$	$7$	$D_5$	$SO(10)$	$1$	$1$	$2$	$3$	$5$

$I_2^{*ns}$	$8$	$B_5$	$SO(11)$	$1$	$1$	$3$	$3$	$5$

$I_2^{*s}$	$8$	$D_6$	$SO(12)$	$1$	$1$	$3$	$3$	$5$

$I_{2k-3}^{*ns}$	$2k+3$	$B_{2k}$	$SO(4k+1)$	$1$	$1$	$k$	$k+1$	$2k$

$I_{2k-3}^{*s}$	$2k+3$	$D_{2k+1}$	$SO(4k+2)$	$1$	$1$	$k$	$k+1$	$2k+1$

$I_{2k-2}^{*ns}$	$2k+4$	$B_{2k+1}$	$SO(4k+3)$	$1$	$1$	$k+1$	$k+1$	$2k+1$

$I_{2k-2}^{*s}$	$2k+4$	$D_{2k+2}$	$SO(4k+4)$	$1$	$1$	$k+1$	$k+1$	$2k+1$

$IV^{*ns}$	$8$	$F_4$	$F_4$	$1$	$2$	$2$	$3$	$4$

$IV^{*s}$	$8$	$E_6$	$E_6$	$1$	$2$	$2$	$3$	$5$

$III^*$	$9$	$E_7$	$E_7$	$1$	$2$	$3$	$3$	$5$

$II^*$	$10$	$E_8$	$E_8$	$1$	$2$	$3$	$4$	$5$

non-min.	$12$			$1$	$2$	$3$	$4$	$6$

... over concrete bases

For some F-theory constructions, one fixes a particular base space. We support the following constructors:

global_tate_model — Method

global_tate_model(base::AbstractNormalToricVariety)

This method constructs a global Tate model over a given toric base 3-fold. The Tate sections $a_i$ are taken with (pseudo) random coefficients.

Examples

julia> t = global_tate_model(test_base())
Global Tate model over a concrete base

julia> is_smooth(toric_ambient_space(t))
false

global_tate_model_over_projective_space — Method

global_tate_model_over_projective_space()

This method constructs a global Tate model over the 3-dimensional projective space.

Examples

julia> global_tate_model_over_projective_space()
Global Tate model over a concrete base

global_tate_model — Method

global_tate_model(ais::Vector{T}, base::AbstractNormalToricVariety) where {T<:MPolyRingElem{QQFieldElem}}

This method operates analogously to global_tate_model(base::AbstractNormalToricVariety). The only difference is that the Tate sections $a_i$ can be specified with non-generic values.

Examples

julia> base = test_base()
Normal toric variety

julia> a1 = sum([rand(Int) * b for b in basis_of_global_sections(anticanonical_bundle(base))]);

julia> a2 = sum([rand(Int) * b for b in basis_of_global_sections(anticanonical_bundle(base)^2)]);

julia> a3 = sum([rand(Int) * b for b in basis_of_global_sections(anticanonical_bundle(base)^3)]);

julia> a4 = sum([rand(Int) * b for b in basis_of_global_sections(anticanonical_bundle(base)^4)]);

julia> a6 = sum([rand(Int) * b for b in basis_of_global_sections(anticanonical_bundle(base)^6)]);

julia> t = global_tate_model([a1, a2, a3, a4, a6], base)
Global Tate model over a concrete base

julia> is_smooth(toric_ambient_space(t))
false

... over not fully specified bases

This method constructs a global Tate model over a base space that is not fully specified. Rather, it assumes that a base space exists such that the Tate sections $a_i$ are well-defined so that the Tate model in question is well-defined.

For many practical applications, one wishes to assume a further factorization of the Tate sections $a_i$. This has the advantage that one can engineer singularity loci or even the singularity type over a specific locus. This is the backbone of many F-theory constructions.

To this end, this method accepts a polynomial ring whose variables are the sections used in the desired factorization of the Tate sections $a_i$. For example, if we desired a factorization:

$a_1 = a_{10} w^0$,

$a_2 = a_{21} w^1$,

$a_3 = a_{32} w^2$,

$a_4 = a_{43} w^3$,

$a_6 = a_{65} w^5$,

then the polynomial ring in question is the ring with indeterminates $a_{10}$, $a_{21}$, $a_{32}$, $a_{43}$, $a_{65}$ and $w$.

In theory, one can consider these indeterminates as local coordinate of the base space. For the computer implementation, this ring will therefore serve as the coordinate ring of an auxiliary toric base space, namely an affine space with those coordinates.

Such geometries can be constructed with the following constructor:

global_tate_model — Method

global_tate_model(ais::Vector{T}, auxiliary_base_ring::MPolyRing, d::Int) where {T<:MPolyRingElem{QQFieldElem}}

This method constructs a global Tate model over a base space that is not fully specified. The following example exemplifies this approach.

Examples

julia> auxiliary_base_ring, (a10, a21, a32, a43, a65, w) = QQ["a10", "a21", "a32", "a43", "a65", "w"];

julia> a1 = a10;

julia> a2 = a21 * w;

julia> a3 = a32 * w^2;

julia> a4 = a43 * w^3;

julia> a6 = a65 * w^5;

julia> ais = [a1, a2, a3, a4, a6];

julia> t = global_tate_model(ais, auxiliary_base_ring, 3)
Global Tate model over a not fully specified base

julia> tate_polynomial(t)
-a10*x*y*z + a21*w*x^2*z^2 - a32*w^2*y*z^3 + a43*w^3*x*z^4 + a65*w^5*z^6 + x^3 - y^2

julia> toric_base_space(t)
Normal toric variety

julia> toric_ambient_space(t)
Normal, simplicial toric variety

julia> dim(toric_ambient_space(t))
5

Standard constructions

Certain Tate models have been studied in the physics literature over and over again. Thereby, these constructions became famous and some were given special names. We aim to provide support for such standard constructions. Currently, we have support for the following:

su5_tate_model_over_arbitrary_3d_base — Method

su5_tate_model_over_arbitrary_3d_base()

Return the SU(5) Tate model over an arbitrary 3-dimensional base space. For more details see e.g. Timo Weigand (2018) and references therein.

julia> su5_tate_model_over_arbitrary_3d_base()
Global Tate model over a not fully specified base

Attributes

For all Tate models, we support the following attributes:

tate_section_a1 — Method

tate_section_a1(t::GlobalTateModel)

Return the Tate section $a_1$.

julia> t = su5_tate_model_over_arbitrary_3d_base()
Global Tate model over a not fully specified base

julia> tate_section_a1(t);

tate_section_a2 — Method

tate_section_a2(t::GlobalTateModel)

Return the Tate section $a_2$.

julia> t = su5_tate_model_over_arbitrary_3d_base()
Global Tate model over a not fully specified base

julia> tate_section_a2(t);

tate_section_a3 — Method

tate_section_a3(t::GlobalTateModel)

Return the Tate section $a_3$.

julia> t = su5_tate_model_over_arbitrary_3d_base()
Global Tate model over a not fully specified base

julia> tate_section_a3(t);

tate_section_a4 — Method

tate_section_a4(t::GlobalTateModel)

Return the Tate section $a_4$.

julia> t = su5_tate_model_over_arbitrary_3d_base()
Global Tate model over a not fully specified base

julia> tate_section_a4(t);

tate_section_a6 — Method

tate_section_a6(t::GlobalTateModel)

Return the Tate section $a_6$.

julia> t = su5_tate_model_over_arbitrary_3d_base()
Global Tate model over a not fully specified base

julia> tate_section_a6(t);

tate_polynomial — Method

tate_polynomial(t::GlobalTateModel)

Return the Tate polynomial of the global Tate model.

julia> t = su5_tate_model_over_arbitrary_3d_base()
Global Tate model over a not fully specified base

julia> tate_polynomial(t);

cy_hypersurface — Method

cy_hypersurface(t::GlobalTateModel)

Return the Calabi-Yau hypersurface in the toric ambient space which defines the global Tate model.

julia> t = su5_tate_model_over_arbitrary_3d_base()
Global Tate model over a not fully specified base

julia> cy_hypersurface(t)
Closed subvariety of a normal toric variety

global_weierstrass_model — Method

global_weierstrass_model(t::GlobalTateModel)

Return the global Weierstrass model which is equivalent to the given Tate model.

julia> t = su5_tate_model_over_arbitrary_3d_base()
Global Tate model over a not fully specified base

julia> global_weierstrass_model(t)
Global Weierstrass model over a not fully specified base

discriminant — Method

discriminant(t::GlobalTateModel)

Return the discriminant of the global Tate model.

julia> t = su5_tate_model_over_arbitrary_3d_base()
Global Tate model over a not fully specified base

julia> discriminant(t);

In case the Tate model is constructed over a not fully specified base, it is nonetheless possible to construct an auxiliary base space as well as an auxiliary ambient space. The (auxiliary) base and ambient space can be accessed with the following functions:

toric_base_space — Method

toric_base_space(t::GlobalTateModel)

Return the toric base space of the global Tate model.

julia> t = global_tate_model(test_base())
Global Tate model over a concrete base

julia> toric_base_space(t)
Normal toric variety without torusfactor

toric_ambient_space — Method

toric_ambient_space(t::GlobalTateModel)

Return the toric ambient space of the global Tate model.

julia> t = global_tate_model(test_base())
Global Tate model over a concrete base

julia> toric_ambient_space(t)
Normal, simplicial toric variety

julia> is_smooth(toric_ambient_space(t))
false

To tell if they are auxiliary or not, one can use base_fully_specified(t), which returns true in case the Tate model is defined over a concrete base and false otherwise.

The user can decide to get an information whenever an auxiliary base space, auxiliary ambient space or auxiliary hypersurface have been computed. To this end, one invokes set_verbosity_level(:GlobalTateModel, 1). More background information is available here.

Properties

base_fully_specified — Method

base_fully_specified(t::GlobalTateModel)

Return true is the Tate model has a concrete base space and false otherwise.

julia> t = su5_tate_model_over_arbitrary_3d_base()
Global Tate model over a not fully specified base

julia> base_fully_specified(t)
false

Singular loci

For applications in F-theory, singular elliptic fibrations are key (cf. Timo Weigand (2018) and references therein). The general approach is to not work with the singular space directly. Rather, one resolves the singularities in order to obtain a smooth space instead. Subsequently, one performs computations on this smooth space.

In this sense, knowledge of the singular loci is a first step to a successful analysis of such a geometry. For this, we provide the following functionality.

singular_loci — Method

singular_loci(t::GlobalTateModel)

Return the singular loci of the global Tate model, along with the order of vanishing of $(f, g, \Delta)$` at each locus and the refined Tate fiber type.

For the time being, we either explicitly or implicitly focus on toric varieties as base spaces. Explicitly, in case the user provides such a variety as base space, and implicitly, in case we work over a non-fully specified base. This has the advantage that we can "filter out" trivial singular loci.

Specifically, recall that every closed subvariety of a simplicial toric variety is of the form $V(I)$, where $I$ is a homogeneous ideal of the Cox ring. Let $B$ be the irrelevant ideal of this toric variety. Then, by proposition 5.2.6. of David A. Cox, John B. Little, Henry K. Schenck (2011), $V(I)$ is trivial/empty iff $B^l \subseteq I$ for a suitable $l \geq 0$. This can be checked by checking if the saturation $I:B^\infty$ is the ideal generated by $1$.

By treating a non-fully specified base space implicitly as a toric space, we can extend this result straightforwardly to this situation also. This is the reason for constructing this auxiliary base space.

Let us demonstrate the functionality by computing the singular loci of a Type $III$ Tate model Sheldon Katz, David R. Morrison, Sakura Schafer-Nameki, James Sully (2011). In this case, we will consider Global Tate model over a non-fully specified base. The Tate sections are factored as follows:

$a_1 = a_{11} w^1$,

$a_2 = a_{21} w^1$,

$a_3 = a_{31} w^1$,

$a_4 = a_{41} w^1$,

$a_6 = a_{62} w^2$.

For this factorization, we expect a singularity of Kodaira type $III$ over the divisor $W = {w = 0}$, as desired. So this should be one irreducible component of the discriminant. Moreover, we should find that the discriminant vanishes to order 3 on $W = {w = 0}$, while the Weierstrass sections $f$ and $g$ vanish to orders 1 and 2, respectively. Let us verify this.

julia> auxiliary_base_ring, (a11, a21, a31, a41, a62, w) = QQ["a10", "a21", "a32", "a43", "a65", "w"];

julia> a1 = a11 * w;

julia> a2 = a21 * w;

julia> a3 = a31 * w;

julia> a4 = a41 * w;

julia> a6 = a62 * w^2;

julia> ais = [a1, a2, a3, a4, a6];

julia> t = global_tate_model(ais, auxiliary_base_ring, 3)
Global Tate model over a not fully specified base

julia> length(singular_loci(t))
2

julia> singular_loci(t)[2]
(ideal(w), (1, 2, 3), "III")