Groups of automorphisms

Inner automorphisms

Groups of automorphisms

Groups of automorphisms over a group G have parametric type AutomorphismGroup{T}, where T is the type of G. The group of automorphisms over a group G is defined by the following instruction:

AutomorphismGroup{T}
automorphism_group(G)

The evaluation of the automorphism f in the element x is analogous to the homomorphism evaluation: it can be obtained by typing either f(x) or x^f.

It is possible to turn an automorphism f into a homomorphism by typing hom(f). The viceversa is also possible: if g is a bijective homomorphism from the group G to itself and A is the automorphism group of G, then the instruction A(g) returns g as automorphism of G. This is the standard way to build explicitly an automorphism (another way, available for inner automorphisms, is shown in Section Inner_automorphisms).

automorphism_group — Method

automorphism_group(G::Group) -> A::AutomorphismGroup{T}

Return the full automorphism group of G. If f is an object of type $GAPGroupHomomorphism$ and it is bijective from G to itself, then A(f) return the embedding of f in A.

Elements of A can be multiplied with other elements of A or by elements of type GAPGroupHomomorphism; in this last case, the result has type GAPGroupHomomorphism.

Example:

julia> S=symmetric_group(4);

julia> A=automorphism_group(S);

julia> g=hom(S,S,x->x^S[1]);

julia> g in A
false

julia> au=A(g);

julia> au in A
true

julia> g==hom(au)
true

julia> x=cperm(S,[1,2,3]);

julia> au(x)
(2,3,4)

julia> g(x)==au(x)
true

In Oscar it is possible to multiply homomorphisms and automorphisms (whenever it makes sense); in such cases, the output is always a variable of type GAPGroupHomomorphism{S,T}.

julia> S=symmetric_group(4);

julia> A=automorphism_group(S);

julia> g=hom(S,S,x->x^S[1]);

julia> f=A(g);

julia> typeof(g*f)
GAPGroupHomomorphism{PermGroup,PermGroup}

The following functions are available for automorphisms, some of them similar to the corresponding functions for homomorphisms of groups.

isinvariant — Method

isinvariant(f::GAPGroupElem{AutomorphismGroup{T}}, H::T)

Return whether f(H) == H.

restrict_automorphism — Method

restrict_automorphism(f::GAPGroupElem{AutomorphismGroup{T}}, H::T)

If H is invariant under f, returns the restriction of f to H as automorphism of H; otherwise it returns ERROR.

induced_automorphism — Method

induced_automorphism(f::GAPGroupHomomorphism, g::GAPGroupHomomorphism)
induced_automorphism(f::GAPGroupHomomorphism, g::GAPGroupElem{AutomorphismGroup{T}})

Return the automorphism h of the image of f such that h(f) == f(g), where g is an automorphism of a group G and f is a group homomorphism defined over G such that the kernel of f is invariant under g

hom — Method

hom(f::GAPGroupElem{AutomorphismGroup{T}}) where T

Return the element f of type $GAPGroupHomomorphism{T,T}$.

Inner automorphisms

Oscar disposes of the following functions to handle inner automorphisms of a group.

inner_automorphism — Method

inner_automorphism(g::GAPGroupElem)

Return the inner automorphism in automorphism_group(parent(g)) defined by x -> x^g.

isinner_automorphism — Method

isinner_automorphism(f::GAPGroupHomomorphism)
isinner_automorphism(f::GAPGroupElem{AutomorphismGroup{T}})

Return whether f is an inner automorphism.

inner_automorphisms_group — Method

inner_automorphisms_group(A::AutomorphismGroup{T})

Return the subgroup of A of the inner automorphisms.