Basics

Elements of groups

Operations

Properties of groups

Attributes of groups

Basics

Elements of groups

Given a group G, it is always possible to have access to some particular elements.

GAPGroup — Type

TODO: document this

BasicGAPGroupElem — Type

BasicGAPGroupElem{T<:GAPGroup}

The type BasicGAPGroupElem gathers all types of group elements described only by an underlying GAP object.

elem_type — Method

elem_type(parent)
elem_type(parent_type)

Given a parent object (or its type), return the type of its elements.

Example

julia> S, x = PowerSeriesRing(QQ, 2, "x")
(Univariate power series ring in x over Rationals, x + O(x^3))

julia> elem_type(S) == typeof(x)
true

one — Method

one(G::Group) -> x::GAPGroupElem{typeof(G)}

Return the identity of the group G.

one — Method

one(x::GAPGroupElement{T}) -> x::GAPGroupElem{T}

Return the identity of the parent group of x.

gens — Method

gens(G::Group)

Return an array of generators of the group G. To get a specific generator, use G[i] or gen(G,i) instead of gens(G)[i], as that is more efficient.

ngens — Method

ngens(G::Group) -> Int

Return the length of the array gens(G).

WARNING:

this is NOT, in general, the minimum number of generators for G.

gen — Method

gen(G::Group, i::Integer)

Return the i-th element of the array gens(G). This is equivalent to G[i], and returns gens(G)[i] but may be more efficient than the latter. If i is greater than the length of gens(G), an ERROR is thrown.

rand — Method

rand(G::Group)

Return a random element of the group G.

rand_pseudo — Method

rand_pseudo(G::Group)

Return a random element of the group G. It works faster than rand, but the elements are not necessarily equally distributed.

It is also possible to obtain the generators of G by typing

f1,f2,f3 = gens(G)

This is equivalent to

f1=G[1]; f2=G[2]; f3=G[3];

If the group G has been created as the subgroup of another group generated by a list of elements L, then the generating set returned by the function gens corresponds to L.

Note

The output of gens(G) is not, in general, the minimal list of generators for G.

Operations

Oscar supports the following operations and functions on group elements.

* = multiplication between two elements in a group.

inv(x) = $x^{-1}$ the inverse of x.

x/y = the element $xy^{-1}$.

x^n = the $n$-th power of x. If $n = 0$, the identity of the group is returned; if $n < 0$, the $-n$-th power of the inverse of x is returned.

isone(x): returns whether x is the identity of the group.

conj(x,y) = x^y = the conjugate of x by y, i.e. the element $y^{-1}xy$.

comm(x,y) = the commutator of x and y, i.e. the element $x^{-1}y^{-1}xy$.

Note

In Oscar, the expression x^y^z is equivalent to x^(y^z). In other words, conjugations are evaluated from the right to the left.

comm — Method

comm(x::GAPGroupElem, y::GAPGroupElem)

Return [x,y] = x^-1*y^-1*x*y.

nilpotency_class — Method

nilpotency_class(G::GAPGroup)

Return the nilpotency class of G, that is the smallest integer d such that G has a central series of length n.

Properties of groups

isabelian — Function

isabelian(G::GrpGen) -> Bool

Returns whether $G$ is abelian.

isabelian(K::NfRel) -> Bool

Check if the extension is abelian over the coefficient ring. The function is probabilistic.

isabelian(G::Group)

Return whether G is abelian.

ispgroup — Function

ispgroup(G)

Return ($true$,nothing) if G is the trivial group, ($true$,$p$) if |G| is a non-trivial $p$-power, ($false$,nothing) otherwise.

isnilpotent — Function

isnilpotent(G::GAPGroup)

Return whether the G is nilpotent.

issupersolvable — Function

issupersolvable(G::GAPGroup)

Return whether the G is supersolvable.

issolvable — Function

issolvable(G::GAPGroup)

Return whether the G is solvable.

isperfect — Function

isperfect(G)

Return whether G is a perfect group, i.e., equal to its derived subgroup.

issimple — Function

issimple(L::NumField) -> Bool

Given a number field $L/K$ this function returns whether $L$ is simple, that is, whether $L/K$ is defined by a univariate polynomial$.

issimple(G)

Return whether G is a simple group, i.e., it has not non-trivial normal subgroups.

isalmostsimple — Function

isalmostsimple(G)

Return whether G is an almost simple group, i.e., if S < G < Aut(S) for some non-abelian simple group S.

Attributes of groups

exponent — Method

exponent(G::Group)

Return the exponent of G, i.e. the smallest positive integer e such that g^e=1 for every g in G.