Group homomorphisms

Operations on homomorphisms

Properties of homomorphisms

Subgroups described by homomorphisms

Groups created by isomorphisms

Group homomorphisms

In Oscar, a group homomorphism from G to H is an object of parametric type GAPGroupHomomorphism{S,T}, where S and T are the types of G and H respectively.

A homomorphism from G to H can be defined in two ways.

Writing explicitly the images of the generators of G:

f = hom(G,H,[x1,x2,...],[y1,y2,...])

Here, [x1,x2,...] must be a generating set for G (not necessarily minimal) and [y1,y2,...] is a vector of elements of H of the same length of [x1,x2,...]. This assigns to f the value of the group homomorphism sending x_i into y_i.

An exception is thrown if such a homomorphism does not exist.

Taking an existing function g satisfying the group homomorphism properties:

f = hom(G,H,g)

An exception is thrown if the function g does not define a group homomorphism.

Example: The following procedures define the same homomorphism (conjugation by x) in the two ways explained above.

julia> S=symmetric_group(4);

julia> x=S[1];

julia> f=hom(S,S,gens(S),[S[1]^x,S[2]^x]);

julia> g=hom(S,S,y->y^x);

julia> f==g
true

hom — Method

hom(G::GAPGroup, H::GAPGroup, f::Function)

Return the group homomorphism defined by the function f.

hom — Method

hom(G::GAPGroup, H::GAPGroup, gensG::Vector, imgs::Vector)

Return the group homomorphism defined by gensG[i] -> imgs[i] for every i. In order to work, the elements of gensG must generate G.

image — Method

image(f::GAPGroupHomomorphism, x::GAPGroupElem)
(f::GAPGroupHomomorphism)(x::GAPGroupElem)

Return f(x).

restrict_homomorphism — Method

restrict_homomorphism(f::GAPGroupHomomorphism, H::Group)
restrict_homomorphism(f::GAPGroupElem{AutomorphismGroup{T}}, H::T) where T <: Group

Return the restriction of f to H. An exception is thrown if H is not a subgroup of domain(f).

Oscar has also the following standard homomorphism.

id_hom — Function

id_hom(G::GAPGroup)

Return the identity homomorphism on the group G.

trivial_morphism — Function

trivial_morphism(G::GAPGroup, H::GAPGroup)

Return the homomorphism from G to H sending every element of G into the identity of H. If H is not specified, it is taken equal to G.

To evaluate the homomorphism f in the element x of G, it is possible to use the instruction

image(f,x)

or the more compact notations f(x) and x^f.

Example:

julia> S=symmetric_group(4);

julia> f=hom(S,S,x->x^S[1]);

julia> x=cperm(S,[1,2]);

julia> image(f,x)
(2,3)

julia> f(x)
(2,3)

julia> x^f
(2,3)

A sort of "inverse" of the evaluation is the following

haspreimage — Method

haspreimage(f::GAPGroupHomomorphism, x::GAPGroupElem)

Return (true,y) if there exists y such that f(y) = x; otherwise, return (false,1).

Example:

julia> S=symmetric_group(4);

julia> f=hom(S,S,x->x^S[1]);

julia> x=cperm(S,[1,2]);

julia> haspreimage(f,x)
(true, (1,4))

Warning

Do not confuse haspreimage with the function has_preimage, which works on variable of type GrpGenToGrpGenMor.

A further function that produces isomorphisms is isisomorphic:

isisomorphic — Method

isisomorphic(G::Group, H::Group)

Return (true,f) if G and H are isomorphic groups, where f is a group isomorphism. Otherwise, return (false,f), where f is the trivial homomorphism.

Operations on homomorphisms

Oscar supports the following operations on homomorphisms.

inv(f) = the inverse of f. An exception is thrown if f is not bijective.

f^n = the homomorphism f composed n times with itself. An exception is thrown if the domain and the codomain of f do not coincide (unless n=1). If n is negative, the result is the inverse of f composed n times with itself.

compose(f, g) = composition of f and g. This works only if the codomain of f coincides with the domain of g. Shorter equivalent expressions are f*g and g(f).

Example:

julia> S=symmetric_group(4);

julia> f=hom(S,S,x->x^S[1]);

julia> g=hom(S,S,x->x^S[2]);

julia> f*g==hom(S,S,x->x^(S[1]*S[2]))
true

julia> f==f^-3
true

Note

The composition operation * has to be read from the right to the left. So, (f*g)(x) is equivalent to g(f(x)).

Properties of homomorphisms

Oscar implements the following attributes of homomorphisms, in addition to the usual domain and codomain.

isinjective — Method

isinjective(f::GAPGroupHomomorphism)

Return whether f is injective.

issurjective — Method

issurjective(f::GAPGroupHomomorphism)

Return whether f is surjective.

isbijective — Method

isbijective(f::GAPGroupHomomorphism)

Return whether f is bijective.

isinvertible — Method

isinvertible(f::GAPGroupHomomorphism)

Return whether f is invertible.

isinvariant — Method

isinvariant(f::GAPGroupHomomorphism, H::Group)
isinvariant(f::GAPGroupElem{AutomorphismGroup{T}}, H::T)

Return whether f(H) == H holds. An exception is thrown if domain(f) and codomain(f) are not equal or if H is not contained in domain(f).

Subgroups described by homomorphisms

The following functions compute subgroups or quotients of either the domain or the codomain. Analogously to the functions described in Sections Subgroups and Quotients, the output consists of a pair (H, g), where H is a subgroup (resp. quotient) and g is its embedding (resp. projection) homomorphism.

kernel — Method

kernel(f::GAPGroupHomomorphism)

Return the kernel of f, together with its embedding into domain(f).

image — Method

image(f::GAPGroupHomomorphism)

Return the image of f as subgroup of codomain(f), together with the embedding homomorphism.

image — Method

image(f::GAPGroupHomomorphism{S, T}, H::S) where S <: GAPGroup where T <: GAPGroup
(f::GAPGroupHomomorphism{S, T})(H::S)

Return f(H), together with the embedding homomorphism into codomain(f).

cokernel — Method

cokernel(f::GAPGroupHomomorphism)

Return the cokernel of f, that is, the quotient of the codomain of f by the normal closure of the image.

preimage — Method

preimage(f::GAPGroupHomomorphism{S, T}, H::T) where S <: GAPGroup where T <: GAPGroup

If H is a subgroup of the codomain of f, return the subgroup f^-1(H), together with its embedding homomorphism into the domain of f.

Groups created by isomorphisms

isomorphic_perm_group — Method

isomorphic_perm_group(G::GAPGroup)

Return a permutation group H and an isomorphism f from G to H.

If G is infinite, then no such isomorphism exists and an exception is thrown.

isomorphic_pc_group — Method

isomorphic_pc_group(G::GAPGroup)

Return a group H of type PcGroup and an isomorphism f from G to H.

If G is infinite or not solvable, then no such isomorphism exists and an exception is thrown.

isomorphic_fp_group — Method

isomorphic_fp_group(G::GAPGroup)

Return a group H of type FPGroup and an isomorphism f from G to H.