Group libraries

Transitive permutation groups of small degree

Primitive permutation groups of small degree

Perfect groups of small order

Groups of small order

Group libraries

Transitive permutation groups of small degree

number_transitive_groups — Function

number_transitive_groups(n::Int)

Return the number of transitive groups acting on a set of size n.

transitive_group — Function

transitive_group(deg::Int, i::Int)

Return the i-th group in the catalogue of transitive groups over the set {1,...,deg} in the GAP Small Groups Library. The output is a group of type $PermGroup$.

transitive_identification — Function

transitive_identification(G::PermGroup)

Return (deg, m), where G = transitive_group(deg,m).

all_transitive_groups — Function

all_transitive_groups(L...)

Return the list of all transitive groups (up to isomorphism) satisfying the conditions in L. Here, L is a vector whose arguments are organized as L = [ func1, arg1, func2, arg2, ... ], and the function returns all the groups G satisfying the conditions func1(G) = arg1, func2(G) = arg2, etc. An argument can be omitted if it corresponds to the boolean value $true$.

Examples

julia> all_transitive_groups(degree, 4, isabelian)
2-element Array{PermGroup,1}:
 C(4) = 4
 E(4) = 2[x]2

returns the list of all abelian transitive groups acting on a set of order 4.

The type of the groups is $PermGroup$.

Primitive permutation groups of small degree

number_primitive_groups — Function

number_primitive_groups(n::Int)

Return the number of primitive groups acting on a set of size n.

primitive_group — Function

primitive_group(deg::Int, i::Int)

Return the i-th group in the catalogue of primitive groups over the set {1,...,deg} in the GAP Small Groups Library. The output is a group of type $PermGroup$.

all_primitive_groups — Function

all_primitive_groups(L...)

Return the list of all primitive groups (up to isomorphism) satisfying the conditions in L. Here, L is a vector whose arguments are organized as L = [ func1, arg1, func2, arg2, ... ], and the function returns all the groups G satisfying the conditions func1(G) = arg1, func2(G) = arg2, etc. An argument can be omitted if it corresponds to the boolean value $true$.

Examples

julia> all_primitive_groups(degree, 4, isabelian)
PermGroup[]

returns the list of all abelian primitive groups acting on a set of order 4.

The type of the groups is $PermGroup$.

Perfect groups of small order

number_perfect_groups — Function

number_perfect_groups(n::Int)

Return the number of perfect groups of order n.

perfect_group — Function

perfect_group(n::Int, i::Int)
perfect_group(::Type{T}, n::Int, i::Int)

Return the i-th group of order n and type T in the catalogue of perfect groups in the GAP Small Groups Library. The type T can be either $PermGroup$ or $FPGroup$. The group is given of type $FPGroup$ if T is not specified.

perfect_identification — Function

perfect_identification(G::Group)

Return (n, m), where G = perfect_group(n,m).

Groups of small order

number_small_groups — Function

number_small_groups(n::Int)

Return the number of small groups of order n.

small_group — Function

small_group(n::Int, i::Int)

Return the i-th group of order n in the catalogue of the GAP Small Groups Library. The group is given of type $PcGroup$ if the group is solvable, $PermGroup$ otherwise.

small_group_identification — Function

small_group_identification(G::Group)

Return (n, m), where G = small_group(n,m).

all_small_groups — Function

all_small_groups(n::Int, L...)

Return the list of all groups (up to isomorphism) of order n and satisfying the conditions in L. Here, L is a vector whose arguments are organized as L = [ func1, arg1, func2, arg2, ... ], and the function returns all the groups G satisfying the conditions func1(G) = arg1, func2(G) = arg2, etc. An argument can be omitted if it corresponds to the boolean value $true$.

Examples

julia> all_small_groups(12, cyclic, false, isabelian)

returns the list of all abelian non-cyclic groups of order 12.

The type of the groups is $PcGroup$ if the group is solvable, $PermGroup$ otherwise.