Groups

Basics

Subgroups

Quotients

Products of groups

Permutation groups

Finitely presented groups

Polycyclic groups

Matrix groups

Group homomorphisms

Groups of automorphisms

Group libraries

Groups

Oscar supports the following types of groups:

PermGroup = groups of permutations

MatrixGroup = groups of matrices

FPGroup = finitely presented groups

PcGroup = polycyclic groups

DirectProductOfGroups = direct product of two groups

AutomorphismGroup = group of automorphisms over a group

If x is an element of the group G of type T, then the type of x is GAPGroupElement{T}.

Basics

Elements of groups

Given a group G, it is always possible to have access to some particular elements.

GAPGroup — Type

TODO: document this

BasicGAPGroupElem — Type

BasicGAPGroupElem{T<:GAPGroup}

The type BasicGAPGroupElem gathers all types of group elements described only by an underlying GAP object.

elem_type — Method

elem_type(parent)
elem_type(parent_type)

Given a parent object (or its type), return the type of its elements.

Example

julia> S, x = PowerSeriesRing(QQ, 2, "x")
(Univariate power series ring in x over Rationals, x + O(x^3))

julia> elem_type(S) == typeof(x)
true

one — Method

one(G::Group) -> x::GAPGroupElem{typeof(G)}

Return the identity of the group G.

one — Method

one(x::GAPGroupElement{T}) -> x::GAPGroupElem{T}

Return the identity of the parent group of x.

gens — Method

gens(G::Group)

Return an array of generators of the group G. To get a specific generator, use G[i] or gen(G,i) instead of gens(G)[i], as that is more efficient.

ngens — Method

ngens(G::Group) -> Int

Return the length of the array gens(G).

WARNING:

this is NOT, in general, the minimum number of generators for G.

gen — Method

gen(G::Group, i::Integer)

Return the i-th element of the array gens(G). This is equivalent to G[i], and returns gens(G)[i] but may be more efficient than the latter. If i is greater than the length of gens(G), an ERROR is thrown.

rand — Method

rand(G::Group)

Return a random element of the group G.

rand_pseudo — Method

rand_pseudo(G::Group)

Return a random element of the group G. It works faster than rand, but the elements are not necessarily equally distributed.

It is also possible to obtain the generators of G by typing

f1,f2,f3 = gens(G)

This is equivalent to

f1=G[1]; f2=G[2]; f3=G[3];

If the group G has been created as the subgroup of another group generated by a list of elements L, then the generating set returned by the function gens corresponds to L.

Note

The output of gens(G) is not, in general, the minimal list of generators for G.

Operations

Oscar supports the following operations and functions on group elements.

* = multiplication between two elements in a group.

inv(x) = $x^{-1}$ the inverse of x.

x/y = the element $xy^{-1}$.

x^n = the $n$-th power of x. If $n = 0$, the identity of the group is returned; if $n < 0$, the $-n$-th power of the inverse of x is returned.

isone(x): returns whether x is the identity of the group.

conj(x,y) = x^y = the conjugate of x by y, i.e. the element $y^{-1}xy$.

comm(x,y) = the commutator of x and y, i.e. the element $x^{-1}y^{-1}xy$.

Note

In Oscar, the expression x^y^z is equivalent to x^(y^z). In other words, conjugations are evaluated from the right to the left.

comm — Method

comm(x::GAPGroupElem, y::GAPGroupElem)

Return [x,y] = x^-1*y^-1*x*y.

nilpotency_class — Method

nilpotency_class(G::GAPGroup)

Return the nilpotency class of G, that is the smallest integer d such that G has a central series of length n.

Properties of groups

isabelian — Function

isabelian(G::GrpGen) -> Bool

Returns whether $G$ is abelian.

isabelian(K::NfRel) -> Bool

Check if the extension is abelian over the coefficient ring. The function is probabilistic.

isabelian(G::Group)

Return whether G is abelian.

ispgroup — Function

ispgroup(G)

Return ($true$,nothing) if G is the trivial group, ($true$,$p$) if |G| is a non-trivial $p$-power, ($false$,nothing) otherwise.

isnilpotent — Function

isnilpotent(G::GAPGroup)

Return whether the G is nilpotent.

issupersolvable — Function

issupersolvable(G::GAPGroup)

Return whether the G is supersolvable.

issolvable — Function

issolvable(G::GAPGroup)

Return whether the G is solvable.

isperfect — Function

isperfect(G)

Return whether G is a perfect group, i.e., equal to its derived subgroup.

issimple — Function

issimple(L::NumField) -> Bool

Given a number field $L/K$ this function returns whether $L$ is simple, that is, whether $L/K$ is defined by a univariate polynomial$.

issimple(G)

Return whether G is a simple group, i.e., it has not non-trivial normal subgroups.

isalmostsimple — Function

isalmostsimple(G)

Return whether G is an almost simple group, i.e., if S < G < Aut(S) for some non-abelian simple group S.

Attributes of groups

exponent — Method

exponent(G::Group)

Return the exponent of G, i.e. the smallest positive integer e such that g^e=1 for every g in G.

Subgroups

The subgroup of a group G generated by the elements x,y,... is defined by the following instruction:

sub(G, [x,y,...]) ;

sub(x,y,...).

This function returns two objects: a group H, that is the subgroup of G generated by the elements x,y,..., and the embedding homomorphism of H into G. The object H has the same type of G, and it has no memory of the "parent" group G: it is an independent group.

Examples:

julia> G = symmetric_group(4); H = sub(G,[cperm([1,2,3]),cperm([2,3,4])]);

julia> H[1] == alternating_group(4)
true

The following functions are available in Oscar for subgroup properties:

issubgroup — Function

issubgroup(G::T, H::T) where T <: GAPGroup

Return (true,f) if H is a subgroup of G, where f is the embedding homomorphism of H into G, otherwise return (false,nothing).

embedding — Method

embedding(G::T, H::T) where T <: GAPGroup

Return the embedding morphism of H into G. It throws ERROR if H is not a subgroup of G.

index — Method

index(G::T, H::T) where T <: GAPGroup

Return the index of H in G.

isnormal — Method

isnormal(G::T, H::T) where T <: GAPGroup

Return whether the subgroup H is normal in G.

ischaracteristic — Method

ischaracteristic(G::T, H::T) where T <: GAPGroup

Return whether the subgroup H is characteristic in G.

Standard subgroups

The following functions are available in Oscar to obtain standard subgroups of a group G. Every such function returns a tuple (H,f), where H is a group of the same type of G and f is the embedding homomorphism of H into G.

trivial_subgroup — Function

trivial_subgroup(G::GAPGroup)

Return the trivial subgroup of G.

centre — Function

centre(G::Group)

Return the centre of G, i.e. the subgroup of all x in G such that xy=yx for every y in G, together with its embedding morphism into G.

sylow_subgroup — Method

sylow_subgroup(G::Group, p::Int64)

Return a Sylow p-subgroup of G.

hall_subgroup — Method

hall_subgroup(G::Group, P::Array{Int64})

Return a Hall P-subgroup of G. It works only if G is solvable.

derived_subgroup — Function

derived_subgroup(G::GAPGroup)

Return the derived subgroup of G, i.e. the subgroup generated by all commutators of G.

fitting_subgroup — Method

fitting_subgroup(G::GAPGroup)

Return the Fitting subgroup of G, the largest nilpotent normal subgroup of G.

frattini_subgroup — Method

frattini_subgroup(G::GAPGroup)

Return the Frattini subgroup of G, the intersection of all maximal subgroups of G.

radical_subgroup — Method

radical_subgroup(G::GAPGroup)

Return the radical subgroup of G, the largest solvable normal subgroup of G.

socle — Method

socle(G::GAPGroup)

Return the socle of G, the subgroup generated by all minimal normal subgroups of G.

pcore — Method

pcore(G::Group, p::Int64)

Return C,f, where C is the p-core (i.e. the largest normal p-subgroup) of G and f is the embedding morphism of C into G.

intersect — Method

intersect(V::T...) where T <: Group
intersect(V::AbstractVector{T}) where T <: Group

If V = [G_1, ... , G_n], return the group intersection K of the groups G_1, ..., G_n, together with the embeddings K -> G_i.

The following functions return a list of subgroups.

subgroups — Method

subgroups(G::Group)

Return the list of subgroups of G.

normal_subgroups — Function

normal_subgroups(G::Group)

Return the list of normal subgroups of G.

maximal_subgroups — Function

maximal_subgroups(G::Group)

Return the list of maximal subgroups of G.

maximal_normal_subgroups — Function

maximal_normal_subgroups(G::Group)

Return the list of maximal normal subgroups of G.

minimal_normal_subgroups — Function

minimal_normal_subgroups(G::Group)

Return the list of minimal normal subgroups of G.

characteristic_subgroups — Function

characteristic_subgroups(G::Group)

Return the list of characteristic subgroups of G, i.e. the subgroups that are invariant under all automorphisms of G.

derived_series — Function

derived_series(G::GAPGroup)

Return the list [G_1, G_2, G_3, ... ], where G_1=G and G_{i+1} = derived_subgroup(G_i).

sylow_system — Function

sylow_system(G::Group)

Return an array of Sylow $p$-subgroups of G, where $p$ runs over the prime factors of |G|, such that every two such subgroups commute each other (as subgroups). It works only if G is solvable.

hall_system — Function

hall_system(G::Group)

Return an array of $P$-Hall subgroups of G, where $P$ runs over the subsets of prime factors of |G|. It works only if G is solvable.

complement_system — Function

complement_system(G::Group)

Return an array of $p'$-Hall subgroups of G, where $p$ runs over the prime factors of |G|. It works only if G is solvable.

Note

When a function returns a list of subgroups, the output consists in the subgroups only; the embeddings are not returned as well. To get the embedding homomorphism of the subgroup H in G, one has to type embedding(G,H)

Conjugation action of elements and subgroups

isconjugate — Method

isconjugate(G::GAPGroup, x::GAPGroupElem, y::GAPGroupElem)

Return whether x and y are conjugate elements in G.

isconjugate — Method

isconjugate(G::GAPGroup, H::GAPGroup, K::GAPGroup)

Return whether H and K are conjugate subgroups in G.

representative_action — Method

representative_action(G::Group, x::GAPGroupElem, y::GAPGroupElem)

If x,y are conjugate in G, return

true, z

where x^z=y; otherwise, return

false, nothing

representative_action — Method

representative_action(G::Group, H::Group, K::Group)

If H,K are conjugate subgroups in G, return

true, z

where H^z=K; otherwise, return

false, nothing

centralizer — Method

centralizer(G::Group, x::GroupElem)

Return the centralizer of x in G, i.e. the subgroup of all g in G such that gx=xg, together with its embedding morphism into G.

centralizer — Method

centralizer(G::Group, H::Group)

Return the centralizer of H in G, i.e. the subgroup of all g in G such that gh=hg for every h in H, together with its embedding morphism into G.

normalizer — Method

normalizer(G::Group, x::GAPGroupElem)

Return N,f, where N is the normalizer of <x> in G and f is the embedding morphism of N into G.

normalizer — Method

normalizer(G::Group, H::Group)

Return N,f, where N is the normalizer of H in G and f is the embedding morphism of N into G.

core — Method

core(G::Group, H::Group)

Return C,f, where C is the normal core of H in G, that is, the largest normal subgroup of G that is contained in H, and f is the embedding morphism of C into G.

normal_closure — Method

normal_closure(G::Group, H::Group)

Return N, f, where N is the normal closure of H in G, that is, the smallest normal subgroup of G that contains H, and f is the embedding morphism of N into G.

Note that H must be a subgroup of G.

GroupConjClass — Type

GroupConjClass

It could be either the conjugacy class of an element or of a subgroup in a group G. It is displayed as

     cc = x ^ G

where G is a group and x = representative(cc) is either an element or a subgroup of G.

conjugacy_class — Method

conjugacy_class(G::Group, g::GAPGroupElem) -> GroupConjClass

Return the conjugacy class cc of g in G, where g = representative(cc).

conjugacy_class — Method

conjugacy_class(G::T, H::T) where T<:Group -> GroupConjClass

Return the subgroup conjugacy class cc of H in G, where H = representative(cc).

conjugacy_classes — Method

conjugacy_classes(G::Group)

Return the array of all conjugacy classes of elements in G. It is guaranteed that the class of the identity is in the first position.

conjugacy_classes_subgroups — Method

conjugacy_classes_subgroups(G::Group)

Return the array of all conjugacy classes of subgroups of G.

conjugacy_classes_maximal_subgroups — Method

conjugacy_classes_maximal_subgroups(G::Group)

Return the array of all conjugacy classes of maximal subgroups of G.

Cosets (left/right/double)

GroupCoset — Type

GroupCoset{T<: Group, S <: GAPGroupElem}

Type of group cosets. It is displayed as H * x (right cosets) or x * H (left cosets), where H is a subgroup of a group G and x is an element of G. Two cosets are equal if, and only if, they are both left (resp. right) and they contain the same elements.

right_coset — Method

right_coset(H::Group, g::GAPGroupElem)
*(H::Group, g::GAPGroupElem)

Return the coset Hg.

left_coset — Method

left_coset(H::Group, g::GAPGroupElem)
*(g::GAPGroupElem, H::Group)

Return the coset gH.

Note

Since GAP supports right cosets only, the underlying GAP object of left_coset(H,g) is the right coset H^(g^-1) * g.

isright — Method

isright(c::GroupCoset)

Return whether the coset c is a right coset of its acting domain.

isleft — Method

isleft(c::GroupCoset)

Return whether the coset c is a left coset of its acting domain.

isbicoset — Method

isbicoset(C::GroupCoset)

Return whether C is simultaneously a right coset and a left coset for the same subgroup H.

acting_domain — Method

acting_domain(C::GroupCoset)

If C = Hx or xH, return H.

representative — Method

representative(C::GroupCoset)

If C = Hx or xH, return x.

right_cosets — Method

right_cosets(G::Group, H::Group)

Return the array of the right cosets of H in G.

left_cosets — Method

left_cosets(G::Group, H::Group)

Return the array of the left cosets of H in G.

right_transversal — Method

right_transversal(G::T, H::T) where T<: Group

Return a vector containing a complete set of representatives for right cosets for H.

left_transversal — Method

left_transversal(G::T, H::T) where T<: Group

Return a vector containing a complete set of representatives for left cosets for H.

GroupDoubleCoset — Type

GroupDoubleCoset{T<: Group, S <: GAPGroupElem}

Group double coset. It is displayed as H * x * K, where H and K are subgroups of a group G and x is an element of G. Two double cosets are equal if, and only if, they contain the same elements.

double_coset — Method

double_coset(H::Group, x::GAPGroupElem, K::Group)
*(H::Group, x::GAPGroupElem, K::Group)

returns the double coset HxK.

double_cosets — Method

double_cosets(G::T, H::T, K::T; NC=false) where T<: GAPGroup

Return the array of all the double cosets HxK for x in G. If NC = true, do not check whether H and K are subgroups of G.

left_acting_group — Method

left_acting_group(C::GroupDoubleCoset)

if C = HxK, returns H

right_acting_group — Method

right_acting_group(C::GroupDoubleCoset)

if C = HxK, returns K

representative — Method

representative(C::GroupDoubleCoset)

if C = HxK, returns x.

order — Method

order(C::Union{GroupCoset,GroupDoubleCoset})

Return the cardinality of the (double) coset C.

rand — Method

rand(C::Union{GroupCoset,GroupDoubleCoset})

Return a random element of the (double) coset C.

intersect — Method

intersect(V::AbstractVector{Union{T, GroupCoset, GroupDoubleCoset}}) where T <: GAPGroup

Return an array containing all elements belonging to all groups and cosets in the vector V.

Quotients

Quotient groups in Oscar can be defined using the instruction quo in two ways.

Quotients by normal subgroups.

quo — Method

quo(G::T, H::T)

Return the quotient group G/H of type PcGroup (if the quotient group is solvable) or PermGroup (otherwise), together with the projection G -> G/H.

Quotients by elements.

quo — Method

quo(G::T, elements::Vector{S})

Return the quotient group G/H of type FPGroup (if T=FPGroup), PcGroup (if the quotient group is solvable) or PermGroup (otherwise), where H is the normal closure of elements in G.

This is the typical way to build finitely presented groups.

Example:

julia> F=free_group(2);

julia> (f1,f2)=gens(F);

julia> G,_=quo(F,[f1^2,f2^3,(f1*f2)^2]);

julia> isfinite(G)
true

julia> isisomorphic(G,symmetric_group(3))[1]
true

Similarly to the subgroups, the output consists of a pair (Q,p), where Q is the quotient group and p is the projection homomorphism of G into Q.

Products of groups

Direct products

DirectProductGroup — Type

DirectProductGroup

Either direct product of two or more groups of any type, or subgroup of a direct product of groups.

direct_product — Method

direct_product(L::AbstractVector{<:GAPGroup}; morphisms)
direct_product(L::GAPGroup...)

Return the direct product of the groups in the collection L.

The parameter morphisms is false by default. If it is set true, then the output is a triple (G, emb, proj), where emb and proj are the vectors of the embeddings (resp. projections) of the direct product G.

inner_direct_product — Method

inner_direct_product(L::AbstractVector{T}; morphisms)
inner_direct_product(L::T...)

Return a direct product of groups of the same type T as a group of type T. It works for T of the following types:

PermGroup, PcGroup, FPGroup.

The parameter morphisms is false by default. If it is set true, then the output is a triple (G, emb, proj), where emb and proj are the vectors of the embeddings (resp. projections) of the direct product G.

number_of_factors — Method

number_of_factors(G::DirectProductGroup)

Return the number of factors of G.

cartesian_power — Method

cartesian_power(G::T, n::Int)

Return the direct product of n copies of G.

inner_cartesian_power — Method

inner_cartesian_power(G::T, n::Int; morphisms)

Return the direct product of n copies of G as group of type T.

The parameter morphisms is false by default. If it is set true, then the output is a triple (G, emb, proj), where emb and proj are the vectors of the embeddings (resp. projections) of the direct product G.

factor_of_direct_product — Method

factor_of_direct_product(G::DirectProductGroup, j::Int)

Return the j-th factor of G.

as_perm_group — Method

as_perm_group(G::DirectProductGroup)

If G is direct product of permutations groups, return G as permutation group.

as_polycyclic_group — Method

as_polycyclic_group(G::DirectProductGroup)

If G is direct product of polycyclic groups, return G as polycyclic group.

embedding — Method

embedding(G::DirectProductGroup, j::Integer)

Return the embedding of the j-th component of G into G, for j = 1,...,#factors of G.

projection — Method

projection(G::DirectProductGroup, j::Integer)

Return the projection of G into the j-th component of G, for j = 1,...,#factors of G.

write_as_full — Method

write_as_full(G::DirectProductGroup)

If G is a subgroup of the direct product G_1 x ... x G_n such that G = H_1 x ... x H_n for H_i subgroup of G_i, return G as full direct product of the H_i. If such H_i do not exist, an ERROR is returned.

isfull_direct_product — Method

isfull_direct_product(G::DirectProductGroup)

Return whether G is direct product of its factors (false if it is a proper subgroup).

Semidirect products

SemidirectProductGroup — Type

SemidirectProductGroup{S,T}

Semidirect product of two groups of type S and T respectively, or subgroup of a semidirect product of groups.

semidirect_product — Method

semidirect_product(N::S, f::GAPGroupHomomorphism, H::T)

Return the semidirect product of N and H, of type $SemidirectProductGroup{S,T}$, where f is a homomorphism from H to Aut(N).

normal_subgroup — Method

normal_subgroup(G::SemidirectProductGroup)

Return N, where G is the semidirect product of the normal subgroup N and H.

acting_subgroup — Method

acting_subgroup(G::SemidirectProductGroup)

Return H, where G is the semidirect product of the normal subgroup N and H.

homomorphism_of_semidirect_product — Method

homomorphism_of_semidirect_product(G::SemidirectProductGroup)

Return f, where G is the semidirect product of the normal subgroup N and the group H acting on N via the homomorphism h.

isfull_semidirect_product — Method

isfull_semidirect_product(G::SemidirectProductGroup)

Return whether G is a semidirect product of two groups, instead of a proper subgroup.

embedding — Method

embedding(G::SemidirectProductGroup, n::Integer)

Return the embedding of the n-th component of G into G, for n = 1,2. It is not defined for proper subgroups of semidirect products.

projection — Method

projection(G::SemidirectProductGroup, n::Integer)

Return the projection of G into the second component of G.

Wreath products

WreathProductGroup — Type

WreathProductGroup

Wreath product of a group G and a group of permutations H, or a generic group H together with the homomorphism a from H to a permutation group.

wreath_product — Method

wreath_product(G::T, H::S, a::GAPGroupHomomorphism{S,PermGroup})
wreath_product(G::T, H::PermGroup) where T<: Group

Return the wreath product of the group G and the group H, where H acts on n copies of G through the homomorphism a from H to a permutation group, and n is the number of moved points of Image(a).

If a is not specified, then H must be a group of permutations. In this case, n is NOT the number of moved points, but the degree of H.

If W is a wreath product of G and H, {g_1, ..., g_n} are elements of G and h in H, the element (g_1, ..., h) of W can be obtained by typing

    W(g_1,...,g_n, h).

normal_subgroup — Method

normal_subgroup(W::WreathProductGroup)

Return G, where W is the wreath product of G and H.

acting_subgroup — Method

acting_subgroup(W::WreathProductGroup)

Return H, where W is the wreath product of G and H.

homomorphism_of_wreath_product — Method

homomorphism_of_wreath_product(G::WreathProductGroup)

If W is the wreath product of G and H, then return the homomorphism f from H to Sym(n), where n is the number of copies of G.

isfull_wreath_product — Method

isfull_wreath_product(G::WreathProductGroup)

Return whether G is a wreath product of two groups, instead of a proper subgroup.

projection — Method

projection(G::WreathProductGroup)

Return the projection of wreath_product(G,H) into the permutation group H.

embedding — Method

embedding(G::WreathProductGroup, n::Integer)

Return the embedding of the n-th component of G into G.

Permutation groups

Permutation groups can be defined as symmetric groups, alternating groups or their subgroups.

PermGroup — Type

PermGroup

Groups of permutations. Every group of this type is the subgroup of Sym(n) for some n.

Examples

symmetric_group(n::Int): the symmetric group Sym(n)

alternating_group(n::Int): the alternating group Alt(n)

subgroups of Sym(n)

dihedral_group(PermGroup, n::Int): the dihedral group D(n) as group of permutations. Same holds replacing dihedral_group by quaternion_group

PermGroupElem — Type

PermGroupElem

Element of a group of permutation. It is displayed as product of disjoint cycles.

Assumptions:

for x,y in Sym(n), the product xy is read from left to right;

for x in Sym(n) and i in {1,...,n}, i^x and x(i) return the image of i under the action of x.

symmetric_group — Function

symmetric_group(n::Int64)
symmetric_group(::Type{T}, n::Int)

Return the full symmetric group over a set of n elements. The group is returned of type T for T in {PermGroup, PcGroup}. If T is not specified, then T is set as PermGroup.

alternating_group — Function

alternating_group(n::Int64)
alternating_group(::Type{T}, n::Int)

Return the full alternating group over a set of n elements. The group is returned of type T for T in {PermGroup, PcGroup}. If T is not specified, then T is set as PermGroup.

In Oscar, every permutation group has a degree n, that corresponds to the size of the set on which G acts.

degree — Method

degree(G::PermGroup)

Return the degree as permutation group, that is the integer n such that G < Sym(n).

Note

This is not the smallest k such that G embeds in Sym(k).

Note

The degree of a group of permutations is not necessarily equal to the largest moved point of the group G. For example, the trivial subgroup of symmetric_group(n) has degree n even though it fixes n.

Permutations

Permutations in Oscar are displayed as products of disjoint cycles, as in GAP. An explicit permutation can be built using the functions perm, gap_perm and cperm.

perm — Function

perm(G::PermGroup, L::AbstractVector{<:Integer})
(G::PermGroup)(L::AbstractVector{<:Integer})

Return the permutation x which maps every i from 1 to length(L) to L[i]. L must contain every integer from 1 to length(L) exactly, otherwise an exception is thrown. The parent of x is G. If x is not contained in G, an ERROR is returned. For gap_perm, the parent group of x is set as Sym(n), where n is the largest moved point of x.

Examples

julia> perm(symmetric_group(6),[2,4,6,1,3,5])
(1,2,4)(3,6,5)

gap_perm — Function

gap_perm(L::AbstractVector)

Return the permutation x which maps every i from 1 to length(L) to L[i]. L must contain every integer from 1 to length(L) exactly, otherwise an exception is thrown. The parent of x is set as Sym(n).

julia> gap_perm([2,4,6,1,3,5])
(1,2,4)(3,6,5)

cperm — Function

cperm(L::AbstractVector{<:T}...) where T <: Union{Base.Integer, fmpz}
cperm(G::PermGroup, L::AbstractVector{<:T}...)

For given lists of positive integers [a_1, a_2, ..., a_n],[b_1, b_2, ... , b_m], ... return the permutation x = (a_1,a_2,...,a_n)(b_1,b_2,...,b_m).... The array [n,n+1,...,n+k] can be replaced by n:n+k.

If a list is empty or contains duplicates, it fails. The parent of x is G. If x is not contained in G, an ERROR is returned. If G is not specified, then the parent of x is set as Sym(n), where n is the largest moved point of x.

Examples

julia> cperm([1,2,3],4:7)
(1,2,3)(4,5,6,7)

julia> cperm([1,2],[2,3])
(1,3,2)

Example:

julia> perm(symmetric_group(6),[2,4,6,1,3,5])
(1,2,4)(3,6,5)

julia> cperm([1,2,3],4:7)
(1,2,3)(4,5,6,7)

julia> cperm([1,2],[2,3])
(1,3,2)

At the moment, the input vectors of the function cperm need not to be disjoint.

Warning

If the function perm is evaluated in a vector of integers without specifying the group G, then the returned value is an element of the AbstractAlgebra.jl type Perm{Int}. For this reason, if one wants a permutation of type GAPGroupElem{PermGroup} without specifying a parent, one has to use the function gap_perm.

Every permutation has always a permutation group as a parent. Two permutations coincide if, and only if, they move the same points and their parent groups have the same degree.

julia> G=symmetric_group(5);

julia> A=alternating_group(5);

julia> x=cperm(G,[1,2,3]);

julia> y=cperm(A,[1,2,3]);

julia> z=cperm([1,2,3]);

julia> x==y
true

julia> x==z
false

In the example above, x and y are equal because both act on a set of cardinality 5, while x and z are different because x belongs to Sym(5) and z belongs to Sym(2).

If G is a group and x is a permutation, it is possible to set G as parent of x simply typing G(x). This returns the permutation x as element of G (or ERROR if x does not embed into G).

julia> G=symmetric_group(5);

julia> x=cperm([1,2,3]);

julia> y=G(x);

julia> parent(x)
Sym( [ 1 .. 3 ] )

julia> parent(y)
Sym( [ 1 .. 5 ] )

The function Vector{T} works in the opposite way with respect to perm:

Vector — Type

Vector{T}(x::PermGroupElem, n::Int = x.parent.deg) where {T}
Vector(x::PermGroupElem, n::Int = x.parent.deg)

Return the list of length n that contains x(i) at position i. If not specified, T is set as Int.

Permutations as functions

A permutation can be viewed as a function on the set {1,...,n}, hence it can be evaluated on integers.

Note

The multiplication between permutations works from the left to the right. So, if x and y are permutations and n is an integer, then (x*y)(n) = (y(x(n)), NOT x(y(n)).

julia> x = cperm([1,2,3,4,5]);

julia> x(2)
3

This works also if the argument is not in the range 1:n; in such a case, the output coincides with the input.

Finitely presented groups

FPGroup — Type

FPGroup

Finitely presented group. It can be defined via the function free_group.

FPGroupElem — Type

TODO: document this

free_group — Method

free_group

There are four ways to define a free group.

free_group(n::Int) -> FPGroup; return the free group of rank n, with generators printed as "f1","f2","f3", etc.

free_group(L::String...) -> FPGroup; return the free group with length(L) generators, printed as L[1], L[2], L[3], etc.

free_group(L::Array{String,1}) -> FPGroup; same as above.

free_group(n::Int, s::String) -> FPGroup; return the free group of rank n, with generators printed as "s1", "s2", "s3", etc.

Note

In every case, it is not defined a variable named as the generators are printed.

relators — Method

relators(G::FPGroup)

Return a list of relators for the finitely presented group, i.e. elements [x_1, ... , x_n] in F = free_group(ngens(G)) such that G = F/[x_1,...,x_n].

Polycyclic groups

PcGroup — Type

PcGroup

Polycyclic group

Examples

cyclic_group(n::Int): cyclic group of order n

abelian_group(v::Vector{Int}): direct product of cyclic groups of order v[1],v[2],...,v[length(v)]

PcGroupElem — Type

PcGroupElem

Element of a polycyclic group.

Julia has the following functions that allow to generate polycyclic groups:

abelian_group — Method

abelian_group(::Type{T}, v::Vector{Int}) where T <: Group -> PcGroup

Return the direct product of cyclic groups of order v[1] x v[2] x ... x v[n], as group of type T. Here, T must be of type PermGroup, FPGroup or PcGroup.

cyclic_group — Function

cyclic_group(::Type{T}, n::Int)

Return the cyclic group of order n and type T. If the type is not specified, the group is returned of type PcGroup.

dihedral_group — Function

dihedral_group(n::Int)
dihedral_group(::Type{T}, n::Int)

Return the dihedral group of order n of type T, where T is in {PcGroup,PermGroup,FPGroup}. In the first case, the type is set as PcGroup.

quaternion_group — Function

quaternion_group(n::Int)
quaternion_group(::Type{T}, n::Int)

Return the quaternion group of order n of type T, where T is in {PcGroup,PermGroup,FPGroup}. In the first case, the type is set as PcGroup.

Warning

The type need to be specified in the input of the function abelian_group, otherwise a group of type GrpAbFinGen is returned, which is not a GAP group type. In future versions of Oscar, this may change.

The generators of a polycyclic group are displayed as f1, f2, f3, etc., and every element of a polycyclic group is displayed as product of such generators.

Example:

julia> G=abelian_group(PcGroup, [2,4]);

julia> G[1], G[2]
(f1, f2)

julia> G[2]*G[1]
f1*f2

Note that this does not define Julia variables named f1, f2, etc.! To get the generators of the group G, use gens(G); for convenience they can also be accessed as G[1], G[2], as shown in Section Elements of groups.

Matrix groups

MatrixGroup — Type

MatrixGroup{RE<:RingElem, T<:MatElem{RE}} <: GAPGroup

Type of groups G of n x n matrices over the ring R, where n = degree(G) and R = base_ring(G).

At the moment, only rings of type FqNmodFiniteField are supported.

MatrixGroupElem — Type

MatrixGroupElem{RE<:RingElem, T<:MatElem{RE}} <: AbstractMatrixGroupElem

Elements of a group of type MatrixGroup{RE<:RingElem, T<:MatElem{RE}}

base_ring — Method

base_ring(G::MatrixGroup)

Return the base ring of the matrix group G.

degree — Method

degree(G::MatrixGroup)

Return the degree of the matrix group G, i.e. the number of rows of its matrices.

centralizer — Method

centralizer(G::MatrixGroup{T}, x::MatrixGroupElem{T})

Return (C,f), where C is the centralizer of x in C and f is the embedding of C into G. If G = GL(n,F) or SL(n,F), then f = nothing. In this case, to get the embedding homomorphism of C into G, use

issubgroup(G,C)[2]

Elements of matrix groups

matrix — Method

matrix(x::MatrixGroupElem)

Return the underlying AbstractAlgebra matrix of x.

base_ring — Method

base_ring(x::MatrixGroupElem)

Return the base ring of x.

nrows — Method

nrows(x::MatrixGroupElem)

Return the number of rows of the given matrix.

det — Method

det(x::MatrixGroupElem)

Return the determinant of x.

trace — Method

trace(x::MatrixGroupElem)
tr(x::MatrixGroupElem)

Return the trace of x.

multiplicative_jordan_decomposition — Method

multiplicative_jordan_decomposition(M::MatrixGroupElem)

Return S and U in the group G = parent(M) such that S is semisimple, U is unipotent and M = SU = US.

WARNING:

this is NOT, in general, the same output returned when M has type MatElem.

issemisimple — Method

issemisimple(x::MatrixGroupElem{T}) where T <: FinFieldElem

Return whether x is semisimple, i.e. has order coprime with the characteristic of its base ring.

isunipotent — Method

isunipotent(x::MatrixGroupElem{T}) where T <: FinFieldElem

Return whether x is unipotent, i.e. its order is a power of the characteristic of its base ring.

Sesquilinear forms

SesquilinearForm — Type

SesquilinearForm{T<:RingElem}

Type of groups G of n x n matrices over the ring R, where n = degree(G) and R = base_ring(G). At the moment, only rings of type FqNmodFiniteField are supported.

isalternating_form — Method

isalternating_form(f::SesquilinearForm)

Return whether the form f is an alternating form.

ishermitian_form — Method

ishermitian_form(f::SesquilinearForm)

Return whether the form f is a hermitian form.

isquadratic_form — Method

isquadratic_form(f::SesquilinearForm)

Return whether the form f is a quadratic form.

issymmetric_form — Method

issymmetric_form(f::SesquilinearForm)

Return whether the form f is a symmetric form.

alternating_form — Method

alternating_form(B::MatElem{T})

Return the alternating form with Gram matrix B.

symmetric_form — Method

symmetric_form(B::MatElem{T})

Return the symmetric form with Gram matrix B.

hermitian_form — Method

hermitian_form(B::MatElem{T})

Return the hermitian form with Gram matrix B.

quadratic_form — Method

quadratic_form(B::MatElem{T})

Return the quadratic form with Gram matrix B.

quadratic_form — Method

quadratic_form(f::MPolyElem{T}; check=true)

Return the quadratic form described by the polynomial f. Here, f must be a homogeneous polynomial of degree 2. If check is set as false, it does not check whether the polynomial is homogeneous of degree 2. To define quadratic forms of dimension 1, f can also have type PolyElem{T}.

corresponding_bilinear_form — Method

corresponding_bilinear_form(Q::SesquilinearForm)

Given a quadratic form Q, return the bilinear form B defined by B(u,v) = Q(u+v)-Q(u)-Q(v).

corresponding_quadratic_form — Method

corresponding_quadratic_form(Q::SesquilinearForm)

Given a symmetric form f, returns the quadratic form Q defined by Q(v) = f(v,v)/2. It is defined only in odd characteristic.

gram_matrix — Method

gram_matrix(B::SesquilinearForm)

Return the Gram matrix of a sesquilinear or quadratic form B.

defining_polynomial — Method

defining_polynomial(f::SesquilinearForm)

Return the polynomial that defines the quadratic form f.

radical — Method

 radical(A::AbsAlgAss) -> AbsAlgAssIdl

Returns the Jacobson-Radical of $A$.

radical(f::SesquilinearForm{T})

Return the radical of the sesquilinear form f, i.e. the subspace of all v such that f(u,v)=0 for all u. The radical of a quadratic form Q is the set of vectors v such that Q(v)=0 and v lies in the radical of the corresponding bilinear form.

witt_index — Method

witt_index(f::SesquilinearForm{T})

Return the Witt index of the form induced by f on V/Rad(f). The Witt Index is the dimension of a maximal totally isotropic (singular for quadratic forms) subspace.

isdegenerate — Method

isdegenerate(f::SesquilinearForm{T})

Return whether f is degenerate, i.e. f has nonzero radical. A quadratic form is degenerate if the corresponding bilinear form is.

issingular — Method

issingular(Q::SesquilinearForm{T})

For a quadratic form Q, return whether Q is singular, i.e. Q has nonzero radical.

iscongruent — Method

iscongruent(f::SesquilinearForm{T}, g::SesquilinearForm{T}) where T <: RingElem

If f and g are sesquilinear forms, return (true, C) if there exists a matrix C such that f^C = g, or equivalently, CBC* = A, where A and B are the Gram matrices of f and g respectively, and C* is the transpose-conjugate matrix of C. If such C does not exist, then return (false, nothing). If f and g are quadratic forms, return (true, C) if there exists a matrix C such that f^A = ag for some scalar a. If such C does not exist, then return (false, nothing).

Invariant forms

invariant_bilinear_forms — Method

invariant_bilinear_forms(G::MatrixGroup)

Return a generating set for the vector spaces of bilinear forms preserved by the group G.

Note:

At the moment, elements of the generating set are returned of type mat_elem_type(G).

invariant_sesquilinear_forms — Method

invariant_sesquilinear_forms(G::MatrixGroup)

Return a generating set for the vector spaces of sesquilinear non-bilinear forms preserved by the group G. It works only if base_ring(G) is a finite field with even degree on its prime subfield.

Note:

At the moment, elements of the generating set are returned of type mat_elem_type(G).

invariant_quadratic_forms — Method

invariant_quadratic_forms(G::MatrixGroup)

Return a generating set for the vector spaces of quadratic forms preserved by the group G.

Note:

At the moment, elements of the generating set are returned of type mat_elem_type(G).

invariant_symmetric_forms — Method

invariant_symmetric_forms(G::MatrixGroup)

Return a generating set for the vector spaces of symmetric forms preserved by the group G.

Note:

At the moment, elements of the generating set are returned of type mat_elem_type(G).

Note:

Work properly only in odd characteristic. In even characteristic, only alternating forms are found.

invariant_alternating_forms — Method

invariant_alternating_forms(G::MatrixGroup)

Return a generating set for the vector spaces of alternating forms preserved by the group G.

Note:

At the moment, elements of the generating set are returned of type mat_elem_type(G).

invariant_hermitian_forms — Method

invariant_hermitian_forms(G::MatrixGroup)

Return a generating set for the vector spaces of hermitian forms preserved by the group G. It works only if base_ring(G) is a finite field with even degree on its prime subfield.

Note:

At the moment, elements of the generating set are returned of type mat_elem_type(G).

invariant_bilinear_form — Method

function invariant_bilinear_form(G::MatrixGroup)

Return an invariant bilinear form for the group G. It works only if the module induced by the action of G is absolutely irreducible.

Note:

At the moment, the output is returned of type mat_elem_type(G).

invariant_sesquilinear_form — Method

function invariant_sesquilinear_form(G::MatrixGroup)

Return an invariant sesquilinear (non bilinear) form for the group G. It works only if the module induced by the action of G is absolutely irreducible.

Note:

At the moment, the output is returned of type mat_elem_type(G).

invariant_quadratic_form — Method

function invariant_quadratic_form(G::MatrixGroup)

Return an invariant bilinear form for the group G. It works only if the module induced by the action of G is absolutely irreducible.

Note:

At the moment, the output is returned of type mat_elem_type(G).

preserved_quadratic_forms — Method

preserved_quadratic_forms(G::MatrixGroup)

Uses random methods to find all of the quadratic forms preserved by G up to a scalar (i.e. such that G is a group of similarities for the forms). Since the procedure relies on a pseudo-random generator, the user may need to execute the operation more than once to find all invariant quadratic forms.

preserved_sesquilinear_forms — Method

preserved_sesquilinear_forms(G::MatrixGroup)

Uses random methods to find all of the sesquilinear forms preserved by G up to a scalar (i.e. such that G is a group of similarities for the forms). Since the procedure relies on a pseudo-random generator, the user may need to execute the operation more than once to find all invariant sesquilinear forms.

isometry_group — Method

isometry_group(f::SesquilinearForm{T})

Return the group of isometries for the sesquilinear form f.

Utilities for matrices (replace by available functions, or document elsewhere?)

pol_elementary_divisors — Method

pol_elementary_divisors(x::MatElem)
pol_elementary_divisors(x::MatrixGroupElem)

Return a list of pairs (f_i,m_i), for irreducible polynomials f_i and positive integers m_i, where the f_i^m_i are the elementary divisors of x.

generalized_jordan_block — Method

generalized_jordan_block(f::T, n::Int) where T<:PolyElem

Return the Jordan block of dimension n corresponding to the polynomial f.

generalized_jordan_form — Method

generalized_jordan_form(A::MatElem{T}; with_pol=false) where T

Return (J,Z), where Z^-1*J*Z = A and J is a diagonal join of Jordan blocks (corresponding to irreducible polynomials).

matrix — Method

matrix(A::Array{AbstractAlgebra.Generic.FreeModuleElem{T},1})

Return the matrix whose rows are the vectors in A. All vectors in A must have the same length and the same base ring.

upper_triangular_matrix — Method

upper_triangular_matrix(L)

Return the upper triangular matrix whose entries on and above the diagonal are the elements of L.

An error is returned whenever the length of L is not n*(n+1)/2 for some integer n.

lower_triangular_matrix — Method

lower_triangular_matrix(L)

Return the upper triangular matrix whose entries on and below the diagonal are the elements of L.

An error is returned whenever the length of L is not n*(n+1)/2 for some integer n.

conjugate_transpose — Method

conjugate_transpose(x::MatElem{T}) where T <: FinFieldElem

If the base ring of x is GF(q^2), return the matrix transpose( map ( y -> y^q, x) ). An error is signalled if the base ring does not have even degree.

complement — Method

complement(V::AbstractAlgebra.Generic.FreeModule{T}, W::AbstractAlgebra.Generic.Submodule{T})

Return a complement for W in V, i.e. a subspace U of V such that V is direct sum of U and W.

permutation_matrix — Method

permutation_matrix(F::Ring, Q::AbstractVector{T}) where T <: Int
permutation_matrix(F::Ring, p::PermGroupElem)

Return the permutation matrix over the ring R corresponding to the sequence Q or to the permutation p. If Q is a sequence, then Q must contain exactly once every integer from 1 to some n.

isskewsymmetric_matrix — Method

isskewsymmetric_matrix(B::MatElem{T}) where T <: Ring

Return whether the matrix B is skew-symmetric, i.e. B = -transpose(B) and B has zeros on the diagonal. Returns false if B is not a square matrix.

ishermitian_matrix — Method

ishermitian_matrix(B::MatElem{T}) where T <: FinFieldElem

Return whether the matrix B is hermitian, i.e. B = conjugate_transpose(B). Returns false if B is not a square matrix, or the field has not even degree.

Classical groups

general_linear_group — Method

general_linear_group(n::Int, q::Int)
general_linear_group(n::Int, F::FqNmodFiniteField)
GL = general_linear_group

Return the general linear group of dimension n either over the field F or the field GF(q).

special_linear_group — Method

special_linear_group(n::Int, q::Int)
special_linear_group(n::Int, F::FqNmodFiniteField)
SL = special_linear_group

Return the special linear group of dimension n either over the field F or the field GF(q).

symplectic_group — Method

symplectic_group(n::Int, q::Int)
symplectic_group(n::Int, F::FqNmodFiniteField)
Sp = symplectic_group

Return the symplectic group of dimension n either over the field F or the field GF(q). The dimension n must be even.

orthogonal_group — Method

orthogonal_group(e::Int, n::Int, F::Ring)
orthogonal_group(e::Int, n::Int, q::Int)
GO = orthogonal_group

Return the orthogonal group of dimension n either over the field F or the field GF(q) of type e, where e in {+1,-1} for n even and e=0 for n odd. If n is odd, e can be omitted.

special_orthogonal_group — Method

special_orthogonal_group(e::Int, n::Int, F::Ring)
special_orthogonal_group(e::Int, n::Int, q::Int)
SO = special_orthogonal_group

Return the special orthogonal group of dimension n either over the field F or the field GF(q) of type e, where e in {+1,-1} for n even and e=0 for n odd. If n is odd, e can be omitted.

omega_group — Method

omega_group(e::Int, n::Int, F::Ring)
omega_group(e::Int, n::Int, q::Int)

Return the Omega group of dimension n over the field GF(q) of type e, where e in {+1,-1} for n even and e=0 for n odd. If n is odd, e can be omitted.

unitary_group — Method

unitary_group(n::Int, q::Int)
GU = unitary_group

Return the unitary group of dimension n over the field GF(q^2).

special_unitary_group — Method

special_unitary_group(n::Int, q::Int)
SU = special_unitary_group

Return the special unitary group of dimension n over the field GF(q^2).

matrix_group — Method

matrix_group(V::T...) where T<:MatrixGroup
matrix_group(V::AbstractVector{T}) where T<:MatrixGroup

Return the matrix group generated by elements in the vector V.

Group homomorphisms

In Oscar, a group homomorphism from G to H is an object of parametric type GAPGroupHomomorphism{S,T}, where S and T are the types of G and H respectively.

A homomorphism from G to H can be defined in two ways.

Writing explicitly the images of the generators of G:

f = hom(G,H,[x1,x2,...],[y1,y2,...])

Here, [x1,x2,...] must be a generating set for G (not necessarily minimal) and [y1,y2,...] is a vector of elements of H of the same length of [x1,x2,...]. This assigns to f the value of the group homomorphism sending x_i into y_i. If such a homomorphism does not exist, an error is returned.

Taking an existing function g satisfying the group homomorphism properties:

f = hom(G,H,g)

If the function g does not satisfy the group homomorphism properties, an error is returned.

Example: The following procedures define the same homomorphism (conjugation by x) in the two ways explained above.

julia> S=symmetric_group(4);

julia> x=S[1];

julia> f=hom(S,S,gens(S),[S[1]^x,S[2]^x]);

julia> g=hom(S,S,y->y^x);

julia> f==g
true

hom — Method

hom(G::GAPGroup, H::GAPGroup, f::Function)

Return the group homomorphism defined by the function $f$.

hom — Method

hom(G::GAPGroup, H::GAPGroup, gensG::Vector, imgs::Vector)

Return the group homomorphism defined by gensG[i] -> imgs[i] for every i. In order to work, the elements of gensG must generate G.

image — Method

image(f::GAPGroupHomomorphism, x::GAPGroupElem)
(f::GAPGroupHomomorphism)(x::GAPGroupElem)

Return f(x).

restrict_homomorphism — Method

restrict_homomorphism(f::GAPGroupHomomorphism, H::Group)
restrict_homomorphism(f::GAPGroupElem{AutomorphismGroup{T}}, H::T) where T <: Group

Return the restriction of f to H; otherwise it return ERROR.

Oscar has also the following standard homomorphism.

id_hom — Function

id_hom(G::GAPGroup)

Return the identity homomorphism on the group G.

trivial_morphism — Function

trivial_morphism(G::GAPGroup, H::GAPGroup)

Return the homomorphism from G to H sending every element of G into the identity of H. If H is not specified, it is taken equal to G.

To evaluate the homomorphism f in the element x of G, it is possible to use the instruction

image(f,x)

or the more compact notations f(x) and x^f.

Example:

julia> S=symmetric_group(4);

julia> f=hom(S,S,x->x^S[1]);

julia> x=cperm(S,[1,2]);

julia> image(f,x)
(2,3)

julia> f(x)
(2,3)

julia> x^f
(2,3)

A sort of "inverse" of the evaluation is the following

haspreimage — Method

haspreimage(f::GAPGroupHomomorphism, x::GAPGroupElem)

Return (true,y) if there exists y such that f(y) = x; otherwise, return (false,1).

Example:

julia> S=symmetric_group(4);

julia> f=hom(S,S,x->x^S[1]);

julia> x=cperm(S,[1,2]);

julia> haspreimage(f,x)
(true, (1,4))

Warning

Do not confuse haspreimage with the function has_preimage, which works on variable of type GrpGenToGrpGenMor.

A further function that produces isomorphisms is isisomorphic:

isisomorphic — Method

isisomorphic(G::Group, H::Group)

Return (true,f) if G and H are isomorphic groups, where f is a group isomorphism. Otherwise, return (false,f), where f is the trivial homomorphism.

Operations on homomorphisms

Oscar supports the following operations on homomorphisms.

inv(f) = the inverse of f. An error is returned if f is not bijective.

f^n = the homomorphism f composed n times with itself. An error is returned if the domain and the codomain of f do not coincide (unless n=1). If n is negative, the result is the inverse of f composed n times with itself.

compose(g,f) = composition of g and f. This works only if the codomain of g coincide with the domain of f. Shorter equivalent expressions are g*f and f(g).

Example:

julia> S=symmetric_group(4);

julia> f=hom(S,S,x->x^S[1]);

julia> g=hom(S,S,x->x^S[2]);

julia> f*g==hom(S,S,x->x^(S[1]*S[2]))
true

julia> f==f^-3
true

Note

The composition operation * has to be read from the right to the left. So, (f*g)(x) is equivalent to g(f(x)).

Properties of homomorphisms

Oscar implements the following attributes of homomorphisms.

domain — Function

domain(f::GAPGroupHomomorphism)

Return the domain of f.

function domain(F::AlgHom)

Return the domain of F.

codomain — Function

codomain(f::GAPGroupHomomorphism)

Return the codomain of f.

function codomain(F::AlgHom)

Return the codomain of F.

isinjective — Method

isinjective(f::GAPGroupHomomorphism)

Return whether f is injective.

issurjective — Method

issurjective(f::GAPGroupHomomorphism)

Return whether f is surjective.

isbijective — Method

isbijective(f::GAPGroupHomomorphism)

Return whether f is bijective.

isinvertible — Method

isinvertible(f::GAPGroupHomomorphism)

Return whether f is invertible.

isinvariant — Method

isinvariant(f::GAPGroupHomomorphism, H::Group)
isinvariant(f::GAPGroupElem{AutomorphismGroup{T}}, H::T)

Return whether f(H) == H.

Subgroups described by homomorphisms

The following functions compute subgroups or quotients of either the domain or the codomain. Analogously to the functions described in Sections Subgroups and Quotients, the output consists of a pair (H, g), where H is a subgroup (resp. quotient) and g is its embedding (resp. projection) homomorphism.

kernel — Method

kernel(f::GAPGroupHomomorphism)

Return the kernel of f, together with its embedding into domain(f).

image — Method

image(f::GAPGroupHomomorphism)

Return the image of f as subgroup of codomain(f), together with the embedding homomorphism.

image — Method

image(f::GAPGroupHomomorphism{S, T}, H::S) where S <: GAPGroup where T <: GAPGroup
(f::GAPGroupHomomorphism{S, T})(H::S)

Return f(H), together with the embedding homomorphism into codomain(f).

cokernel — Method

cokernel(f::GAPGroupHomomorphism)

Return the cokernel of f, that is, the quotient of the codomain of f by the normal closure of the image.

preimage — Method

preimage(f::GAPGroupHomomorphism{S, T}, H::T) where S <: GAPGroup where T <: GAPGroup

If H is a subgroup of the codomain of f, return the subgroup f^-1(H), together with its embedding homomorphism into the domain of f.

Groups created by isomorphisms

isomorphic_perm_group — Method

isomorphic_perm_group(G::GAPGroup)

Return a permutation group H and an isomorphism f from G to H.

If G is infinite, then no such isomorphism exists and an exception is thrown.

isomorphic_pc_group — Method

isomorphic_pc_group(G::GAPGroup)

Return a group H of type PcGroup and an isomorphism f from G to H.

If G is infinite or not solvable, then no such isomorphism exists and an exception is thrown.

isomorphic_fp_group — Method

isomorphic_fp_group(G::GAPGroup)

Return a group H of type FPGroup and an isomorphism f from G to H.

Groups of automorphisms

Groups of automorphisms over a group G have parametric type AutomorphismGroup{T}, where T is the type of G. The group of automorphisms over a group G is defined by the following instruction:

AutomorphismGroup{T}
automorphism_group(G)

The evaluation of the automorphism f in the element x is analogous to the homomorphism evaluation: it can be obtained by typing either f(x) or x^f.

It is possible to turn an automorphism f into a homomorphism by typing hom(f). The viceversa is also possible: if g is a bijective homomorphism from the group G to itself and A is the automorphism group of G, then the instruction A(g) returns g as automorphism of G. This is the standard way to build explicitly an automorphism (another way, available for inner automorphisms, is shown in Section Inner_automorphisms).

automorphism_group — Method

automorphism_group(G::Group) -> A::AutomorphismGroup{T}

Return the full automorphism group of G. If f is an object of type $GAPGroupHomomorphism$ and it is bijective from G to itself, then A(f) return the embedding of f in A.

Elements of A can be multiplied with other elements of A or by elements of type GAPGroupHomomorphism; in this last case, the result has type GAPGroupHomomorphism.

Example:

julia> S=symmetric_group(4);

julia> A=automorphism_group(S);

julia> g=hom(S,S,x->x^S[1]);

julia> g in A
false

julia> au=A(g);

julia> au in A
true

julia> g==hom(au)
true

julia> x=cperm(S,[1,2,3]);

julia> au(x)
(2,3,4)

julia> g(x)==au(x)
true

In Oscar it is possible to multiply homomorphisms and automorphisms (whenever it makes sense); in such cases, the output is always a variable of type GAPGroupHomomorphism{S,T}.

julia> S=symmetric_group(4);

julia> A=automorphism_group(S);

julia> g=hom(S,S,x->x^S[1]);

julia> f=A(g);

julia> typeof(g*f)
GAPGroupHomomorphism{PermGroup,PermGroup}

The following functions are available for automorphisms, some of them similar to the corresponding functions for homomorphisms of groups.

isinvariant — Method

isinvariant(f::GAPGroupElem{AutomorphismGroup{T}}, H::T)

Return whether f(H) == H.

restrict_automorphism — Method

restrict_automorphism(f::GAPGroupElem{AutomorphismGroup{T}}, H::T)

If H is invariant under f, returns the restriction of f to H as automorphism of H; otherwise it returns ERROR.

induced_automorphism — Method

induced_automorphism(f::GAPGroupHomomorphism, g::GAPGroupHomomorphism)
induced_automorphism(f::GAPGroupHomomorphism, g::GAPGroupElem{AutomorphismGroup{T}})

Return the automorphism h of the image of f such that h(f) == f(g), where g is an automorphism of a group G and f is a group homomorphism defined over G such that the kernel of f is invariant under g

hom — Method

hom(f::GAPGroupElem{AutomorphismGroup{T}}) where T

Return the element f of type $GAPGroupHomomorphism{T,T}$.

Inner automorphisms

Oscar disposes of the following functions to handle inner automorphisms of a group.

inner_automorphism — Method

inner_automorphism(g::GAPGroupElem)

Return the inner automorphism in automorphism_group(parent(g)) defined by x -> x^g.

isinner_automorphism — Method

isinner_automorphism(f::GAPGroupHomomorphism)
isinner_automorphism(f::GAPGroupElem{AutomorphismGroup{T}})

Return whether f is an inner automorphism.

inner_automorphisms_group — Method

inner_automorphisms_group(A::AutomorphismGroup{T})

Return the subgroup of A of the inner automorphisms.

Group libraries

Transitive permutation groups of small degree

number_transitive_groups — Function

number_transitive_groups(n::Int)

Return the number of transitive groups acting on a set of size n.

transitive_group — Function

transitive_group(deg::Int, i::Int)

Return the i-th group in the catalogue of transitive groups over the set {1,...,deg} in the GAP Small Groups Library. The output is a group of type $PermGroup$.

transitive_identification — Function

transitive_identification(G::PermGroup)

Return (deg, m), where G = transitive_group(deg,m).

all_transitive_groups — Function

all_transitive_groups(L...)

Return the list of all transitive groups (up to isomorphism) satisfying the conditions in L. Here, L is a vector whose arguments are organized as L = [ func1, arg1, func2, arg2, ... ], and the function returns all the groups G satisfying the conditions func1(G) = arg1, func2(G) = arg2, etc. An argument can be omitted if it corresponds to the boolean value $true$.

Examples

julia> all_transitive_groups(degree, 4, isabelian)
2-element Array{PermGroup,1}:
 C(4) = 4
 E(4) = 2[x]2

returns the list of all abelian transitive groups acting on a set of order 4.

The type of the groups is $PermGroup$.

Primitive permutation groups of small degree

number_primitive_groups — Function

number_primitive_groups(n::Int)

Return the number of primitive groups acting on a set of size n.

primitive_group — Function

primitive_group(deg::Int, i::Int)

Return the i-th group in the catalogue of primitive groups over the set {1,...,deg} in the GAP Small Groups Library. The output is a group of type $PermGroup$.

all_primitive_groups — Function

all_primitive_groups(L...)

Return the list of all primitive groups (up to isomorphism) satisfying the conditions in L. Here, L is a vector whose arguments are organized as L = [ func1, arg1, func2, arg2, ... ], and the function returns all the groups G satisfying the conditions func1(G) = arg1, func2(G) = arg2, etc. An argument can be omitted if it corresponds to the boolean value $true$.

Examples

julia> all_primitive_groups(degree, 4, isabelian)
PermGroup[]

returns the list of all abelian primitive groups acting on a set of order 4.

The type of the groups is $PermGroup$.

Perfect groups of small order

number_perfect_groups — Function

number_perfect_groups(n::Int)

Return the number of perfect groups of order n.

perfect_group — Function

perfect_group(n::Int, i::Int)
perfect_group(::Type{T}, n::Int, i::Int)

Return the i-th group of order n and type T in the catalogue of perfect groups in the GAP Small Groups Library. The type T can be either $PermGroup$ or $FPGroup$. The group is given of type $FPGroup$ if T is not specified.

perfect_identification — Function

perfect_identification(G::Group)

Return (n, m), where G = perfect_group(n,m).

Groups of small order

number_small_groups — Function

number_small_groups(n::Int)

Return the number of small groups of order n.

small_group — Function

small_group(n::Int, i::Int)

Return the i-th group of order n in the catalogue of the GAP Small Groups Library. The group is given of type $PcGroup$ if the group is solvable, $PermGroup$ otherwise.

small_group_identification — Function

small_group_identification(G::Group)

Return (n, m), where G = small_group(n,m).

all_small_groups — Function

all_small_groups(n::Int, L...)

Return the list of all groups (up to isomorphism) of order n and satisfying the conditions in L. Here, L is a vector whose arguments are organized as L = [ func1, arg1, func2, arg2, ... ], and the function returns all the groups G satisfying the conditions func1(G) = arg1, func2(G) = arg2, etc. An argument can be omitted if it corresponds to the boolean value $true$.

Examples

julia> all_small_groups(12, cyclic, false, isabelian)

returns the list of all abelian non-cyclic groups of order 12.

The type of the groups is $PcGroup$ if the group is solvable, $PermGroup$ otherwise.