Permutation groups

Permutations

Permutations as functions

Permutation groups

Permutation groups can be defined as symmetric groups, alternating groups or their subgroups.

PermGroup — Type

PermGroup

Groups of permutations. Every group of this type is the subgroup of Sym(n) for some n.

Examples

symmetric_group(n::Int): the symmetric group Sym(n)

alternating_group(n::Int): the alternating group Alt(n)

subgroups of Sym(n)

dihedral_group(PermGroup, n::Int): the dihedral group D(n) as group of permutations. Same holds replacing dihedral_group by quaternion_group

PermGroupElem — Type

PermGroupElem

Element of a group of permutation. It is displayed as product of disjoint cycles.

Assumptions:

for x,y in Sym(n), the product xy is read from left to right;

for x in Sym(n) and i in {1,...,n}, i^x and x(i) return the image of i under the action of x.

symmetric_group — Function

symmetric_group(n::Int64)
symmetric_group(::Type{T}, n::Int)

Return the full symmetric group over a set of n elements. The group is returned of type T for T in {PermGroup, PcGroup}. If T is not specified, then T is set as PermGroup.

alternating_group — Function

alternating_group(n::Int64)
alternating_group(::Type{T}, n::Int)

Return the full alternating group over a set of n elements. The group is returned of type T for T in {PermGroup, PcGroup}. If T is not specified, then T is set as PermGroup.

In Oscar, every permutation group has a degree n, that corresponds to the size of the set on which G acts.

degree — Method

degree(G::PermGroup)

Return the degree as permutation group, that is the integer n such that G < Sym(n).

Note

This is not the smallest k such that G embeds in Sym(k).

Note

The degree of a group of permutations is not necessarily equal to the largest moved point of the group G. For example, the trivial subgroup of symmetric_group(n) has degree n even though it fixes n.

Permutations

Permutations in Oscar are displayed as products of disjoint cycles, as in GAP. An explicit permutation can be built using the functions perm, gap_perm and cperm.

perm — Function

perm(G::PermGroup, L::AbstractVector{<:Integer})
(G::PermGroup)(L::AbstractVector{<:Integer})

Return the permutation x which maps every i from 1 to length(L) to L[i]. L must contain every integer from 1 to length(L) exactly, otherwise an exception is thrown. The parent of x is G. If x is not contained in G, an ERROR is returned. For gap_perm, the parent group of x is set as Sym(n), where n is the largest moved point of x.

Examples

julia> perm(symmetric_group(6),[2,4,6,1,3,5])
(1,2,4)(3,6,5)

gap_perm — Function

gap_perm(L::AbstractVector)

Return the permutation x which maps every i from 1 to length(L) to L[i]. L must contain every integer from 1 to length(L) exactly, otherwise an exception is thrown. The parent of x is set as Sym(n).

julia> gap_perm([2,4,6,1,3,5])
(1,2,4)(3,6,5)

cperm — Function

cperm(L::AbstractVector{<:T}...) where T <: Union{Base.Integer, fmpz}
cperm(G::PermGroup, L::AbstractVector{<:T}...)

For given lists of positive integers [a_1, a_2, ..., a_n],[b_1, b_2, ... , b_m], ... return the permutation x = (a_1,a_2,...,a_n)(b_1,b_2,...,b_m).... The array [n,n+1,...,n+k] can be replaced by n:n+k.

If a list is empty or contains duplicates, it fails. The parent of x is G. If x is not contained in G, an ERROR is returned. If G is not specified, then the parent of x is set as Sym(n), where n is the largest moved point of x.

Examples

julia> cperm([1,2,3],4:7)
(1,2,3)(4,5,6,7)

julia> cperm([1,2],[2,3])
(1,3,2)

Example:

julia> perm(symmetric_group(6),[2,4,6,1,3,5])
(1,2,4)(3,6,5)

julia> cperm([1,2,3],4:7)
(1,2,3)(4,5,6,7)

julia> cperm([1,2],[2,3])
(1,3,2)

At the moment, the input vectors of the function cperm need not to be disjoint.

Warning

If the function perm is evaluated in a vector of integers without specifying the group G, then the returned value is an element of the AbstractAlgebra.jl type Perm{Int}. For this reason, if one wants a permutation of type GAPGroupElem{PermGroup} without specifying a parent, one has to use the function gap_perm.

Every permutation has always a permutation group as a parent. Two permutations coincide if, and only if, they move the same points and their parent groups have the same degree.

julia> G=symmetric_group(5);

julia> A=alternating_group(5);

julia> x=cperm(G,[1,2,3]);

julia> y=cperm(A,[1,2,3]);

julia> z=cperm([1,2,3]);

julia> x==y
true

julia> x==z
false

In the example above, x and y are equal because both act on a set of cardinality 5, while x and z are different because x belongs to Sym(5) and z belongs to Sym(2).

If G is a group and x is a permutation, it is possible to set G as parent of x simply typing G(x). This returns the permutation x as element of G (or ERROR if x does not embed into G).

julia> G=symmetric_group(5);

julia> x=cperm([1,2,3]);

julia> y=G(x);

julia> parent(x)
Sym( [ 1 .. 3 ] )

julia> parent(y)
Sym( [ 1 .. 5 ] )

The function Vector{T} works in the opposite way with respect to perm:

Vector — Type

Vector{T}(x::PermGroupElem, n::Int = x.parent.deg) where {T}
Vector(x::PermGroupElem, n::Int = x.parent.deg)

Return the list of length n that contains x(i) at position i. If not specified, T is set as Int.

Permutations as functions

A permutation can be viewed as a function on the set {1,...,n}, hence it can be evaluated on integers.

Note

The multiplication between permutations works from the left to the right. So, if x and y are permutations and n is an integer, then (x*y)(n) = (y(x(n)), NOT x(y(n)).

julia> x = cperm([1,2,3,4,5]);

julia> x(2)
3

This works also if the argument is not in the range 1:n; in such a case, the output coincides with the input.