Subgroups

Standard subgroups

Conjugation action of elements and subgroups

Cosets (left/right/double)

Subgroups

The subgroup of a group G generated by the elements x,y,... is defined by the following instruction:

sub(G, [x,y,...]) ;

sub(x,y,...).

This function returns two objects: a group H, that is the subgroup of G generated by the elements x,y,..., and the embedding homomorphism of H into G. The object H has the same type of G, and it has no memory of the "parent" group G: it is an independent group.

Examples:

julia> G = symmetric_group(4); H = sub(G,[cperm([1,2,3]),cperm([2,3,4])]);

julia> H[1] == alternating_group(4)
true

The following functions are available in Oscar for subgroup properties:

issubgroup — Function

issubgroup(G::T, H::T) where T <: GAPGroup

Return (true,f) if H is a subgroup of G, where f is the embedding homomorphism of H into G, otherwise return (false,nothing).

embedding — Method

embedding(G::T, H::T) where T <: GAPGroup

Return the embedding morphism of H into G. It throws ERROR if H is not a subgroup of G.

index — Method

index(G::T, H::T) where T <: GAPGroup

Return the index of H in G.

isnormal — Method

isnormal(G::T, H::T) where T <: GAPGroup

Return whether the subgroup H is normal in G.

ischaracteristic — Method

ischaracteristic(G::T, H::T) where T <: GAPGroup

Return whether the subgroup H is characteristic in G.

Standard subgroups

The following functions are available in Oscar to obtain standard subgroups of a group G. Every such function returns a tuple (H,f), where H is a group of the same type of G and f is the embedding homomorphism of H into G.

trivial_subgroup — Function

trivial_subgroup(G::GAPGroup)

Return the trivial subgroup of G.

centre — Function

centre(G::Group)

Return the centre of G, i.e. the subgroup of all x in G such that xy=yx for every y in G, together with its embedding morphism into G.

sylow_subgroup — Method

sylow_subgroup(G::Group, p::Int64)

Return a Sylow p-subgroup of G.

hall_subgroup — Method

hall_subgroup(G::Group, P::Array{Int64})

Return a Hall P-subgroup of G. It works only if G is solvable.

derived_subgroup — Function

derived_subgroup(G::GAPGroup)

Return the derived subgroup of G, i.e. the subgroup generated by all commutators of G.

fitting_subgroup — Method

fitting_subgroup(G::GAPGroup)

Return the Fitting subgroup of G, the largest nilpotent normal subgroup of G.

frattini_subgroup — Method

frattini_subgroup(G::GAPGroup)

Return the Frattini subgroup of G, the intersection of all maximal subgroups of G.

radical_subgroup — Method

radical_subgroup(G::GAPGroup)

Return the radical subgroup of G, the largest solvable normal subgroup of G.

socle — Method

socle(G::GAPGroup)

Return the socle of G, the subgroup generated by all minimal normal subgroups of G.

pcore — Method

pcore(G::Group, p::Int64)

Return C,f, where C is the p-core (i.e. the largest normal p-subgroup) of G and f is the embedding morphism of C into G.

intersect — Method

intersect(V::T...) where T <: Group
intersect(V::AbstractVector{T}) where T <: Group

If V = [G_1, ... , G_n], return the group intersection K of the groups G_1, ..., G_n, together with the embeddings K -> G_i.

The following functions return a list of subgroups.

subgroups — Method

subgroups(G::Group)

Return the list of subgroups of G.

normal_subgroups — Function

normal_subgroups(G::Group)

Return the list of normal subgroups of G.

maximal_subgroups — Function

maximal_subgroups(G::Group)

Return the list of maximal subgroups of G.

maximal_normal_subgroups — Function

maximal_normal_subgroups(G::Group)

Return the list of maximal normal subgroups of G.

minimal_normal_subgroups — Function

minimal_normal_subgroups(G::Group)

Return the list of minimal normal subgroups of G.

characteristic_subgroups — Function

characteristic_subgroups(G::Group)

Return the list of characteristic subgroups of G, i.e. the subgroups that are invariant under all automorphisms of G.

derived_series — Function

derived_series(G::GAPGroup)

Return the list [G_1, G_2, G_3, ... ], where G_1=G and G_{i+1} = derived_subgroup(G_i).

sylow_system — Function

sylow_system(G::Group)

Return an array of Sylow $p$-subgroups of G, where $p$ runs over the prime factors of |G|, such that every two such subgroups commute each other (as subgroups). It works only if G is solvable.

hall_system — Function

hall_system(G::Group)

Return an array of $P$-Hall subgroups of G, where $P$ runs over the subsets of prime factors of |G|. It works only if G is solvable.

complement_system — Function

complement_system(G::Group)

Return an array of $p'$-Hall subgroups of G, where $p$ runs over the prime factors of |G|. It works only if G is solvable.

Note

When a function returns a list of subgroups, the output consists in the subgroups only; the embeddings are not returned as well. To get the embedding homomorphism of the subgroup H in G, one has to type embedding(G,H)

Conjugation action of elements and subgroups

isconjugate — Method

isconjugate(G::GAPGroup, x::GAPGroupElem, y::GAPGroupElem)

Return whether x and y are conjugate elements in G.

isconjugate — Method

isconjugate(G::GAPGroup, H::GAPGroup, K::GAPGroup)

Return whether H and K are conjugate subgroups in G.

representative_action — Method

representative_action(G::Group, x::GAPGroupElem, y::GAPGroupElem)

If x,y are conjugate in G, return

true, z

where x^z=y; otherwise, return

false, nothing

representative_action — Method

representative_action(G::Group, H::Group, K::Group)

If H,K are conjugate subgroups in G, return

true, z

where H^z=K; otherwise, return

false, nothing

centralizer — Method

centralizer(G::Group, x::GroupElem)

Return the centralizer of x in G, i.e. the subgroup of all g in G such that gx=xg, together with its embedding morphism into G.

centralizer — Method

centralizer(G::Group, H::Group)

Return the centralizer of H in G, i.e. the subgroup of all g in G such that gh=hg for every h in H, together with its embedding morphism into G.

normalizer — Method

normalizer(G::Group, x::GAPGroupElem)

Return N,f, where N is the normalizer of <x> in G and f is the embedding morphism of N into G.

normalizer — Method

normalizer(G::Group, H::Group)

Return N,f, where N is the normalizer of H in G and f is the embedding morphism of N into G.

core — Method

core(G::Group, H::Group)

Return C,f, where C is the normal core of H in G, that is, the largest normal subgroup of G that is contained in H, and f is the embedding morphism of C into G.

normal_closure — Method

normal_closure(G::Group, H::Group)

Return N, f, where N is the normal closure of H in G, that is, the smallest normal subgroup of G that contains H, and f is the embedding morphism of N into G.

Note that H must be a subgroup of G.

GroupConjClass — Type

GroupConjClass

It could be either the conjugacy class of an element or of a subgroup in a group G. It is displayed as

     cc = x ^ G

where G is a group and x = representative(cc) is either an element or a subgroup of G.

conjugacy_class — Method

conjugacy_class(G::Group, g::GAPGroupElem) -> GroupConjClass

Return the conjugacy class cc of g in G, where g = representative(cc).

conjugacy_class — Method

conjugacy_class(G::T, H::T) where T<:Group -> GroupConjClass

Return the subgroup conjugacy class cc of H in G, where H = representative(cc).

conjugacy_classes — Method

conjugacy_classes(G::Group)

Return the array of all conjugacy classes of elements in G. It is guaranteed that the class of the identity is in the first position.

conjugacy_classes_subgroups — Method

conjugacy_classes_subgroups(G::Group)

Return the array of all conjugacy classes of subgroups of G.

conjugacy_classes_maximal_subgroups — Method

conjugacy_classes_maximal_subgroups(G::Group)

Return the array of all conjugacy classes of maximal subgroups of G.

Cosets (left/right/double)

GroupCoset — Type

GroupCoset{T<: Group, S <: GAPGroupElem}

Type of group cosets. It is displayed as H * x (right cosets) or x * H (left cosets), where H is a subgroup of a group G and x is an element of G. Two cosets are equal if, and only if, they are both left (resp. right) and they contain the same elements.

right_coset — Method

right_coset(H::Group, g::GAPGroupElem)
*(H::Group, g::GAPGroupElem)

Return the coset Hg.

left_coset — Method

left_coset(H::Group, g::GAPGroupElem)
*(g::GAPGroupElem, H::Group)

Return the coset gH.

Note

Since GAP supports right cosets only, the underlying GAP object of left_coset(H,g) is the right coset H^(g^-1) * g.

isright — Method

isright(c::GroupCoset)

Return whether the coset c is a right coset of its acting domain.

isleft — Method

isleft(c::GroupCoset)

Return whether the coset c is a left coset of its acting domain.

isbicoset — Method

isbicoset(C::GroupCoset)

Return whether C is simultaneously a right coset and a left coset for the same subgroup H.

acting_domain — Method

acting_domain(C::GroupCoset)

If C = Hx or xH, return H.

representative — Method

representative(C::GroupCoset)

If C = Hx or xH, return x.

right_cosets — Method

right_cosets(G::Group, H::Group)

Return the array of the right cosets of H in G.

left_cosets — Method

left_cosets(G::Group, H::Group)

Return the array of the left cosets of H in G.

right_transversal — Method

right_transversal(G::T, H::T) where T<: Group

Return a vector containing a complete set of representatives for right cosets for H.

left_transversal — Method

left_transversal(G::T, H::T) where T<: Group

Return a vector containing a complete set of representatives for left cosets for H.

GroupDoubleCoset — Type

GroupDoubleCoset{T<: Group, S <: GAPGroupElem}

Group double coset. It is displayed as H * x * K, where H and K are subgroups of a group G and x is an element of G. Two double cosets are equal if, and only if, they contain the same elements.

double_coset — Method

double_coset(H::Group, x::GAPGroupElem, K::Group)
*(H::Group, x::GAPGroupElem, K::Group)

returns the double coset HxK.

double_cosets — Method

double_cosets(G::T, H::T, K::T; NC=false) where T<: GAPGroup

Return the array of all the double cosets HxK for x in G. If NC = true, do not check whether H and K are subgroups of G.

left_acting_group — Method

left_acting_group(C::GroupDoubleCoset)

if C = HxK, returns H

right_acting_group — Method

right_acting_group(C::GroupDoubleCoset)

if C = HxK, returns K

representative — Method

representative(C::GroupDoubleCoset)

if C = HxK, returns x.

order — Method

order(C::Union{GroupCoset,GroupDoubleCoset})

Return the cardinality of the (double) coset C.

rand — Method

rand(C::Union{GroupCoset,GroupDoubleCoset})

Return a random element of the (double) coset C.

intersect — Method

intersect(V::AbstractVector{Union{T, GroupCoset, GroupDoubleCoset}}) where T <: GAPGroup

Return an array containing all elements belonging to all groups and cosets in the vector V.