Basics

Creation of spaces

quadratic_space — Method

quadratic_space(K::NumField, n::Int) -> QuadSpace

Create the quadratic space over K with dimension n and Gram matrix equal to the identity matrix.

quadratic_space — Method

quadratic_space(K::NumField, G::Int) -> QuadSpace

Create the quadratic space over K with Gram matrix G. The matrix G must be square and symmetric.

Attributes

rank — Method

rank(V::AbsSpace) -> Int

Return the rank of the quadratic space V.

dim — Method

dim(V::AbsSpace) -> Int

Return the dimension of the space V.

gram_matrix — Method

gram_matrix(V::AbsSpace) -> MatElem

Return the Gram matrix of V.

involution — Method

involution(V::AbsSpace) -> NumField

Return the involution of V.

isregular — Method

isregular(V::AbsSpace) -> Bool

Return whether V is regular, that is, if the Gram matrix has full rank.

det — Method

det(V::AbsSpace) -> FieldElem

Returns the determinant of the space V as an element of the fixed field.

discriminant — Method

discriminant(V::AbsSpace) -> FieldElem

Returns the discriminant of the space V as an element of the fixed field.

Inner products and diagonalization

gram_matrix — Method

gram_matrix(V::AbsSpace, M::MatElem) -> MatElem

Returns the gram matrix of the rows of M.

gram_matrix — Method

gram_matrix(V::AbsSpace, S::Vector{Vector}) -> MatElem

Returns the gram matrix of the sequence S.

inner_product — Method

inner_product(V::AbsSpace, v::Vector, w::Vector) -> FieldElem

Returns the inner product of v and w with respect to V.

orthogonal_basis — Method

orthogonal_basis(V::AbsSpace) -> MatElem

Returns a matrix M, such that the rows of M form an orthgonal basis of V.

diagonal — Method

diagonal(V::AbsSpace) -> Vector{FieldElem}

Returns a vector of elements $a_1,\dotsc,a_n$ such that V is isometric to the diagonal space $\langle a_1,\dotsc,a_n \rangle$.

The elements will be contained in the fixed field of V.

Equivalence

hasse_invariant — Method

hasse_invariant(V::QuadSpace, p::Union{InfPlc, NfOrdIdl}) -> Int

Returns the Hasse invariant of the quadratic space V at p. This is equal to the product of local Hilbert symbols $(a_i, a_j)_p$, $i < j$, where $V$ is isometric to $\langle a_1,\dotsc,a_n\rangle$.

witt_invariant — Method

witt_invariant(V::QuadSpace, p::Union{InfPlc, NfOrdIdl}) -> Int

Returns the Witt invariant of the quadratic space V at p.

See [Definition 3.2.1, Kir16].

isequivalent — Method

isequivalent(L::AbsSpace, M::AbsSpace, p::Union{InfPlc, NfOrdIdl}) -> Bool

Returns whether L and M are equivalent over the completion at p.

invariants — Method

invariants(M::QuadSpace)
      -> FieldElem, Dict{NfOrdIdl, Int}, Vector{Tuple{InfPlc, Int}}

Returns a tuple (n, k, d, H, I) of invariants of M, which determine the equivalence class completely. Here n is the dimension. The dimension of the kernel is k. The element d is the determinant of a Gram matrix of the non-degenerate part, H contains the non-trivial Hasse invariants and I contains for each real place the negative index of inertia.

Note that d is determined only modulo squares.

isequivalent — Method

isequivalent(M::QuadSpace, L::QuadSpace) -> Bool

Tests if M and L are equivalent.