Polyhedra

Introduction

Construction

Polyhedron and polymake's Polytope

Polyhedra

Introduction

A set $P \subseteq \mathbb{R}^n$ is called a (convex) polyhedron if it can be written as the intersection of finitely many closed affine halfspaces in $\mathbb{R}^n$. That is, there exists a matrix $A$ and a vector $b$ such that $P = P(A,b) = \{ x \in \mathbb{R}^n \mid Ax \leq b\}.$ Writing $P$ as above is called an $H$-representation of $P$.

When a polyhedron $P \subset \mathbb{R}^n$ is bounded, it is called a polytope and the fundamental theorem of polytopes states that it may be written as the convex hull of finitely many points. That is $P = \textrm{conv}(p_1,\ldots,p_N), p_i \in \mathbb{R}^n.$ Writing $P$ in this way is called a $V$-representation. Polytopes are necessarily compact, i.e., they form convex bodies.

Each polytope has a unique $V$-representation which is minimal with respect to inclusion (or cardinality). Conversely, a polyhedron which is full-dimensional, has a unique minimal $H$-representation. If the polyhedron is not full-dimensional, then there is no canonical choice of an $H$-representation.

Construction

Based on the definition of an $H$-representation, the constructor of Polyhedron can be used:

Polyhedron — Method

Polyhedron(A::Union{Oscar.MatElem,AbstractMatrix}, b)

The (convex) polyhedron defined by

\[P(A,b) = \{ x | Ax ≤ b \}.\]

see Def. 3.35 and Section 4.1. of Michael Joswig, Thorsten Theobald (2013)

Arguments

A::Matrix: Matrix corresponding to the linear coefficients of the inequalilites that describe P.

b::Vector: Vector corresponding to the constant term of the inequalilites that describe P.

Examples

The following lines define the square $[0,1]^2 \subset \mathbb{R}^2$:

julia> A = [1 0; 0 1; -1 0 ; 0 -1];
julia> b = [1, 1, 0, 0];
julia> Polyhedron(A,b)
A polyhedron in ambient dimension 2

The following defines a polytope, as a Polyhedron, via a $V$-representation by calling the following function:

convex_hull — Method

convex_hull(V [, R [, L]])

The polytope given as the convex hull of the rows of a set of points.

see Def. 2.11 and Def. 3.1 of Michael Joswig, Thorsten Theobald (2013)

Arguments

V::Matrix: Points whose convex hull is to be computed; encoded as row vectors.

R::Matrix: Rays completing the set of points; encoded row-wise as representative vectors.

L::Matrix: Generators of the Lineality space; encoded as row vectors.

Examples

The following lines define the square $[0,1]^2 \subset \mathbb{R}^2$:

julia> Square = convex_hull([0 0; 0 1; 1 0; 1 1])
A polyhedron in ambient dimension 2

To construct the positive orthant, rays have to be passed:

julia> V = [0 0];
julia> R = [1 0; 0 1];
julia> PO = convex_hull(V, R)
A polyhedron in ambient dimension 2

The closed-upper half plane can be constructed by passing rays and a lineality space:

julia> V = [0 0];
julia> R = [0 1];
julia> L = [1 0];
julia> UH = convex_hull(V, R, L)
A polyhedron in ambient dimension 2

To obtain the x-axis in $\mathbb{R}^2$:

julia> V = [0 0];
julia> R = [];
julia> L = [1 0];
julia> XA = convex_hull(V, R, L)
A polyhedron in ambient dimension 2

Computing convex hulls

This is a standard triangle, defined via a (redundant) $V$-representation and its unique minimal $H$-representation:

julia> T = convex_hull([ 0 0 ; 1 0 ; 0 1; 0 1/2 ])A polyhedron
julia> facets_as_halfspace_matrix_pair(T)(A = pm::Matrix<pm::Rational>
-1 0
0 -1
1 1
, b = pm::Vector<pm::Rational>
0 0 1)

Polyhedron and polymake's Polytope

Many polyhedral computations are done through polymake. This is visible in the structure Polyhedron via a pointer pm_polytope to the corresponding polymake object. This allows to apply all functionality of polymake by calling a suitable Polymake.jl function on this pointer.

To allow both Oscar's and polymake's functionality to be applicable to a polyhedron object, it can be converted back and forth:

Polyhedron — Method

Polyhedron(P::Polymake.BigObject)

Construct a Polyhedron corresponding to a polymake Polytope.

pm_polytope — Method

pm_polytope(P::Polyhedron)

Get the underlying polymake Polytope.

There are several design differences between polymake and Oscar.

Warning

Polyhedra in polymake and Polymake.jl use homogeneous coordinates. The polyhedra in Oscar use affine coordinates.