Normal Toric Varieties

Introduction

Constructors

Properties of Toric Varieties

Operations for Toric Varieties

Auxillary Methods

Normal Toric Varieties

Introduction

We introduce two main types of normal toric varieties, distinguishing between the affine and non-affine case:

AffineNormalToricVariety is the toric variety associated to a cone $\sigma$, denoted by $U_{\sigma}$ in David A. Cox, John B. Little, Henry K. Schenck (2011)

NormalToricVariety is the toric variety associated to a polyhedral fan $\Sigma$, denoted by $X_{\Sigma}$ in David A. Cox, John B. Little, Henry K. Schenck (2011)

Warning

The lattice is always assumed to be the standard lattice $\mathbb{Z}^n$. Transformations for non-standard lattices will have to be done by the user.

Constructors

Affine Toric Varieties

AffineNormalToricVariety — Method

AffineNormalToricVariety(C::Cone)

Construct the affine normal toric variety $U_{C}$ corresponding to a polyhedral cone C.

Examples

Set C to be the positive orthant in two dimensions.

julia> C = positive_hull([1 0; 0 1])
A polyhedral cone in ambient dimension 2

julia> antv = AffineNormalToricVariety(C)
A normal toric variety corresponding to a polyhedral fan in ambient dimension 2

NormalToricVariety — Method

NormalToricVariety(C::Cone)

Construct the (affine) normal toric variety $X_{\Sigma}$ corresponding to a polyhedral fan $\Sigma = C$ consisting only of the cone C.

Examples

Set C to be the positive orthant in two dimensions.

julia> C = positive_hull([1 0; 0 1])
A polyhedral cone in ambient dimension 2
julia> ntv = NormalToricVariety(C)
A normal toric variety corresponding to a polyhedral fan in ambient dimension 2

AffineNormalToricVariety — Method

AffineNormalToricVariety(v::NormalToricVariety)

For internal design, we make a strict distinction between normal toric varieties and affine toric varieties. Given an affine, normal toric variety v, this method turns it into an affine toric variety.

Examples

julia> v = NormalToricVariety(positive_hull([1 0; 0 1]))
A normal toric variety corresponding to a polyhedral fan in ambient dimension 2

julia> affineVariety = AffineNormalToricVariety(v)
A normal toric variety corresponding to a polyhedral fan in ambient dimension 2

Normal Toric Varieties

NormalToricVariety — Method

NormalToricVariety(PF::PolyhedralFan)

Construct the normal toric variety $X_{PF}$ corresponding to a polyhedral fan PF.

Examples

Take PF to be the normal fan of the square.

julia> square = cube(2)
A polyhedron in ambient dimension 2

julia> nf = normal_fan(square)
A polyhedral fan in ambient dimension 2

julia> ntv = NormalToricVariety(nf)
A normal toric variety corresponding to a polyhedral fan in ambient dimension 2

NormalToricVariety — Method

NormalToricVariety(P::Polyhedron)

Construct the normal toric variety $X_{\Sigma_P}$ corresponding to the normal fan $\Sigma_P$ of the given polyhedron P.

Note that this only coincides with the projective variety associated to P, if P is normal.

Examples

Set P to be a square.

julia> square = cube(2)
A polyhedron in ambient dimension 2

julia> ntv = NormalToricVariety(square)
A normal toric variety corresponding to a polyhedral fan in ambient dimension 2

Famous Toric Vareties

del_pezzo — Method

del_pezzo(b::Int)

Constructs the delPezzo surface with b blowups for b at most 3.

Examples

julia> del_pezzo(3)
A normal toric variety corresponding to a polyhedral fan in ambient dimension 2

hirzebruch_surface — Method

hirzebruch_surface(r::Int)

Constructs the r-th Hirzebruch surface.

Examples

julia> hirzebruch_surface(5)
A normal toric variety corresponding to a polyhedral fan in ambient dimension 2

toric_projective_space — Method

toric_projective_space(d::Int)

Construct the projective space of dimension d.

Examples

julia> toric_projective_space(2)
A normal toric variety corresponding to a polyhedral fan in ambient dimension 2

Further Constructions

blowup_on_ith_minimal_torus_orbit — Method

blowup_on_ith_minimal_torus_orbit(v::AbstractNormalToricVariety, n::Int)

Computes the blowup of the normal toric variety v on its i-th minimal torus orbit.

Examples

julia> P2 = toric_projective_space(2)
A normal toric variety corresponding to a polyhedral fan in ambient dimension 2

julia> blowup_on_ith_minimal_torus_orbit(P2,1)
A normal toric variety corresponding to a polyhedral fan in ambient dimension 2

* — Method

Base.:*(v::AbstractNormalToricVariety, w::AbstractNormalToricVariety)

Computes the Cartesian/direct product of two normal toric varieties v and w.

Examples

julia> P2 = toric_projective_space(2)
A normal toric variety corresponding to a polyhedral fan in ambient dimension 2

julia> P2 * P2
A normal toric variety corresponding to a polyhedral fan in ambient dimension 4

Properties of Toric Varieties

hastorusfactor — Method

hastorusfactor(v::AbstractNormalToricVariety)

Checks if the normal toric variety v has a torus factor.

Examples

julia> hastorusfactor(toric_projective_space(2))
false

isaffine — Method

isaffine(v::AbstractNormalToricVariety)

Checks if the normal toric variety v is affine.

Examples

julia> isaffine(toric_projective_space(2))
false

iscomplete — Method

iscomplete(v::AbstractNormalToricVariety)

Checks if the normal toric variety v is complete.

Examples

julia> iscomplete(toric_projective_space(2))
true

isfano — Method

isfano(v::AbstractNormalToricVariety)

Checks if the normal toric variety v is fano.

Examples

julia> isfano(toric_projective_space(2))
true

isgorenstein — Method

isgorenstein(v::AbstractNormalToricVariety)

Checks if the normal toric variety v is Gorenstein.

Examples

julia> isgorenstein(toric_projective_space(2))
true

issimplicial — Method

issimplicial(v::AbstractNormalToricVariety)

Checks if the normal toric variety v is simplicial. Hence, this function works just as isorbifold. It is implemented for user convenience.

Examples

julia> issimplicial(toric_projective_space(2))
true

issmooth — Method

issmooth(v::AbstractNormalToricVariety)

Checks if the normal toric variety v is smooth.

Examples

julia> issmooth(toric_projective_space(2))
true

isnormal — Method

isnormal(v::AbstractNormalToricVariety)

Checks if the normal toric variety v is normal. (This function is somewhat tautological at this point.)

Examples

julia> isnormal(toric_projective_space(2))
true

isorbifold — Method

isorbifold(v::AbstractNormalToricVariety)

Checks if the normal toric variety v is an orbifold.

Examples

julia> isorbifold(toric_projective_space(2))
true

isprojective — Method

isprojective(v::AbstractNormalToricVariety)

Checks if the normal toric variety v is projective, i.e. if the fan of v is the the normal fan of a polytope.

Examples

julia> isprojective(toric_projective_space(2))
true

isprojective_space — Method

isprojective_space(v::AbstractNormalToricVariety)

Decides if the normal toric varieties v is a projective space.

Examples

julia> H5 = hirzebruch_surface(5)
A normal toric variety corresponding to a polyhedral fan in ambient dimension 2

julia> isprojective_space(H5)
false

julia> isprojective_space(toric_projective_space(2))
true

isq_gorenstein — Method

isq_gorenstein(v::AbstractNormalToricVariety)

Checks if the normal toric variety v is Q-Gorenstein.

Examples

julia> isq_gorenstein(toric_projective_space(2))
true

Operations for Toric Varieties

Affine Open Covering

affine_open_covering — Method

affine_open_covering(v::AbstractNormalToricVariety)

Computes an affine open cover of the normal toric variety v, i.e. returns a list of affine toric varieties.

Examples

julia> p2 = toric_projective_space(2)
A normal toric variety corresponding to a polyhedral fan in ambient dimension 2

julia> affine_open_covering(p2)
3-element Vector{AffineNormalToricVariety}:
 A normal toric variety corresponding to a polyhedral fan in ambient dimension 2
 A normal toric variety corresponding to a polyhedral fan in ambient dimension 2
 A normal toric variety corresponding to a polyhedral fan in ambient dimension 2

Characters, Weil Divisors, Cartier Divisors, Class Group and Picard Group

cartier_divisor_group — Method

cartier_divisor_group(v::AbstractNormalToricVariety)

Computes the Cartier divisor group of an abstract normal toric variety v.

Examples

julia> p2 = toric_projective_space(2)
A normal toric variety corresponding to a polyhedral fan in ambient dimension 2

julia> cartier_divisor_group(p2)
GrpAb: Z^3

character_lattice — Method

character_lattice(v::AbstractNormalToricVariety)

Computes the character lattice of a normal toric variety v.

Examples

julia> p2 = toric_projective_space(2)
A normal toric variety corresponding to a polyhedral fan in ambient dimension 2

julia> character_lattice(p2)
GrpAb: Z^2

class_group — Method

class_group(v::AbstractNormalToricVariety)

Computes the class group of the normal toric variety v.

Examples

julia> p2 = toric_projective_space(2)
A normal toric variety corresponding to a polyhedral fan in ambient dimension 2

julia> class_group(p2)
(General) abelian group with relation matrix
[1 0 -1 0 1 -1]

map_from_cartier_divisor_group_to_torus_invariant_divisor_group — Method

map_from_cartier_divisor_group_to_torus_invariant_divisor_group(v::AbstractNormalToricVariety)

Computes the embedding of the group of Cartier divisors into the group of torus-invariant Weil divisors of an abstract normal toric variety v.

Examples

julia> p2 = toric_projective_space(2)
A normal toric variety corresponding to a polyhedral fan in ambient dimension 2

julia> map_from_cartier_divisor_group_to_torus_invariant_divisor_group(p2)
Map with following data
Domain:
=======
Abelian group with structure: Z^3
Codomain:
=========
Abelian group with structure: Z^3

map_from_character_to_principal_divisors — Method

map_from_character_to_principal_divisors(v::AbstractNormalToricVariety)

Computes the map from the character lattice to the group of principal divisors of a normal toric variety v.

Examples

julia> p2 = toric_projective_space(2)
A normal toric variety corresponding to a polyhedral fan in ambient dimension 2

julia> map_from_character_to_principal_divisors(p2)
Map with following data
Domain:
=======
Abelian group with structure: Z^2
Codomain:
=========
Abelian group with structure: Z^3

map_from_weil_divisors_to_class_group — Method

map_from_weil_divisors_to_class_group(v::AbstractNormalToricVariety)

Computes the map from the group of Weil divisors to the class of group of a normal toric variety v.

Examples

julia> map_from_weil_divisors_to_class_group(p2)
Map with following data
Domain:
=======
Abelian group with structure: Z^3
Codomain:
=========
(General) abelian group with relation matrix
[0 0 0 0 0 0 0 0 0 1 0 -1 0 1 -1]
with structure of Abelian group with structure: Z

picard_group — Method

picard_group(v::AbstractNormalToricVariety)

Computes the Picard group of an abstract normal toric variety v.

Examples

julia> p2 = toric_projective_space(2)
A normal toric variety corresponding to a polyhedral fan in ambient dimension 2

julia> picard_group(p2)
GrpAb: Z

torusinvariant_divisor_group — Method

torusinvariant_divisor_group(v::AbstractNormalToricVariety)

Computes the torusinvariant divisor group of a normal toric variety v.

Examples

julia> p2 = toric_projective_space(2)
A normal toric variety corresponding to a polyhedral fan in ambient dimension 2

julia> torusinvariant_divisor_group(p2)
GrpAb: Z^3

torusinvariant_prime_divisors — Method

torusinvariant_prime_divisors(v::AbstractNormalToricVariety)

Computes the list of all torus invariant prime divisors in a normal toric variety v.

Examples

julia> p2 = toric_projective_space(2)
A normal toric variety corresponding to a polyhedral fan in ambient dimension 2

julia> torusinvariant_prime_divisors(p2)
Free module of rank 3 over Integer Ring

Cones and Fans

fan — Method

fan(v::AbstractNormalToricVariety)

Computes the fan of an abstract normal toric variety v.

Examples

julia> p2 = toric_projective_space(2)
A normal toric variety corresponding to a polyhedral fan in ambient dimension 2

julia> fan(p2)
A polyhedral fan in ambient dimension 2

cone — Method

cone(v::AffineNormalToricVariety)

Returns the cone of the affine normal toric variety v.

Examples

```jdoctest julia> cone(AffineNormalToricVariety(Oscar.positive_hull([1 1; -1 1]))) A polyhedral cone in ambient dimension 2

mori_cone — Method

mori_cone(v::NormalToricVariety)

Computes the mori cone of the normal toric variety v.

Examples

julia> pp = toric_projective_space(2)
A normal toric variety corresponding to a polyhedral fan in ambient dimension 2

julia> mori = mori_cone(pp)
A polyhedral cone in ambient dimension 1

julia> dim(mori)
1

nef_cone — Method

nef_cone(v::NormalToricVariety)

Computes the nef cone of the normal toric variety v.

Examples

julia> pp = toric_projective_space(2)
A normal toric variety corresponding to a polyhedral fan in ambient dimension 2

julia> nef = nef_cone(pp)
A polyhedral cone in ambient dimension 1

julia> dim(nef)
1

Dimensions

dim — Method

dim(v::AbstractNormalToricVariety)

Computes the dimension of the normal toric variety v.

dim_of_torusfactor — Method

dim_of_torusfactor(v::AbstractNormalToricVariety)

Computes the dimension of the torus factor of the normal toric variety v.

euler_characteristic — Method

euler_characteristic(v::AbstractNormalToricVariety)

Computes the Euler characteristic of the normal toric variety v.

ith_betti_number — Method

ith_betti_number(v::AbstractNormalToricVariety, i::Int)

Compute the i-th Betti number of the normal toric variety v.

Rings and ideals

cox_ring — Method

cox_ring(v::AbstractNormalToricVariety)

Computes the Cox ring of the normal toric variety v.

Examples

julia> p2 = toric_projective_space(2)
A normal toric variety corresponding to a polyhedral fan in ambient dimension 2

julia> cox_ring(p2)
(Multivariate Polynomial Ring in x[1], x[2], x[3] over Rational Field graded by 
  x[1] -> [0 0 1]
  x[2] -> [0 0 1]
  x[3] -> [0 0 1], MPolyElem_dec{fmpq, fmpq_mpoly}[x[1], x[2], x[3]])

irrelevant_ideal — Method

irrelevant_ideal(v::AbstractNormalToricVariety)

Computes the irrelevant ideal of a normal toric variety v.

Examples

julia> p2 = toric_projective_space(2)
A normal toric variety corresponding to a polyhedral fan in ambient dimension 2

julia> length(irrelevant_ideal(p2).gens)
1

stanley_reisner_ideal — Method

stanley_reisner_ideal(v::AbstractNormalToricVariety)

Computes the Stanley-Reisner ideal of a normal toric variety v.

Examples

julia> p2 = toric_projective_space(2)
A normal toric variety corresponding to a polyhedral fan in ambient dimension 2

julia> ngens(stanley_reisner_ideal(P2))
1

toric_ideal — Method

toric_ideal(antv::AffineNormalToricVariety)

Return the toric ideal defining the affine normal toric variety.

Examples

Take the cone over the square at height one. The resulting toric variety has one defining equation. In projective space this corresponds to $\mathbb{P}^1\times\mathbb{P}^1$. Note that this cone is self-dual, the toric ideal comes from the dual cone.

julia> C = positive_hull([1 0 0; 1 1 0; 1 0 1; 1 1 1])
A polyhedral cone in ambient dimension 3

julia> antv = AffineNormalToricVariety(C)
A normal toric variety corresponding to a polyhedral fan in ambient dimension 3

julia> toric_ideal(antv)
ideal(-x[1]*x[2] + x[3]*x[4])

Auxillary Methods

binomial_exponents_to_ideal — Method

binomial_exponents_to_ideal(binoms::Union{AbstractMatrix, fmpz_mat})

This function converts the rows of a matrix to binomials. Each row $r$ is written as $r=u-v$ with $u,v\ge 0$ by splitting into positive and negative entries. Then the row $r$ corresponds to $x^u-x^v$. The resulting ideal is returned.

Examples

julia> A = [-1 -1 0 2; 2 3 -2 -1]
2×4 Matrix{Int64}:
 -1  -1   0   2
  2   3  -2  -1

julia> binomial_exponents_to_ideal(A)
ideal(-x[1]*x[2] + x[4]^2, x[1]^2*x[2]^3 - x[3]^2*x[4])

toric_ideal — Method

toric_ideal(pts::AbstractMatrix)

Return the toric ideal generated from the linear relations between the points pts. This is the ideal generated by the set of binomials $\{x^u-x^v\ |\ u,v\in\mathbb{Z}^n_{\ge 0}\ (pts)^T\cdot(u-v)=0\}$

Examples

julia> C = positive_hull([-2 5; 1 0]);

julia> H = hilbert_basis(C).m;

julia> toric_ideal(H)
ideal(x[2]*x[3] - x[4]^2, -x[1]*x[3] + x[2]^2*x[4], -x[1]*x[4] + x[2]^3, -x[1]*x[3]^2 + x[2]*x[4]^3, -x[1]*x[3]^3 + x[4]^5)