Normal Toric Varieties

Introduction

Constructors

Properties of Toric Varieties

Operations for Toric Varieties

Auxillary Methods

Normal Toric Varieties

Introduction

We introduce two main types of normal toric varieties, distinguishing between the affine and non-affine case:

AffineNormalToricVariety is the toric variety associated to a cone $\sigma$, denoted by $U_{\sigma}$ in David A. Cox, John B. Little, Henry K. Schenck (2011)

NormalToricVariety is the toric variety associated to a polyhedral fan $\Sigma$, denoted by $X_{\Sigma}$ in David A. Cox, John B. Little, Henry K. Schenck (2011)

Warning

The lattice is always assumed to be the standard lattice $\mathbb{Z}^n$. Transformations for non-standard lattices will have to be done by the user.

Constructors

Affine Toric Varieties

AffineNormalToricVariety — Method

AffineNormalToricVariety(C::Cone)

Construct the affine normal toric variety $U_{C}$ corresponding to a polyhedral cone C.

Examples

Set C to be the positive orthant in two dimensions.

julia> C = positive_hull([1 0; 0 1])
A polyhedral cone in ambient dimension 2

julia> antv = AffineNormalToricVariety(C)
A normal, affine toric variety

NormalToricVariety — Method

NormalToricVariety(C::Cone)

Construct the (affine) normal toric variety $X_{\Sigma}$ corresponding to a polyhedral fan $\Sigma = C$ consisting only of the cone C.

Examples

Set C to be the positive orthant in two dimensions.

julia> C = positive_hull([1 0; 0 1])
A polyhedral cone in ambient dimension 2

julia> ntv = NormalToricVariety(C)
A normal, affine toric variety

AffineNormalToricVariety — Method

AffineNormalToricVariety(v::NormalToricVariety)

For internal design, we make a strict distinction between normal toric varieties and affine toric varieties. Given an affine, normal toric variety v, this method turns it into an affine toric variety.

Examples

julia> v = NormalToricVariety(positive_hull([1 0; 0 1]))
A normal, affine toric variety

julia> affineVariety = AffineNormalToricVariety(v)
A normal, affine toric variety

Normal Toric Varieties

NormalToricVariety — Method

NormalToricVariety(PF::PolyhedralFan)

Construct the normal toric variety $X_{PF}$ corresponding to a polyhedral fan PF.

Examples

Take PF to be the normal fan of the square.

julia> square = cube(2)
A polyhedron in ambient dimension 2

julia> nf = normal_fan(square)
A polyhedral fan in ambient dimension 2

julia> ntv = NormalToricVariety(nf)
A normal toric variety

NormalToricVariety — Method

NormalToricVariety(P::Polyhedron)

Construct the normal toric variety $X_{\Sigma_P}$ corresponding to the normal fan $\Sigma_P$ of the given polyhedron P.

Note that this only coincides with the projective variety associated to P from the affine relations of the lattice points in P, if P is very ample.

Examples

Set P to be a square.

julia> square = cube(2)
A polyhedron in ambient dimension 2

julia> ntv = NormalToricVariety(square)
A normal toric variety

Famous Toric Vareties

affine_space — Method

affine_space(::Type{NormalToricVariety}, d::Int)

Constructs the (toric) affine space of dimension d.

Examples

julia> affine_space(NormalToricVariety, 2)
A normal, affine, 2-dimensional toric variety

del_pezzo — Method

del_pezzo(b::Int)

Constructs the delPezzo surface with b blowups for b at most 3.

Examples

julia> del_pezzo(3)
A normal, non-affine, smooth, projective, gorenstein, fano, 2-dimensional toric variety without torusfactor

hirzebruch_surface — Method

hirzebruch_surface(r::Int)

Constructs the r-th Hirzebruch surface.

Examples

julia> hirzebruch_surface(5)
A normal, non-affine, smooth, projective, gorenstein, non-fano, 2-dimensional toric variety without torusfactor

projective_space — Method

projective_space(::Type{NormalToricVariety}, d::Int)

Construct the projective space of dimension d.

Examples

julia> projective_space(NormalToricVariety, 2)
A normal, non-affine, smooth, projective, gorenstein, fano, 2-dimensional toric variety without torusfactor

Further Constructions

blowup_on_ith_minimal_torus_orbit — Method

blowup_on_ith_minimal_torus_orbit(v::AbstractNormalToricVariety, n::Int, coordinate_name::String)

Return the blowup of the normal toric variety v on its i-th minimal torus orbit.

Examples

julia> P2 = projective_space(NormalToricVariety, 2)
A normal, non-affine, smooth, projective, gorenstein, fano, 2-dimensional toric variety without torusfactor

julia> bP2 = blowup_on_ith_minimal_torus_orbit(P2,1,"e")
A normal toric variety over QQ

julia> cox_ring(bP2)
Multivariate Polynomial Ring in x2, x3, x1, e over Rational Field graded by 
  x2 -> [1 0]
  x3 -> [0 1]
  x1 -> [1 0]
  e -> [-1 1]

* — Method

Base.:*(v::AbstractNormalToricVariety, w::AbstractNormalToricVariety)

Return the Cartesian/direct product of two normal toric varieties v and w.

Examples

julia> P2 = projective_space(NormalToricVariety, 2)
A normal, non-affine, smooth, projective, gorenstein, fano, 2-dimensional toric variety without torusfactor

julia> P2 * P2
A normal toric variety

Properties of Toric Varieties

hastorusfactor — Method

hastorusfactor(v::AbstractNormalToricVariety)

Checks if the normal toric variety v has a torus factor.

Examples

julia> hastorusfactor(projective_space(NormalToricVariety, 2))
false

isaffine — Method

isaffine(v::AbstractNormalToricVariety)

Checks if the normal toric variety v is affine.

Examples

julia> isaffine(projective_space(NormalToricVariety, 2))
false

iscomplete — Method

iscomplete(v::AbstractNormalToricVariety)

Checks if the normal toric variety v is complete.

Examples

julia> iscomplete(projective_space(NormalToricVariety, 2))
true

isfano — Method

isfano(v::AbstractNormalToricVariety)

Checks if the normal toric variety v is fano.

Examples

julia> isfano(projective_space(NormalToricVariety, 2))
true

isgorenstein — Method

isgorenstein(v::AbstractNormalToricVariety)

Checks if the normal toric variety v is Gorenstein.

Examples

julia> isgorenstein(projective_space(NormalToricVariety, 2))
true

issimplicial — Method

issimplicial(v::AbstractNormalToricVariety)

Checks if the normal toric variety v is simplicial. Hence, this function works just as isorbifold. It is implemented for user convenience.

Examples

julia> issimplicial(projective_space(NormalToricVariety, 2))
true

issmooth — Method

issmooth(v::AbstractNormalToricVariety)

Checks if the normal toric variety v is smooth.

Examples

julia> issmooth(projective_space(NormalToricVariety, 2))
true

isnormal — Method

isnormal(v::AbstractNormalToricVariety)

Checks if the normal toric variety v is normal. (This function is somewhat tautological at this point.)

Examples

julia> isnormal(projective_space(NormalToricVariety, 2))
true

isorbifold — Method

isorbifold(v::AbstractNormalToricVariety)

Checks if the normal toric variety v is an orbifold.

Examples

julia> isorbifold(projective_space(NormalToricVariety, 2))
true

isprojective — Method

isprojective(v::AbstractNormalToricVariety)

Checks if the normal toric variety v is projective, i.e. if the fan of v is the the normal fan of a polytope.

Examples

julia> isprojective(projective_space(NormalToricVariety, 2))
true

is_projective_space — Method

is_projective_space(v::AbstractNormalToricVariety)

Decides if the normal toric varieties v is a projective space.

Examples

julia> H5 = hirzebruch_surface(5)
A normal, non-affine, smooth, projective, gorenstein, non-fano, 2-dimensional toric variety without torusfactor

julia> is_projective_space(H5)
false

julia> is_projective_space(projective_space(NormalToricVariety, 2))
true

is_q_gorenstein — Method

is_q_gorenstein(v::AbstractNormalToricVariety)

Checks if the normal toric variety v is Q-Gorenstein.

Examples

julia> is_q_gorenstein(projective_space(NormalToricVariety, 2))
true

Operations for Toric Varieties

Affine Open Covering

affine_open_covering — Method

affine_open_covering(v::AbstractNormalToricVariety)

Compute an affine open cover of the normal toric variety v, i.e. returns a list of affine toric varieties.

Examples

julia> p2 = projective_space(NormalToricVariety, 2)
A normal, non-affine, smooth, projective, gorenstein, fano, 2-dimensional toric variety without torusfactor

julia> affine_open_covering(p2)
3-element Vector{AffineNormalToricVariety}:
 A normal, affine toric variety
 A normal, affine toric variety
 A normal, affine toric variety

Characters, Weil Divisors, Cartier Divisors, Class Group and Picard Group

cartier_divisor_group — Method

cartier_divisor_group(v::AbstractNormalToricVariety)

Return the Cartier divisor group of an abstract normal toric variety v.

Examples

julia> p2 = projective_space(NormalToricVariety, 2)
A normal, non-affine, smooth, projective, gorenstein, fano, 2-dimensional toric variety without torusfactor

julia> cartier_divisor_group(p2)
GrpAb: Z^3

character_lattice — Method

character_lattice(v::AbstractNormalToricVariety)

Return the character lattice of a normal toric variety v.

Examples

julia> p2 = projective_space(NormalToricVariety, 2);

julia> character_lattice(p2)
GrpAb: Z^2

class_group — Method

class_group(v::AbstractNormalToricVariety)

Return the class group of the normal toric variety v.

Examples

julia> p2 = projective_space(NormalToricVariety, 2);

julia> class_group(p2)
GrpAb: Z

map_from_cartier_divisor_group_to_torus_invariant_divisor_group — Method

map_from_cartier_divisor_group_to_torus_invariant_divisor_group(v::AbstractNormalToricVariety)

Return the embedding of the group of Cartier divisors into the group of torus-invariant Weil divisors of an abstract normal toric variety v.

Examples

julia> p2 = projective_space(NormalToricVariety, 2)
A normal, non-affine, smooth, projective, gorenstein, fano, 2-dimensional toric variety without torusfactor

julia> map_from_cartier_divisor_group_to_torus_invariant_divisor_group(p2)
Identity map with

Domain:
=======
GrpAb: Z^3

map_from_cartier_divisor_group_to_picard_group — Method

map_from_cartier_divisor_group_to_picard_group(v::AbstractNormalToricVariety)

Return the map from the Cartier divisors to the Picard group of an abstract normal toric variety v.

Examples

julia> p2 = projective_space(NormalToricVariety, 2)
A normal, non-affine, smooth, projective, gorenstein, fano, 2-dimensional toric variety without torusfactor

julia> map_from_cartier_divisor_group_to_picard_group(p2)
Map with following data
Domain:
=======
Abelian group with structure: Z^3
Codomain:
=========
Abelian group with structure: Z

map_from_character_to_principal_divisors — Method

map_from_character_to_principal_divisors(v::AbstractNormalToricVariety)

Return the map from the character lattice to the group of principal divisors of a normal toric variety v.

Examples

julia> p2 = projective_space(NormalToricVariety, 2);

julia> map_from_character_to_principal_divisors(p2)
Map with following data
Domain:
=======
Abelian group with structure: Z^2
Codomain:
=========
Abelian group with structure: Z^3

map_from_weil_divisors_to_class_group — Method

map_from_weil_divisors_to_class_group(v::AbstractNormalToricVariety)

Return the map from the group of Weil divisors to the class of group of a normal toric variety v.

Examples

julia> p2 = projective_space(NormalToricVariety, 2);

julia> map_from_weil_divisors_to_class_group(p2)
Map with following data
Domain:
=======
Abelian group with structure: Z^3
Codomain:
=========
Abelian group with structure: Z

picard_group — Method

picard_group(v::AbstractNormalToricVariety)

Return the Picard group of an abstract normal toric variety v.

Examples

julia> p2 = projective_space(NormalToricVariety, 2)
A normal, non-affine, smooth, projective, gorenstein, fano, 2-dimensional toric variety without torusfactor

julia> picard_group(p2)
GrpAb: Z

torusinvariant_divisor_group — Method

torusinvariant_divisor_group(v::AbstractNormalToricVariety)

Return the torusinvariant divisor group of a normal toric variety v.

Examples

julia> p2 = projective_space(NormalToricVariety, 2);

julia> torusinvariant_divisor_group(p2)
GrpAb: Z^3

torusinvariant_prime_divisors — Method

torusinvariant_prime_divisors(v::AbstractNormalToricVariety)

Return the list of all torus invariant prime divisors in a normal toric variety v.

Examples

julia> p2 = projective_space(NormalToricVariety, 2);

julia> torusinvariant_prime_divisors(p2)
3-element Vector{ToricDivisor}:
 A torus-invariant, prime divisor on a normal toric variety
 A torus-invariant, prime divisor on a normal toric variety
 A torus-invariant, prime divisor on a normal toric variety

Cones and Fans

fan — Method

fan(v::AbstractNormalToricVariety)

Return the fan of an abstract normal toric variety v.

Examples

julia> p2 = projective_space(NormalToricVariety, 2)
A normal, non-affine, smooth, projective, gorenstein, fano, 2-dimensional toric variety without torusfactor

julia> fan(p2)
A polyhedral fan in ambient dimension 2

cone — Method

cone(v::AffineNormalToricVariety)

Return the cone of the affine normal toric variety v.

Examples

julia> cone(AffineNormalToricVariety(Oscar.positive_hull([1 1; -1 1])))
A polyhedral cone in ambient dimension 2

mori_cone — Method

mori_cone(v::NormalToricVariety)

Return the mori cone of the normal toric variety v.

Examples

julia> pp = projective_space(NormalToricVariety, 2)
A normal, non-affine, smooth, projective, gorenstein, fano, 2-dimensional toric variety without torusfactor

julia> mori = mori_cone(pp)
A polyhedral cone in ambient dimension 1

julia> dim(mori)
1

nef_cone — Method

nef_cone(v::NormalToricVariety)

Return the nef cone of the normal toric variety v.

Examples

julia> pp = projective_space(NormalToricVariety, 2)
A normal, non-affine, smooth, projective, gorenstein, fano, 2-dimensional toric variety without torusfactor

julia> nef = nef_cone(pp)
A polyhedral cone in ambient dimension 1

julia> dim(nef)
1

Dimensions

dim — Method

dim(v::AbstractNormalToricVariety)

Return the dimension of the normal toric variety v.

dim_of_torusfactor — Method

dim_of_torusfactor(v::AbstractNormalToricVariety)

Return the dimension of the torus factor of the normal toric variety v.

euler_characteristic — Method

euler_characteristic(v::AbstractNormalToricVariety)

Return the Euler characteristic of the normal toric variety v.

betti_number — Method

betti_number(v::AbstractNormalToricVariety, i::Int)

Compute the i-th Betti number of the normal toric variety v. Specifically, this method returns the dimension of the i-th simplicial homology group (with rational coefficients) of v. The employed algorithm is derived from theorem 12.3.12 in David A. Cox, John B. Little, Henry K. Schenck (2011). Note that this theorem requires that the normal toric variety v is both complete and simplicial.

Coefficient ring and coordinate names

We support the Cox ring (also termed the "total coordinate ring" in David A. Cox, John B. Little, Henry K. Schenck (2011)) and the following ideals:

irrelevant ideal,

Stanley-Reisner ideal

For their computation, names for the indeterminates of the Cox ring must be chosen. The user is free to change these until the Cox ring is computed for the first time.

coordinate_names — Method

coordinate_names(v::AbstractNormalToricVariety)

This method returns the names of the homogeneous coordinates of the normal toric variety v. If they are not yet set an error is returned.

set_coordinate_names — Method

set_coordinate_names(v::AbstractNormalToricVariety, coordinate_names::Vector{String})

Allows to set the names of the homogeneous coordinates. If the Cox ring of the variety has already been computed, we do not allow changes of the coordinate names. In this case, an error is triggered.

Likewise, a coefficient ring for the Cox ring must be chosen. The user can change it until either the Cox ring or the toric ideal is computed for the first time.

coefficient_ring — Method

coefficient_ring(v::AbstractNormalToricVariety)

This method returns the coefficient_ring of the normal toric variety v. An error is triggered if it is not yet set.

set_coefficient_ring — Method

set_coefficient_ring(v::AbstractNormalToricVariety, coefficient_ring::AbstractAlgebra.Ring)

Allows to set the coefficient_ring. If the Cox ring of the variety has already been computed, we do not allow this to be changed. In this case an error is triggered.

If no choice is made, we invoke the following default values:

coefficient_ring is chosen as the field of rational numbers,

coordinate_names is chosen as [x1, x2, ... ].

Rings and ideals

cox_ring — Method

cox_ring(v::AbstractNormalToricVariety)

Computes the Cox ring of the normal toric variety v. Note that David A. Cox, John B. Little, Henry K. Schenck (2011) refers to this ring as the "total coordinate ring".

Examples

julia> p2 = projective_space(NormalToricVariety, 2);

julia> set_coordinate_names(p2, ["y1", "y2", "y3"])

julia> set_coefficient_ring(p2, ZZ)

julia> cox_ring(p2)
Multivariate Polynomial Ring in y1, y2, y3 over Integer Ring graded by 
  y1 -> [1]
  y2 -> [1]
  y3 -> [1]

cox_ring — Method

cox_ring(R::MPolyRing, v::AbstractNormalToricVariety)

Computes the Cox ring of the normal toric variety v, in this case by adding the Cox grading to the given ring R. Note that David A. Cox, John B. Little, Henry K. Schenck (2011) refers to this ring as the "total coordinate ring".

Examples

julia> p2 = projective_space(NormalToricVariety, 2);

julia> R,_ = PolynomialRing(QQ, 3);

julia> cox_ring(R, p2)
Multivariate Polynomial Ring in x1, x2, x3 over Rational Field graded by 
  x1 -> [1]
  x2 -> [1]
  x3 -> [1]

irrelevant_ideal — Method

irrelevant_ideal(v::AbstractNormalToricVariety)

Return the irrelevant ideal of a normal toric variety v.

Examples

julia> p2 = projective_space(NormalToricVariety, 2);

julia> length(gens(irrelevant_ideal(p2)))
3

irrelevant_ideal — Method

irrelevant_ideal(R::MPolyRing, v::AbstractNormalToricVariety)

Return the irrelevant ideal of a normal toric variety v as an ideal in R.

Examples

julia> p2 = projective_space(NormalToricVariety, 2);

julia> R,_ = PolynomialRing(QQ, 3);

julia> length(gens(irrelevant_ideal(R, p2)))
3

stanley_reisner_ideal — Method

stanley_reisner_ideal(v::AbstractNormalToricVariety)

Return the Stanley-Reisner ideal of a normal toric variety v.

Examples

julia> p2 = projective_space(NormalToricVariety, 2);

julia> ngens(stanley_reisner_ideal(p2))
1

stanley_reisner_ideal — Method

stanley_reisner_ideal(R::MPolyRing, v::AbstractNormalToricVariety)

Return the Stanley-Reisner ideal of a normal toric variety v as an ideal of R.

Examples

julia> p2 = projective_space(NormalToricVariety, 2);

julia> R,_ = PolynomialRing(QQ, 3);

julia> ngens(stanley_reisner_ideal(R, p2))
1

toric_ideal — Method

toric_ideal(antv::AffineNormalToricVariety)

Return the toric ideal defining the affine normal toric variety.

Examples

Take the cone over the square at height one. The resulting toric variety has one defining equation. In projective space this corresponds to $\mathbb{P}^1\times\mathbb{P}^1$. Note that this cone is self-dual, the toric ideal comes from the dual cone.

julia> C = positive_hull([1 0 0; 1 1 0; 1 0 1; 1 1 1])
A polyhedral cone in ambient dimension 3

julia> antv = AffineNormalToricVariety(C)
A normal, affine toric variety

julia> toric_ideal(antv)
ideal(-x1*x2 + x3*x4)

toric_ideal — Method

toric_ideal(R::MPolyRing, antv::AffineNormalToricVariety)

Return the toric ideal defining the affine normal toric variety as an ideal in R.

Examples

Take the cone over the square at height one. The resulting toric variety has one defining equation. In projective space this corresponds to $\mathbb{P}^1\times\mathbb{P}^1$. Note that this cone is self-dual, the toric ideal comes from the dual cone.

julia> C = positive_hull([1 0 0; 1 1 0; 1 0 1; 1 1 1])
A polyhedral cone in ambient dimension 3

julia> antv = AffineNormalToricVariety(C)
A normal, affine toric variety

julia> R,_ = PolynomialRing(QQ, 4);

julia> toric_ideal(R, antv)
ideal(-x1*x2 + x3*x4)

Auxillary Methods

binomial_exponents_to_ideal — Method

binomial_exponents_to_ideal(binoms::Union{AbstractMatrix, fmpz_mat})

This function converts the rows of a matrix to binomials. Each row $r$ is written as $r=u-v$ with $u,v\ge 0$ by splitting into positive and negative entries. Then the row $r$ corresponds to $x^u-x^v$. The resulting ideal is returned.

Examples

julia> A = [-1 -1 0 2; 2 3 -2 -1]
2×4 Matrix{Int64}:
 -1  -1   0   2
  2   3  -2  -1

julia> binomial_exponents_to_ideal(A)
ideal(-x1*x2 + x4^2, x1^2*x2^3 - x3^2*x4)

toric_ideal — Method

toric_ideal(pts::fmpz_mat)

Return the toric ideal generated from the linear relations between the points pts. This is the ideal generated by the set of binomials $\{x^u-x^v\ |\ u,v\in\mathbb{Z}^n_{\ge 0}\ (pts)^T\cdot(u-v)=0\}$

Examples

julia> C = positive_hull([-2 5; 1 0]);

julia> H = hilbert_basis(C);

julia> toric_ideal(H)
ideal(x2*x3 - x4^2, -x1*x3 + x2^2*x4, -x1*x4 + x2^3, -x1*x3^2 + x2*x4^3, -x1*x3^3 + x4^5)