Toric Line Bundles

Constructors

Attributes

Toric Line Bundles

Constructors

ToricLineBundle — Method

ToricLineBundle(v::AbstractNormalToricVariety, c::Vector{fmpz})

Construct the line bundle on the abstract normal toric variety v with class c. ```

ToricLineBundle — Method

ToricLineBundle(v::AbstractNormalToricVariety, c::Vector{Int})

Convenience method for ToricLineBundle(v::AbstractNormalToricVariety, c::Vector{fmpz}).

Examples

julia> v = toric_projective_space(2)
A normal toric variety corresponding to a polyhedral fan in ambient dimension 2

julia> l = ToricLineBundle( v, [ 2 ] )
A line bundle on a normal toric variety corresponding to a polyhedral fan in ambient dimension 2

Attributes

divisor_class — Method

divisor_class(l::ToricLineBundle)

Returns the divisor class which defines the toric line bundle l.

variety — Method

variety(l::ToricLineBundle)

Returns the toric variety over which the toric line bundle l is defined.